% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{add_species}
\alias{add_species}
\title{Add more species into an ecosystem with scaling background species.}
\usage{
add_species(params, species_params, biomass = NA,
  min_w_observed = species_params$w_mat, rfac = 10, effort = 0)
}
\arguments{
\item{params}{A mizer params object for the original system. The A slot holds
1's for foreground species we wish to hold fixed the abundance multiplier
of, and NA's for for background species that we shall vary the abundance
multiplier of.}

\item{species_params}{The species parameters of the foreground species we
want to add to the system.}

\item{biomass}{The total biomass of members of the newly added species which
have weights greater than min_w_observed. If not provided, the abundance
of new species will be chosen so that the biomass of its mature fish is
the same as that of the species with the next smaller maturity size.}

\item{min_w_observed}{The minimum weight of fish of the new species that
contribute to the biomass which we setup for the new species. Default value
is the maturity size of the new species.}

\item{rfac}{Default value is 10.}

\item{effort}{Default value is 0.}
}
\value{
An object of type \code{MizerParams}
}
\description{
Takes a \linkS4class{MizerParams} object and adds additional species with
given parameters to the ecosystem.
}
\examples{
\dontrun{
params <- set_scaling_model(max_w_inf = 5000)
params@A[] <- NA
a_m <- 0.0085
b_m <- 3.11
L_inf_m <- 24.3
L_mat <- 11.1
species_params <- data.frame(
    species = "mullet",
    w_min = 0.001, 
    w_inf = a_m*L_inf_m^b_m, 
    w_mat = a_m*L_mat^b_m, 
    beta = 283, 
    sigma = 1.8, 
    z0 = 0,
    alpha = 0.6, 
    erepro = 0.1, 
    sel_func = "knife_edge", 
    knife_edge_size = 100, 
    gear = "knife_edge_gear",
    k = 0,
    r_max = Inf,
    k_vb = 0.6,
    a = a_m,
    b = b_m
)
params_out <- add_species(params, species_params, biomass = 3070953023, 
                          min_w_observed = 0)
sim <- project(params_out, t_max = 5, effort = 0)
plot(sim)
}
}
\seealso{
\linkS4class{MizerParams}
}
