% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_mat.R
\name{age_mat}
\alias{age_mat}
\title{Calculate age at maturity}
\usage{
age_mat(params, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{...}{Currently unused.}
}
\value{
A named vector. The names are the species names and the values are
the ages at maturity.
}
\description{
Uses the size-dependent growth rate and the size at maturity to calculate
the age at maturity.
}
\details{
Using that by definition of the growth rate \eqn{g(w) = dw/dt} we have that
\deqn{\mathrm{age_{mat}} = \int_0^{w_{mat}.}\frac{dw}{g(w)}}{age_mat = \int_0^w_mat 1/g(w) dw.}
In the implementation this integral is approximated on the model size grid by
summing \code{dw / g(w)} over all size bins with \code{w < w_mat}.
}
\examples{
age_mat(NS_params)
}
\concept{helper}
