% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{array_log_y}
\alias{array_log_y}
\title{The logarithmic y axis an array's type calls for}
\usage{
array_log_y(x, log_y, log, given)
}
\arguments{
\item{x}{A mizer array object.}

\item{log_y}{The \code{log_y} argument of the plot method.}

\item{log}{The \code{log} argument of the plot method.}

\item{given}{Whether the caller supplied \code{log_y} (i.e. \code{!missing(log_y)}).}
}
\value{
The \code{log_y} to use.
}
\description{
A proportion belongs on a linear axis, so a plot of one turns \code{log_y} off
unless the caller asked for a particular axis. Called before
\code{\link[=parsePlotLog]{parsePlotLog()}}, which is why it also has to check \code{log}.
}
\keyword{internal}
