% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R
\name{array_type}
\alias{array_type}
\title{The type of a mizer array}
\usage{
array_type(x)
}
\arguments{
\item{x}{A mizer array object.}
}
\value{
One of \link{array_types}.
}
\description{
The type of a mizer array
}
\keyword{internal}
