% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanModel.R
\name{as_series_matrix}
\alias{as_series_matrix}
\title{Bring what a value function returned into a time by series matrix}
\usage{
as_series_matrix(x, default_name = "Value")
}
\arguments{
\item{x}{The return value of a \code{value_func}.}

\item{default_name}{The column name to use for a single unnamed series.}
}
\value{
A matrix with time in the rows and the series in the columns.
}
\description{
Bring what a value function returned into a time by series matrix
}
\keyword{internal}
