% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Benguela_params}
\alias{Benguela_params}
\title{Example MizerParams object for the Benguela current}
\format{A MizerParams object}
\source{
{N.S. Jacobsen, M. Burgess and K.H. Andersen (2017): 
Efficiency of fisheries is increasing at the ecosystem level. 
Fish and Fisheries 18(2) 199- 211. doi:10.1111/faf.12171.}
}
\description{
Created with purely size-based predation, i.e., no species-specific
interactions. Set up with three fishing gears targeting small, medium and
large species. Vulnerabilities are represented by changing the clearance rate
constant (gamma) between species. Calibrated to efforts
\code{effort = c(small = 0.13, medium = 0.05, large = 0.45)}.
}
\examples{
\dontrun{
sim = project(Benguela_params, effort = c(small = 0.13, medium = 0.05, large = 0.45))
plot(sim)
}
}
\seealso{
Other example parameter objects: 
\code{\link{Baltic_params}},
\code{\link{Barents_params}},
\code{\link{NEUSCS_params}},
\code{\link{NS_params}},
\code{\link{NorthSea_params}}
}
\concept{example parameter objects}
