% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{biomass_callback}
\alias{biomass_callback}
\title{Callback function for plotting biomass in real-time}
\usage{
biomass_callback(...)
}
\arguments{
\item{...}{Other arguments passed to the callback (such as \code{ylim}, \code{species},
\code{use_cutoff}, etc.).}
}
\value{
A callback function that can be passed to \code{\link[=project]{project()}}.
}
\description{
This function can be passed to the \code{callback} argument of \code{\link[=project]{project()}} to
plot the biomasses of the species in real-time as the simulation runs.
}
\examples{
\dontrun{
params <- NS_params
# Open an external graphics window to see the real-time plot in RStudio
x11()
sim <- project(params, callback = biomass_callback())
}
}
