% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{biomass_cutoff_min_w}
\alias{biomass_cutoff_min_w}
\title{The minimum weights given by the \code{biomass_cutoff} species parameter}
\usage{
biomass_cutoff_min_w(params)
}
\arguments{
\item{params}{A MizerParams object.}
}
\value{
A numeric vector with one minimum weight for each species.
}
\description{
Internal helper for \code{\link[=getBiomass]{getBiomass()}}. Returns the \code{biomass_cutoff} column of
the species parameters, with any NAs replaced by the smallest weight in the
model.
}
\concept{helper}
\keyword{internal}
