% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{calc_retain_prob}
\alias{calc_retain_prob}
\title{Calculate the proportion of fish retained instead of discarded}
\usage{
calc_retain_prob(params)
}
\arguments{
\item{params}{A \code{MizerParams} object}
}
\value{
A matrix (gear x species x size) with the retention probability.
}
\description{
This function calculates the proportion of fish retained instead of discarded
by a fishing gear. The retention probability is given by a sigmoidal function.
The parameters of the sigmoid are \code{retain_l50} and \code{retain_l25}. The sigmoid
is defined as
\deqn{p(w) = \frac{1}{1 + \exp(-\frac{w - l_{50}}{l_{25}})}}
where \eqn{w} is the weight of the fish and \eqn{l_{50}} and \eqn{l_{25}} are
the parameters of the sigmoid. The sigmoid is defined such that 50\% of the
fish are retained at \eqn{l_{50}} and 25\% of the fish are retained at
\eqn{l_{50} + l_{25}}.
}
\examples{
retain_prob <- getRetainProb(NS_params)
retain_prob["Pelagic", "Herring", 100]
}
\concept{helper}
