% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_dynamics.R
\name{carrion_dynamics}
\alias{carrion_dynamics}
\title{Carrion dynamics}
\usage{
carrion_dynamics(params, n, n_pp, B, rates, dt,
  carrion_external = params@resource_params$carrion_external, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of current species abundances (species x size)}

\item{n_pp}{A vector of current plankton abundance by size}

\item{B}{A vector of current resource biomasses}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step}

\item{carrion_external}{External inflow rate of carrion biomass}

\item{...}{Unused}
}
\value{
Biomass of carrion at next time step
}
\description{
Calculates the biomass of carrion (dead animals) at the next timestep from
the current biomass.
}
\details{
The inflow of carrion biomass consists of
\itemize{
\item An external inflow at a rate given by \code{carrion_external}.
\item Discards from fishing.
\item Animals killed by fishing gear.
\item Animals that have died by causes other than predation.
}

The outflow of biomass is due to
\itemize{
\item consumption by scavenger species, where the encounter rate is
specified by \code{params@rho[, "carrion", ]}.
}
}
