% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cdf_y_label}
\alias{cdf_y_label}
\title{Y-axis label for a cumulative-distribution plot}
\usage{
cdf_y_label(power, normalise)
}
\arguments{
\item{power}{The power of weight that the abundance was multiplied by.}

\item{normalise}{Whether the cumulative distribution is normalised.}
}
\value{
A character string for the y-axis label.
}
\description{
Y-axis label for a cumulative-distribution plot
}
\keyword{internal}
