% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{changeBMort}
\alias{changeBMort}
\title{Set background mortality rate}
\usage{
changeBMort(params, mu_b = NULL, z0pre = 0.6, z0exp = params@n - 1)
}
\arguments{
\item{params}{MizerParams}

\item{mu_b}{Optional. An array (species x size) holding the background
mortality rate.}

\item{z0pre}{If \code{z0}, the mortality from other sources, is not a column
in the species data frame, it is calculated as z0pre * w_inf ^ z0exp.
Default value is 0.6.}

\item{z0exp}{If \code{z0}, the mortality from other sources, is not a column
in the species data frame, it is calculated as \code{z0pre * w_inf ^ z0exp}.
Default value is \code{n-1}.}
}
\value{
MizerParams
}
\description{
Set background mortality rate
}
\section{Setting background mortality rate}{

The background mortality is all the mortality that is not due to either
predation or fishing. 

If the \code{mu_b} argument is not supplied, then the background mortality
is assumed to depend only on the asymptotic size of the species, not on the
size of the individual: \eqn{\mu_{b.i}(w) = z_{0.i}}. The value of the
constant \eqn{z_0} for each species is taken from the \code{z0} column of the
species_params data frame, if that column exists. Otherwise it is calculated
as \deqn{z_{0.i} = {\tt z0pre}_i\, w_{inf}^{\tt z0exp}.}{z_{0.i} = z0pre_i w_{inf}^{z0exp}.}
}

