% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{changeInitial}
\alias{changeInitial}
\title{Change initial abundances}
\usage{
changeInitial(params, initial_n = params@initial_n,
  initial_n_pp = params@initial_n_pp, initial_B = params@initial_B,
  sim)
}
\arguments{
\item{params}{A \code{MizerParams} object}

\item{initial_n}{The initial abundances of species. A matrix with dimensions
species x size. The order of species must be the same as in the MizerParams
argument. Optional. Ignored if `sim` is supplied.}

\item{initial_n_pp}{The initial abundances of plankton. A numeric vector.
Optional. Ignored if `sim` is supplied.}

\item{initial_B}{The initial biomasses of the unstructured resources. A named
vector with one entry for each resource. Optional. Ignored if `sim` is
supplied.}

\item{sim}{A \code{MizerSim} object. If supplied, the `initial_n`, 
`initial_n_pp` and `initial_B` arguments are ignored and the information
is taken from the last timestep of the simulation in `sim`.}
}
\description{
Changes the slots in the \code{MizerParams} object holding the initial
abundances, \code{initial_n}, \code{initial_n_pp} and \code{initial_B}.
}
\seealso{
Other functions for changing parameters: \code{\link{changeBMort}},
  \code{\link{changeFishing}},
  \code{\link{changeIntakeMax}},
  \code{\link{changeInteraction}},
  \code{\link{changeMetab}}, \code{\link{changeParams}},
  \code{\link{changePlankton}},
  \code{\link{changePredKernel}},
  \code{\link{changeReproduction}},
  \code{\link{changeResourceDynamics}},
  \code{\link{changeResourceEncounter}},
  \code{\link{changeSearchVolume}}
}
\concept{functions for changing parameters}
