% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{changePlankton}
\alias{changePlankton}
\title{Set up plankton}
\usage{
changePlankton(params, kappa = params@kappa, lambda = params@lambda,
  r_pp = 10, w_pp_cutoff = 10,
  plankton_dynamics = plankton_semichemostat)
}
\arguments{
\item{params}{A MizerParams object}

\item{kappa}{Carrying capacity of the plankton spectrum.}

\item{lambda}{Exponent of the plankton spectrum.}

\item{r_pp}{Growth rate of the primary productivity. Default is 10 g/year.}

\item{w_pp_cutoff}{The upper cut off size of the plankton spectrum. 
Default is 10 g.}

\item{plankton_dynamics}{Function that determines plankton dynamics by
calculating the plankton spectrum at the next time step from the current
state. The default is \code{"\link{plankton_semichemostat}"}.}
}
\value{
A MizerParams object
}
\description{
Set up plankton
}
\section{Setting plankton dynamics}{

Still need to document
}

