% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{changeResourceEncounter}
\alias{changeResourceEncounter}
\title{Set resource encounter rate}
\usage{
changeResourceEncounter(params, rho = NULL, n = params@n)
}
\arguments{
\item{params}{A MizerParams object}

\item{rho}{Optional. An array (species x resource x size)
holding the rate at which a consumer of a particular size feeds on each
resource. Described in the section "Setting resource encounter rate".}

\item{n}{Scaling exponent of the intake rate.}
}
\value{
A MizerParams object
}
\description{
Set resource encounter rate
}
\section{Setting resource encounter rate}{

The resource encounter rate \eqn{\rho_{id}(w)} determines the rate at which
an individual of species \eqn{i} encounters biomass of resource \eqn{d},
\deqn{\sum_d\rho_{id}(w) B_d,} where \eqn{B_d} is the biomass of the d-th
unstructured resource component. 
See \code{\link{resource_dynamics}} for more details.

The ordering of the entries in the array \code{rho} is important. The order
of the species in the first array dimension needs to be the same as that in
the species parameter dataframe. The order of the resources in the second
array dimension must be the same as in the list of resource dynamics. The
third dimension, if given, is the size dimension.

If the \code{rho} array is not supplied, then the resource encounter rate is
set to a power law
\deqn{\rho_{id}(w) = \rho_{id} w^n.}
The coefficients \eqn{\rho_{id}} are parameters in the species_params dataframe.
For example if there is a resource called "detritus" then the species_params
data frame needs to have a column called \code{rho_detritus} and similarly
for each other resource.
}

