% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{changeResources}
\alias{changeResources}
\title{Set resource dynamics}
\usage{
changeResources(params, resource_dynamics, resource_params)
}
\arguments{
\item{params}{A MizerParams object}

\item{resource_dynamics}{A named list of functions that determine the
dynamics of the unstructured resources by calculating their biomasses at
the next time step from the current state. See
\code{\link{resource_dynamics}} for details. An empty list if the model
does not have unstructured resources.}

\item{resource_params}{A list of parameters needed by the
\code{resource_dynamics} functions. An empty list if no parameters are
needed.}
}
\description{
Set resource dynamics
}
\section{Setting resource dynamics}{

Still need to document
}

