% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{changeSearchVolume}
\alias{changeSearchVolume}
\title{Set search volume}
\usage{
changeSearchVolume(params, search_vol = NULL, q = params@q)
}
\arguments{
\item{params}{MizerParams}

\item{search_vol}{Optional. An array (species x size) holding the search volume
for each species at size. If not supplied, a default is set as described in
the section "Setting search volume".}

\item{q}{Exponent of the allometric search volume. Not needed if a
\code{search_vol} array is specified.}
}
\value{
MizerParams
}
\description{
Set search volume
}
\section{Setting search volume}{

The search volume \eqn{\gamma_i(w)} of an individual of species \eqn{i}
and weight \eqn{w} multiplies the predation kernel when
calculating the encounter rate in \code{\link{getEncounter}} and the 
predation rate in \code{\link{getPredRate}}.

If the \code{search_vol} argument is not supplied, then the search volume is set to
\deqn{\gamma_i(w) = \gamma_i w^q.} The values of \eqn{\gamma_i} are taken from
the \code{gamma} column in the species parameter dataframe. If the \code{gamma}
column is not supplied in the species parameter dataframe, a default is
calculated by the \code{\link{get_gamma_default}} function.
}

\examples{
\dontrun{
params <- set_trait_model()
params@species_params$gamma[3] <- 1000
params <- changeSearchVolume(params)
}
}
