% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{check_band_vars}
\alias{check_band_vars}
\title{Check that a data frame has the variables a band style needs}
\usage{
check_band_vars(var_names, style)
}
\arguments{
\item{var_names}{The names of the variables in the data frame.}

\item{style}{The style that was requested.}
}
\value{
Nothing; called for the error.
}
\description{
Check that a data frame has the variables a band style needs
}
\keyword{internal}
