% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_species.R
\name{checkAddSpeciesParams}
\alias{checkAddSpeciesParams}
\title{Validate Added Species Parameters}
\usage{
checkAddSpeciesParams(params, species_params, gear_params, interaction)
}
\arguments{
\item{params}{A mizer params object for the original system.}

\item{species_params}{The species parameters for the new species.}

\item{gear_params}{The gear parameters for the new species.}

\item{interaction}{Optional interaction matrix.}
}
\value{
A list with validated parameters.
}
\description{
Checks and processes parameters needed by addSpecies().
}
