% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{checkSpeciesParams}
\alias{checkSpeciesParams}
\title{Check that species parameter dataframe has essential variables}
\usage{
checkSpeciesParams(species_params)
}
\description{
Checks for the presence of the following variables:
\itemize{
\item \code{species} The names of the species
\item \code{w_inf}   The maximum sizes of species in grams
\item \code{w_min}   The egg sizes of species in grams
}
}
