% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{clearExtensionChain}
\alias{clearExtensionChain}
\title{Clear the registered extension chain}
\usage{
clearExtensionChain()
}
\value{
Invisibly, an empty character vector.
}
\description{
Clears the session's extension registry. You can then create a new
extension chain with \code{\link[=registerExtensions]{registerExtensions()}}.
}
\seealso{
The \href{https://sizespectrum.org/mizer/articles/guide-use-extension-packages.html}{guide to using mizer extension packages}

Other extension tools:
\code{\link[=NOther]{NOther()}},
\code{\link[=coerceToExtensionClass]{coerceToExtensionClass()}},
\code{\link[=getRegisteredExtensions]{getRegisteredExtensions()}},
\code{\link[=initialNOther<-]{initialNOther<-()}},
\code{\link[=other_mort]{other_mort()}},
\code{\link[=recordExtension]{recordExtension()}},
\code{\link[=registerExtension]{registerExtension()}},
\code{\link[=registerExtensions]{registerExtensions()}},
\code{\link[=setComponent]{setComponent()}},
\code{\link[=setRateFunction]{setRateFunction()}}
}
\concept{extension tools}
