% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_species.R
\name{combineGearParams}
\alias{combineGearParams}
\title{Combine Gear Parameters}
\usage{
combineGearParams(params, gear_params, species_params)
}
\arguments{
\item{params}{A mizer params object for the original system.}

\item{gear_params}{The gear parameters for the new species.}

\item{species_params}{The new species parameters.}
}
\value{
A data frame with combined gear parameters.
}
\description{
Merges new gear parameters with existing ones for addSpecies().
}
