% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_species.R
\name{combineSpeciesParams}
\alias{combineSpeciesParams}
\title{Combine Species Parameters}
\usage{
combineSpeciesParams(params, given_species_params, species_params)
}
\arguments{
\item{params}{A mizer params object for the original system.}

\item{given_species_params}{User-given original species parameters.}

\item{species_params}{The new species parameters.}
}
\value{
A list containing combined parameters.
}
\description{
Combines existing and new species parameters for addSpecies().
}
