% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady.R
\name{constant_other}
\alias{constant_other}
\title{Helper function to keep other components constant}
\usage{
constant_other(params, n_other, component, ...)
}
\description{
Helper function to keep other components constant
}
