% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerScan-class.R
\name{copy_scan_attributes}
\alias{copy_scan_attributes}
\title{Copy the metadata of a MizerScan onto another object}
\usage{
copy_scan_attributes(to, from)
}
\arguments{
\item{to}{The object to receive the attributes.}

\item{from}{The MizerScan to take them from.}
}
\value{
\code{to} with the attributes and class of a MizerScan.
}
\description{
Copy the metadata of a MizerScan onto another object
}
\keyword{internal}
