% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPredKernel.R
\name{default_pred_kernel_params}
\alias{default_pred_kernel_params}
\title{Set defaults for predation kernel parameters}
\usage{
default_pred_kernel_params(object)
}
\arguments{
\item{object}{Either a MizerParams object or a species parameter data frame}
}
\value{
The \code{object} with updated columns in the species params data frame.
}
\description{
If the predation kernel type has not been specified for a species, then it
is set to "lognormal" and the default values are set for the parameters
\code{beta} and \code{sigma}.
}
\concept{helper}
\keyword{internal}
