% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{defaultBMort}
\alias{defaultBMort}
\title{Set default background mortality rate}
\usage{
defaultBMort(params, z0pre, z0exp)
}
\arguments{
\item{params}{MizerParams}

\item{z0pre}{If \code{z0}, the mortality from other sources, is not a column
in the species data frame, it is calculated as z0pre * w_inf ^ z0exp.
Default value is 0.6.}

\item{z0exp}{If \code{z0}, the mortality from other sources, is not a column
in the species data frame, it is calculated as \code{z0pre * w_inf ^ z0exp}.
Default value is \code{n-1}.}
}
\value{
MizerParams
}
\description{
Still need to document
}
