% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{defaultIntakeMax}
\alias{defaultIntakeMax}
\title{Set default maximum intake rate}
\usage{
defaultIntakeMax(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
MizerParams
}
\description{
Still need to document
}
