% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{defaultMetab}
\alias{defaultMetab}
\title{Set default metabolic rate}
\usage{
defaultMetab(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
MizerParams
}
\description{
Still need to document
}
