% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{defaultPredKernel}
\alias{defaultPredKernel}
\title{Set default feeding kernel}
\usage{
defaultPredKernel(params, store_kernel)
}
\arguments{
\item{params}{MizerParams}

\item{store_kernel}{A boolean flag that determines whether the full
feeding kernel is stored. If FALSE, only its Fourier transforms are stored.
The default is TRUE if the number of size bins is no larger than 100 and
FALSE otherwise.}
}
\value{
MizerParams
}
\description{
Still need to document
}
