% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{defaultResourceEncounter}
\alias{defaultResourceEncounter}
\title{Set default resource encounter rate}
\usage{
defaultResourceEncounter(params, rho)
}
\arguments{
\item{params}{A MizerParams object}

\item{rho}{An array (species x resource) that gives the
rate \eqn{\rho_{id}} that determines the rate at which species \eqn{i}
encounters biomass of resource \eqn{d}.}
}
\value{
A MizerParams object
}
\description{
This sets up an allometric rate
\deqn{\rho_{id}(w) = \rho_{id} w^n.}
The total contribution of the resources to the rate at which an individual of
species \eqn{i} and weight \eqn{w} encounters biomass is
\deqn{\sum_d\rho_{id} w^n B_d,} where \eqn{B_d} is the biomass of the d-th
unstructured resource component. See \code{\link{resource_dynamics}} help for
more details.
}
\details{
The ordering of the rows and columns of the array \code{rho} is important.
The order of the species in the first array dimension (columns) needs to
be the same as that in the species parameter dataframe. The order of the
resources in the second array dimension (rows) must be the same as in the
list of resource dynamics.
}
