% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults_edition.R
\name{defaults_edition}
\alias{defaults_edition}
\title{Default editions}
\usage{
defaults_edition(edition = NULL)
}
\arguments{
\item{edition}{NULL or a numerical value.}
}
\value{
If \code{edition} is \code{NULL}, the currently active edition number. If
\code{edition} is supplied, the function sets the global
\code{mizer_defaults_edition} option, emits a message, and returns the supplied
value invisibly.
}
\description{
Function to set and get which edition of default choices is being used.
}
\details{
The mizer functions for creating new models make a lot of choices for default
values for parameters that are not provided by the user. Sometimes we find
better ways to choose the defaults and update mizer accordingly. When we do
this, we will increase the edition number.

If you call \code{defaults_edition()} without an argument it returns the
currently active edition. Otherwise it sets the active edition to the
given value.

Users who want their existing code for creating models not to change
behaviour when run with future versions of mizer should explicitly set the
desired defaults edition at the top of their code.

The most recent edition is edition 2. It will become the default in the
next release. The current default is edition 1. The following defaults
are changed in edition 2:
\itemize{
\item \code{catchability} = 0.3 instead of 1
\item \verb{initial effort} = 1 instead of 0
\item \code{gamma} is set to ensure a feeding level of \code{f0} for larvae with the
current value of \code{interaction_resource} instead of
interaction_resource = 1`.
\item \code{initial_n} is set using \code{\link[=get_steady_state_n]{get_steady_state_n()}} instead of the rather
arbitrary old choice.
}
}
