% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{defaultSearchVol}
\alias{defaultSearchVol}
\title{Set default search volume}
\usage{
defaultSearchVol(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
MizerParams
}
\description{
Still need to document
}
