% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_dynamics.R
\name{detritus_dynamics}
\alias{detritus_dynamics}
\title{Detritus dynamics}
\usage{
detritus_dynamics(params, n, n_pp, B, rates, dt, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{The biomass of detritus}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step}

\item{...}{Unused}
}
\value{
Biomass of detritus at next time step
}
\description{
Calculates the detritus biomass at the next timestep from the current
detritus biomass.
}
\details{
The inflow of biomass consists of
\itemize{
\item an external inflow at a rate given by
\code{params@resource_params$detritus_external} and
\item an inflow of feces, calculated as a proportion
\code{params@resource_params$detritus_proportion} of the biomass consumed.
}

The outflow of biomass is due to
\itemize{
\item consumption by detritivorous species, where the encounter rate is
specified by \code{params@rho[, "detritus", ]}.
}

This function can be used as the project_resources function in the case
where there is only a single unstructured resource that represents
detritus, or as part of a more comprehensive function when there are also
other unstructured resource components. It is used for example in the
\code{\link[=dead_matter_dynamics]{dead_matter_dynamics()}} function.
}
