% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{different}
\alias{different}
\title{Check whether two objects are different}
\usage{
different(a, b)
}
\arguments{
\item{a}{First object}

\item{b}{Second object}
}
\value{
TRUE or FALSE
}
\description{
Check whether two objects are numerically different, ignoring all attributes
}
