% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeIntegral.R
\name{dim_extents}
\alias{dim_extents}
\title{Collect the extent of each labelled dimension}
\usage{
dim_extents(arrays, labels)
}
\arguments{
\item{arrays}{A list of arrays.}

\item{labels}{A list of the corresponding label vectors.}
}
\value{
A named integer vector giving the extent of each label.
}
\description{
Internal helper for \code{\link[=sizeIntegral]{sizeIntegral()}} that checks that arrays sharing a
dimension label agree on its extent.
}
\concept{helper}
\keyword{internal}
