% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{display_frames}
\alias{display_frames}
\title{Display frames}
\usage{
display_frames(f1, f2, params, y_ticks = 6)
}
\arguments{
\item{f1}{Data frame for left plot}

\item{f2}{Data frame for right plot}

\item{params}{A MizerParams object}

\item{y_ticks}{The approximate number of ticks desired on the y axis}
}
\value{
ggplot2 object
}
\description{
Display frames
}
