% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{displayFrames}
\alias{displayFrames}
\title{Display frames}
\usage{
displayFrames(f1, f2, params, xlab = NA, ylab = NA, y_ticks = 6)
}
\arguments{
\item{f1}{Data frame for left plot}

\item{f2}{Data frame for right plot}

\item{params}{A MizerParams object}

\item{xlab}{Label for x-axis. Defaults to first variable name.}

\item{ylab}{Label for y-axis. Defaults to third variable name.}

\item{y_ticks}{The approximate number of ticks desired on the y axis}
}
\value{
A plot
}
\description{
Takes two data frames with plotting data and displays them side-by-side,
using the same axes and legend.
}
\details{
The two data frames each need to have the same three variables. The first
variable will go on the x-axis, the third on the y-axis with a logarithmic
scale. The second variable should be the species and will be used to group
the data and display with the linetype and linecolour specified by the
\code{linetype} and \code{linecolour} slots of the \code{params} object.
}
\seealso{
Other frame functions: 
\code{\link{getBiomassFrame}()}
}
\concept{frame functions}
