% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.hasSlot}
\alias{.hasSlot}
\title{S4/S3 compatibility .hasSlot helper}
\usage{
.hasSlot(object, name)
}
\arguments{
\item{object}{An object (S4 or S3).}

\item{name}{Name of the slot / list element.}
}
\value{
Logical.
}
\description{
S4/S3 compatibility .hasSlot helper
}
\keyword{internal}
