% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{.tpo2_g}
\alias{.tpo2_g}
\title{Oxygen downregulation helper}
\usage{
.tpo2_g(params, t, f, h_t, M_A_star)
}
\arguments{
\item{params}{A \code{MizerParams} object with the tpo2 extension enabled.}

\item{t}{Numeric time.}

\item{f}{Raw feeding level.}

\item{h_t}{Temperature-adjusted maximum intake rate.}

\item{M_A_star}{Intended active metabolism.}
}
\value{
A list with \code{g}, \code{D_approx}, \code{pO2_int_approx}, \code{M_M} and \code{G_O}.
}
\description{
Oxygen downregulation helper
}
\keyword{internal}
