% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{.tpo2_get_env}
\alias{.tpo2_get_env}
\title{Get temperature and oxygen forcing}
\usage{
.tpo2_get_env(params, t)
}
\arguments{
\item{params}{A \code{MizerParams} object with the tpo2 extension enabled.}

\item{t}{Numeric time.}
}
\value{
A list with scalar entries \code{T_t} and \code{pO2_env_t}.
}
\description{
Get temperature and oxygen forcing
}
\keyword{internal}
