% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{.tpo2_go}
\alias{.tpo2_go}
\title{Oxygen conductance term}
\usage{
.tpo2_go(params, t)
}
\arguments{
\item{params}{A \code{MizerParams} object with the tpo2 extension enabled.}

\item{t}{Numeric time.}
}
\value{
A species x size array.
}
\description{
Oxygen conductance term
}
\keyword{internal}
