% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{.tpo2_ma_max}
\alias{.tpo2_ma_max}
\title{Temperature-adjusted active-metabolism asymptote}
\usage{
.tpo2_ma_max(params, t)
}
\arguments{
\item{params}{A \code{MizerParams} object with the tpo2 extension enabled.}

\item{t}{Numeric time.}
}
\value{
A species x size array.
}
\description{
Temperature-adjusted active-metabolism asymptote
}
\keyword{internal}
