% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{.tpo2_mm}
\alias{.tpo2_mm}
\title{Temperature-adjusted standard metabolism}
\usage{
.tpo2_mm(params, t)
}
\arguments{
\item{params}{A \code{MizerParams} object with the tpo2 extension enabled.}

\item{t}{Numeric time.}
}
\value{
A species x size array.
}
\description{
Temperature-adjusted standard metabolism
}
\keyword{internal}
