% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{.tpo2_theta_q10}
\alias{.tpo2_theta_q10}
\title{Q10 temperature multiplier}
\usage{
.tpo2_theta_q10(T_t, T_ref, Q10)
}
\arguments{
\item{T_t}{Temperature.}

\item{T_ref}{Reference temperature.}

\item{Q10}{Q10 coefficient.}
}
\value{
Numeric Q10 multiplier.
}
\description{
Q10 temperature multiplier
}
\keyword{internal}
