% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{.tpo2_u_from_f}
\alias{.tpo2_u_from_f}
\title{Intended activity scalar from feeding level}
\usage{
.tpo2_u_from_f(f)
}
\arguments{
\item{f}{Feeding level.}
}
\value{
Feeding-level-shaped intended activity scalar clipped to \link{0, 1}.
}
\description{
Intended activity scalar from feeding level
}
\keyword{internal}
