% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{effort}
\alias{effort}
\title{Fishing effort used in simulation}
\usage{
effort(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
An array (time x gear) that stores the fishing effort by time and 
  gear.
}
\description{
Fishing effort used in simulation
}
