% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{emptyParams}
\alias{emptyParams}
\title{Basic constructor that creates empty MizerParams object of the right size}
\usage{
emptyParams(object, min_w = 0.001, max_w = 1000, no_w = 100,
  min_w_pp = 1e-10, no_w_pp = NA, species_names = 1:object,
  gear_names = species_names)
}
\arguments{
\item{object}{Number of species}

\item{min_w}{Smallest weight}

\item{max_w}{Largest weight}

\item{no_w}{Number of weight brackets}

\item{min_w_pp}{Smallest plankton weight}

\item{no_w_pp}{No longer used}

\item{species_names}{Names of species}

\item{gear_names}{Names of gears}
}
\value{
An empty but valid MizerParams object
}
\description{
Basic constructor that creates empty MizerParams object of the right size
}
