% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{ensureExtensionNamespaces}
\alias{ensureExtensionNamespaces}
\title{Load (and optionally install) namespaces for all non-NA extensions}
\usage{
ensureExtensionNamespaces(extensions, install = FALSE)
}
\arguments{
\item{extensions}{Named character vector of extensions.}

\item{install}{Logical. If \code{TRUE}, install or upgrade missing/outdated
packages via \code{\link[pak:pkg_install]{pak::pkg_install()}}.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Load (and optionally install) namespaces for all non-NA extensions
}
\keyword{internal}
