% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_species.R
\name{expandSizeGrid}
\alias{expandSizeGrid}
\alias{expandSizeGrid.MizerParams}
\title{Expand the size grid}
\usage{
expandSizeGrid(params, ...)

\method{expandSizeGrid}{MizerParams}(
  params,
  new_min_w = min(params@w),
  new_max_w = max(params@w),
  preserve_species = params@species_params$species,
  ...
)
}
\arguments{
\item{params}{A \link{MizerParams} object.}

\item{...}{Additional arguments (currently unused).}

\item{new_min_w}{The new minimum size in the grid. Defaults to the current minimum.}

\item{new_max_w}{The new maximum size in the grid. Defaults to the current maximum.}

\item{preserve_species}{A vector of species names for which all rate arrays
should be copied over to the new params object rather than being
re-calculated from the species parameters. If missing, all species are
preserved.}
}
\value{
A new \link{MizerParams} object with the updated size grid.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function expands the size grid in a \link{MizerParams} object to the desired
min and max size, preserving all existing species. The function is deprecated
because you can achieve the same more flexibly with \code{adjustSizeGrid()}.
}
\concept{deprecated}
