% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{final_n}
\alias{final_n}
\alias{final_n_pp}
\alias{final_idx}
\title{Consumer size spectra at end of simulation}
\usage{
final_n(sim)

final_n_pp(sim)

final_idx(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
For \code{final_n()}: An array (species x size) holding the consumer
number densities at the end of the simulation

For \code{final_n_pp()}: A vector holding the plankton number densities at
the end of the simulation for all size classes

For \code{final_idx()}: An integer giving the index for extracting the
results for the final time step
}
\description{
Consumer size spectra at end of simulation

Plankton size spectrum at end of simulation

Time index at end of simulation
}
\examples{
\dontrun{
sim <- project(NS_params, t_max = 12, t_save = 0.5)
idx <- final_idx(sim)
idx
# This coincides with
length(times(sim))
# and corresponds to the final time
times(sim)[idx]
# We can use this index to extract the result at the final time
identical(n(sim)[idx, , ], final_n(sim))
identical(n_pp(sim)[idx, ], final_n_pp(sim))
}
}
