% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{final_n_other}
\alias{final_n_other}
\title{Values of other ecosystem components at end of simulation}
\usage{
final_n_other(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
A named list holding the values of other ecosystem components at the
  end of the simulation
}
\description{
Values of other ecosystem components at end of simulation
}
