% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{final_n_pp}
\alias{final_n_pp}
\title{Plankton size spectrum at end of simulation}
\usage{
final_n_pp(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
A vector holding the plankton number densities at the end of the
  simulation for all size classes
}
\description{
Plankton size spectrum at end of simulation
}
