% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{finalN}
\alias{finalN}
\alias{finalNResource}
\title{Size spectra at end of simulation}
\usage{
finalN(sim)

finalNResource(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
For \code{finalN()}: An array (species x size) holding the consumer
number densities at the end of the simulation

For \code{finalNResource()}: A vector holding the resource number
densities at the end of the simulation for all size classes
}
\description{
Size spectra at end of simulation
}
\examples{
str(finalN(NS_sim))

# This could also be obtained using `N()` and `idxFinalT()`
identical(N(NS_sim)[idxFinalT(NS_sim), , ], finalN(NS_sim))
str(finalNResource(NS_sim))
}
\seealso{
\code{\link[=idxFinalT]{idxFinalT()}}
}
