% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{finalNOther}
\alias{finalNOther}
\title{Values of other ecosystem components at end of simulation}
\usage{
finalNOther(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
A named list holding the values of other ecosystem components at the
end of the simulation
}
\description{
Values of other ecosystem components at end of simulation
}
