% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{finalParams}
\alias{finalParams}
\title{Extract the final state from a simulation}
\usage{
finalParams(sim)
}
\arguments{
\item{sim}{A \code{MizerSim} object.}
}
\value{
A \code{MizerParams} object with initial values taken from the final time
step of the simulation.
}
\description{
Returns the \code{MizerParams} object underlying the simulation with its initial
abundances set to the abundances at the \emph{last} saved time step of the
simulation. This is a convenience wrapper around \code{\link[=getParams]{getParams()}} with no
\code{time_range} argument (the default).
}
\examples{
sim <- project(NS_params, t_max = 20, effort = 0.5)
params_end <- finalParams(sim)
}
\seealso{
\code{\link[=getParams]{getParams()}}, \code{\link[=initialParams]{initialParams()}}
}
