% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{flux_units}
\alias{flux_units}
\title{Units string for the flux returned by \code{\link[=getFlux]{getFlux()}}}
\usage{
flux_units(power)
}
\arguments{
\item{power}{The power of weight the flux was multiplied by.}
}
\value{
A character string with the units of the flux.
}
\description{
Units string for the flux returned by \code{\link[=getFlux]{getFlux()}}
}
\keyword{internal}
