% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_setters.R
\name{free_gear_name}
\alias{free_gear_name}
\title{A gear name that the model is not already using}
\usage{
free_gear_name(params)
}
\arguments{
\item{params}{A MizerParams object.}
}
\value{
A string naming a gear that does not exist in \code{params}.
}
\description{
A gear name that the model is not already using
}
\keyword{internal}
