% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_setters.R
\name{gear_names}
\alias{gear_names}
\title{Every gear name a model uses}
\usage{
gear_names(params)
}
\arguments{
\item{params}{A MizerParams object.}
}
\value{
A character vector of gear names.
}
\description{
Every gear name a model uses
}
\keyword{internal}
