% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPredKernel.R
\name{get_phi}
\alias{get_phi}
\title{Get values from feeding kernel function}
\usage{
get_phi(species_params, ppmr)
}
\arguments{
\item{species_params}{A species parameter data frame}

\item{ppmr}{Values of the predator/prey mass ratio at which to evaluate the
predation kernel function}
}
\value{
An array (species x ppmr) with the values of the predation kernel
function
}
\description{
This involves finding the feeding kernel function for each species, using the
pred_kernel_type parameter in the species_params data frame, checking that it
is valid and all its arguments are contained in the species_params data
frame, and then calling this function with the ppmr vector.
}
\concept{helper}
\keyword{internal}
