% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{get_required_reproduction}
\alias{get_required_reproduction}
\title{Determine reproduction rate needed for initial egg abundance}
\usage{
get_required_reproduction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A vector of reproduction rates for all species
}
\description{
Determine reproduction rate needed for initial egg abundance
}
