% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{get_size_range_array}
\alias{get_size_range_array}
\title{Get size range array}
\usage{
get_size_range_array(params, min_w = min(params@w),
  max_w = max(params@w), min_l = NULL, max_l = NULL, ...)
}
\arguments{
\item{params}{MizerParams object}

\item{min_w}{Smallest weight in size range. Defaults to smallest weight in
the model.}

\item{max_w}{Largest weight in size range. Defaults to largest weight in the
model.}

\item{min_l}{Smallest length in size range. If supplied, this takes
precedence over \code{min_w}.}

\item{max_l}{Largest length in size range. If supplied, this takes precedence
over \code{max_w}.}
}
\value{
Boolean array (species x size)
}
\description{
Helper function that returns an array (species x size) of boolean values
indicating whether that size bin is within the size limits specified by the
arguments.
}
