% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getAvailEnergy}
\alias{getAvailEnergy}
\title{Get available energy}
\usage{
getAvailEnergy(object, n, n_pp)
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}
}
\value{
A two dimensional array (predator species x predator size)
}
\description{
Calculates the amount \eqn{E_{a,i}(w)} of food exposed to each predator as
a function of predator size.
}
\details{
This function is used by the \code{\link{project}} method for
performing simulations.

The function returns values also for sizes outside the size-range of the
species. These values should not be trusted, as they are meaningless.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
n <- sim@n[21,,]
n_pp <- sim@n_pp[21,]
getAvailEnergy(params,n,n_pp)
}
}
\seealso{
\code{\link{project}}
}
