% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{getComponent}
\alias{getComponent}
\title{Get information about other ecosystem components}
\usage{
getComponent(params, component)
}
\arguments{
\item{params}{A MizerParams object}

\item{component}{Name of the component of interest. If missing, a list of
all components will be returned.}
}
\value{
A list with the entries \code{initial_value}, \code{dynamics_fun},
\code{encounter_fun}, \code{mort_fun}, \code{component_params} for the requested
component. If the requested component does not exist, \code{NULL} is returned.
If no \code{component} argument is given, then a list of lists for all
components is returned.
}
\description{
Get information about other ecosystem components
}
