% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{getDiffusion}
\alias{getDiffusion}
\title{Get diffusion rate from predation}
\usage{
getDiffusion(
  object,
  n = initialN(object),
  n_pp = initialNResource(object),
  n_other = initialNOther(object),
  t = 0,
  ...
)
}
\arguments{
\item{object}{A \linkS4class{MizerParams} or \linkS4class{MizerSim} object}

\item{n}{A matrix of species abundances (species x size). Defaults to the
initial abundances in \code{object}.}

\item{n_pp}{A vector of the resource abundance by size. Defaults to the
initial resource abundances in \code{object}.}

\item{n_other}{A list of abundances for other dynamical components.}

\item{t}{The time for which to do the calculation.}

\item{...}{Unused}
}
\value{
An array of dimensions species x size holding the diffusion rates.
}
\description{
Calculates the diffusion rate \eqn{D_i(w)} (grams^2/year) for each species.
This is the rate at which the abundance density is diffused along the
size axis due to the variability in prey sizes. This is the diffusion
term from the jump-growth equation.
}
\references{
Datta, S., Delius, G. W. and Law, R. (2010). A jump-growth model for
predator-prey dynamics: derivation and application to marine ecosystems.
Bulletin of Mathematical Biology, 72(6):1361–1382
}
\seealso{
Other rate functions:
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFlux]{getFlux()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getRates]{getRates()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
