% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{getDiscardsByGear}
\alias{getDiscardsByGear}
\title{Calculate the rate at which biomass of each species is discarded by each gear}
\usage{
getDiscardsByGear(object)
}
\arguments{
\item{object}{An object of class \code{MizerParams} or \code{MizerSim}.}
}
\value{
If called with a MizerParams object, an array (gear x species) with
the discard rate in grams per year from each gear for each species in the
model. If called with a MizerSim object, an array (time x gear x species)
containing the discard rate at each time step.
}
\description{
For details of how the discard rate is defined see the help page of
\code{\link[=getDiscards]{getDiscards()}}.
}
\examples{
discards <- getDiscardsByGear(NS_sim)
discards["1972", "Herring", "Herring"]
# (In this example MizerSim object each species was set up with its own gear)
}
\seealso{
\code{\link[=getDiscards]{getDiscards()}}

Other summary functions: 
\code{\link{getBiomass}()},
\code{\link{getDiet}()},
\code{\link{getGrowthCurves}()},
\code{\link{getN}()},
\code{\link{getSSB}()},
\code{\link{getYield}()},
\code{\link{getYieldGear}()}
}
\concept{summary functions}
\concept{summary_function}
