% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getFeedingLevel}
\alias{getFeedingLevel}
\title{Get feeding level}
\usage{
getFeedingLevel(object, ...)
}
\arguments{
\item{object}{A \linkS4class{MizerParams} or \linkS4class{MizerSim} object.}

\item{...}{Additional arguments that depend on the class of \code{object}.

\strong{For a \linkS4class{MizerParams} object:}
\describe{
\item{\code{n}}{A matrix of species abundances (species x size). Defaults to
the initial abundances stored in \code{object}.}
\item{\code{n_pp}}{A vector of the resource abundance by size. Defaults to the
initial resource abundance stored in \code{object}.}
\item{\code{n_other}}{A named list of the abundances of other dynamical
components. Defaults to the initial values stored in \code{object}.}
\item{\code{t}}{The time for which to do the calculation. Defaults to 0.}
}

\strong{For a \linkS4class{MizerSim} object:}
\describe{
\item{\code{time_range}}{The time range over which to return the rates. Either
a vector of values, a vector of min and max time, or a single value.
Defaults to the whole time range of the simulation.}
\item{\code{drop}}{If \code{TRUE} then any dimension of length 1 is removed from the
returned array.}
}}
}
\value{
\itemize{
\item \code{MizerParams}: An \code{ArraySpeciesBySize} object (predator species x predator
size) with the feeding level.
\item \code{MizerSim}: An \code{ArrayTimeBySpeciesBySize} object (time step x predator
species x predator size) with the feeding level at every time step.
If \code{drop = TRUE} then dimensions of length 1 will be removed.
}
}
\description{
Returns the feeding level.
By default this function uses \code{\link[=mizerFeedingLevel]{mizerFeedingLevel()}} to calculate
the feeding level, but this can be overruled via \code{\link[=setRateFunction]{setRateFunction()}}.
}
\section{Feeding level}{

The feeding level \eqn{f_i(w)} is the
proportion of its maximum intake rate at which the predator is actually
taking in fish. It is calculated from the encounter rate \eqn{E_i} and the
maximum intake rate \eqn{h_i(w)} as
\deqn{f_i(w) = \frac{E_i(w)}{E_i(w)+h_i(w)}.}{E_i(w)/(E_i(w)+h_i(w)).}
The encounter rate \eqn{E_i} is passed as an argument or calculated with
\code{\link[=getEncounter]{getEncounter()}}. The maximum intake rate \eqn{h_i(w)} is
taken from the \code{params} object, and is set with
\code{\link[=setMaxIntakeRate]{setMaxIntakeRate()}}.
As a consequence of the above expression for the feeding level,
\eqn{1-f_i(w)} is the proportion of the food available to it that the
predator actually consumes.
}

\section{Your own feeding level function}{

By default \code{\link[=getFeedingLevel]{getFeedingLevel()}} calls \code{\link[=mizerFeedingLevel]{mizerFeedingLevel()}}. However you can
replace this with your own alternative feeding level function. If
your function is called \code{"myFeedingLevel"} then you register it in a MizerParams
object \code{params} with

\if{html}{\out{<div class="sourceCode">}}\preformatted{params <- setRateFunction(params, "FeedingLevel", "myFeedingLevel")
}\if{html}{\out{</div>}}

Your function will then be called instead of \code{\link[=mizerFeedingLevel]{mizerFeedingLevel()}}, with the
same arguments.
}

\examples{
\donttest{
params <- NS_params
# Get initial feeding level
fl <- getFeedingLevel(params)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the feeding level at all saved time steps
fl <- getFeedingLevel(sim)
# Get the feeding level for years 15 - 20
fl <- getFeedingLevel(sim, time_range = c(15, 20))
}
}
\seealso{
Other rate functions: 
\code{\link{getDiffusion}()},
\code{\link{getEGrowth}()},
\code{\link{getERepro}()},
\code{\link{getEReproAndGrowth}()},
\code{\link{getEncounter}()},
\code{\link{getFMort}()},
\code{\link{getFMortGear}()},
\code{\link{getFlux}()},
\code{\link{getFluxGradient}()},
\code{\link{getMort}()},
\code{\link{getPredMort}()},
\code{\link{getPredRate}()},
\code{\link{getRDD}()},
\code{\link{getRDI}()},
\code{\link{getRates}()},
\code{\link{getResourceMort}()}
}
\concept{rate functions}
