% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getFlux}
\alias{getFlux}
\alias{getFlux.MizerParams}
\alias{getFlux.MizerSim}
\title{Get flux into size bins}
\usage{
getFlux(object, ..., power = 0)
}
\arguments{
\item{object}{A \linkS4class{MizerParams} or \linkS4class{MizerSim} object.}

\item{...}{Additional arguments that depend on the class of \code{object}.

\strong{For a \linkS4class{MizerParams} object:}
\describe{
\item{\code{n}}{A matrix of species abundances (species x size). Defaults to
the initial abundances stored in \code{object}.}
\item{\code{n_pp}}{A vector of the resource abundance by size. Defaults to the
initial resource abundance stored in \code{object}.}
\item{\code{n_other}}{A named list of the abundances of other dynamical
components. Defaults to the initial values stored in \code{object}.}
\item{\code{t}}{The time for which to do the calculation. Defaults to 0.}
}

\strong{For a \linkS4class{MizerSim} object:}
\describe{
\item{\code{time_range}}{The time range over which to return the rates. Either
a vector of values, a vector of min and max time, or a single value.
Defaults to the whole time range of the simulation.}
\item{\code{drop}}{If \code{TRUE} then any dimension of length 1 is removed from the
returned array.}
}}

\item{power}{The flux at weight \eqn{w} is multiplied by \eqn{w} raised to
\code{power}. The default \code{power = 0} gives the flux of individuals
(numbers/year), whereas \code{power = 1} gives the flux of biomass
(grams/year).}
}
\value{
\itemize{
\item \code{MizerParams}: An \code{ArraySpeciesBySize} object (species x size) with the
flux entering each size class. The units are \code{numbers/year} when
\code{power = 0} and \code{g^power/year} otherwise.
\item \code{MizerSim}: An \code{ArrayTimeBySpeciesBySize} object (time step x species x
size) with the flux at every time step. If \code{drop = TRUE} then dimensions
of length 1 will be removed.
}
}
\description{
Calculates the flux \eqn{J_i(w)} (numbers/year) entering each size class
from the one below it. This is composed of an advective flux from somatic
growth and a diffusive flux from the redistribution of individuals.
}
\details{
At the recruitment size, the flux is simply the recruitment rate
\eqn{R_{dd,i}} (see \code{\link[=getRDD]{getRDD()}}). For sizes below the recruitment size
the flux is zero.

The flux at weight \eqn{w} is multiplied by \eqn{w} raised to the \code{power}
given by the \code{power} argument, similar to the \code{power} argument of
\code{\link[=plotSpectra]{plotSpectra()}}. The default \code{power = 0} returns the flux of individuals
(numbers/year). With \code{power = 1} the result is the flux of biomass
(grams/year).
}
\examples{
\donttest{
params <- NS_params
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the flux at a particular time step
flux <- getFlux(params, n = N(sim)[15, , ], n_pp = NResource(sim)[15, ], t = 15)
# Flux for Sprat of size 2g
flux["Sprat", "2"]
}
}
\seealso{
\code{\link[=getEGrowth]{getEGrowth()}}, \code{\link[=getRDD]{getRDD()}}

Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFluxGradient]{getFluxGradient()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getRates]{getRates()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
