% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getFluxGradient}
\alias{getFluxGradient}
\alias{getFluxGradient.MizerParams}
\title{Get flux gradient}
\usage{
getFluxGradient(object, ...)
}
\arguments{
\item{object}{A \linkS4class{MizerParams} or \linkS4class{MizerSim} object.}

\item{...}{Additional arguments that depend on the class of \code{object}.

\strong{For a \linkS4class{MizerParams} object:}
\describe{
\item{\code{n}}{A matrix of species abundances (species x size). Defaults to
the initial abundances stored in \code{object}.}
\item{\code{n_pp}}{A vector of the resource abundance by size. Defaults to the
initial resource abundance stored in \code{object}.}
\item{\code{n_other}}{A named list of the abundances of other dynamical
components. Defaults to the initial values stored in \code{object}.}
\item{\code{t}}{The time for which to do the calculation. Defaults to 0.}
}

\strong{For a \linkS4class{MizerSim} object:}
\describe{
\item{\code{time_range}}{The time range over which to return the rates. Either
a vector of values, a vector of min and max time, or a single value.
Defaults to the whole time range of the simulation.}
\item{\code{drop}}{If \code{TRUE} then any dimension of length 1 is removed from the
returned array.}
}}
}
\value{
An \code{ArraySpeciesBySize} object (species x size) giving the flux
divergence in each size bin, in units of \eqn{g^{-1} \, \text{year}^{-1}}.
}
\description{
Calculates the flux divergence \eqn{(J_{j+1} - J_j)/\Delta w_j} that
appears as the second term in the discretised size-spectrum transport
equation
\deqn{\frac{\partial N_j}{\partial t} + \frac{J_{j+1} - J_j}{\Delta w_j}
      = -\mu_j N_j.}
The bin-boundary fluxes \eqn{J_j} are obtained from \code{\link[=getFlux]{getFlux()}}, which
uses the advective-flux scheme stored in the \code{flux} entry of the
\code{\link{second_order_w}} slot of \code{params}.  The flux leaving
the largest size class through the upper boundary (\eqn{J_{K+1}}) is
evaluated with the same scheme using the boundary condition
\eqn{N_{K+1} = 0}.
}
\examples{
\donttest{
params <- NS_params
fg <- getFluxGradient(params)
}
}
\seealso{
\code{\link[=getFlux]{getFlux()}}, \code{\link[=second_order_w]{second_order_w()}}

Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFlux]{getFlux()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getRates]{getRates()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
