% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInteraction.R
\name{getInteraction}
\alias{getInteraction}
\title{Deprecated function to get interaction matrix}
\usage{
getInteraction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\description{
You should now use \code{\link[=interaction_matrix]{interaction_matrix()}} instead.
}
\keyword{internal}
