% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{getMaxPPMR}
\alias{getMaxPPMR}
\title{Get maximum predator/prey mass ratio for each species}
\usage{
getMaxPPMR(species_params)
}
\arguments{
\item{species_params}{}
}
\value{
Vector with the maximum predator/prey ratio for each species
}
\description{
Because of how we cut of the Gaussian feeding kernel, the maximum
predator/prey mass ratio is \eqn{\beta * \exp(3 \sigma)}.
}
\details{
This is used in \code{\link{emptyParams}} when setting up the size grid to
determine the smallest relevant plankton size. The reason we have a function
to calculate this is that it makes it easier for users to modify when they
choose their own feeding kernel.
}
