% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getPlanktonMort}
\alias{getPlanktonMort}
\title{Get predation mortality rate for plankton}
\usage{
getPlanktonMort(
  params,
  n = params@initial_n,
  n_pp = params@initial_n_pp,
  n_other = params@initial_n_other,
  pred_rate = getPredRate(params, n = n, n_pp = n_pp, n_other = n_other)
)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the plankton abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{pred_rate}{An array of predation rates of dimension no. sp x no.
community size bins x no. of size bins in whole spectra (i.e. community +
plankton, the w_full slot). The array is optional. If it is not provided
it is calculated by the \code{\link{getPredRate}} function.}
}
\value{
A vector of mortality rate by plankton size.
}
\description{
Calculates the predation mortality rate \eqn{\mu_p(w)} on the plankton
spectrum by plankton size (in units 1/year).
}
\details{
Used by the \code{project} function for running size based simulations.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- newMultispeciesParams(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get plankton mortality at one time step
n <- sim@n[21,,]
n_pp <- sim@n_pp[21,]
getPlanktonMort(params,n,n_pp)
}
}
\seealso{
Other rate functions: 
\code{\link{getEGrowth}()},
\code{\link{getEReproAndGrowth}()},
\code{\link{getERepro}()},
\code{\link{getEncounter}()},
\code{\link{getFMortGear}()},
\code{\link{getFMort}()},
\code{\link{getFeedingLevel}()},
\code{\link{getMort}()},
\code{\link{getPredMort}()},
\code{\link{getPredRate}()},
\code{\link{getRDD}()},
\code{\link{getRDI}()},
\code{\link{getRates}()},
\code{\link{getStarvMort}()}
}
\concept{rate functions}
