% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPredKernel.R
\name{getPredKernel}
\alias{getPredKernel}
\title{Get predation kernel}
\usage{
getPredKernel(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
An array (predator species x predator_size x prey_size)
}
\description{
If no explicit predation kernel \eqn{\phi_i(w, w_p)} is stored in the params
object, then this function calculates it from the information in the species
parameter data frame in the params object.
}
\details{
For more detail about the predation kernel see \code{\link[=setPredKernel]{setPredKernel()}}.
}
