% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getRDI}
\alias{getRDI}
\title{Get density independent rate of egg production}
\usage{
getRDI(object, ...)
}
\arguments{
\item{object}{A \linkS4class{MizerParams} or \linkS4class{MizerSim} object.}

\item{...}{Additional arguments that depend on the class of \code{object}.

\strong{For a \linkS4class{MizerParams} object:}
\describe{
\item{\code{n}}{A matrix of species abundances (species x size). Defaults to
the initial abundances stored in \code{object}.}
\item{\code{n_pp}}{A vector of the resource abundance by size. Defaults to the
initial resource abundance stored in \code{object}.}
\item{\code{n_other}}{A named list of the abundances of other dynamical
components. Defaults to the initial values stored in \code{object}.}
\item{\code{t}}{The time for which to do the calculation. Defaults to 0.}
}

\strong{For a \linkS4class{MizerSim} object:}
\describe{
\item{\code{time_range}}{The time range over which to return the rates. Either
a vector of values, a vector of min and max time, or a single value.
Defaults to the whole time range of the simulation.}
}}
}
\value{
\itemize{
\item \code{MizerParams}: A numeric vector the length of the number of species.
\item \code{MizerSim}: An \code{ArrayTimeBySpecies} object (time x species).
}
}
\description{
Calculates the density-independent rate of total egg production
\eqn{R_{di}}{R_di} (units 1/year) before density dependence, by species.
}
\details{
This rate is obtained by taking the per capita rate \eqn{E_r(w)\psi(w)} at
which energy is invested in reproduction, as calculated by \code{\link[=getERepro]{getERepro()}},
multiplying it by the number of individuals\eqn{N(w)} and integrating over
all sizes \eqn{w} and then multiplying by the reproductive efficiency
\eqn{\epsilon} and dividing by the egg size \code{w_min}, and by a factor of two
to account for the two sexes:
\deqn{R_{di} = \frac{\epsilon}{2 w_{min}} \int N(w)  E_r(w) \psi(w) \, dw}{R_di = (\epsilon/(2 w_min)) \int N(w)  E_r(w) \psi(w) dw}

Used by \code{\link[=getRDD]{getRDD()}} to calculate the actual, density dependent rate.
See \code{\link[=setReproduction]{setReproduction()}} for more details.
}
\section{Your own reproduction function}{

By default \code{\link[=getRDI]{getRDI()}} calls \code{\link[=mizerRDI]{mizerRDI()}}. However you can
replace this with your own alternative reproduction function. If
your function is called \code{"myRDI"} then you register it in a MizerParams
object \code{params} with

\if{html}{\out{<div class="sourceCode">}}\preformatted{params <- setRateFunction(params, "RDI", "myRDI")
}\if{html}{\out{</div>}}

Your function will then be called instead of \code{\link[=mizerRDI]{mizerRDI()}}, with the
same arguments. For an example of an alternative reproduction function
see \code{constantEggRDI()}.
}

\examples{
\donttest{
params <- NS_params
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the density-independent reproduction rate at a particular time step
getRDI(params, n = N(sim)[15, , ], n_pp = NResource(sim)[15, ], t = 15)
}
}
\seealso{
\code{\link[=getRDD]{getRDD()}}

Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFlux]{getFlux()}},
\code{\link[=getFluxGradient]{getFluxGradient()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRates]{getRates()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
