% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{getRegisteredExtensions}
\alias{getRegisteredExtensions}
\title{Get the registered mizer extension chain}
\usage{
getRegisteredExtensions()
}
\value{
A named character vector giving the maximal extension chain
registered for this R session.
}
\description{
Get the registered mizer extension chain
}
\seealso{
The \href{https://sizespectrum.org/mizer/articles/guide-use-extension-packages.html}{guide to using mizer extension packages}

Other extension tools:
\code{\link[=NOther]{NOther()}},
\code{\link[=clearExtensionChain]{clearExtensionChain()}},
\code{\link[=coerceToExtensionClass]{coerceToExtensionClass()}},
\code{\link[=initialNOther<-]{initialNOther<-()}},
\code{\link[=other_mort]{other_mort()}},
\code{\link[=recordExtension]{recordExtension()}},
\code{\link[=registerExtension]{registerExtension()}},
\code{\link[=registerExtensions]{registerExtensions()}},
\code{\link[=setComponent]{setComponent()}},
\code{\link[=setRateFunction]{setRateFunction()}}
}
\concept{extension tools}
