% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setResourceSemichemostat.R
\name{getResourceLevel}
\alias{getResourceLevel}
\title{Get resource level}
\usage{
getResourceLevel(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A vector of the same length as \code{w_full(params)} with the resource
level in each size class.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} The resource level is the ratio between
the current resource number density and the resource carrying capacity.
Where both are zero, the resource level is NaN.
}
\examples{
getResourceLevel(NS_params)

# The resource level can be changed without changing the steady state:
params <- setResourceSemichemostat(NS_params, resource_level = 0.9)
getResourceLevel(params)
}
\concept{resource dynamics}
