% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getResourceMort}
\alias{getResourceMort}
\title{Get predation mortality rate for resource}
\usage{
getResourceMort(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  t = 0,
  ...
)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{...}{Unused}
}
\value{
A vector of mortality rate by resource size.
}
\description{
Calculates the predation mortality rate \eqn{\mu_p(w)} on the resource
spectrum by resource size (in units 1/year).
}
\section{Your own resource mortality function}{

By default \code{\link[=getResourceMort]{getResourceMort()}} calls \code{\link[=mizerResourceMort]{mizerResourceMort()}}. However you can
replace this with your own alternative resource mortality function. If
your function is called \code{"myResourceMort"} then you register it in a MizerParams
object \code{params} with\preformatted{params <- setRateFunction(params, "ResourceMort", "myResourceMort")
}

Your function will then be called instead of \code{\link[=mizerResourceMort]{mizerResourceMort()}}, with the
same arguments.
}

\examples{
\dontrun{
params <- newMultispeciesParams(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get resource mortality at one time step
getResourceMort(params, n = N(sim)[15, , ], n_pp = NResource(sim)[15, ])
}
}
\seealso{
Other rate functions: 
\code{\link{getEGrowth}()},
\code{\link{getEReproAndGrowth}()},
\code{\link{getERepro}()},
\code{\link{getEncounter}()},
\code{\link{getFMortGear}()},
\code{\link{getFMort}()},
\code{\link{getFeedingLevel}()},
\code{\link{getMort}()},
\code{\link{getPredMort}()},
\code{\link{getPredRate}()},
\code{\link{getRDD}()},
\code{\link{getRDI}()},
\code{\link{getRates}()}
}
\concept{rate functions}
