% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{getRetainProbGear}
\alias{getRetainProbGear}
\title{Calculate the proportion of fish retained instead of discarded}
\usage{
getRetainProbGear(params)
}
\arguments{
\item{params}{A \code{MizerParams} object}
}
\value{
A matrix (gear x species x size) with the retention probability.
}
\description{
This function calculates the proportion of fish retained instead of discarded
by a fishing gear. The retention probability is given by a sigmoidal
function.
}
\details{
The probability that a fish of length \eqn{l} is retained is given by the
logistic function \deqn{p(l) = \frac{1}{1 + \exp\left(\log(3)\frac{l50
-l}{l50 - l25}\right)}}{p(l) = 1/(1 + exp(log(3)*(l50 -l) / (l50 - l25)))}
where \eqn{l_{50}} and \eqn{l_{25}} are the parameters of the sigmoid. The
sigmoid is defined such that 50\% of the fish are retained at length
\eqn{l_{50}} and 25\% of the fish are retained at length \eqn{l_{25}}. For
each species-gear pair, these parameters are stored in the \code{gear_params} data
frame in the \code{retain_l50} and \code{retain_l25} columns. If the \code{retain_l25} value
is missing, it is assumed to be 1.5 times the \code{retain_l50} value. If the
\code{retain_l50} value is also missing, the retain probability is set to zero.

As the mizer model is weight based, and this selectivity function is length
based, it uses the length-weight parameters \code{a} and \code{b} to convert between
length and weight \deqn{l = \left(\frac{w}{a}\right)^{1/b}}{l = (w/a)^(1/b)}
}
\examples{
retain_prob <- getRetainProb(NS_params)
retain_prob["Pelagic", "Herring", 100]
}
\concept{helper}
