% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{getSSB}
\alias{getSSB}
\title{Calculate the SSB of species}
\usage{
getSSB(object)
}
\arguments{
\item{object}{An object of class \code{MizerParams} or \code{MizerSim}.}
}
\value{
If called with a MizerParams object, a named vector with the SSB in
grams for each species in the model. If called with a MizerSim object, a
\code{ArrayTimeBySpecies} object (time x species) containing the SSB in grams
at each time step for all species.
}
\description{
Calculates the spawning stock biomass (SSB) for each species. For a
\code{MizerSim} object this is returned for every saved time; for a
\code{MizerParams} object it is calculated from the initial state. SSB is the
total mass of all mature individuals.
}
\examples{
ssb <- getSSB(NS_sim)
ssb[c("1972", "2010"), c("Herring", "Cod")]
}
\seealso{
Other summary functions:
\code{\link[=getBiomass]{getBiomass()}},
\code{\link[=getDiet]{getDiet()}},
\code{\link[=getGrowthCurves]{getGrowthCurves()}},
\code{\link[=getN]{getN()}},
\code{\link[=getSteadyResidual]{getSteadyResidual()}},
\code{\link[=getTrophicLevel]{getTrophicLevel()}},
\code{\link[=getTrophicLevelBySpecies]{getTrophicLevelBySpecies()}},
\code{\link[=getYield]{getYield()}},
\code{\link[=getYieldGear]{getYieldGear()}}
}
\concept{summary functions}
\concept{summary_function}
