% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getStarvMort}
\alias{getStarvMort}
\title{Get starvation mortality}
\usage{
getStarvMort(
  params,
  n = params@initial_n,
  n_pp = params@initial_n_pp,
  n_other = params@initial_n_other,
  e = getEReproAndGrowth(params, n = n, n_pp = n_pp, n_other)
)
}
\arguments{
\item{e}{The energy available for reproduction and growth (optional). A
matrix of size no. species x no. size bins. If not supplied, is calculated
internally using the \code{getEReproAndGrowth()} method.}
}
\value{
A two dimensional array of instantaneous starvation mortality
  (species x size).
}
\description{
Calculates starvation mortality based on the equation in the original mizer
vignette. Starvation mortality is proportional to the energy deficiency and
is inversely proportional to body weight (and therefore also lipid reserves).
\eqn{\mu_s(w)} The weight proportionality constant is currently set to 0.1,
but could be a separate parameter. The 0.1 constant means that the
instantaneous starvation mortality is 1 when energy deficit is equal
individual's body mass.
}
\note{
The default value of starv_coef is 10, which scales how energy intake
  deficit (e) translates to starvation mortality. When instantaneous energy
  intake rate deficit per year is 10% of body weight it will give
  instantaneous starvation mortality rate of 1/year, i.e. an average
  individual can survive for about a year. If starv_coef = 0, there is no
  starvation mortality and negative intake just gets converted to 0 (there is
  no cost to having negative energy intake, which is unrealistic!). If
  starv_coef = 20, the cost of negative intake increases, and 10% energy
  deficit will lead to instantaneous starvation mortality of 2/year.
}
\seealso{
Other rate functions: 
\code{\link{getEGrowth}()},
\code{\link{getEReproAndGrowth}()},
\code{\link{getERepro}()},
\code{\link{getEncounter}()},
\code{\link{getFMortGear}()},
\code{\link{getFMort}()},
\code{\link{getFeedingLevel}()},
\code{\link{getMort}()},
\code{\link{getPlanktonMort}()},
\code{\link{getPredMort}()},
\code{\link{getPredRate}()},
\code{\link{getRDD}()},
\code{\link{getRDI}()},
\code{\link{getRates}()}
}
\concept{rate functions}
