% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{getTimes}
\alias{getTimes}
\title{Times for which simulation results are available}
\usage{
getTimes(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
A numeric vectors of the times (in years) at which simulation results
have been stored in the MizerSim object.
}
\description{
Times for which simulation results are available
}
