% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R,
%   R/ArraySpeciesByTime-class.R
\name{ggplotly.ArraySpeciesBySize}
\alias{ggplotly.ArraySpeciesBySize}
\alias{ggplotly.ArraySpeciesByTime}
\title{Interactive ggplotly version of a plot method}
\usage{
\method{ggplotly}{ArraySpeciesBySize}(p, ...)

\method{ggplotly}{ArraySpeciesByTime}(p, ...)
}
\arguments{
\item{p}{An object with a \code{plot()} method.}

\item{...}{Further arguments passed to the \code{plot()} method.}
}
\value{
A plotly object.
}
\description{
Converts the ggplot2 output of the \code{plot()} method into an interactive
plotly figure using \code{\link[=ggplotly]{ggplotly()}}. All arguments are forwarded to
\code{plot()}.
}
\examples{
\donttest{
ggplotly(getEncounter(NS_params))
}
\donttest{
ggplotly(getBiomass(NS_sim))
}
}
