% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{initial_n<-}
\alias{initial_n<-}
\alias{initial_n}
\title{Initial values for fish spectra}
\usage{
initial_n(params) <- value

initial_n(params)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A matrix with dimensions species x size holding the initial
number densities for the fish spectra.}
}
\description{
Values used as starting values for simulations with `project()`.
}
