% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{initial_n_other<-}
\alias{initial_n_other<-}
\alias{initial_n_other}
\title{Initial values for other ecosystem components}
\usage{
initial_n_other(params) <- value

initial_n_other(params)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A named list with the initial values of other ecosystem components}
}
\description{
Values used as starting values for simulations with \code{project()}.
}
