% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{initial_n_pp<-}
\alias{initial_n_pp<-}
\alias{initial_n_pp}
\title{Initial value for plankton spectrum}
\usage{
initial_n_pp(params) <- value

initial_n_pp(params)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A vector with the initial number densities for the plankton
spectrum}
}
\description{
Value used as starting value for simulations with \code{project()}.
}
