% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues.R
\name{initialN<-}
\alias{initialN<-}
\alias{initialN}
\alias{initialN.MizerParams}
\alias{initialN.MizerSim}
\title{Initial values for fish spectra}
\usage{
initialN(params) <- value

initialN(object)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A matrix with dimensions species x size holding the initial
number densities for the fish spectra.}

\item{object}{An object of class MizerParams or MizerSim}
}
\value{
An \code{ArraySpeciesBySize} object with dimensions species x size holding
the initial number densities for the fish spectra.
}
\description{
Values used as starting values for simulations with \code{project()}.
}
\examples{
# Doubling abundance of Cod in the initial state of the North Sea model
params <- NS_params
initialN(params)["Cod", ] <- 2 * initialN(params)["Cod", ]
}
\seealso{
\code{\link[=initialNResource]{initialNResource()}}, \code{\link[=initialNOther]{initialNOther()}}
}
