% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{initialNOther<-}
\alias{initialNOther<-}
\alias{initialNOther}
\title{Initial values for other ecosystem components}
\usage{
initialNOther(params) <- value

initialNOther(object)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A named list with the initial values of other ecosystem
components}

\item{object}{An object of class MizerParams or MizerSim}
}
\value{
A named list with the initial values of other ecosystem
components
}
\description{
Values used as starting values for simulations with \code{project()}.
}
\seealso{
\code{\link[=initialNResource]{initialNResource()}}, \code{\link[=initialN]{initialN()}}

Other extension tools:
\code{\link[=NOther]{NOther()}},
\code{\link[=coerceToExtensionClass]{coerceToExtensionClass()}},
\code{\link[=other_mort]{other_mort()}},
\code{\link[=recordExtension]{recordExtension()}},
\code{\link[=setComponent]{setComponent()}},
\code{\link[=setRateFunction]{setRateFunction()}}
}
\concept{extension tools}
