% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues.R
\name{initialNPlankton<-}
\alias{initialNPlankton<-}
\alias{initialNPlankton}
\title{Initial value for plankton spectrum}
\usage{
initialNPlankton(params) <- value

initialNPlankton(object)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A vector with the initial number densities for the plankton
spectrum}

\item{object}{An object of class MizerParams or MizerSim}
}
\description{
Value used as starting value for simulations with \code{project()}.
}
