% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues.R
\name{initialNResource<-}
\alias{initialNResource<-}
\alias{initialNResource}
\title{Initial value for resource spectrum}
\usage{
initialNResource(params) <- value

initialNResource(object)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A vector with the initial number densities for the resource
spectrum}

\item{object}{An object of class MizerParams or MizerSim}
}
\value{
A vector with the initial number densities for the resource
spectrum
}
\description{
Value used as starting value for simulations with \code{project()}.
}
\examples{
# Doubling resource abundance in the initial state of the North Sea model
params <- NS_params
initialNResource(params) <- 2 * initialNResource(params)
# Of course this initial state will no longer be a steady state
params <- steady(params)
}
