% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{initialParams}
\alias{initialParams}
\title{Extract the initial state from a simulation}
\usage{
initialParams(sim)
}
\arguments{
\item{sim}{A \code{MizerSim} object.}
}
\value{
A \code{MizerParams} object with initial state of the simulation.
}
\description{
Returns the \code{MizerParams} object underlying the simulation with its initial
abundances set to the abundances at the initial time of the
simulation.
}
\examples{
sim <- project(NS_params, t_max = 20, effort = 0.5)
params_start <- initialParams(sim)
}
\seealso{
\code{\link[=getParams]{getParams()}}, \code{\link[=finalParams]{finalParams()}}
}
