% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inter}
\alias{inter}
\title{Alias for \code{NS_interaction}}
\format{
A 12 x 12 matrix.
}
\source{
{Blanchard et al.}
}
\usage{
inter
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
An alias provided for backward compatibility with mizer version <= 2.3
}
\concept{deprecated}
\keyword{datasets}
