% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayResourceBySize-class.R
\name{is.ArrayResourceBySize}
\alias{is.ArrayResourceBySize}
\title{Test if an object is an ArrayResourceBySize}
\usage{
is.ArrayResourceBySize(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{ArrayResourceBySize} object, \code{FALSE} otherwise.
}
\description{
Test if an object is an ArrayResourceBySize
}
\examples{
is.ArrayResourceBySize(getResourceMort(NS_params))
is.ArrayResourceBySize(1:4)
}
