% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R
\name{is.ArraySpeciesBySize}
\alias{is.ArraySpeciesBySize}
\title{Test if an object is a ArraySpeciesBySize}
\usage{
is.ArraySpeciesBySize(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{ArraySpeciesBySize} object, \code{FALSE} otherwise.
}
\description{
Test if an object is a ArraySpeciesBySize
}
\examples{
is.ArraySpeciesBySize(getEncounter(NS_params))
is.ArraySpeciesBySize(matrix(1:4, nrow = 2))
}
