% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesByTime-class.R
\name{is.ArraySpeciesByTime}
\alias{is.ArraySpeciesByTime}
\title{Test if an object is a ArraySpeciesByTime}
\usage{
is.ArraySpeciesByTime(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{ArraySpeciesByTime} object, \code{FALSE} otherwise.
}
\description{
Test if an object is a ArraySpeciesByTime
}
\examples{
is.ArraySpeciesByTime(getBiomass(NS_sim))
is.ArraySpeciesByTime(matrix(1:4, nrow = 2))
}
