% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayResourceBySize-class.R
\name{is.ArrayTimeByResourceBySize}
\alias{is.ArrayTimeByResourceBySize}
\title{Test if an object is an ArrayTimeByResourceBySize}
\usage{
is.ArrayTimeByResourceBySize(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{ArrayTimeByResourceBySize} object, \code{FALSE}
otherwise.
}
\description{
Test if an object is an ArrayTimeByResourceBySize
}
\examples{
is.ArrayTimeByResourceBySize(NResource(NS_sim))
is.ArrayTimeByResourceBySize(matrix(1:4, nrow = 2))
}
