% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayTimeBySpecies-class.R
\name{is.ArrayTimeBySpecies}
\alias{is.ArrayTimeBySpecies}
\title{Test if an object is a ArrayTimeBySpecies}
\usage{
is.ArrayTimeBySpecies(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{ArrayTimeBySpecies} object, \code{FALSE} otherwise.
}
\description{
Test if an object is a ArrayTimeBySpecies
}
\examples{
is.ArrayTimeBySpecies(getBiomass(NS_sim))
is.ArrayTimeBySpecies(matrix(1:4, nrow = 2))
}
