% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayTimeBySpeciesBySize-class.R
\name{is.ArrayTimeBySpeciesBySize}
\alias{is.ArrayTimeBySpeciesBySize}
\title{Test if an object is an ArrayTimeBySpeciesBySize}
\usage{
is.ArrayTimeBySpeciesBySize(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{ArrayTimeBySpeciesBySize} object, \code{FALSE}
otherwise.
}
\description{
Test if an object is an ArrayTimeBySpeciesBySize
}
\examples{
is.ArrayTimeBySpeciesBySize(getFMort(NS_sim))
is.ArrayTimeBySpeciesBySize(array(1:8, dim = c(2, 2, 2)))
}
