% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFishing.R
\name{is.gear_params}
\alias{is.gear_params}
\title{Test if an object is a gear_params object}
\usage{
is.gear_params(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{gear_params} object, \code{FALSE} otherwise.
}
\description{
Test if an object is a gear_params object
}
