% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerRate-class.R
\name{is.MizerRate}
\alias{is.MizerRate}
\title{Test if an object is a MizerRate}
\usage{
is.MizerRate(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{MizerRate} object, \code{FALSE} otherwise.
}
\description{
Test if an object is a MizerRate
}
\examples{
is.MizerRate(getEncounter(NS_params))
is.MizerRate(matrix(1:4, nrow = 2))
}
