% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_params.R
\name{is.species_params}
\alias{is.species_params}
\title{Test if an object is a species_params object}
\usage{
is.species_params(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{species_params} object, \code{FALSE} otherwise.
}
\description{
Test if an object is a species_params object
}
