% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_funcs.R
\name{knife_edge_length}
\alias{knife_edge_length}
\title{Length based knife-edge selectivity function}
\usage{
knife_edge_length(w, knife_edge_length, species_params, ...)
}
\arguments{
\item{w}{Vector of sizes (weights).}

\item{knife_edge_length}{The length at which the knife-edge operates.}

\item{species_params}{A list with the species params for the current species.
Used to get at the length-weight parameters \code{a} and \code{b}.}

\item{...}{Unused}
}
\value{
Vector of selectivities at the given sizes.
}
\description{
A knife-edge selectivity function where individuals with a length greater or
equal to \code{knife_edge_length} are fully selected and no fish shorter than this
length are selected.
}
\details{
You would not usually call this function directly. Instead, set the \code{sel_func}
column in \code{\link[=gear_params]{gear_params()}} to \code{"knife_edge_length"} and provide
\code{knife_edge_length} as an additional column. \code{\link[=setFishing]{setFishing()}} will then call
this function automatically when calculating the selectivity array.

As the mizer model is weight based, and this selectivity function is length
based, it uses the length-weight parameters \code{a} and \code{b} to convert the
cut-off length to a weight:
\deqn{w_{\text{cut}} = a \cdot l_{\text{cut}}^b}
}
\examples{
# Knife-edge at 20 cm using length-weight parameters a = 0.01, b = 3
sp <- list(a = 0.01, b = 3)
knife_edge_length(w = c(1, 10, 100, 1000), knife_edge_length = 20,
                  species_params = sp)
}
\seealso{
\code{\link[=gear_params]{gear_params()}} for setting the \code{knife_edge_length} parameter.

Other selectivity functions:
\code{\link[=double_sigmoid_length]{double_sigmoid_length()}},
\code{\link[=knife_edge]{knife_edge()}},
\code{\link[=sigmoid_length]{sigmoid_length()}},
\code{\link[=sigmoid_weight]{sigmoid_weight()}}
}
\concept{selectivity functions}
