% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_funcs.R
\name{large_predation}
\alias{large_predation}
\title{Extra Mortality From Large Predators}
\usage{
large_predation(res, WW = max(params@species_params$w_inf))
}
\arguments{
\item{res}{the mizer params object.}
}
\description{
Computes the predation mortality that would be induced by extra predators
beyond the maximum size of a fish included in the simulation, under the assumption
that these new large predators have an abundance as described by the 
idealized power law,
}
