% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{log_breaks}
\alias{log_breaks}
\title{Helper function to produce nice breaks on logarithmic axes}
\usage{
log_breaks(n = 6)
}
\arguments{
\item{n}{Approximate number of ticks}
}
\value{
A function that can be used as the break argument in calls to
    scale_y_continuous() or scale_x_continuous()
}
\description{
This is needed when the logarithmic y-axis spans less than one order of
magnitude, in which case the ggplot2 default produces no ticks.
Thanks to Heather Turner at
https://stackoverflow.com/questions/14255533/pretty-ticks-for-log-normal-scale-using-ggplot2-dynamic-not-manual
}
