% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_kernel_funcs.R
\name{lognormal_max_ppmr}
\alias{lognormal_max_ppmr}
\title{Maximum predator/prey mass ratio for lognormal kernel}
\usage{
lognormal_max_ppmr(species_params)
}
\arguments{
\item{species_params}{A species parameter dataframe from which this function
uses the \code{beta} and \code{sigma} slots.}
}
\value{
Vector with the maximum predator/prey ratio for each species
}
\description{
Gives the maximum predator/prey mass ratio at which the predation kernel
computed by \code{\link{lognormal_pred_kernel}} is non-zero.
This is used in \code{\link{set_multispecies_model}} to determine the 
smallest relevant plankton size if the \code{min_w_pp} argument is not
specified there.
}
\details{
Because of how we cut off the Gaussian feeding kernel in
\code{\link{lognormal_pred_kernel}}, the maximum predator/prey mass ratio is
\eqn{\beta  \exp(3 \sigma)}{\beta exp(3 \sigma)}.
}
