% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{markBackground}
\alias{markBackground}
\title{Designate species as background species}
\usage{
markBackground(object, species = NULL)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerParams} or
\linkS4class{MizerSim}.}

\item{species}{The species to be selected. Optional. By default all target
species are selected. A vector of species names, or a
numeric vector with the species indices, or a logical vector indicating for
each species whether it is to be selected (TRUE) or not.}
}
\value{
An object of the same class as the \code{object} argument
}
\description{
Marks the specified set of species as background species by setting the
\code{is_background} column in their species parameters to \code{TRUE}. Background
species are handled differently in plots (displayed in grey) and their
abundances can be automatically adjusted to keep the community close to the
Sheldon spectrum (see \code{adjustBackgroundSpecies()} in the mizerExperimental
package).
}
\examples{
params <- markBackground(NS_params,
                         species = c("Sprat", "Sandeel", "N.pout"))
any(species_params(params)$is_background)
}
\seealso{
\code{\link[=removeBackgroundSpecies]{removeBackgroundSpecies()}}
}
