% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_to}
\alias{match_to}
\title{Match a quantity to observations species by species}
\usage{
match_to(params, species = NULL, to = c("biomass", "number"), fname)
}
\arguments{
\item{params}{A MizerParams object.}

\item{species}{The species to be affected, in any of the forms accepted by
\code{\link[=valid_species_arg]{valid_species_arg()}}.}

\item{to}{The type of observation, either "biomass" or "number".}

\item{fname}{The name of the calling function, used when reporting that the
model has been moved off its steady state.}
}
\value{
A MizerParams object.
}
\description{
Internal implementation shared by \code{\link[=matchBiomasses]{matchBiomasses()}} and \code{\link[=matchNumbers]{matchNumbers()}}.
Multiplies the abundance density of each selected species at all sizes by
the factor that brings the modelled quantity onto the observation. Species
that were not selected, or that have no positive observation, are left
alone.
}
\concept{helper}
\keyword{internal}
