\docType{package}
\name{mizer}
\alias{mizer}
\alias{mizer,}
\alias{mizer-package}
\title{Multi-species size-based modelling in R}
\description{
  mizer implements multi-species size-based modelling in R.
  Specifically for marine ecosystems. Food dependent
  growth. Explore impact of fishing pressure on ecosystem.
  Community and species specific metrics.

  mizer
}
\details{
  Usage

  Set up parameters Set initial population Set fishing
  effort Project forward through time using
  \code{\link{project}} method Interrogate with summaries,
  metrics and plots
}

