% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{mizer_needs_upgrading}
\alias{mizer_needs_upgrading}
\title{Whether the core mizer slots of an object need upgrading}
\usage{
mizer_needs_upgrading(params)
}
\arguments{
\item{params}{A MizerParams object.}
}
\value{
TRUE or FALSE.
}
\description{
Whether the core mizer slots of an object need upgrading
}
\keyword{internal}
