% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{mizerDiffusion}
\alias{mizerDiffusion}
\alias{projectDiffusion}
\alias{projectDiffusion.MizerParams}
\title{Calculate diffusion rate}
\usage{
projectDiffusion(params, n, n_pp, n_other, t = 0, feeding_level, ...)

\method{projectDiffusion}{MizerParams}(params, n, n_pp, n_other, t = 0, feeding_level, ...)

mizerDiffusion(params, n, n_pp, n_other, t = 0, feeding_level, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions.)}

\item{feeding_level}{An array (species x size) with the feeding level.
If not provided, it is calculated from the given abundances.}

\item{...}{Unused}
}
\value{
A two dimensional array (species x size) holding the diffusion rate.
}
\description{
Calculates the diffusion rate \eqn{D_i(w)} (grams^2/year) for each species.
This diffusion rate has two components:
\enumerate{
\item The diffusion due due to the variability in prey sizes. This is the
diffusion term from the jump-growth equation.
\item Any externally specified diffusion, which is added via \code{\link[=setExtDiffusion]{setExtDiffusion()}}
}

You would not usually call this function directly but instead use
\code{\link[=getDiffusion]{getDiffusion()}}, which then calls this function unless an alternative
diffusion rate function has been registered, see \code{\link[=setRateFunction]{setRateFunction()}}.
}
