% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.r
\docType{methods}
\name{MizerParams,data.frame,matrix-method}
\alias{MizerParams,data.frame,matrix-method}
\title{Constructor that takes the species_params data.frame and the interaction matrix}
\usage{
\S4method{MizerParams}{data.frame,matrix}(object, interaction, n = 2/3,
  p = 0.7, q = 0.8, r_pp = 10, kappa = 1e+11, lambda = (2 + q - n),
  w_pp_cutoff = 10, max_w = max(object$w_inf) * 1.1, f0 = 0.6,
  z0pre = 0.6, z0exp = n - 1, ...)
}
\description{
Constructor that takes the species_params data.frame and the interaction matrix
}

