% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.r
\docType{methods}
\name{MizerParams,numeric,missing-method}
\alias{MizerParams,numeric,missing-method}
\title{Basic constructor with only the number of species as dispatching argument}
\usage{
\S4method{MizerParams}{numeric,missing}(object, min_w = 0.001, max_w = 1000,
  no_w = 100, min_w_pp = 1e-10, no_w_pp = round(no_w) * 0.3,
  species_names = 1:object, gear_names = species_names)
}
\description{
Only really used to make MizerParams of the right size and shouldn't be used
by user
}

