% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\docType{class}
\name{MizerParamsVariablePPMR-class}
\alias{MizerParamsVariablePPMR-class}
\title{A class to hold the parameters for a size based model with variable PPMR}
\description{
A subclass of the \linkS4class{MizerParams} class with an additional slot for
a predation kernel that depends on both the predator size and the prey
size.
}
\details{
When the predator/prey mass ratio is size dependent, then the integrals
for calculating available energy and predation death are not convolution
integrals and hence can not be calculated with fast Fourier transforms.
Instead they are calculated by sweeps over a 3-dimensional array holding
the preference function.

The class inherits all slots from the \linkS4class{MizerParams}
class and has one additional slot:
}
\section{Slots}{

\describe{
\item{\code{pred_kernel}}{An array (species x predator size x prey size) that holds
the predation coefficient of each predator at size on each prey size}
}}

