% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{multispeciesParams}
\alias{multispeciesParams}
\title{Construct MizerParams object for multispecies model}
\usage{
multispeciesParams(object, interaction, min_w = 0.001, max_w = NA,
  no_w = 100, min_w_pp = NA, no_w_pp = NA, n = 2/3, p = 0.7,
  q = 0.8, r_pp = 10, kappa = 1e+11, lambda = (2 + q - n),
  w_pp_cutoff = 10, f0 = 0.6, z0pre = 0.6, z0exp = n - 1,
  store_kernel = (no_w <= 100),
  plankton_dynamics = plankton_semichemostat, interaction_p = rep(1,
  nrow(object)), rho = NULL, resource_dynamics = list(),
  resource_params = list())
}
\arguments{
\item{object}{A data frame of species specific parameter values (see notes
below).}

\item{interaction}{Optional argument to specify the interaction matrix of the
species (predator by prey). If missing a default interaction is used where
all interactions between species are set to 1. Note that any dimnames of
the interaction matrix argument are ignored by the constructor. The
dimnames of the interaction matrix in the returned \code{MizerParams}
object are taken from the species names in the \code{species_params} slot.
This means that the order of the columns and rows of the interaction matrix
argument should be the same as the species name in the
\code{species_params} slot.}

\item{min_w}{The smallest size of the community spectrum. By default his is
set to the smallest w_min given in the species_params data frame, or to
0.001g if the w_min slot is missing.}

\item{max_w}{The largest size of the community spectrum. By default this is
set to the largest w_inf specified in the species_params data frame.}

\item{no_w}{The number of size bins in the community spectrum.}

\item{min_w_pp}{The smallest size of the plankton spectrum. By default this
is set so the smallest value at which any of the consumers can feed.}

\item{no_w_pp}{Obsolete argument that is no longer used because the number
of plankton size bins is determined because all size bins have to
be logarithmically equally spaced.}

\item{n}{Scaling of the intake. Default value is 2/3.}

\item{p}{Scaling of the standard metabolism. Default value is 0.7.}

\item{q}{Exponent of the search volume. Default value is 0.8.}

\item{r_pp}{Growth rate of the primary productivity. Default value is 10.}

\item{kappa}{Carrying capacity of the plankton spectrum. Default value is
1e11.}

\item{lambda}{Exponent of the plankton spectrum. Default value is (2+q-n).}

\item{w_pp_cutoff}{The cut off size of the plankton spectrum. Default value
is 10.}

\item{f0}{Average feeding level. Used to calculated \code{h} and \code{gamma}
if those are not columns in the species data frame. Also requires
\code{k_vb} (the von Bertalanffy K parameter) to be a column in the species
data frame. If \code{h} and \code{gamma} are supplied then this argument is
ignored. Default is 0.6.}

\item{z0pre}{If \code{z0}, the mortality from other sources, is not a column
in the species data frame, it is calculated as z0pre * w_inf ^ z0exp.
Default value is 0.6.}

\item{z0exp}{If \code{z0}, the mortality from other sources, is not a column
in the species data frame, it is calculated as \code{z0pre * w_inf ^ z0exp}.
Default value is \code{n-1}.}

\item{store_kernel}{A boolean flag that determines whether the full
feeding kernel is stored. If FALSE, only its Fourier transforms are stored.
The default is TRUE if the number of size bins is no larger than 100 and
FALSE otherwise.}

\item{plankton_dynamics}{Function that determines plankton dynamics by
calculating the plankton spectrum at the next time step from the current
state. The default is \code{"\link{plankton_semichemostat}"}.}

\item{interaction_p}{Vector specifying for each species its interaction with
plankton, similar to what the interaction matrix does for the interaction
with other species. Entries should be numbers between 0 and 1. The 
default is a vector of 1s.}

\item{rho}{Either NULL (default), in which case there will be no unstructured
resources in the model, or an array (species x resource) that gives the
rate \eqn{\rho_{id}} that determines the rate at which species \eqn{i}
encounters biomass of resource \eqn{d}. The rate is assumed to scale
allometrically with size with exponent \code{n}. So the total contribution
of the resources to the rate at which the individual encounters biomass is
\deqn{\sum_d\rho_{id} w^n B_d,}
where \eqn{B_d} is the biomass of the d-th unstructured resource component.
See \code{\link{resource_dynamics}} help for more details.}

\item{resource_dynamics}{A named list of functions that determine the
dynamics of the unstructured resources by calculating their biomasses at
the next time step from the current state. See
\code{\link{resource_dynamics}} for details. Ignored if \code{rho} is NULL.}

\item{resource_params}{A list of parameters needed by the
\code{resource_dynamics} functions. Ignored if \code{rho} is NULL.}

\item{...}{Additional arguments.}
}
\value{
An object of type \linkS4class{MizerParams}
}
\description{
Provides default functional forms for all slots in the
\linkS4class{MizerParams} object based on user-provided species parameters.
}
\note{
The only essential argument is a data frame which contains the species
  data. The data frame is arranged species by parameter, so each column of
  the parameter data frame is a parameter and each row has the parameters for
  one of the species in the model.
  
  There are some essential columns that must be included in the species parameter
  data.frame and that do not have default values. Other columns do have
  default values, so that if they are not included in the species parameter
  data frame, they will be automatically added when the \code{MizerParams}
  object is created. See the accompanying vignette for details of these
  columns.
  
\if{html}{
The following table gives different components of a species parameters data.frame.
\tabular{lll}{
  Column name \tab Description \tab Default value \cr
  species \tab Name of the species \tab Compulsory (no default) \cr
  w_inf \tab The asymptotic mass of the species \tab Compulsory (no default) \cr
  w_mat \tab The maturation mass of the species \tab Compulsory (no default) \cr
  beta \tab Preferred predator prey mass ratio \tab Compulsory (no default) \cr
  sigma \tab Width of prey size preference \tab Compulsory (no default) \cr
  h \tab Maximum food intake rate. If this is not provided, it is calculated using the k_vb column. Therefore, either h or k_vb must be provided. \tab Optional (no default) \cr
  k_vb \tab The von Bertalanffy K parameter. Only used to calculate h if that column is not provided \tab Optional (no default) \cr
  gamma \tab Volumetric search rate. If this is not provided, it is calculated using the h column and other parameters. \tab Optional (no default) \cr
  ks \tab Standard metabolism coefficient \tab h*0.2 \cr
  z0 \tab Background mortality (constant for all sizes). If this is not provided then z0 is calculated as z0pre ∗ w_inf^z0exp. Also z0pre and z0exp have default values of 0.6 and -1/3 respectively. \tab Optional (no default) \cr
  k \tab Activity coefficient \tab 0 \cr
  alpha \tab Assimilation efficiency \tab 0.6 \cr
  erepro \tab Reproductive efficiency \tab 1 \cr
  w_min \tab The size class that recruits are placed in. \tab smallest size class of the species size spectrum
}}
}
\examples{
data(NS_species_params_gears)
data(inter)
params <- multispeciesParams(NS_species_params_gears, inter)
}
\seealso{
\code{\link{project}} \linkS4class{MizerSim}
}
