% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R, R/setInitialValues.R
\name{N}
\alias{N}
\alias{N.MizerParams}
\alias{N<-}
\alias{N<-.MizerParams}
\alias{NResource}
\alias{NResource.MizerParams}
\alias{NResource<-}
\alias{NResource<-.MizerParams}
\alias{initialN}
\alias{initialN.MizerParams}
\alias{initialN.MizerSim}
\alias{initialN<-}
\alias{initialNResource}
\alias{initialNResource.MizerParams}
\alias{initialNResource.MizerSim}
\alias{initialNResource<-}
\title{Consumer and resource size spectra}
\usage{
N(object)

N(object) <- value

NResource(object)

NResource(object) <- value

initialN(object)

initialN(object) <- value

initialNResource(object)

initialNResource(object) <- value
}
\arguments{
\item{object}{A MizerSim or MizerParams object}

\item{value}{A vector with the initial number densities for the resource
spectrum.}
}
\value{
For \code{N()} on a \code{MizerSim}: An \code{ArrayTimeBySpeciesBySize} object
(time x species x size) with the number density of consumers.

For \code{N()} on a \code{MizerParams}: An \code{ArraySpeciesBySize} object
(species x size) with the initial number densities of consumers.

For \code{NResource()} on a \code{MizerSim}: An array (time x size) with the
number density of resource.

For \code{NResource()} on a \code{MizerParams}: A vector with the initial
number densities for the resource spectrum.
}
\description{
Get or set the consumer or resource number densities. For a \code{MizerSim}
object, \code{N()} and \code{NResource()} return the full time series. For a
\code{MizerParams} object they return (and \verb{N<-} / \verb{NResource<-} set) the
initial spectra used as starting values for \code{\link[=project]{project()}}.
}
\details{
\code{initialN()}, \verb{initialN<-()}, \code{initialNResource()}, and
\verb{initialNResource<-()} are plain aliases provided for backwards
compatibility.
}
\examples{
str(N(NS_sim))
str(N(NS_params))
str(NResource(NS_sim))
str(NResource(NS_params))
}
\seealso{
\code{\link[=NResource]{NResource()}}, \code{\link[=initialNOther]{initialNOther()}}
}
