% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{n_other}
\alias{n_other}
\title{Time series of other components}
\usage{
n_other(sim, component)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
A list array (time x component) that stores the projected
  values for other ecosystem components.
}
\description{
Fetch the simulation results for other components over time.
}
