% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{n_pp}
\alias{n_pp}
\title{Time series of plankton size spectrum}
\usage{
n_pp(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
An array (time x size) with the number density of plankton
}
\description{
Fetch the simulation results for the plankton size spectrum over time.
}
