% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{needs_upgrading}
\alias{needs_upgrading}
\title{Determine whether a MizerParams or MizerSim object needs to be upgraded}
\usage{
needs_upgrading(object)
}
\arguments{
\item{object}{A MizerParams or MizerSim object}
}
\value{
TRUE or FALSE
}
\description{
Looks at the mizer version that was used to last update the object and
returns TRUE if changes since that version require an upgrade of the object.
You would not usually have to call this function. Upgrades are initiated
automatically by \code{validParams} and \code{validSim} when necessary.
}
\concept{helper}
