% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{newSheldonParams}
\alias{newSheldonParams}
\title{Set up parameters for a single-species in a Sheldon power-law background}
\usage{
newSheldonParams(
  w_inf = 100,
  w_min = 0.001,
  eta = 10^(-0.6),
  w_mat = w_inf * eta,
  no_w = log10(w_inf/w_min) * 50 + 1,
  n = 3/4,
  p = n,
  lambda = 2.05,
  kappa = 0.005,
  alpha = 0.4,
  ks = 4,
  k_vb = 1,
  beta = 100,
  sigma = 1.3,
  f0 = 0.6,
  gamma = NA,
  bmort_prop = 0,
  rfac = 4,
  ...
)
}
\arguments{
\item{w_inf}{Asymptotic size of species}

\item{w_min}{Egg size of species}

\item{eta}{Ratio between maturity size \code{w_mat} and asymptotic size
\code{w_inf}. Default is 10^(-0.6), approximately 1/4.. Ignored if
\code{w_mat} is supplied explicitly.}

\item{w_mat}{Maturity size of species. Default value is 
\code{eta * w_inf}.}

\item{no_w}{The number of size bins in the community spectrum. These bins 
will be equally spaced on a logarithmic scale. Default value
is such that there are 50 bins for each factor of 10 in weight.}

\item{n}{Scaling exponent of the maximum intake rate. Default value is 2/3.}

\item{p}{Scaling exponent of the standard metabolic rate. By default this is
equal to the exponent \code{n}.}

\item{lambda}{Exponent of the abundance power law.}

\item{kappa}{Coefficient in abundance power law. Default value is
0.005.}

\item{alpha}{The assimilation efficiency of the community. The default value
is 0.4.}

\item{ks}{Standard metabolism coefficient. Default value is 4.}

\item{beta}{Preferred predator prey mass ratio. Default value is 100.}

\item{sigma}{Width of prey size preference. Default value is 1.3.}

\item{f0}{Expected average feeding level. Used to set \code{gamma}, the
coefficient in the search rate. The default value is 0.6. Ignored if 
\code{gamma} is given explicitly.}

\item{gamma}{Volumetric search rate. If not provided, default is determined
by \code{\link{get_gamma_default}} using the value of \code{f0}.}

\item{bmort_prop}{The proportion of the total mortality that comes from
background mortality, i.e., from sources other than predation or fishing. A
number in the interval [0, 1). Default 0.}

\item{rfac}{The factor such that \code{R_max = rfac * R}, where \code{R_max}
is the maximum recruitment allowed and \code{R} is the steady-state
recruitment. Thus the larger \code{rfac} the less the impact of the
non-linear stock-recruitment curve. The default is 4.}

\item{...}{Other arguments to pass to the \code{MizerParams} constructor.}
}
\value{
An object of type \code{MizerParams}
}
\description{
This functions creates a \code{MizerParams} object with a single 
species. This species is embedded in a fixed power-law community spectrum 
\deqn{N_c(w) = \kappa w^{-\lambda}} 
This community provides the food income for the species. Cannibalism is
switched off. The predation mortality arises only from the predators in the
power-law community and it is assumed that the predators in the community
have the same feeding parameters as the foreground species. The function has
many arguments, all of which have default values.
}
\details{
In addition to setting up the parameters, this function also sets up an
initial condition that is close to steady state, under the assumption of
no fishing.

Although the steady state is often stable without imposing a stock
recruitment relationship, the function can set a Beverton-Holt type stock
recruitment relationship that imposes a maximal reproduction rate that is a
multiple of the recruitment rate at steady state. That multiple is set by the
argument \code{rfac}.
}
\examples{
\dontrun{
params <- newSheldonParams()
sim <- project(params, t_max = 5, effort = 0)
plotSpectra(sim)
}
}
\seealso{
Other functions for setting up models: 
\code{\link{newCommunityParams}()},
\code{\link{newMultispeciesParams}()},
\code{\link{newTraitParams}()}
}
\concept{functions for setting up models}
