% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{noRDD}
\alias{noRDD}
\title{Give density-independent reproduction rate}
\usage{
noRDD(rdi, ...)
}
\arguments{
\item{rdi}{Vector of density-independent reproduction rates \eqn{R_p} for
all species.}

\item{...}{Not used.}
}
\value{
Vector of density-dependent reproduction rates.
}
\description{
Simply returns its \code{rdi} argument.
}
\seealso{
Other functions calculating density-dependent reproduction rate: 
\code{\link{BevertonHoltRDD}()},
\code{\link{RickerRDD}()},
\code{\link{SheperdRDD}()},
\code{\link{constantRDD}()}
}
\concept{functions calculating density-dependent reproduction rate}
