% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{NOther}
\alias{NOther}
\alias{finalNOther}
\title{Time series of other components}
\usage{
NOther(sim)

finalNOther(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
For \code{NOther}: A list array indexed by time and component that stores the projected
values for other ecosystem components.

For \code{finalNOther}: A named list holding the values of other ecosystem components at the
end of the simulation
}
\description{
Fetch the simulation results for other components over time.
}
\seealso{
Other extension tools:
\code{\link[=clearExtensionChain]{clearExtensionChain()}},
\code{\link[=coerceToExtensionClass]{coerceToExtensionClass()}},
\code{\link[=getRegisteredExtensions]{getRegisteredExtensions()}},
\code{\link[=initialNOther<-]{initialNOther<-()}},
\code{\link[=other_mort]{other_mort()}},
\code{\link[=recordExtension]{recordExtension()}},
\code{\link[=registerExtension]{registerExtension()}},
\code{\link[=registerExtensions]{registerExtensions()}},
\code{\link[=setComponent]{setComponent()}},
\code{\link[=setRateFunction]{setRateFunction()}}
}
\concept{extension tools}
