% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NS_interaction}
\alias{NS_interaction}
\title{Example interaction matrix for the North Sea example}
\format{
A 12 x 12 matrix.
}
\source{
{Blanchard et al.}
}
\usage{
NS_interaction
}
\description{
The interaction coefficient between predator and prey species
in the North Sea.
}
\examples{
\dontrun{
params <- MizerParams(NS_species_params_gears,
                      interaction = NS_inter)
sim = project(params, effort = c(Industrial = 0, Pelagic = 1, 
                                 Beam = 0.5, Otter = 0.5))
plot(sim)
}
}
\keyword{datasets}
