% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NS_params}
\alias{NS_params}
\title{Example MizerParams object for the North Sea example}
\format{A MizerParams object}
\source{
{Blanchard et al.}
}
\description{
A MizerParams object created from the \code{NS_species_params_gears} species
parameters and the \code{inter} interaction matrix together with an initial
condition corresponding to the steady state obtained from fishing with an
effort 
\code{effort = c(Industrial = 0, Pelagic = 1, Beam = 0.5, Otter = 0.5)}.
}
