% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NS_sim}
\alias{NS_sim}
\title{Example MizerSim object for the North Sea example}
\format{
A MizerSim object
}
\source{
\url{https://sizespectrum.org/mizer/articles/a_multispecies_model_of_the_north_sea.html}
}
\usage{
NS_sim
}
\description{
A MizerSim object containing a simulation with historical fishing
mortalities from the North Sea, as created in the tutorial
"A Multi-Species Model of the North Sea".
}
\examples{
plotBiomass(NS_sim)
}
\seealso{
Other example parameter objects: 
\code{\link{NS_params}}
}
\concept{example parameter objects}
\keyword{datasets}
