% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NS_species_params_gears}
\alias{NS_species_params_gears}
\title{Example species parameter set based on the North Sea with different gears}
\format{
A data frame with 12 rows and 8 columns. Each row is a species.
\describe{
\item{species}{Name of the species}
\item{w_inf}{The von Bertalanffy W_infinity parameter}
\item{w_mat}{Size at maturity}
\item{beta}{Size preference ratio}
\item{sigma}{Width of the size-preference}
\item{R_max}{Maximum reproduction rate}
\item{k_vb}{The von Bertalanffy k parameter}
\item{gear}{Name of the fishing gear}
}
}
\source{
{Blanchard et al.}
}
\description{
This data set is based on species in the North Sea (Blanchard et al.).
It is similar to the data set \code{NS_species_params} except that
this one has an additional column specifying the fishing gear that
operates on each species.
}
\examples{
\dontrun{
params <- MizerParams(NS_species_params_gears)
sim = project(params, effort = c(Industrial = 0, Pelagic = 1, 
                                 Beam = 0.5, Otter = 0.5))
plot(sim)
}
}
