% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{params}
\alias{params}
\title{Extract the parameter object underlying a simulation}
\usage{
params(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
The MizerParams object that was used to run the simulation
}
\description{
Extract the parameter object underlying a simulation
}
