% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankton_dynamics.R
\name{plankton_semichemostat}
\alias{plankton_semichemostat}
\title{Project plankton using semichemostat model}
\usage{
plankton_semichemostat(params, n, n_pp, n_other, rates, t, dt, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{n_other}{A list with the abundances of other components}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{t}{The current time}

\item{dt}{Time step}

\item{...}{Unused}
}
\value{
Vector containing plankton spectrum at next timestep
}
\description{
Project plankton using semichemostat model
}
\examples{
\dontrun{
params <- newMultispeciesParams(NS_species_params_gears, inter,
                                plankton_dynamics = "plankton_semichemostat")
}
}
