% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesByTime-class.R
\name{plot.ArraySpeciesByTime}
\alias{plot.ArraySpeciesByTime}
\title{Plot a ArraySpeciesByTime object}
\usage{
\method{plot}{ArraySpeciesByTime}(
  x,
  params = NULL,
  species = NULL,
  highlight = NULL,
  return_data = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{ArraySpeciesByTime} object.}

\item{params}{A \code{MizerParams} object. Used for species colours and
linetypes. If \code{NULL}, a basic plot is produced.}

\item{species}{Character vector of species to include. \code{NULL} (default) means
all species.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{return_data}{If \code{TRUE}, return the data frame instead of the plot.}

\item{...}{Further arguments (currently unused).}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with columns 'time', 'value', 'Species' is returned.
}
\description{
Plots the value against time for each species, using species colours and
linetypes from the MizerParams object.
}
\examples{
\donttest{
plot(getBiomass(NS_sim), NS_params)
plot(getYield(NS_sim), NS_params, species = c("Cod", "Herring"))
}
}
