% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayResourceBySize-class.R
\name{plot.ArrayTimeByResourceBySize}
\alias{plot.ArrayTimeByResourceBySize}
\title{Plot method for \code{ArrayTimeByResourceBySize} objects}
\usage{
\method{plot}{ArrayTimeByResourceBySize}(x, time = NULL, ...)
}
\arguments{
\item{x}{An \code{ArrayTimeByResourceBySize} object.}

\item{time}{The time to display. Default (\code{NULL}) is the final time
step.}

\item{...}{Passed on to \code{\link[=plot.ArrayResourceBySize]{plot.ArrayResourceBySize()}}.}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a
data frame is returned.
}
\description{
See \code{\link[=plot]{plot()}} for an overview of the mizer plotting system. This method
plots a single time slice, by first extracting it as an
\code{ArrayResourceBySize} object and delegating to
\code{\link[=plot.ArrayResourceBySize]{plot.ArrayResourceBySize()}}, which the further arguments in \code{...} are
passed on to.
}
\examples{
\donttest{
plot(NResource(NS_sim))
}
}
\keyword{internal}
