% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_size_x_var}
\alias{plot_size_x_var}
\title{Name of the x-variable for a given size axis}
\usage{
plot_size_x_var(size_axis)
}
\arguments{
\item{size_axis}{Either \code{"w"} (weight) or \code{"l"} (length).}
}
\value{
\code{"l"} for a length axis, otherwise \code{"w"}.
}
\description{
Name of the x-variable for a given size axis
}
\keyword{internal}
