% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_size_xlim}
\alias{plot_size_xlim}
\title{Choose the x-axis limits for a given size axis}
\usage{
plot_size_xlim(wlim, size_axis, llim = c(NA, NA))
}
\arguments{
\item{wlim}{Numeric vector of length two giving the weight limits.}

\item{size_axis}{Either \code{"w"} (weight) or \code{"l"} (length).}

\item{llim}{Numeric vector of length two giving the length limits.}
}
\value{
\code{llim} for a length axis, otherwise \code{wlim}.
}
\description{
Choose the x-axis limits for a given size axis
}
\keyword{internal}
