% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot biomass or abundance distribution}
\usage{
plotDistribution(params, sp)
}
\arguments{
\item{params}{A \code{MizerParams} object}

\item{sp}{Name or index of species to be plotted}
}
\description{
Where \code{\link{plotSpectra}} plots the density function, this function 
plots the cummulative distribution function. Work in progress.
}
\seealso{
\code{\link{plotting_functions}}

Other plotting functions: \code{\link{displayFrames}},
  \code{\link{plot,MizerSim,missing-method}},
  \code{\link{plotBiomass}}, \code{\link{plotFMort}},
  \code{\link{plotFeedingLevel}},
  \code{\link{plotGrowthCurves}},
  \code{\link{plotPredMort}}, \code{\link{plotSpectra}},
  \code{\link{plotYieldGear}}, \code{\link{plotYield}},
  \code{\link{plotlyBiomass}}, \code{\link{plotlyFMort}},
  \code{\link{plotlyFeedingLevel}},
  \code{\link{plotlyGrowthCurves}},
  \code{\link{plotlyPredMort}},
  \code{\link{plotlySpectra}},
  \code{\link{plotlyYieldGear}}, \code{\link{plotlyYield}},
  \code{\link{plotting_functions}}
}
\concept{plotting functions}
