% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R,
%   R/ArraySpeciesByTime-class.R
\name{plotly}
\alias{plotly}
\alias{plotly.ArraySpeciesBySize}
\alias{plotly.ArraySpeciesByTime}
\title{Interactive plotly version of a plot method}
\usage{
plotly(x, ...)

\method{plotly}{ArraySpeciesBySize}(x, ...)

\method{plotly}{ArraySpeciesByTime}(x, ...)
}
\arguments{
\item{x}{An object with a \code{plot()} method.}

\item{...}{Further arguments passed to the \code{plot()} method.}
}
\value{
A plotly object.
}
\description{
Converts the ggplot2 output of the \code{plot()} method into an interactive
plotly figure using \code{\link[plotly:ggplotly]{plotly::ggplotly()}}. All arguments are forwarded to
\code{plot()}.
}
\examples{
\donttest{
plotly(getEncounter(NS_params))
}
\donttest{
plotly(getBiomass(NS_sim))
}
}
