% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotlyFeedingLevel}
\alias{plotlyFeedingLevel}
\title{Plot the feeding level of species by size with plotly}
\usage{
plotlyFeedingLevel(object, species = NULL, time_range,
  highlight = NULL, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or \linkS4class{MizerParams}.}

\item{species}{Name or vector of names of the species to be plotted. By
default all species are plotted.}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step. Ignored when called with a \linkS4class{MizerParams}
object.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{...}{Other arguments (currently unused)}
}
\value{
A ggplot2 object
}
\description{
After running a projection, plot the feeding level of each species by size. 
The feeding level is averaged over the specified time range (a single value
for the time range can be used).
}
\details{
When called with a \linkS4class{MizerSim} object, the feeding level is averaged
over the specified time range (a single value for the time range can be used
to plot a single time step). When called with a \linkS4class{MizerParams}
object the initial feeding level is plotted.
}
\seealso{
Other plotting functions: \code{\link{displayFrames}},
  \code{\link{plot,MizerSim,missing-method}},
  \code{\link{plotBiomass}}, \code{\link{plotDiet}},
  \code{\link{plotFMort}}, \code{\link{plotFeedingLevel}},
  \code{\link{plotGrowthCurves}},
  \code{\link{plotPredMort}}, \code{\link{plotSpectra}},
  \code{\link{plotYieldGear}}, \code{\link{plotYield}},
  \code{\link{plotlyBiomass}}, \code{\link{plotlyFMort}},
  \code{\link{plotlyGrowthCurves}},
  \code{\link{plotlyPredMort}},
  \code{\link{plotlySpectra}},
  \code{\link{plotlyYieldGear}}, \code{\link{plotlyYield}},
  \code{\link{plotting_functions}}
}
\concept{plotting functions}
