% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotlyFMort}
\alias{plotlyFMort}
\title{Plot total fishing mortality of each species by size with plotly}
\usage{
plotlyFMort(sim, species = dimnames(sim@n)$sp,
  time_range = max(as.numeric(dimnames(sim@n)$time)), print_it = FALSE,
  ...)
}
\arguments{
\item{sim}{An object of class \linkS4class{MizerSim}}

\item{species}{Name or vector of names of the species to be plotted. By
default all species are plotted.}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step.}

\item{print_it}{Display the plot, or just return the ggplot2 object.
Defaults to FALSE}

\item{...}{Other arguments to pass to \code{getM2} method.}
}
\value{
A ggplot2 object
}
\description{
After running a projection, plot the predation mortality rate of each species
by size. The mortality rate is averaged over the specified time range (a
single value for the time range can be used to plot a single time step).
}
\seealso{
\code{\link{getM2}}
}
