% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotlyGrowthCurves}
\alias{plotlyGrowthCurves}
\title{Plot growth curves giving weight as a function of age with plotly}
\usage{
plotlyGrowthCurves(object, species, max_age = 20, percentage = FALSE,
  print_it = FALSE)
}
\arguments{
\item{species}{Name or vector of names of the species to be plotted. By
default all species are plotted.}

\item{print_it}{Display the plot, or just return the ggplot2 object.
Defaults to FALSE}
}
\value{
A ggplot2 object
}
\description{
After running a projection, plot the predation mortality rate of each species
by size. The mortality rate is averaged over the specified time range (a
single value for the time range can be used to plot a single time step).
}
\seealso{
\code{\link{getM2}}
}
