% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotlySpectra}
\alias{plotlySpectra}
\title{Plotly plot of the abundance spectra}
\usage{
plotlySpectra(object, species = NULL, time_range, wlim = c(NA, NA),
  ylim = c(NA, NA), power = 1, biomass = TRUE, total = FALSE,
  plankton = TRUE, background = TRUE, highlight = NULL, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or \linkS4class{MizerParams}.}

\item{species}{Name or vector of names of the species to be plotted. By
default all species are plotted.}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step. Ignored when called with a \linkS4class{MizerParams}
object.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the w axis. Use NA to refer to the existing minimum or maximum.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the y axis. Use NA to refer to the existing minimum or maximum. Any
values below 1e-20 are always cut off.}

\item{power}{The abundance is plotted as the number density times the weight
raised to \code{power}. The default \code{power = 1} gives the biomass 
density, whereas \code{power = 2} gives the biomass density with respect
to logarithmic size bins.}

\item{biomass}{Obsolete. Only used if \code{power} argument is missing. Then
\code{biomass = TRUE} is equivalent to \code{power=1} and 
\code{biomass = FALSE} is equivalent to \code{power=0}}

\item{total}{A boolean value that determines whether the total over all
species in the system is plotted as well. Default is FALSE}

\item{plankton}{A boolean value that determines whether plankton is included.
Default is TRUE.}

\item{background}{A boolean value that determines whether background species
are included. Ignored if the model does not contain background species.
Default is TRUE.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{...}{Other arguments (currently unused)}
}
\value{
A ggplot2 object
}
\description{
Plots the number density multiplied by a power of the weight, with the power
specified by the \code{power} argument.
}
\details{
When called with a \linkS4class{MizerSim} object, the abundance is averaged
over the specified time range (a single value for the time range can be used
to plot a single time step). When called with a \linkS4class{MizerParams}
object the initial abundance is plotted.
}
\seealso{
Other plotting functions: \code{\link{displayFrames}},
  \code{\link{plot,MizerSim,missing-method}},
  \code{\link{plotBiomass}}, \code{\link{plotDiet}},
  \code{\link{plotFMort}}, \code{\link{plotFeedingLevel}},
  \code{\link{plotGrowthCurves}},
  \code{\link{plotPredMort}}, \code{\link{plotSpectra}},
  \code{\link{plotYieldGear}}, \code{\link{plotYield}},
  \code{\link{plotlyBiomass}}, \code{\link{plotlyFMort}},
  \code{\link{plotlyFeedingLevel}},
  \code{\link{plotlyGrowthCurves}},
  \code{\link{plotlyPredMort}},
  \code{\link{plotlyYieldGear}}, \code{\link{plotlyYield}},
  \code{\link{plotting_functions}}
}
\concept{plotting functions}
