% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotlyYield}
\alias{plotlyYield}
\title{Plot the total yield of species through time with plotly}
\usage{
plotlyYield(sim, sim2, species = dimnames(sim@n)$sp, print_it = FALSE,
  total = FALSE, log = TRUE, ...)
}
\arguments{
\item{sim}{An object of class \linkS4class{MizerSim}}

\item{sim2}{An optional second object of class \linkS4class{MizerSim}. If
this is provided its yields will be shown on the same plot in bolder lines.}

\item{species}{Name or vector of names of the species to be plotted. By
default all species contained in \code{sim} are plotted.}

\item{print_it}{Display the plot, or just return the ggplot2 object.
Defaults to FALSE.}

\item{total}{A boolean value that determines whether the total yield from
all species in the system is plotted as well. Default is FALSE.}

\item{log}{Boolean whether yield should be plotted on a logarithmic axis. 
Defaults to true.}

\item{...}{Other arguments to pass to \code{\link{getYield}} method.}
}
\value{
A ggplot2 object
}
\description{
After running a projection, the total yield of each species across all 
fishing gears can be plotted against time. The yield is obtained with
\code{\link{getYield}}.
}
\seealso{
\code{\link{getYield}}
}
