% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotResource}
\alias{plotResource}
\title{Plot the proportion of the resource spectrum(s) compared to
their carrying capacity}
\usage{
plotResource(object, return_data = FALSE)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is FALSE}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the three variables 'w', 'value', 'Resource' is returned.
}
\description{
Plot the proportion of the resource spectrum(s) compared to
their carrying capacity
}
\examples{
\donttest{
plotResource(NS_params)

# Returning the data frame
fr <- plotResource(NS_params, return_data = TRUE)
str(fr)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions: 
\code{\link{animateSpectra}()},
\code{\link{plot,MizerSim,missing-method}},
\code{\link{plotBiomass}()},
\code{\link{plotDeath}()},
\code{\link{plotDiet}()},
\code{\link{plotEnergyBudget}()},
\code{\link{plotFMort}()},
\code{\link{plotFeedingLevel}()},
\code{\link{plotGrowthCurves}()},
\code{\link{plotPredMort}()},
\code{\link{plotResourcePred}()},
\code{\link{plotSpectra}()},
\code{\link{plotYieldGear}()},
\code{\link{plotYieldVsSize}()},
\code{\link{plotYield}()},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
