% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotYieldvsAbundance}
\alias{plotYieldvsAbundance}
\alias{plotlyYieldvsAbundance}
\title{Plot the modelled abundance and catch yield per species through size}
\usage{
plotYieldvsAbundance(
  object,
  species = NULL,
  catch = NULL,
  catch_x = "Weight",
  return_data = FALSE
)

plotlyYieldvsAbundance(
  object,
  species = NULL,
  catch = NULL,
  catch_x = "Weight",
  ...
)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{species}{The name of the predator species for which to plot the mortality.}

\item{catch}{Optional. Data frame holding binned observed catch data. The data can
be binned either into length bins or weight bins. In the former case the data
frame should have columns \code{length} and \code{dl} holding the start of
the size bins in cm and the width of the size bins in cm respectively. In
the latter case the data frame should have columns \code{weight} and
\code{dw} holding the start of the size bins in grams and the width of the
size bins in grams. The data frame also needs to have the columns
\code{species} (the name of the species), \code{catch} (the number of
individuals of a particular species caught in a size bin).}

\item{catch_x}{Choose to display the values using weight data ("Weight") or
size data ("Size"). Default is "Weight".}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is FALSE}

\item{...}{Other arguments (currently unused)}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a list composed of
two slots is returned. First slot is a data frame with the six variables 'w', 'l', 'catch_w',
'catch_l', 'Type', 'Species and the second slot is a data frame with the four variables 'w_mat',
'Species', 'y_coord', 'Type' (to plot vertical lines).
}
\description{
Plot the modelled abundance and catch yield per species through size
}
\examples{
\donttest{
plotYieldvsAbundance(NS_params, species = "Cod")

# Returning the data frame
fr <- plotYieldvsAbundance(NS_params, species = "Cod", return_data = TRUE)
str(fr)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions: 
\code{\link{animateSpectra}()},
\code{\link{plot,MizerSim,missing-method}},
\code{\link{plotBiomass}()},
\code{\link{plotDeath}()},
\code{\link{plotDiet}()},
\code{\link{plotEnergyBudget}()},
\code{\link{plotFMort}()},
\code{\link{plotFeedingLevel}()},
\code{\link{plotGrowthCurves}()},
\code{\link{plotPredMort}()},
\code{\link{plotResourcePred}()},
\code{\link{plotResource}()},
\code{\link{plotSpectra}()},
\code{\link{plotYieldGear}()},
\code{\link{plotYield}()},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
