% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_species.R
\name{prepareInteractionMatrix}
\alias{prepareInteractionMatrix}
\title{Prepare Interaction Matrix}
\usage{
prepareInteractionMatrix(params, species_params, interaction)
}
\arguments{
\item{params}{A mizer params object for the original system.}

\item{species_params}{The species parameters for the new species.}

\item{interaction}{Optional interaction matrix.}
}
\value{
A matrix of interaction coefficients.
}
\description{
Prepares the interaction matrix for addSpecies().
}
