% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_funcs.R
\name{project_dead_animals}
\alias{project_dead_animals}
\title{Project dead animal biomass}
\usage{
project_dead_animals(params, n, n_pp, B, rates, dt)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{The biomass of detritus}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step}

\item{...}{Unused}
}
\value{
Biomass of dead animals at next time step
}
\description{
Calculates the biomass of dead animals at the next timestep from the current
biomass.
}
\details{
The inflow of dead animal biomass consists of
\itemize{
\item An external inflow at a rate given by
\code{params@resource_params$dead_animals_external}.
\item Discards from fishing.
\item Animals killed by fishing gear.
\item Animals that have died by causes other than predation.
}

The outflow of biomass is due to
\itemize{
\item consumption by detritivorous species, where the encounter rate is
specified by \code{params@rho[, "dead_animals"]}.
}

This function can be used as the project_resources function in the case
where there is only a single unstructured resource that represents
dead animals, or as part of a more comprehensive function when there are also
other unstructured resource components. It is used for example in the
\code{\link[=project_dead_matter]{project_dead_matter()}} function.
}
