% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_funcs.R
\name{project_dead_matter}
\alias{project_dead_matter}
\title{Project two-component dead-matter biomass}
\usage{
project_dead_matter(params, n, n_pp, B, rates, dt)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{Vector of biomasses with named components "detritus" and
"dead_animals".}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step}

\item{...}{Unused}
}
\value{
Named vector of biomasses at next time step
}
\description{
This function can be used as the \code{project_resources} function in the case
where there are two unstructured resource components named "detritus"
and "dead_animals". It gives the biomass of these components at the next
time step.
}
\details{
This function in turn calls \code{\link[=project_detritus]{project_detritus()}} and
\code{\link[=project_dead_animals]{project_dead_animals()}}.
}
