% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{project_detritus}
\alias{project_detritus}
\title{Project detritus biomass by one timestep}
\usage{
project_detritus(params, n, n_pp, B, rates, dt, t = NA)
}
\arguments{
\item{params}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{rates}{A list of rates as returned by \code{\link{getRates}}}

\item{dt}{Time step}

\item{t}{Current time}
}
\value{
Biomass of detritus at next time step
}
\description{
This function can be used as the project_resources function in the case
where there is only a single unstructured resource that represents 
detritus.
}
