% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{project_methods}
\alias{project_methods}
\title{Functions for calculating rates used for projecting}
\description{
The functions defined in the file project_methods calculate the various
quantities needed to project the size-spectra forward in time, using the
model described in section 3 of the mizer vignette.
}
\section{List of functions}{

In this list we relate the functions in this file to the quantities named in
the mizer vignette.
\tabular{llll}{
  Function \tab Expression \tab Description \tab Section in vignette\cr
  \code{\link{mizerEncounter}} \tab \eqn{E_{e.i}(w)} \tab Encounter rate \tab 3.2 \cr
  \code{\link{mizerFeedingLevel}} \tab \eqn{f_i(w)} \tab Feeding level \tab 3.3 \cr
  \code{\link{mizerPredRate}} \tab \eqn{\phi_i(w_p/w) (1-f_i(w)) \gamma_i w^q N_i(w) dw} \tab Predation \tab 3.7 \cr
  \code{\link{mizerPredMort}} \tab \eqn{\mu_{p.i}(w)} \tab Predation mortality \tab 3.7 \cr
  \code{\link{mizerPlanktonMort}} \tab \eqn{\mu_{p}(w)} \tab Mortality on plankton \tab 3.8 \cr
  \code{\link{mizerFMortGear}} \tab \eqn{F_{g,i}(w)} \tab Fishing mortality by gear \tab 8.3 \cr
  \code{\link{mizerFMort}} \tab \eqn{\mu_{f.i}(w)} \tab Total fishing mortality \tab 8.3 \cr
  \code{\link{mizerMort}} \tab \eqn{\mu_{i}(w)} \tab Total mortality \tab 3.7 \cr
  \code{\link{mizerEReproAndGrowth}} \tab \eqn{E_{r.i}(w)} \tab Energy put into growth and reproduction \tab 3.4 \cr
  \code{\link{mizerERepro}} \tab \eqn{\psi_i(w)E_{r.i}(w)} \tab Energy put reproduction\tab 3.5 \cr
  \code{\link{mizerEGrowth}} \tab \eqn{g_i(w)} \tab Energy put growth \tab 3.4 \cr
  \code{\link{mizerRDI}} \tab \eqn{R_{p.i}} \tab Density-independent reproduction rate \tab 3.5 \cr
  \code{\link{BevertonHoltRDD}} \tab \eqn{R_i} \tab Density-dependent reproduction rate \tab 3.6 \cr
}
}

