% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{projectRDD}
\alias{projectRDD}
\alias{projectRDD.MizerParams}
\title{Get density-dependent reproduction rate during projection}
\usage{
projectRDD(params, rdi, species_params = params@species_params, t = 0, ...)

\method{projectRDD}{MizerParams}(params, rdi, species_params = params@species_params, t = 0, ...)
}
\arguments{
\item{params}{A MizerParams object.}

\item{rdi}{Vector of density-independent reproduction rates
\eqn{R_{di}}{R_di} for all species.}

\item{species_params}{A species parameter dataframe. Must contain a column
\code{R_max} holding the maximum reproduction rate \eqn{R_{max}}{R_max} for each
species.}

\item{t}{The time for which to do the calculation.}

\item{...}{Unused}
}
\value{
Vector of density-dependent reproduction rates.
}
\description{
S3 generic used by extension-aware projections to calculate the
density-dependent reproduction rate. The base method calls the selected
density-dependence function in \code{params@rates_funcs$RDD}.
}
