% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{pruneSpecies}
\alias{pruneSpecies}
\title{Removes species with abundance below a threshold}
\usage{
pruneSpecies(params, cutoff = 0.001)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{cutoff}{Species with an abundance at maturity size that is less than 
cutoff times community abundance will be removed. Default 1e-3.}
}
\value{
An object of type \code{MizerParams}
}
\description{
This species simply removes the low-abundance species from the params object. 
It does not recalculate the steady state for the remaining species or
retune their reproductive efficiencies.
}
