% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_species.R
\name{renameGear}
\alias{renameGear}
\title{Rename gears}
\usage{
renameGear(params, replace, ...)
}
\arguments{
\item{params}{A mizer params object}

\item{replace}{A named character vector, with new names as values, and old
names as names.}

\item{...}{Additional arguments}
}
\value{
An object of type \linkS4class{MizerParams}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Changes the names of gears in a MizerParams object. This involves for
example changing the gear dimension names of selectivity and catchability
arrays appropriately.
}
\examples{
replace <- c(Industrial = "Trawl", Otter = "Beam_Trawl")
params <- renameGear(NS_params, replace)
gear_params(params)$gear
}
