% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{renameSpecies}
\alias{renameSpecies}
\title{Rename species}
\usage{
renameSpecies(params, replace)
}
\arguments{
\item{params}{A mizer params object}

\item{replace}{A named character vector, with new names as values, and old 
names as names.}
}
\value{
An object of type \linkS4class{MizerParams}
}
\description{
Changes the names of species in a MizerParams object
}
\examples{
\dontrun{
replace <- c(Cod = "Kabeljau", Haddock = "Schellfisch")
params <- renameSpecies(NS_params, replace)
params@species_params$species
}
}
