% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{resetMizerSession}
\alias{resetMizerSession}
\title{Reset the registered extension chain for the current R session}
\usage{
resetMizerSession()
}
\value{
Invisibly, an empty character vector.
}
\description{
Clears the session's extension registry. You can then create a new
extension chain with \code{\link[=registerExtensions]{registerExtensions()}}.
}
\seealso{
Other extensions:
\code{\link[=NOther]{NOther()}},
\code{\link[=coerceToExtensionClass]{coerceToExtensionClass()}},
\code{\link[=getRegisteredExtensions]{getRegisteredExtensions()}},
\code{\link[=initialNOther<-]{initialNOther<-()}},
\code{\link[=registerExtension]{registerExtension()}},
\code{\link[=registerExtensions]{registerExtensions()}},
\code{\link[=setComponent]{setComponent()}},
\code{\link[=setRateFunction]{setRateFunction()}}
}
\concept{extensions}
