% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_dynamics.R
\name{resource_dynamics}
\alias{resource_dynamics}
\alias{resource_dynamics<-}
\alias{getResourceDynamics}
\title{Set or get resource dynamics function}
\usage{
resource_dynamics(params)

resource_dynamics(params) <- value

getResourceDynamics(params)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A string with the name of the resource dynamics function.}
}
\description{
Mizer can model the dynamics of the resource spectrum in different ways,
depending on the choice of the resource dynamics function. This function is
called at each time step to calculate the value of the resource spectrum at
the next time step. Mizer has three built-in resource dynamics functions:
\code{\link[=resource_semichemostat]{resource_semichemostat()}}, \code{\link[=resource_logistic]{resource_logistic()}} and \code{\link[=resource_constant]{resource_constant()}}.
But you can write your own function.
}
\examples{
params <- NS_params
resource_dynamics(params)
resource_dynamics(params) <- "resource_constant"
}
\seealso{
Other resource parameters: 
\code{\link{resource_params}()},
\code{\link{setResourceLogistic}()},
\code{\link{setResourceSemichemostat}()}
}
\concept{resource parameters}
