% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{resource_encounter}
\alias{resource_encounter}
\title{Dummy function used during testing only}
\usage{
resource_encounter(params, n, n_pp, n_other, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list with the abundances of other components}

\item{...}{Unused}
}
\description{
Dummy function used during testing only
}
