% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setResource.R
\name{resource_rate}
\alias{resource_rate}
\alias{resource_rate<-}
\title{Deprecated way of directly setting resource rate}
\usage{
resource_rate(params)

resource_rate(params) <- value
}
\arguments{
\item{value}{.}
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This is deprecated because it does not preserve the steady state. Use
\code{\link[=setResourceSemichemostat]{setResourceSemichemostat()}} instead.
}
\keyword{internal}
