% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{retuneBackground}
\alias{retuneBackground}
\title{Retunes abundance of background species.}
\usage{
retuneBackground(params)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}
}
\value{
An object of type \code{MizerParams}
}
\description{
Rescales all background species in such a way that the total community
spectrum is as close to the Sheldon power law as possible. Background
species that are no longer needed are removed. The reproductive efficiencies
of all species are retuned.
}
\seealso{
markBackground
}
