% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{retuneReproductionEfficiency}
\alias{retuneReproductionEfficiency}
\title{Retune reproduction efficiency to maintain initial egg abundances}
\usage{
retuneReproductionEfficiency(params, species = params@species_params$species)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{species}{A vector of the names of the species to be affected or a
boolean vector indicating for each species whether it is to be affected
(TRUE) or not. By default all species are affected}
}
\value{
A MizerParams object
}
\description{
Sets the reproductive efficiency for all species so that the rate of egg
production exactly compensates for the loss from the first size class due
to growth and mortality. Sets the identical stock recruitment function.
}
