% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerScan-class.R
\name{scan_plot_params}
\alias{scan_plot_params}
\title{The params object to use when plotting a MizerScan}
\usage{
scan_plot_params(x, plot_dat)
}
\arguments{
\item{x}{A MizerScan object.}

\item{plot_dat}{The data frame that will be plotted.}
}
\value{
A MizerParams object with a colour for every series in \code{plot_dat}.
}
\description{
Series that are not species have no colour in the model, and
\code{\link[=plotDataFrame]{plotDataFrame()}} silently drops any legend level it cannot find a colour
for. So any such series is given a colour here, using the ordinary
\code{\link[=setColours]{setColours()}} interface, which also leaves the user free to choose a
different one.
}
\keyword{internal}
