% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerScan-class.R
\name{scan_x_var}
\alias{scan_x_var}
\alias{scan_y_var}
\title{The x and y variables of a MizerScan}
\usage{
scan_x_var(x)

scan_y_var(x)
}
\arguments{
\item{x}{A MizerScan object.}
}
\value{
The name of the column holding the scanned value / the measured
value.
}
\description{
The x and y variables of a MizerScan
}
\keyword{internal}
