% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerScan-class.R
\name{scan_y_label}
\alias{scan_y_label}
\alias{scan_x_label}
\alias{label_with_units}
\title{Axis labels for a MizerScan}
\usage{
scan_y_label(x, default = "Value")

scan_x_label(x, default = "Scan value")

label_with_units(name, units)
}
\arguments{
\item{x}{A MizerScan object.}

\item{default}{The label to use when the name is missing.}

\item{name}{The name of the quantity.}

\item{units}{The units, possibly NULL.}
}
\value{
A string.
}
\description{
Assembles \code{"<name> [<units>]"}, the same way \code{array_y_label()} does for the
array classes.
}
\keyword{internal}
