% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{semichemostat}
\alias{semichemostat}
\title{Dummy function used during testing only}
\usage{
semichemostat(params, n_other, rates, dt, component, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n_other}{A list with the abundances of other components}

\item{rates}{A list of rates as returned by \code{\link[=mizerRates]{mizerRates()}}}

\item{dt}{Time step}

\item{component}{Name of the component that is being updated}

\item{...}{Unused}
}
\description{
Dummy function used during testing only
}
