% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{set_scaling_model}
\alias{set_scaling_model}
\title{Sets up parameters for a scale free trait-based model}
\usage{
set_scaling_model(f0 = 0.6, alpha = 0.4, r_pp = 0.1, n = 2/3, q = 3/4,
  kappa = 7e+10, beta = 100, sigma = 1.3, h = 30, ks = 4,
  no_sp = 11, min_egg = 10^(-4), min_w_pp = min_egg/(beta * exp(5 *
  sigma)), max_w_inf = 10^3, min_w_inf = 10, min_w_mat = 10^(0.4),
  no_w = log10(max_w_inf/min_egg) * 100 + 1, knife_edge_size = 100,
  gear_names = "knife_edge_gear", fac = 11, ...)
}
\arguments{
\item{f0}{Expected average feeding level. Used to set \code{gamma}, the
factor for the search volume. The default value is 0.5.}

\item{alpha}{The assimilation efficiency of the community. The default value
is 0.6}

\item{r_pp}{Growth rate of the primary productivity. Default value is 4.}

\item{n}{Scaling of the intake. Default value is 2/3.}

\item{q}{Exponent of the search volume. Default value is 0.9.}

\item{kappa}{Carrying capacity of the resource spectrum. Default value is
0.005.}

\item{beta}{Preferred predator prey mass ratio. Default value is 100.}

\item{sigma}{Width of prey size preference. Default value is 1.3.}

\item{h}{Maximum food intake rate. Default value is 30.}

\item{ks}{Standard metabolism coefficient. Default value is 4.}

\item{no_sp}{The number of species in the model. The default value is 10. The
more species, the longer takes to run.}

\item{min_egg}{The smallest size of the the egg of a species.}

\item{min_w_pp}{The smallest size of the background spectrum.}

\item{max_w_inf}{The asymptotic size of the largest species in the community.}

\item{min_w_inf}{The asymptotic size of the smallest species in the
community.}

\item{no_w}{The number of size bins in the community spectrum.}

\item{knife_edge_size}{The minimum size at which the gear or gears select
species. Must be of length 1 or no_sp.}

\item{gear_names}{The names of the fishing gears. A character vector, the
same length as the number of species. Default is 1 - no_sp.}

\item{fac}{The factor such that Rmax = fac*RDD, where RDD is the measures 
the inflow of eggs, after the Rmax based stock recruitment relationship 
has been implemented. The default is 11.}

\item{...}{Other arguments to pass to the \code{MizerParams} constructor.}

\item{z0pre}{The coefficient of the background mortality of the community. z0
= z0pre * w_inf ^ (n-1). The default value is 0.6.}
}
\value{
An object of type \code{MizerParams}
}
\description{
This functions creates a \code{MizerParams} object so that scale free trait-based-type 
models can be easily set up and run. The scale free trait-based size spectrum model can
be derived as a simplification of the general size-based model used in
\code{mizer}. The scale free trait-based model is similar to the 
standard trait-based model, with three main differences: 
(1) We have an exact equation for a steady state of this system which is often 
stable, even when we include no extra stabilization effects like density dependence 
or stock recruitment relationships.
(2) The egg size is proportional to the maturity size for each species
(3) The parameters are chosen so that R_0 (the expected number of 
offspring produced by an individual over a lifetime) is close to 1
for each species.
}
\details{
All the species-specific parameters are the same, except for
the egg size, maturity size and asymptotic size. These differ over 
the species, but the ratio of egg size to maturity size, and the ratio of 
egg size to asymptotic size, are the same for each species. The egg sizes 
or maturity sizes or asymptotic sizes) of the species are spread evenly on
a logarithmic scale. See the \code{mizer} vignette
for more details and examples of the scale free trait-based model.

The function has many arguments, all of which have default values. Of
particular interest to the user are the number of species in the model and
the minimum and maximum asymptotic sizes.

The characteristic weights of the different species are defined by 
min_egg,min_w_mat, min_w_inf, max_w_inf and no_sp, in the sense 
that the egg weights of the no_sp species are logarithmically 
evenly spaced, ranging from min_w=min_egg to max_w=max_w_inf. 
The maturity weights of the species can be obtained by muliplying 
the egg_weights by min_w_mat/min_egg. The asymptotic weights 
of the species can be obtained by mulitiplying the egg weights by 
min_w_inf/min_egg.
The no_w weights, which we keep track of the abundance of fish at,
form a logarithmically evenly spaced vector w, ranging from 
min_w to max_w. The vector w_full is obtained by extending w 
down to the size range min_w_pp, where min_w_pp is chosen by default
to be so small that almost no fish can eat it.

Also the scale free trait based model's default steady state is often stable 
without imposing a stock recruitment relationship, by default we do impose a 
Beverton-Holt type stock recruitment relationship. The values of the maximum 
recruitment rmax and the reproduction efficiency erepro 
are chosen so that, at the steady state we construct, the stock recruitment 
adjusted reproduction output equals a given factor, fac, multiplied by rmax.

In addition to setting up the parameters, this code also evaluates 
our analytic expression for a steady state of the scale free trait-based model 
and stores it in the initial_n slot.

The factor for the search volume, \code{gamma}, is calculated using the
expected feeding level, \code{f0}.

The option of including fishing is given, but the steady state may lose its 
natural stability if too much fishing is included. In such a case the user may 
which to include stablizing effects (like Rmax and chi) to ensure 
our steady state is stable. Fishing selectivity is modelled as a knife-edge function with one parameter,
\code{knife_edge_size}, which is the size at which species are selected. Each
species can either be fished by the same gear (\code{knife_edge_size} has a
length of 1) or by a different gear (the length of \code{knife_edge_size} has
the same length as the number of species and the order of selectivity size is
that of the asymptotic size).

The resulting \code{MizerParams} object can be projected forward using
\code{project()} like any other \code{MizerParams} object. When projecting
the community model it may be necessary to reduce \code{dt} to 0.1 to avoid
any instabilities with the solver. You can check this by plotting the biomass
or abundance through time after the projection.
}
\examples{
\dontrun{
s_params <- set_scaling_model()
t_max <- 5
sim <- project(s_params, t_max=t_max, dt=0.01, t_save=t_max/100 ,effort = 0, 
              initial_n = s_params@initial_n, initial_n_pp = s_params@initial_n_pp)
plot(sim)
}
}
\references{
K. H. Andersen and M. Pedersen, 2010, Damped trophic cascades
  driven by fishing in model marine ecosystems. Proceedings of the Royal
  Society V, Biological Sciences, 1682, 795-802.
}
\seealso{
\linkS4class{MizerParams}
}
