% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEmigration.R
\name{setEmigration}
\alias{setEmigration}
\alias{emigration}
\alias{emigration<-}
\title{Set emigration rate}
\usage{
setEmigration(params, emigration = NULL, ...)

emigration(params)

emigration(params) <- value
}
\arguments{
\item{params}{MizerParams}

\item{emigration}{Optional. An array (species x size) holding the emigration rate. If
not supplied, the emigration rate is left unchanged. Initially it is set to 0.}

\item{...}{Unused}

\item{value}{emigration}
}
\value{
\code{setEmigration()}: A MizerParams object with updated emigration rate.

\code{emigration()}: An array (species x size) with the emigration rate.
}
\description{
Set emigration rate
}
\section{Setting emigration rate}{

The emigration rate is the rate at which the abundance density is decreased over
time beyond the decrease described by the per-capita mortality.

The \code{emigration} argument allows you to specify an emigration rate that depends on
species and body size.
}

\seealso{
Other functions for setting parameters: 
\code{\link{gear_params}()},
\code{\link{setExtEncounter}()},
\code{\link{setExtMort}()},
\code{\link{setFishing}()},
\code{\link{setInitialValues}()},
\code{\link{setInteraction}()},
\code{\link{setMaxIntakeRate}()},
\code{\link{setMetabolicRate}()},
\code{\link{setParams}()},
\code{\link{setPredKernel}()},
\code{\link{setReproduction}()},
\code{\link{setSearchVolume}()},
\code{\link{species_params}()}
}
\concept{functions for setting parameters}
