% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\docType{methods}
\name{setErepro}
\alias{setErepro}
\alias{setErepro,MizerParams-method}
\title{Method for setting reproductive efficiency}
\usage{
setErepro(params, ...)

\S4method{setErepro}{MizerParams}(params, erepro)
}
\arguments{
\item{params}{An object of class \code{MizerParams}.}

\item{erepro}{The reproductive efficiency. Either a scalar or a vector with
the length equal to the number of species or, if the reproductive
efficiency is weight-dependent, a matrix with columns corresponding to
species and rows corresponding to the sizes in params@w.}
}
\value{
A \code{MizerParams} object
}
\description{
Method for setting reproductive efficiency

Set the reproductive efficiency
}
