% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setFeedingKernel}
\alias{setFeedingKernel}
\title{Set feeding kernel}
\usage{
setFeedingKernel(params, store_kernel)
}
\arguments{
\item{params}{MizerParams}
}
\value{
MizerParams
}
\description{
Still need to document
}
