% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setInitial}
\alias{setInitial}
\title{Set initial abundances}
\usage{
setInitial(
  params,
  sim,
  initial_n = params@initial_n,
  initial_n_pp = params@initial_n_pp,
  initial_n_other = params@initial_n_other
)
}
\arguments{
\item{params}{A \code{\link{MizerParams}} object}

\item{sim}{A \code{MizerSim} object. If supplied, the `initial_n`, 
`initial_n_pp` and `initial_n_pp` arguments are ignored and the information
is taken from the last timestep of the simulation in `sim`.}

\item{initial_n}{The initial abundances of species. A matrix with dimensions
species x size. The order of species must be the same as in the MizerParams
argument. Optional. Ignored if `sim` is supplied.}

\item{initial_n_pp}{The initial abundances of plankton. A numeric vector.
Optional. Ignored if `sim` is supplied.}

\item{initial_n_other}{The initial abundances of the other dynamic ecosystem
components. A named list with one entry for each component. Optional.
Ignored if `sim` is supplied.}
}
\value{
A MizerParams object
}
\description{
Sets the slots in the \code{MizerParams} object holding the initial
abundances, \code{initial_n}, \code{initial_n_pp} and \code{initial_n_other}.
}
\examples{
\dontrun{
params <- NS_params
sim <- project(params, t_max = 20, effort = 0.5)
params <- setInitial(params, sim)
}
}
\seealso{
Other functions for setting parameters: 
\code{\link{setBMort}()},
\code{\link{setFishing}()},
\code{\link{setIntakeMax}()},
\code{\link{setInteraction}()},
\code{\link{setMetab}()},
\code{\link{setParams}()},
\code{\link{setPlankton}()},
\code{\link{setPredKernel}()},
\code{\link{setReproduction}()},
\code{\link{setSearchVolume}()}
}
\concept{functions for setting parameters}
