% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues.R
\name{setInitialValues}
\alias{setInitialValues}
\title{Set initial values to values from a simulation}
\usage{
setInitialValues(params, sim, time_range, geometric_mean = FALSE, ...)
}
\arguments{
\item{params}{A \code{MizerParams} object in which to set the initial values}

\item{sim}{A \code{MizerSim} object from which to take the values.}

\item{time_range}{The time range to average the abundances over. Can be a
vector of values, a vector of min and max time, or a single value. Only the
range of times is relevant, i.e., all times between the smallest and
largest will be selected.  Default is the final time step.}

\item{geometric_mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
If TRUE then the average of the abundances over the
time range is a geometric mean instead of the default arithmetic mean. This
does not affect the average of the effort or of other components, which is
always arithmetic.}

\item{...}{Additional arguments passed to the method.}
}
\value{
The \code{params} object with updated initial values and initial effort.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated. Use \code{\link[=getParams]{getParams()}}, \code{\link[=initialParams]{initialParams()}}, or
\code{\link[=finalParams]{finalParams()}} instead. These functions return a \code{MizerParams} object
with the ecosystem state extracted from a simulation.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\donttest{
params <- NS_params
sim <- project(params, t_max = 20, effort = 0.5)
params <- setInitialValues(params, sim)
}
}
\concept{deprecated}
