% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setIntakeMax}
\alias{setIntakeMax}
\title{Set maximum intake rate}
\usage{
setIntakeMax(params)
}
\arguments{
\item{params}{MizerParams}

\item{f0}{Optional feeding level}
}
\value{
MizerParams
}
\description{
Still need to document
}
