% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMaxIntakeRate.R
\name{setMaxIntakeRate}
\alias{setMaxIntakeRate}
\alias{intake_max}
\alias{intake_max<-}
\title{Set maximum intake rate}
\usage{
setMaxIntakeRate(params, intake_max = NULL, reset = FALSE, ...)

intake_max(params)

intake_max(params) <- value
}
\arguments{
\item{params}{MizerParams}

\item{intake_max}{Optional. An array (species x size) holding the maximum
intake rate for each species at size. If not supplied, a default is set as
described in the section "Setting maximum intake rate".}

\item{reset}{If set to TRUE, then the intake rate will be reset to the value
calculated from the species parameters, even if it was previously
overwritten with a custom value. If set to FALSE (default) then a
recalculation from the species parameters will take place only if no
custom value has been set.}

\item{...}{Unused}

\item{value}{intake_max}
}
\value{
\code{setMaxIntakeRate()}: A MizerParams object with updated maximum
intake rate.

\code{intake_max()}: An \code{ArraySpeciesBySize} object (species x size) with
the maximum intake rate.
}
\description{
You will usually not need to call this function directly. Instead change
the \code{h} and \code{n} species parameters with \verb{given_species_params(params) <-}
and let mizer recalculate the maximum intake rate for you. Call
\code{setMaxIntakeRate()} directly only if you want to impose a different
functional form for the size dependence of the intake rate. See
\code{vignette("guide-change-parameters")} for a full explanation of when
to reach for which level of the model.
}
\section{Setting maximum intake rate}{

The maximum intake rate \eqn{h_i(w)} of an individual of species \eqn{i} and
weight \eqn{w} determines the feeding level, calculated with
\code{\link[=getFeedingLevel]{getFeedingLevel()}}. It is measured in grams/year.

If the \code{intake_max} argument is not supplied, then the maximum intake
rate is set to \deqn{h_i(w) = h_i w^{n_i}.}
The values of \eqn{h_i} (the maximum intake rate of an individual of size 1
gram) and \eqn{n_i} (the allometric exponent for the intake rate) are taken
from the \code{h} and \code{n} columns in the species parameter dataframe. If
the \code{h} column is not supplied in the species parameter dataframe, it is
calculated by the \code{\link[=get_h_default]{get_h_default()}} function. If the \code{n} column is not
supplied, a default of \eqn{n_i = 3/4} is used.

If \eqn{h_i} is set to \code{Inf}, fish of species i will consume all encountered
food.

If the \code{intake_max} slot has a comment and \code{reset = FALSE}, then a
recalculation from the species parameters is suppressed and a message is
issued if the recalculated values would differ from the stored ones.
}

\examples{
# Inspect the current maximum intake rate
intake_max(NS_params)["Cod", 1:5]

# Increase intake rate for Cod by 50\%
im <- intake_max(NS_params)
im["Cod", ] <- im["Cod", ] * 1.5
params <- setMaxIntakeRate(NS_params, intake_max = im)
intake_max(params)["Cod", 1:5]
}
\seealso{
Other functions for setting parameters:
\code{\link[=gear_params]{gear_params()}},
\code{\link[=setExtDiffusion]{setExtDiffusion()}},
\code{\link[=setExtEncounter]{setExtEncounter()}},
\code{\link[=setExtMort]{setExtMort()}},
\code{\link[=setFishing]{setFishing()}},
\code{\link[=setInteraction]{setInteraction()}},
\code{\link[=setMetabolicRate]{setMetabolicRate()}},
\code{\link[=setParams]{setParams()}},
\code{\link[=setPredKernel]{setPredKernel()}},
\code{\link[=setReproduction]{setReproduction()}},
\code{\link[=setSearchVolume]{setSearchVolume()}},
\code{\link[=species_params]{species_params()}},
\code{\link[=use_predation_diffusion]{use_predation_diffusion()}}
}
\concept{functions for setting parameters}
