% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{setPlankton}
\alias{setPlankton}
\title{Set up plankton}
\usage{
setPlankton(params, r_pp = 10, w_pp_cutoff = 10,
  plankton_dynamics = plankton_semichemostat, interaction_p = rep(1,
  nrow(params@species_params)))
}
\arguments{
\item{params}{A MizerParams object}

\item{r_pp}{Growth rate of the primary productivity. Default is 10 g/year.}

\item{w_pp_cutoff}{The upper cut off size of the plankton spectrum. 
Default is 10 g.}

\item{plankton_dynamics}{Function that determines plankton dynamics by
calculating the plankton spectrum at the next time step from the current
state. The default is \code{"\link{plankton_semichemostat}"}.}

\item{interaction_p}{Vector specifying for each species its interaction with
plankton, similar to what the interaction matrix does for the interaction
with other species. Entries should be numbers between 0 and 1. The 
default is a vector of 1s.}
}
\value{
A MizerParams object
}
\description{
Set up plankton
}
