% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setResourceEncounter}
\alias{setResourceEncounter}
\title{Set resource encounter rate}
\usage{
setResourceEncounter(params, rho)
}
\arguments{
\item{params}{A MizerParams object}

\item{rho}{Either a 3-dim array (predator species x resource x predator size)
or a 2-dim array (predator species x resource). In the latter case the
size dependence is given by allometry.}
}
\value{
A MizerParams object
}
\description{
This sets the rate \eqn{\rho_{id}(w)} that determines the rate at which an
individual of species \eqn{i} encounters biomass of resource \eqn{d},
\deqn{\sum_d\rho_{id}(w) B_d,} where \eqn{B_d} is the biomass of the d-th
unstructured resource component. 
See \code{\link{resource_dynamics}} for more details.
}
\details{
If only a two-dimensional array (species x resource) is given for the
\code{rho} argument, then the size dependence is set to a power-law:
\deqn{\rho_{id}(w) = \rho_{id} w^n.}

The ordering of the entries in the array \code{rho} is important. The order
of the species in the first array dimension needs to be the same as that in
the species parameter dataframe. The order of the resources in the second
array dimension must be the same as in the list of resource dynamics. The
third dimension, if given, is the size dimension.
}
