% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setResourceSemichemostat.R
\name{setResourceSemichemostat}
\alias{setResourceSemichemostat}
\title{Tune the density dependence in reproduction}
\usage{
setResourceSemichemostat(
  params,
  resource_rate,
  resource_capacity,
  resource_level
)
}
\arguments{
\item{params}{A MizerParams object}

\item{resource_rate}{Optional. Vector of resource intrinsic birth rates}

\item{resource_capacity}{Optional. Vector of resource intrinsic carrying
capacity}

\item{resource_level}{Sets \code{resource_capacity} to the current resource
number density divided by \code{resource_level}.}
}
\value{
A MizerParams object with updated resource parameters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Takes a MizerParams object \code{params} with arbitrary density dependence in
the resource dynamics and
returns a MizerParams object with the resource parameters chosen in such a
way that the
Hence if you have tuned your \code{params} object to describe a
particular steady state, then setting the resource parameters
with this function will leave you with the exact same steady state.
}
\details{
If you do not provide a value for any of the resource parameter
arguments, then the resource rate will be kept at its current value.
If you do provide one of the reproduction
parameters, this can be either a vector with the same length as
\code{w_full(params)}, giving a
value for every size class, or it can be a single value that is taken as the
value at 1g with the values at other sizes calculated according to a
power-law.

The values for \code{resource_capacity} must be larger than the current resource number
density. If a smaller value is requested a warning is issued and the
value is increased to twice the current resource number density.

The values for the \code{resource_level} must be positive and less than 1.
Otherwise an error is raised.

The values for \code{resource_rate} must be large enough to allow the required
resource production rate. If a smaller value is requested a warning is issued
and the value is increased to the smallest possible value.

As can be seen in the graph above, choosing a lower value for
\code{resource_capacity} or a higher value for \code{resource_rate} means that near the
steady state the resource abundance will be less sensitive to a change in the
predation mortality and hence less sensitive to changes in the fish
abundances.
}
\concept{resource dynamics}
