% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{slot<-}
\alias{slot<-}
\title{S4/S3 compatibility slot assignment helper}
\usage{
slot(object, name, check = TRUE) <- value
}
\arguments{
\item{object}{An object (S4 or S3).}

\item{name}{Name of the slot / list element.}

\item{check}{Logical.}

\item{value}{Value to assign.}
}
\value{
The modified object.
}
\description{
S4/S3 compatibility slot assignment helper
}
\keyword{internal}
