% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{slot}
\alias{slot}
\title{S4/S3 compatibility slot access helper}
\usage{
slot(object, name, ...)
}
\arguments{
\item{object}{An object (S4 or S3).}

\item{name}{Name of the slot / list element.}

\item{...}{Additional arguments.}
}
\value{
The value of the slot or list element.
}
\description{
S4/S3 compatibility slot access helper
}
\keyword{internal}
