% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{slotNames}
\alias{slotNames}
\title{S4/S3 compatibility slotNames helper}
\usage{
slotNames(x)
}
\arguments{
\item{x}{An object (S4 or S3) or class name.}
}
\value{
Character vector of slot / element names.
}
\description{
S4/S3 compatibility slotNames helper
}
\keyword{internal}
