\docType{data}
\name{species_params_gears}
\alias{species_params_gears}
\title{Test parameters data set with fishing gears}
\description{
  Interaction data with fishing gears for the North Sea.
  Where did these values come from? See NS Paper.
}
\references{
  The North Sea paper (Blanchard et al)
}
\keyword{datasets}
\keyword{species_params_gears}

