\docType{data}
\name{species_params_nogears}
\alias{species_params_nogears}
\title{Test parameters data set with no fishing gears}
\description{
  Interaction data with no fishing gears for the North Sea.
  Where did these values come from? See NS Paper.
}
\references{
  The North Sea paper (Blanchard et al)
}
\keyword{datasets}
\keyword{species_params_nogears}

