% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spectrum_density_wrt}
\alias{spectrum_density_wrt}
\title{The density measure of a \code{power}-based spectrum}
\usage{
spectrum_density_wrt(per_log_size)
}
\arguments{
\item{per_log_size}{Whether the spectrum is a density with respect to
logarithmic size.}
}
\value{
\code{"log_w"} or \code{"w"}.
}
\description{
The density measure of a \code{power}-based spectrum
}
\keyword{internal}
