% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{srrBevertonHolt}
\alias{srrBevertonHolt}
\title{Beverton Holt stock-recruitment function}
\usage{
srrBevertonHolt(rdi, species_params)
}
\arguments{
\item{rdi}{x}

\item{species_params}{x}
}
\value{
rdd
}
\description{
Beverton Holt stock-recruitment function
}
