% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{srrConstant}
\alias{srrConstant}
\title{Set the recruitment function for constant recruitment}
\usage{
srrConstant(rdi, species_params, ...)
}
\arguments{
\item{rdi}{Vector of egg production rates \eqn{R_p} for all species.}

\item{species_params}{A species parameter dataframe. Must contain a column
`constant_recruitment`.}

\item{...}{Unused}
}
\value{
Vector `species_params$constant_recruitment`
}
\description{
Simply returns the value from `species_params$constant_recruitment`
}
\seealso{
Other stock-recruitment functions: 
\code{\link{srrBevertonHolt}()},
\code{\link{srrNone}()},
\code{\link{srrRicker}()},
\code{\link{srrSheperd}()}
}
\concept{stock-recruitment functions}
