% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{srrNone}
\alias{srrNone}
\title{Identity stock-recruitment function}
\usage{
srrNone(rdi, ...)
}
\arguments{
\item{rdi}{Vector of density independent egg production rates \eqn{R_p} for
all species.}

\item{...}{Not used.}
}
\value{
Vector of density-dependent larvae production rates.
}
\description{
Simply returns its \code{rdi} argument.
}
\seealso{
Other stock-recruitment functions: 
\code{\link{srrBevertonHolt}()},
\code{\link{srrConstant}()},
\code{\link{srrRicker}()},
\code{\link{srrSheperd}()}
}
\concept{stock-recruitment functions}
