% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{srrRicker}
\alias{srrRicker}
\title{Ricker stock-recruitment function}
\usage{
srrRicker(rdi, species_params)
}
\arguments{
\item{rdi}{Vector of density independent egg production rates \eqn{R_p} for
all species.}

\item{species_params}{A species parameter dataframe. Must contain a column
\code{ricker_b} holding the coefficients b.}
}
\value{
Vector of density-dependent larvae production rates.
}
\description{
Takes the rates \eqn{R_p} of egg production and returns reduced,
density-dependent rates \eqn{R} given as
\deqn{R = R_p \exp{- b R_p}}
}
\seealso{
Other stock-recruitment functions: \code{\link{srrBevertonHolt}},
  \code{\link{srrSheperd}}
}
\concept{stock-recruitment functions}
