% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{srrSheperd}
\alias{srrSheperd}
\title{Sheperd stock-recruitment function}
\usage{
srrSheperd(rdi, species_params)
}
\arguments{
\item{rdi}{Vector of density independent egg production rates \eqn{R_p} for
all species.}

\item{species_params}{A species parameter dataframe. Must contain columns
\code{sheperd_b} and \code{sheperd_c} with the parameters b and c.}
}
\value{
Vector of density-dependent larvae production rates.
}
\description{
Takes the rates \eqn{R_p} of egg production and returns reduced,
density-dependent rates \eqn{R} as
\deqn{R = \frac{R_p}{1+(b\ R_p)^c}}{R = R_p / (1 + (B R_p)^c)}
}
\seealso{
Other stock-recruitment functions: \code{\link{srrBevertonHolt}},
  \code{\link{srrRicker}}
}
\concept{stock-recruitment functions}
