% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{str.MizerSim}
\alias{str.MizerSim}
\title{Display structure of MizerSim object}
\usage{
\method{str}{MizerSim}(object, max.level = NA, ...)
}
\arguments{
\item{object}{A \code{MizerSim} object.}

\item{max.level}{Maximum level of nesting to print. Defaults to \code{NA} (no limit).}

\item{...}{Other arguments passed to \code{\link[utils:str]{utils::str()}}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Prints a clean, compact summary of the slots in a \code{MizerSim} object.
}
