% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{str}
\alias{str}
\alias{str.ArraySpeciesBySize}
\alias{str.ArrayTimeBySpecies}
\alias{str.ArrayTimeBySpeciesBySize}
\title{Display the structure of mizer objects}
\arguments{
\item{object}{The object to display the structure of.}

\item{...}{Further arguments. They are passed to the default \code{str()} method.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Mizer provides \code{str()} methods for \code{\link[=MizerParams]{MizerParams()}} and \code{\link[=MizerSim]{MizerSim()}} objects,
as well as \code{\link[=ArraySpeciesBySize]{ArraySpeciesBySize()}}, \code{\link[=ArrayTimeBySpecies]{ArrayTimeBySpecies()}} and \code{\link[=ArrayTimeBySpeciesBySize]{ArrayTimeBySpeciesBySize()}}
objects. These methods produce a clean, compact overview of the object's structure
without polluting the console with large amounts of internal data.
}
\examples{
\donttest{
str(NS_params)
str(NS_sim)
str(getEncounter(NS_params))
}
}
\seealso{
\code{\link[=MizerParams]{MizerParams()}}, \code{\link[=MizerSim]{MizerSim()}}, \code{\link[=ArraySpeciesBySize]{ArraySpeciesBySize()}},
\code{\link[=ArrayTimeBySpecies]{ArrayTimeBySpecies()}}, \code{\link[=ArrayTimeBySpeciesBySize]{ArrayTimeBySpeciesBySize()}}
}
