% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\docType{methods}
\name{summary,MizerSim-method}
\alias{summary,MizerSim-method}
\title{Summarize MizerSim object}
\usage{
\S4method{summary}{MizerSim}(object, ...)
}
\arguments{
\item{object}{A \code{MizerSim} object.}

\item{...}{Other arguments (currently not used).}
}
\description{
Outputs a general summary of the structure and content of the object
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears,inter)
sim <- project(params, effort=1, t_max=5)
summary(sim)
}
}
