% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.MizerParams}
\alias{summary.MizerParams}
\title{Summarize MizerParams object}
\usage{
\method{summary}{MizerParams}(object, ...)
}
\arguments{
\item{object}{A \code{MizerParams} object.}

\item{...}{Other arguments (currently not used).}
}
\value{
The MizerParams object, invisibly
}
\description{
Outputs a general summary of the structure and content of the object
}
\examples{
summary(NS_params)
}
\concept{summary_function}
