% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.MizerSim}
\alias{summary.MizerSim}
\title{Summarize MizerSim object}
\usage{
\method{summary}{MizerSim}(object, ...)
}
\arguments{
\item{object}{A \code{MizerSim} object.}

\item{...}{Other arguments (currently not used).}
}
\value{
The MizerSim object, invisibly
}
\description{
Outputs a general summary of the structure and content of the object
}
\examples{
summary(NS_sim)
}
\concept{summary_function}
