% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary_methods}
\alias{summary_methods}
\title{Summary methods}
\description{
Here we describe methods to extract information about a simulation from a MizerSim object.
}
\details{
A list of available methods is given in the table below.

\tabular{lll}{
  Method name \tab Expression \tab Description \cr
  \code{\link{getAvailEnergy}} \tab \eqn{E_{a.i}(w)} \tab Available energy \cr
  \code{\link{getFeedingLevel}} \tab \eqn{f_i(w)} \tab Feeding level \cr
  \code{\link{getPredRate}} \tab \eqn{\phi_i(w_p/w) (1-f_i(w)) \gamma_i w^q N_i(w) dw} \tab Predation \cr
  \code{\link{getPredMort}} \tab \eqn{\mu_{p.i}(w)} \tab Predation mortality \cr
  \code{\link{getPlanktonMort}} \tab \eqn{\mu_{p}(w)} \tab Mortality on plankton \cr
  \code{\link{getFMortGear}} \tab \eqn{F_{g,i}(w)} \tab Fishing mortality by gear \cr
  \code{\link{getFMort}} \tab \eqn{\mu_{f.i}(w)} \tab Total fishing mortality \cr
  \code{\link{getMort}} \tab \eqn{\mu_{i}(w)} \tab Total mortality \cr
  \code{\link{getEReproAndGrowth}} \tab \eqn{E_{r.i}(w)} \tab Energy put into growth and reproduction \cr
  \code{\link{getERepro}} \tab \eqn{\psi_i(w)E_{r.i}(w)} \tab Energy put reproduction \cr
  \code{\link{getEGrowth}} \tab \eqn{g_i(w)} \tab Energy put growth \cr
  \code{\link{getRDI}} \tab \eqn{R_{p.i}} \tab Egg production \cr
  \code{\link{getRDD}} \tab \eqn{R_i} \tab Recruitment \cr
}
}
