% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{test_dyn}
\alias{test_dyn}
\title{Dummy functions used during testing only}
\usage{
test_dyn(params, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{...}{Other parameters}
}
\description{
Dummy functions used during testing only
}
