% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{tpo2EReproAndGrowth}
\alias{tpo2EReproAndGrowth}
\title{Energy available for reproduction and growth under the tpo2 extension}
\usage{
tpo2EReproAndGrowth(params, n, n_pp, n_other, t, encounter, feeding_level, ...)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{encounter}{An array (species x size) with the encounter rate as
calculated by \code{\link[=getEncounter]{getEncounter()}}.}

\item{feeding_level}{An array (species x size) with the feeding level as
calculated by \code{\link[=getFeedingLevel]{getFeedingLevel()}}.}

\item{...}{Unused}
}
\value{
A species x size array with energy available for reproduction and growth.
}
\description{
Energy available for reproduction and growth under the tpo2 extension
}
\seealso{
Other mizer rate functions: 
\code{\link{mizerEGrowth}()},
\code{\link{mizerEReproAndGrowth}()},
\code{\link{mizerERepro}()},
\code{\link{mizerEncounter}()},
\code{\link{mizerFMortGear}()},
\code{\link{mizerFMort}()},
\code{\link{mizerFeedingLevel}()},
\code{\link{mizerMort}()},
\code{\link{mizerPredMort}()},
\code{\link{mizerPredRate}()},
\code{\link{mizerRDI}()},
\code{\link{mizerRates}()},
\code{\link{mizerResourceMort}()},
\code{\link{tpo2FeedingLevel}()},
\code{\link{tpo2PredRate}()}
}
\concept{mizer rate functions}
\keyword{internal}
