% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{tpo2FeedingLevel}
\alias{tpo2FeedingLevel}
\title{Feeding level with temperature-adjusted intake rate}
\usage{
tpo2FeedingLevel(params, n, n_pp, n_other, t, encounter, ...)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{encounter}{A two dimensional array (predator species x predator size)
with the encounter rate.}

\item{...}{Unused}
}
\value{
A species x size array with the raw feeding level.
}
\description{
Feeding level with temperature-adjusted intake rate
}
\seealso{
Other mizer rate functions: 
\code{\link{mizerEGrowth}()},
\code{\link{mizerEReproAndGrowth}()},
\code{\link{mizerERepro}()},
\code{\link{mizerEncounter}()},
\code{\link{mizerFMortGear}()},
\code{\link{mizerFMort}()},
\code{\link{mizerFeedingLevel}()},
\code{\link{mizerMort}()},
\code{\link{mizerPredMort}()},
\code{\link{mizerPredRate}()},
\code{\link{mizerRDI}()},
\code{\link{mizerRates}()},
\code{\link{mizerResourceMort}()},
\code{\link{tpo2EReproAndGrowth}()},
\code{\link{tpo2PredRate}()}
}
\concept{mizer rate functions}
\keyword{internal}
