% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpo2_community.R
\name{tpo2PredRate}
\alias{tpo2PredRate}
\title{Predation rate with oxygen-downregulated realised feeding level}
\usage{
tpo2PredRate(params, n, n_pp, n_other, t, feeding_level, ...)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{feeding_level}{An array (species x size) with the feeding level as
calculated by \code{\link[=getFeedingLevel]{getFeedingLevel()}}.}

\item{...}{Unused}
}
\value{
A predator species x prey size array.
}
\description{
Predation rate with oxygen-downregulated realised feeding level
}
\seealso{
Other mizer rate functions: 
\code{\link{mizerEGrowth}()},
\code{\link{mizerEReproAndGrowth}()},
\code{\link{mizerERepro}()},
\code{\link{mizerEncounter}()},
\code{\link{mizerFMortGear}()},
\code{\link{mizerFMort}()},
\code{\link{mizerFeedingLevel}()},
\code{\link{mizerMort}()},
\code{\link{mizerPredMort}()},
\code{\link{mizerPredRate}()},
\code{\link{mizerRDI}()},
\code{\link{mizerRates}()},
\code{\link{mizerResourceMort}()},
\code{\link{tpo2EReproAndGrowth}()},
\code{\link{tpo2FeedingLevel}()}
}
\concept{mizer rate functions}
\keyword{internal}
