% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{tuneParams}
\alias{tuneParams}
\title{Launch shiny gadget for tuning parameters}
\usage{
tuneParams(p, catch = NULL)
}
\arguments{
\item{p}{MizerParams object to tune. If missing, the gadget tries to recover
information from log files left over from aborted previous runs.}

\item{catch}{Data frame holding binned observed catch data. The data can
be binned either into length bins or weight bins. In the former the data
frame should have columns \code{length} and \code{dl} holding the start of
the size bins in cm and the width of the size bins in cm respectively. In
the latter case the data frame should have columns \code{weight} and
\code{dw} holding the start of the size bins in grams and the width of the
size bins in grams. The data frame also needs to have the columns
\code{species} (the name of the species), \code{catch} (the number of
individuals of a particular species caught in a size bin).}
}
\value{
The tuned MizerParams object
}
\description{
The shiny gadget has sliders for the model parameters and tabs with a
variety of plots to visualise the steady-state
}
