% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade.MizerParams}
\alias{upgrade.MizerParams}
\title{Upgrade the core slots of a MizerParams object}
\usage{
\method{upgrade}{MizerParams}(object, ...)
}
\arguments{
\item{object}{An old MizerParams object to be upgraded}

\item{...}{Unused.}
}
\value{
The upgraded MizerParams object
}
\description{
This is the \code{MizerParams} method of the \code{\link[=upgrade]{upgrade()}} generic and performs the
\emph{core mizer} upgrade only. It is called from the orchestrator
\code{\link[=runExtensionUpgrades]{runExtensionUpgrades()}}, which also invokes any registered extension upgrade
methods.
You should never need to call it directly; use \code{validParams()} (or
\code{readParams()}) instead.
}
\seealso{
\code{\link[=validParams]{validParams()}}
}
\concept{helper}
\keyword{internal}
