% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade.MizerSim}
\alias{upgrade.MizerSim}
\title{Upgrade a MizerSim object from earlier versions}
\usage{
\method{upgrade}{MizerSim}(object, ...)
}
\arguments{
\item{object}{An old MizerSim object to be upgraded}

\item{...}{Unused.}
}
\value{
The upgraded MizerSim object
}
\description{
This is the \code{MizerSim} method of the \code{\link[=upgrade]{upgrade()}} generic. It rebuilds the
simulation around an upgraded params object; the params upgrade (core mizer
and any extensions) is performed by the \code{validParams()} call below. You
should never need to call it directly; use \code{validSim()} (or \code{readSim()}).
}
\concept{helper}
\keyword{internal}
