% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{upgrade_s4_to_s3}
\alias{upgrade_s4_to_s3}
\title{Convert an S4 MizerParams or MizerSim object to an S3 object}
\usage{
upgrade_s4_to_s3(object)
}
\arguments{
\item{object}{An S4 MizerParams or MizerSim object.}
}
\value{
An S3 MizerParams or MizerSim object (a list).
}
\description{
Convert an S4 MizerParams or MizerSim object to an S3 object
}
\keyword{internal}
