% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgradeParams}
\alias{upgradeParams}
\title{Back-compatible wrapper for the core MizerParams upgrade}
\usage{
upgradeParams(params)
}
\arguments{
\item{params}{An old MizerParams object.}
}
\value{
The upgraded MizerParams object.
}
\description{
Retained because it was an (unexported) internal entry point. New code should
rely on \code{validParams()} / \code{readParams()}, which orchestrate the core upgrade
together with any extension upgrades.
}
\keyword{internal}
