% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgradeSim}
\alias{upgradeSim}
\title{Back-compatible wrapper for the MizerSim upgrade}
\usage{
upgradeSim(sim)
}
\arguments{
\item{sim}{An old MizerSim object.}
}
\value{
The upgraded MizerSim object.
}
\description{
Retained because it was an (unexported) internal entry point. New code should
rely on \code{validSim()} / \code{readSim()}.
}
\keyword{internal}
