% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady.R
\name{valid_gears_arg}
\alias{valid_gears_arg}
\title{Helper function to assure validity of gears argument}
\usage{
valid_gears_arg(object, gears = NULL, error_on_empty = FALSE)
}
\arguments{
\item{object}{A MizerSim or MizerParams object from which the gears
should be selected.}

\item{gears}{The gears to be selected. Optional. By default all gears
are selected. A vector of gear names.}

\item{error_on_empty}{Whether to throw an error if there are zero valid
gears. Default FALSE.}
}
\value{
A vector of gear names.
}
\description{
If the gears argument contains invalid gears, then these are
ignored but a warning is issued.
}
\concept{helper}
