% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R
\name{validate_array_type}
\alias{validate_array_type}
\title{Validate the type of a mizer array}
\usage{
validate_array_type(type)
}
\arguments{
\item{type}{One of \link{array_types}.}
}
\value{
The validated type.
}
\description{
Validate the type of a mizer array
}
\keyword{internal}
