% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{validate_effort_vector}
\alias{validate_effort_vector}
\title{Check that an effort vector is specified correctly}
\usage{
validate_effort_vector(params, effort)
}
\description{
Check that an effort vector is specified correctly
}
