% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFishing.R
\name{validEffortVector}
\alias{validEffortVector}
\title{Make a valid effort vector}
\usage{
validEffortVector(effort, params)
}
\arguments{
\item{effort}{A vector or scalar with the initial fishing effort, see Details
below.}

\item{params}{A MizerParams object.}
}
\description{
Make a valid effort vector
}
\details{
A valid effort vector is a named vector with one effort value for each gear.
However you can also supply the effort value in different ways:
\itemize{
\item a scalar, which is then replicated for each gear
\item an unnamed vector, which is then assumed to be in the same order as the
gears in the params object
\item a named vector in which the gear names have a different order than in the
params object. This is then sorted correctly.
\item a named vector which only supplies values for some of the gears.
The effort for the other gears is then set to the default effort returned
by \code{\link[=validEffortVector]{validEffortVector()}}, which depends on the defaults edition.
}

These conversions are done by the function \code{validEffortVector()}.

An \code{effort} argument will lead to an error if it is either
\itemize{
\item unnamed and of the wrong length
\item named but where some names do not match any of the gears
\item not numeric
}
}
\seealso{
\code{\link[=initial_effort]{initial_effort()}}
}
\concept{helper}
