% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validObject}
\alias{validObject}
\title{Validity checker compatibility helper}
\usage{
validObject(object, ...)
}
\arguments{
\item{object}{A MizerParams or MizerSim object.}

\item{...}{Additional arguments.}
}
\value{
TRUE invisibly if valid, or stops with error.
}
\description{
Validity checker compatibility helper
}
\keyword{internal}
