% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{validSim}
\alias{validSim}
\title{Validate MizerSim object and upgrade if necessary}
\usage{
validSim(sim)
}
\arguments{
\item{sim}{The MizerSim object to validate}
}
\value{
A valid MizerSim object
}
\description{
Checks that the given MizerSim object is valid and upgrades it if
necessary by calling \code{\link[=upgradeSim]{upgradeSim()}}.
}
\details{
In particular, it makes sure that the MizerParams object inside the
MizerSim object is valid by calling \code{\link[=validParams]{validParams()}}.
}
