% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R
\name{addPlot}
\alias{addPlot}
\alias{addPlot.ArraySpeciesBySize}
\title{Add values to an existing plot}
\usage{
addPlot(plot, x, ...)

\method{addPlot}{ArraySpeciesBySize}(
  plot,
  x,
  species = NULL,
  all.sizes = FALSE,
  wlim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  colour = NULL,
  linetype = "dashed",
  linewidth = 0.8,
  alpha = 1,
  ...
)
}
\arguments{
\item{plot}{A ggplot2 object to which the new values should be added.}

\item{x}{An object containing the values to add.}

\item{...}{Further arguments passed to methods.}

\item{species}{Character vector of species to include. \code{NULL} (default) means
all species.}

\item{all.sizes}{If \code{FALSE} (default), values outside a species' size range
(\code{w_min} to \code{w_max}) are removed. Only applies to \code{ArraySpeciesBySize}.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the weight (x) axis. Use \code{NA} to refer to the existing minimum or
maximum. Only applies to \code{ArraySpeciesBySize}.}

\item{total}{A boolean value that determines whether the total over all
selected species is plotted as well. Default is \code{FALSE}.}

\item{background}{A boolean value that determines whether background species
are included. Ignored if the model does not contain background species.
Default is \code{TRUE}.}

\item{colour}{Optional fixed colour for the added lines. If \code{NULL}, the
species colours from the existing plot are used.}

\item{linetype}{Optional fixed line type for the added lines. If \code{NULL}, the
species line types from the existing plot are used.}

\item{linewidth}{Width of the added lines.}

\item{alpha}{Transparency of the added lines.}
}
\value{
A ggplot2 object.
}
\description{
\code{addPlot()} adds another set of values to an existing ggplot. The first
method supports adding an \code{ArraySpeciesBySize} object to a compatible plot,
for example to compare the same rate before and after a model change.
}
\examples{
\donttest{
p <- plot(getEncounter(NS_params), species = "Cod")
addPlot(p, getEncounter(NS_params), species = "Cod")
}
}
\seealso{
Other plotting functions:
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
