% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R,
%   R/ArrayTimeBySpecies-class.R
\name{addPlot}
\alias{addPlot}
\alias{addPlot.ArraySpeciesBySize}
\alias{addPlot.ArrayTimeBySpecies}
\title{Add values to an existing plot}
\usage{
addPlot(plot, x, ...)

\method{addPlot}{ArraySpeciesBySize}(
  plot,
  x,
  species = NULL,
  all.sizes = FALSE,
  wlim = c(NA, NA),
  llim = c(NA, NA),
  size_axis = c("w", "l"),
  total = FALSE,
  background = TRUE,
  colour = NULL,
  linetype = "dashed",
  linewidth = 0.8,
  alpha = 1,
  ...
)

\method{addPlot}{ArrayTimeBySpecies}(
  plot,
  x,
  species = NULL,
  start_time = NULL,
  end_time = NULL,
  ylim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  colour = NULL,
  linetype = "dashed",
  linewidth = 0.8,
  alpha = 1,
  ...
)
}
\arguments{
\item{plot}{A ggplot2 object to which the new values should be added.}

\item{x}{An object containing the values to add.}

\item{...}{Further arguments passed to methods.}

\item{species}{Character vector of species to include. \code{NULL} (default) means
all species.}

\item{all.sizes}{If \code{FALSE} (default), values outside a species' size range
(\code{w_min} to \code{w_max}) are removed. Only applies to \code{ArraySpeciesBySize}.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the weight (x) axis. Use \code{NA} to refer to the existing minimum or
maximum. Only applies to \code{ArraySpeciesBySize}.}

\item{llim}{A numeric vector of length two providing lower and upper limits
for the length (x) axis when \code{size_axis = "l"}. Use \code{NA} to refer to the
existing minimum or maximum. Only applies to \code{ArraySpeciesBySize}.}

\item{size_axis}{Whether to plot size as weight (\code{"w"}, default) or length
(\code{"l"}), using the allometric weight-length relationship.}

\item{total}{A boolean value that determines whether the total over all
selected species is plotted as well. Default is \code{FALSE}.}

\item{background}{A boolean value that determines whether background species
are included. Ignored if the model does not contain background species.
Default is \code{TRUE}.}

\item{colour}{Optional fixed colour for the added lines. If \code{NULL}, the
species colours from the existing plot are used.}

\item{linetype}{Optional fixed line type for the added lines. If \code{NULL}, the
species line types from the existing plot are used.}

\item{linewidth}{Width of the added lines.}

\item{alpha}{Transparency of the added lines.}

\item{start_time}{The first time to be plotted. Default (\code{NULL}) is the
beginning of the time series. Only applies to \code{ArrayTimeBySpecies}.}

\item{end_time}{The last time to be plotted. Default (\code{NULL}) is the end of
the time series. Only applies to \code{ArrayTimeBySpecies}.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the value (y) axis. Use \code{NA} to refer to the existing minimum or
maximum.}
}
\value{
A ggplot2 object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{addPlot()} adds another set of values to an existing ggplot. The first
method supports adding an \code{ArraySpeciesBySize} object to a compatible plot,
for example to compare the same rate before and after a model change.
The method checks whether the existing plot uses a compatible x variable,
and warns if the y variable or y-axis units appear to differ.
}
\examples{
\donttest{
p <- plot(getEncounter(NS_params), species = "Cod")
addPlot(p, getEncounter(NS_params), species = "Cod")
}
}
\seealso{
Other plotting functions:
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plot2]{plot2()}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotRelative]{plotRelative()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectra2]{plotSpectra2()}},
\code{\link[=plotSpectraRelative]{plotSpectraRelative()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
