% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{animateSpectra}
\alias{animateSpectra}
\title{Animation of the abundance spectra}
\usage{
animateSpectra(
  sim,
  species,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  power = 1,
  total = FALSE,
  plankton = TRUE
)
}
\arguments{
\item{sim}{A MizerSim object}

\item{species}{Name or vector of names of the species to be plotted. By
default all species are plotted.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the w axis. Use NA to refer to the existing minimum or maximum.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the y axis. Use NA to refer to the existing minimum or maximum. Any
values below 1e-20 are always cut off.}

\item{power}{The abundance is plotted as the number density times the weight
raised to \code{power}. The default \code{power = 1} gives the biomass 
density, whereas \code{power = 2} gives the biomass density with respect
to logarithmic size bins.}

\item{total}{A boolean value that determines whether the total over all
species in the system is plotted as well. Default is FALSE}

\item{plankton}{A boolean value that determines whether plankton is included.
Default is TRUE.}
}
\description{
Animation of the abundance spectra
}
\seealso{
Other plotting functions: 
\code{\link{displayFrames}()},
\code{\link{plot,MizerSim,missing-method}},
\code{\link{plotBiomass}()},
\code{\link{plotDiet}()},
\code{\link{plotFMort}()},
\code{\link{plotFeedingLevel}()},
\code{\link{plotGrowthCurves}()},
\code{\link{plotPredMort}()},
\code{\link{plotSpectra}()},
\code{\link{plotYieldGear}()},
\code{\link{plotYield}()},
\code{\link{plotlyBiomass}()},
\code{\link{plotlyFMort}()},
\code{\link{plotlyFeedingLevel}()},
\code{\link{plotlyGrowthCurves}()},
\code{\link{plotlyPredMort}()},
\code{\link{plotlySpectra}()},
\code{\link{plotlyYieldGear}()},
\code{\link{plotlyYield}()},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
