% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFishing.R
\name{calc_selectivity}
\alias{calc_selectivity}
\title{Calculate selectivity from gear parameters}
\usage{
calc_selectivity(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
An array (gear x species x size) with the selectivity values
}
\description{
This function calculates the selectivity for each gear, species and size
from the gear parameters. It is called by \code{\link[=setFishing]{setFishing()}} when the \code{selectivity}
is not set by the user.
}
\examples{
params <- NS_params
calc_selectivity(params)["Pelagic", "Herring", ]
}
\concept{helper}
