% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrateBiomass}
\alias{calibrateBiomass}
\title{Calibrate model biomass to observed biomass}
\usage{
calibrateBiomass(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A MizerParams object
}
\description{
Given a MizerParams object \code{params} for which biomass observations are
available for at least some species via
\code{species_params(params)$biomass_observed}, this function returns an
updated MizerParams object which is rescaled so that the total biomass
of the observed species in the model agrees with the total observed
biomass.
}
\details{
Observed biomasses usually only include individuals above a certain size.
This size should either be specified in
\code{species_params(params)$biomass_cutoff} in grams, or else a cutoff size
of \code{w_mat/20} is assumed.
}
\examples{
params <- NS_params
species_params(params)$biomass_observed <- 
    c(0.8, 61, 12, 35, 1.6, 20, 10, 7.6, 135, 60, 30, 78)
species_params(params)$biomass_cutoff <- 10
params2 <- calibrateBiomass(params)
plotBiomassObservedVsModel(params2)
}
