% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_dynamics.R
\name{carrion_dynamics}
\alias{carrion_dynamics}
\title{Carrion dynamics}
\usage{
carrion_dynamics(params, n, n_pp, B, rates, dt, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{The biomass of carrion}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step}

\item{...}{Unused}
}
\value{
Biomass of carrion at next time step
}
\description{
Calculates the biomass of carrion (dead animals) at the next timestep from
the current biomass.
}
\details{
The inflow of carrion biomass consists of
\itemize{
\item An external inflow at a rate given by
\code{params@resource_params$carrion_external}.
\item Discards from fishing.
\item Animals killed by fishing gear.
\item Animals that have died by causes other than predation.
}

The outflow of biomass is due to
\itemize{
\item consumption by scavenger species, where the encounter rate is
specified by \code{params@rho[, "carrion", ]}.
}

This function can be used as the project_resources function in the case
where there is only a single unstructured resource that represents
carrion, or as part of a more comprehensive function when there are also
other unstructured resource components. It is used for example in the
\code{\link[=dead_matter_dynamics]{dead_matter_dynamics()}} function.
}
