% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{changeResourceEncounter}
\alias{changeResourceEncounter}
\title{Set resource encounter rate}
\usage{
changeResourceEncounter(params, rho = NULL, n = params@n)
}
\arguments{
\item{params}{A MizerParams object}

\item{rho}{Optional. An array (species x resource x size)
holding the rate at which a consumer of a particular size feeds on each
resource. Described in the section "Setting resource encounter rate".}

\item{n}{Scaling exponent of the intake rate.}
}
\value{
A MizerParams object
}
\description{
Set resource encounter rate
}
\section{Setting resource encounter rate}{

The resource encounter rate \eqn{\rho_{id}(w)} (units 1/year) determines the
rate at which an individual of species \eqn{i} encounters biomass of resource
\eqn{d}, so that the contribution from all unstructured resources to the
total encounter rate is
\deqn{E_{u.i}(w) = \sum_d\rho_{id}(w) B_d,} 
where \eqn{B_d} is the biomass of the d-th unstructured resource component.

Resource consumption is subject to satiation in the same way as other food,
so that a consumer only consumes a fraction \eqn{1-f_i(w)} of the encountered
resource biomass, where \eqn{f_i(w)} is the feeding level.

If the \code{rho} array is not supplied, then the resource encounter rate is
set to a power law
\deqn{\rho_{id}(w) = \rho_{id} w^n.}
The coefficients \eqn{\rho_{id}} are parameters in the species_params dataframe.
For example if there is a resource called "detritus" then the species_params
data frame needs to have a column called \code{rho_detritus} and similarly
for each other resource.

If the \code{rho} array is supplied, the ordering of the entries in the array
is important. The order of the species in the first array dimension needs to
be the same as that in the species parameter dataframe. The order of the
resources in the second array dimension must be the same as in the list of
resource dynamics. The third dimension is the size dimension.
}

\seealso{
Other functions for changing parameters: \code{\link{changeBMort}},
  \code{\link{changeFishing}}, \code{\link{changeInitial}},
  \code{\link{changeIntakeMax}},
  \code{\link{changeInteraction}},
  \code{\link{changeMetab}}, \code{\link{changeParams}},
  \code{\link{changePlankton}},
  \code{\link{changePredKernel}},
  \code{\link{changeReproduction}},
  \code{\link{changeResourceDynamics}},
  \code{\link{changeSearchVolume}},
  \code{\link{tuneParams}}
}
\concept{functions for changing parameters}
