% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{clearExtensionChain}
\alias{clearExtensionChain}
\title{Clear the registered extension chain}
\usage{
clearExtensionChain()
}
\value{
Invisibly, an empty character vector.
}
\description{
Clears the session's extension registry. You can then create a new
extension chain with \code{\link[=registerExtensions]{registerExtensions()}}.
}
\seealso{
"Using mizer extension packages":
\code{vignette("using-extension-packages", package = "mizer")}

Other extension tools: 
\code{\link{NOther}()},
\code{\link{coerceToExtensionClass}()},
\code{\link{getRegisteredExtensions}()},
\code{\link{initialNOther<-}()},
\code{\link{recordExtension}()},
\code{\link{registerExtension}()},
\code{\link{registerExtensions}()},
\code{\link{setComponent}()},
\code{\link{setRateFunction}()}
}
\concept{extension tools}
