% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{coerceToExtensionClass}
\alias{coerceToExtensionClass}
\title{Coerce a mizer object to its registered extension class}
\usage{
coerceToExtensionClass(object, extensions = objectExtensions(object))
}
\arguments{
\item{object}{A \code{MizerParams} or \code{MizerSim} object.}

\item{extensions}{Optional extension chain. Defaults to the chain stored in
\code{object}, or in \code{object@params} for \code{MizerSim}.}
}
\value{
The same object coerced to the appropriate marker class, or to the
base class for an empty extension chain.
}
\description{
Coerces a \code{MizerParams} or \code{MizerSim} object to the S4 marker class
corresponding to the object's own extension chain. For \code{MizerSim}, the
extension chain is read from \code{sim@params@extensions}.
}
\seealso{
"Creating a mizer extension package":
\href{https://sizespectrum.org/mizer/articles/guide-create-extension-package.html}{Creating a mizer extension package}

Other extension tools:
\code{\link[=NOther]{NOther()}},
\code{\link[=clearExtensionChain]{clearExtensionChain()}},
\code{\link[=getRegisteredExtensions]{getRegisteredExtensions()}},
\code{\link[=initialNOther<-]{initialNOther<-()}},
\code{\link[=recordExtension]{recordExtension()}},
\code{\link[=registerExtension]{registerExtension()}},
\code{\link[=registerExtensions]{registerExtensions()}},
\code{\link[=setComponent]{setComponent()}},
\code{\link[=setRateFunction]{setRateFunction()}}
}
\concept{extension tools}
