% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{coerceToExtensionClass}
\alias{coerceToExtensionClass}
\title{Coerce a mizer object to its extension class}
\usage{
coerceToExtensionClass(object, extensions = objectExtensions(object))
}
\arguments{
\item{object}{A \code{MizerParams} or \code{MizerSim} object.}

\item{extensions}{Optional extension list or vector. Defaults to the extensions stored in
\code{object}, or in \code{object$params} for \code{MizerSim}.}
}
\value{
The same object coerced to the appropriate S3 class vector, or to the
base class for an empty extension list.
}
\description{
Coerces a \code{MizerParams} or \code{MizerSim} object to the S3 class hierarchy
corresponding to the object's own extension list. For \code{MizerSim}, the
extension names are read from the embedded \code{MizerParams} object.
}
\details{
This is infrastructure for extension package authors. Ordinary users should
load saved models with \code{\link[=readParams]{readParams()}} or \code{\link[=readSim]{readSim()}}; \code{\link[=validParams]{validParams()}} and
\code{\link[=validSim]{validSim()}} repair the class vector of an object that is already in memory.
}
\seealso{
"Creating a mizer extension package":
\href{https://sizespectrum.org/mizer/articles/guide-create-extension-package.html}{Creating a mizer extension package}

Other extension tools:
\code{\link[=NOther]{NOther()}},
\code{\link[=initialNOther<-]{initialNOther<-()}},
\code{\link[=other_mort]{other_mort()}},
\code{\link[=recordExtension]{recordExtension()}},
\code{\link[=setComponent]{setComponent()}},
\code{\link[=setRateFunction]{setRateFunction()}}
}
\concept{extension tools}
