% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareParams.R
\name{compareParams}
\alias{compareParams}
\title{Compare two MizerParams objects and print out differences}
\usage{
compareParams(params1, params2, ...)
}
\arguments{
\item{params1}{First MizerParams object}

\item{params2}{Second MizerParams object}

\item{...}{Additional arguments passed to the method.}
}
\value{
Invisibly returns a character vector of difference messages, one
element per difference. As a side effect, prints the differences in a
human-readable format.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
params1 <- NS_params
params2 <- params1
species_params(params2)$w_mat[1] <- 10
compareParams(params1, params2)
}
