% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{constantRDD}
\alias{constantRDD}
\title{Give constant reproduction rate}
\usage{
constantRDD(rdi, species_params, ...)
}
\arguments{
\item{rdi}{Vector of density-independent reproduction rates \eqn{R_p} for all
species.}

\item{species_params}{A species parameter dataframe. Must contain a column
`constant_recruitment`.}

\item{...}{Unused}
}
\value{
Vector `species_params$constant_recruitment`
}
\description{
Simply returns the value from `species_params$constant_recruitment`.
}
\seealso{
Other functions calculating density-dependent reproduction rate: 
\code{\link{BevertonHoltRDD}()},
\code{\link{RickerRDD}()},
\code{\link{SheperdRDD}()},
\code{\link{noRDD}()}
}
\concept{functions calculating density-dependent reproduction rate}
