% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{constantRDD}
\alias{constantRDD}
\title{Give constant reproduction rate}
\usage{
constantRDD(rdi, species_params, ...)
}
\arguments{
\item{rdi}{Vector of density-independent reproduction rates \eqn{R_p} for all
species.}

\item{species_params}{A species parameter dataframe. Must contain a column
\code{constant_reproduction}.}

\item{...}{Unused}
}
\value{
Vector \code{species_params$constant_reproduction}
}
\description{
Simply returns the value from \code{species_params$constant_reproduction}.
}
\seealso{
Other functions calculating density-dependent reproduction rate: 
\code{\link{BevertonHoltRDD}()},
\code{\link{RickerRDD}()},
\code{\link{SheperdRDD}()},
\code{\link{noRDD}()}
}
\concept{functions calculating density-dependent reproduction rate}
