% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{defaultIntakeMax}
\alias{defaultIntakeMax}
\title{Set default maximum intake rate}
\usage{
defaultIntakeMax(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
MizerParams
}
\description{
Sets the maximum intake rate of an individual of species \eqn{i} and weight
\eqn{w} to \deqn{h_i(w) = h_i w^n} The values of \eqn{h_i} are taken from the
\code{h} column in the species parameter dataframe. The maximum intake rate
determines the feeding level calculated with \code{\link{getFeedingLevel}}.
}
\details{
If the \code{h} column is not supplied in the species parameter dataframe, it
is calculated from \code{f0} and the \code{k_vb} column, if they are
supplied.

It is allowed to set \eqn{h_i} to \code{Inf}, in which case there is no 
limit to the rate at which the fish can take in encountered food.
}
