% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{defaultResourceEncounter}
\alias{defaultResourceEncounter}
\title{Set default resource encounter rate}
\usage{
defaultResourceEncounter(params, rho)
}
\arguments{
\item{params}{A MizerParams object}

\item{rho}{An array (species x resource) that gives the
rate \eqn{\rho_{id}} that determines the rate at which species \eqn{i}
encounters biomass of resource \eqn{d}. The rate is assumed to scale
allometrically with size with exponent \code{n}. So the total contribution
of the resources to the rate at which the individual encounters biomass is
\deqn{\sum_d\rho_{id} w^n B_d,}
where \eqn{B_d} is the biomass of the d-th unstructured resource component.
See \code{\link{resource_dynamics}} help for more details.}
}
\value{
A MizerParams object
}
\description{
Set default resource encounter rate
}
