% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated_accessors}
\alias{deprecated_accessors}
\alias{getCatchability}
\alias{getSelectivity}
\alias{getInitialEffort}
\alias{getPredKernel}
\alias{getSearchVolume}
\alias{getMaxIntakeRate}
\alias{getMetabolicRate}
\alias{getExtMort}
\alias{getExtEncounter}
\alias{getMaturityProportion}
\alias{getReproductionProportion}
\title{Deprecated \code{get}-prefixed accessors for stored rate arrays}
\usage{
getCatchability(params)

getSelectivity(params)

getInitialEffort(params)

getPredKernel(params)

getSearchVolume(params)

getMaxIntakeRate(params)

getMetabolicRate(params)

getExtMort(params)

getExtEncounter(params)

getMaturityProportion(params)

getReproductionProportion(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
The same as the replacement function it forwards to.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Each of these functions had a shorter alias returning exactly the same thing.
The shorter name is the one that also has a replacement function
(\code{catchability(params) <- value} and friends), so that is the name that
survives. Use the replacement given below instead:\tabular{ll}{
   Deprecated \tab Use instead \cr
   \code{getCatchability()} \tab \code{\link[=catchability]{catchability()}} \cr
   \code{getSelectivity()} \tab \code{\link[=selectivity]{selectivity()}} \cr
   \code{getInitialEffort()} \tab \code{\link[=initial_effort]{initial_effort()}} \cr
   \code{getPredKernel()} \tab \code{\link[=pred_kernel]{pred_kernel()}} \cr
   \code{getSearchVolume()} \tab \code{\link[=search_vol]{search_vol()}} \cr
   \code{getMaxIntakeRate()} \tab \code{\link[=intake_max]{intake_max()}} \cr
   \code{getMetabolicRate()} \tab \code{\link[=metab]{metab()}} \cr
   \code{getExtMort()} \tab \code{\link[=ext_mort]{ext_mort()}} \cr
   \code{getExtEncounter()} \tab \code{\link[=ext_encounter]{ext_encounter()}} \cr
   \code{getMaturityProportion()} \tab \code{\link[=maturity]{maturity()}} \cr
   \code{getReproductionProportion()} \tab \code{\link[=repro_prop]{repro_prop()}} \cr
}


The \code{get} prefix is reserved for the functions that \emph{calculate} a rate from
the current state of a model, like \code{\link[=getEncounter]{getEncounter()}} or \code{\link[=getFMort]{getFMort()}}. The
functions above only read back a value stored in the MizerParams object,
which is what the bare names say.
}
\concept{deprecated}
