% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_dynamics.R
\name{detritus_dynamics}
\alias{detritus_dynamics}
\title{Detritus dynamics}
\usage{
detritus_dynamics(params, n, n_pp, B, rates, dt,
  detritus_external = params@resource_params$detritus_external,
  detritus_proportion = params@resource_params$detritus_proportion, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of current species abundances (species x size)}

\item{n_pp}{A vector of current plankton abundance by size}

\item{B}{A vector of current resource biomasses}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step}

\item{detritus_external}{External inflow rate of detritus biomass}

\item{detritus_proportion}{Proportion of consumption by fish that flows into
the detritus component.}

\item{...}{Unused}
}
\value{
Biomass of detritus at next time step
}
\description{
Calculates the detritus biomass at the next timestep from the current
detritus biomass.
}
\details{
The inflow of biomass consists of
\itemize{
\item an external inflow at a rate given by \code{detritus_external} and
\item an inflow of feces, calculated as a proportion \code{detritus_proportion} of the
biomass consumed.
}

The outflow of biomass is due to
\itemize{
\item consumption by detritivorous species, where the encounter rate is
specified by \code{params@rho[, "detritus", ]}.
}
}
