% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_dynamics.R
\name{detritus_dynamics}
\alias{detritus_dynamics}
\title{Detritus dynamics}
\usage{
detritus_dynamics(
  params,
  n,
  n_pp,
  B,
  rates,
  dt,
  detritus_external = params@resource_params$detritus_external,
  detritus_proportion = params@resource_params$detritus_proportion,
  ...
)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of current species abundances (species x size)}

\item{n_pp}{A vector of current plankton abundance by size}

\item{B}{A vector of current resource biomasses}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step}

\item{detritus_external}{Rate of change from external sources}

\item{detritus_proportion}{Proportion of consumption by fish that flows into
the detritus component.}

\item{...}{Unused}
}
\value{
A single number giving the biomass of detritus at next time step
}
\description{
Calculates the detritus biomass at the next timestep from the current
detritus biomass.
}
\details{
The equation for the time evolution of the detritus biomass \eqn{B} is
assumed to be of the form
\deqn{dB/dt = inflow - consumption * B + external}
where
\itemize{
\item \code{inflow} comes from feces, calculated as a proportion
\code{detritus_proportion} of the biomass consumed by all consumers.
\item \code{consumption} is by detritivorous species, where the encounter rate is
specified by \code{params@rho[, "detritus", ]}.
\item \code{external} is an influx from external sources. It can be negative in which
case it represents a loss to external sources.
}

This equation is solved analytically to
\deqn{B(t+dt) = B(t)\exp(-consumption \cdot dt)
  +\frac{inflow + external}{consumption}
  (1-\exp(-consumption \cdot dt)).}{B(t+dt) 
  = B(t) exp(-consumption * dt)
  +(inflow + external)/(consumption) * (1 - exp(-consumption * dt)).}
This avoids the stability problems that would arise if we used the Euler
method to solve the equation numerically.
}
\seealso{
Other resource dynamics functions: 
\code{\link{carrion_dynamics}()}
}
\concept{resource dynamics functions}
