% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{displayFrames}
\alias{displayFrames}
\title{Display frames}
\usage{
displayFrames(f1, f2, params, xlab = NA, ylab = NA, y_ticks = 6)
}
\arguments{
\item{f1}{Data frame for left plot}

\item{f2}{Data frame for right plot}

\item{params}{A MizerParams object}

\item{xlab}{Label for x-axis. Defaults to first variable name.}

\item{ylab}{Label for y-axis. Defaults to third variable name.}

\item{y_ticks}{The approximate number of ticks desired on the y axis}
}
\value{
ggplot2 object
}
\description{
Takes two data frames with plotting data and displays them side-by-side,
using the same axes and legend.
}
\details{
The two data frames each need to have the same three variables. The first
variable will go on the x-axis, the third on the y-axis with a logarithmic
scale. The second variable should be the species and will be used to group
the data and display with the linetype and linecolour specified by the
\code{linetype} and \code{linecolour} slots of the \code{params} object.

The recommended way is to obtain the data frames using one of the supplied
functions, e.g., \code{\link{getBiomassFrame}}, \code{\link{getSSBFrame}}.
}
\examples{
# Set up example MizerParams and MizerSim objects
params <- suppressMessages(newMultispeciesParams(NS_species_params_gears, inter))
sim0 <- project(params, effort=0, t_max=20, progress_bar = FALSE)
sim1 <- project(params, effort=1, t_max=20, progress_bar = FALSE)

# Display biomass from each simulation next to each other
displayFrames(getBiomassFrame(sim0), getBiomassFrame(sim1), params)
}
\seealso{
\code{\link{plotting_functions}},
  \code{\link{getBiomassFrame}}, \code{\link{getSSBFrame}}

Other frame functions: 
\code{\link{getBiomassFrame}()}
}
\concept{frame functions}
