% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{effort}
\alias{effort}
\title{Fishing effort used in simulation}
\usage{
effort(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
An array (time x gear) that stores the fishing effort by time and
gear.
}
\description{
Note that the array returned may not be exactly the same as the \code{effort}
argument that was passed in to \code{project()}. This is because only the saved
effort is stored (the frequency of saving is determined by the argument
\code{t_save}).
}
