% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{emptyParams}
\alias{emptyParams}
\title{Create empty MizerParams object of the right size}
\usage{
emptyParams(species_params, no_w = 100, min_w = 0.001, max_w = NA,
  min_w_pp = NA, no_w_pp = NA)
}
\arguments{
\item{species_params}{A data frame of species-specific parameter values.}

\item{no_w}{The number of size bins in the consumer spectrum.}

\item{min_w}{Sets the size of the eggs of all species for which this is not
given in the \code{w_min} column of the \code{species_params} dataframe.}

\item{max_w}{The largest size of the consumer spectrum. By default this is
set to the largest w_inf specified in the species_params data frame.}

\item{min_w_pp}{The smallest size of the plankton spectrum.}

\item{no_w_pp}{No longer used}
}
\value{
An empty but valid MizerParams object
}
\description{
Sets up a valid \linkS4class{MizerParams} object with all the slots
initialised and given dimension names, but with some slots left empty. This
function is to be used by other functions to set up full parameter objects.
}
\details{
See \code{\link{set_multispecies_model}} for a function that fills the
slots left empty by this function.

A size grid is created so that
the log-sizes are equally spaced. The spacing is chosen so that there will be
`no_w` fish size bins, with the smallest starting at `min_w` and the largest
starting at `max_w`. For `w_full` additional size bins are added below
`min_w`, with the same log size. The number of extra bins is such that
`min_w_pp` comes to lie within the smallest bin.

The \code{species_params} slot of the returned MizerParams object may differ
slightly from the data frame supplied as argument to this function in the
following ways:
\itemize{
  \item Default values are set for \code{w_min, w_inf, alpha, gear, interaction_p}.
  \item The egg sizes in \code{w_min} are rounded down to lie on a grid point.
}
Note that the other characteristic sizes of the species, like \code{w_mat} and
\code{w_inf}, are not modified to lie on grid points.
}
