% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_species.R
\name{expandSizeGrid}
\alias{expandSizeGrid}
\title{Expand the size grid}
\usage{
expandSizeGrid(params, min_w = min(params@w), max_w = max(params@w))
}
\arguments{
\item{params}{A \link{MizerParams} object.}

\item{min_w}{The new minimum size in the grid.}

\item{max_w}{The new maximum size in the grid.}
}
\value{
A new \link{MizerParams} object with the updated size grid.
}
\description{
This function expands the size grid in a \link{MizerParams} object to the desired
min and max size, preserving all existing species.
}
