% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{final_n}
\alias{final_n}
\title{Consumer size spectra at end of simulation}
\usage{
final_n(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
A matrix (species x size) holding the consumer number densities at
  the end of the simulation
}
\description{
Consumer size spectra at end of simulation
}
