% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_params.R
\name{get_gamma_default}
\alias{get_gamma_default}
\title{Get default value for gamma}
\usage{
get_gamma_default(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A vector with the values of gamma for all species
}
\description{
Fills in any missing values for gamma so that if the prey abundance was
described by the power law \eqn{\kappa w^{-\lambda}} then the encounter rate
would lead to the feeding level \eqn{f_0}. Only for internal use.
}
\concept{helper}
\keyword{internal}
