% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{get_h_default}
\alias{get_h_default}
\title{Get default value for h}
\usage{
get_h_default(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A vector with the values of h for all species
}
\description{
Fills in any missing values for h according to the formula
\deqn{h = 3 k_{vb} w_{inf}^{1/3}/ (\alpha f_0)}.
}
\concept{helper}
\keyword{internal}
