% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{get_ks_default}
\alias{get_ks_default}
\title{Get default value for ks}
\usage{
get_ks_default(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A vector with the values of ks for all species
}
\description{
Fills in any missing values for ks according to the formula
\eqn{ks_i = 0.2 h_i}.
}
\concept{helper}
\keyword{internal}
