% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_steady_state_n}
\alias{get_steady_state_n}
\title{Calculate steady state abundance}
\usage{
get_steady_state_n(params, g, mu, N0)
}
\arguments{
\item{params}{A MizerParams object}

\item{g}{A matrix of growth rates (species x size)}

\item{mu}{A matrix of mortality rates (species x size)}

\item{N0}{A vector with the abundance at the smallest size for each species}
}
\value{
A matrix with the steady state abundance
}
\description{
This function calculates the steady state abundance by solving the
transport equation with given growth and mortality rates. It sets up a
tri-diagonal system and solves it.
}
\concept{helper}
