% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_steady_state_n}
\alias{get_steady_state_n}
\title{Helper function to calculate the steady state abundance using the upwind-difference scheme}
\usage{
get_steady_state_n(growth, mort, dw, idx, N0)
}
\arguments{
\item{growth}{A numeric vector of growth rates.}

\item{mort}{A numeric vector of mortality rates.}

\item{dw}{A numeric vector of the size step.}

\item{idx}{A numeric vector of indices.}

\item{N0}{The initial egg density.}
}
\value{
A numeric vector representing the steady state abundances.
}
\description{
Helper function to calculate the steady state abundance using the upwind-difference scheme
}
