% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{get_time_elements}
\alias{get_time_elements}
\title{Get_time_elements}
\usage{
get_time_elements(sim, time_range, slot_name = "n")
}
\arguments{
\item{sim}{A MizerSim object}

\item{time_range}{A vector of times. Only the range of times is relevant,
i.e., all times between the smallest and largest will be selected.
The time_range can be character or numeric.}

\item{slot_name}{Obsolete. Was only needed in early versions of mizer where
the effort slot could have different time dimension from the other slots.}
}
\value{
Named boolean vector indicating for each time whether it is included
in the range or not.
}
\description{
Internal function to get the array element references of the time dimension
for the time based slots of a MizerSim object.
}
\concept{helper}
\keyword{internal}
