% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{get_time_elements}
\alias{get_time_elements}
\title{Get_time_elements}
\usage{
get_time_elements(sim, time_range, slot_name = "n")
}
\arguments{
\item{sim}{A MizerSim object}

\item{time_range}{The time_range can be character or numeric.}

\item{slot_name}{Necessary to include a slot_name argument because the effort
and abundance slots have different time dimensions}
}
\description{
Internal function to get the array element references of the time dimension
for the time based slots of a MizerSim object.
}
\concept{helper}
\keyword{internal}
