% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{getDiet}
\alias{getDiet}
\title{Get diet of predator resolved by prey species}
\usage{
getDiet(params, n, n_pp)
}
\arguments{
\item{params}{A MizerParams object}

\item{n}{An array (species x size) with the abundance density of fish}

\item{n_pp}{A vector with the abundance of plankton}
}
\value{
An array (predator species  x predator size x (prey species + plankton) )
}
\description{
Calculates the rate at which a predator of a particular species and size
consumes biomass of a prey species, where for this purpose we treat the
plankton like another prey species.
}
\details{
This function performs the same integration as
\code{getAvailEnergy()} but does not aggregate over prey species, and
multiplies by (1-feeding_level) to get the consumed biomass rather than the
available biomass. Outside the range of sizes for a predator species the
returned rate is zero.
}
