% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getEGrowth}
\alias{getEGrowth}
\title{Get energy rate available for growth}
\usage{
getEGrowth(object, n = object@initial_n, n_pp = object@initial_n_pp,
  B = object@initial_B, e_repro = getERepro(object, n = n, n_pp = n_pp,
  B = B), e = getEReproAndGrowth(object, n = n, n_pp = n_pp, B = B))
}
\arguments{
\item{object}{A \linkS4class{MizerParams} object.}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{e_repro}{The energy available for reproduction (optional, although if
specified, e must also be specified). A matrix of size no. species x no.
size bins. If not supplied, is calculated internally using the
\code{\link{getERepro}} method.}

\item{e}{The energy available for reproduction and growth (optional, although
if specified, e_repro must also be specified). A matrix of size no.
species x no. size bins. If not supplied, is calculated internally using
the \code{\link{getEReproAndGrowth}} method.}

\item{object}{An \linkS4class{MizerParams} object}
}
\value{
A two dimensional array (prey species x prey size)
}
\description{
Calculates the energy rate \eqn{g_i(w)} available by species and size for
growth after metabolism, movement and reproduction have been accounted for.
Used by the \code{\link{project}} method for performing simulations.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getEGrowth(params,sim@n[21,,],sim@n_pp[21,])
}
}
\seealso{
\code{\link{project}}
}
