% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getEncounter}
\alias{getEncounter}
\title{Get encounter rate}
\usage{
getEncounter(object, ...)
}
\arguments{
\item{object}{A \linkS4class{MizerParams} or \linkS4class{MizerSim} object.}

\item{...}{Additional arguments that depend on the class of \code{object}.

\strong{For a \linkS4class{MizerParams} object:}
\describe{
\item{\code{n}}{A matrix of species abundances (species x size). Defaults to
the initial abundances stored in \code{object}.}
\item{\code{n_pp}}{A vector of the resource abundance by size. Defaults to the
initial resource abundance stored in \code{object}.}
\item{\code{n_other}}{A named list of the abundances of other dynamical
components. Defaults to the initial values stored in \code{object}.}
\item{\code{t}}{The time for which to do the calculation. Defaults to 0.}
}

\strong{For a \linkS4class{MizerSim} object:}
\describe{
\item{\code{time_range}}{The time range over which to return the rates. Either
a vector of values, a vector of min and max time, or a single value.
Defaults to the whole time range of the simulation.}
\item{\code{drop}}{If \code{TRUE} then any dimension of length 1 is removed from the
returned array.}
}}
}
\value{
\itemize{
\item \code{MizerParams}: An \code{ArraySpeciesBySize} object (predator species x predator
size) with the encounter rates.
\item \code{MizerSim}: An \code{ArrayTimeBySpeciesBySize} object (time step x predator
species x predator size) with the encounter rates at every time step.
If \code{drop = TRUE} then dimensions of length 1 will be removed.
}
}
\description{
Returns the rate at which a predator of species \eqn{i} and
weight \eqn{w} encounters food (grams/year).
}
\section{Predation encounter}{

The encounter rate \eqn{E_i(w)} at which a predator of species \eqn{i}
and weight \eqn{w} encounters food has contributions from the encounter of
fish prey and of resource. This is determined by summing over all prey
species and the resource spectrum and then integrating over all prey sizes
\eqn{w_p}, weighted by predation kernel \eqn{\phi(w,w_p)}:
\deqn{
E_i(w) = \gamma_i(w) \int
\left( \theta_{ip} N_R(w_p) + \sum_{j} \theta_{ij} N_j(w_p) \right)
\phi_i(w,w_p) w_p \, dw_p.
}{\gamma_i(w) \int
( \theta_{ip} N_R(w_p) + \sum_{j} \theta_{ij} N_j(w_p) )
\phi_i(w,w_p) w_p dw_p.}
Here \eqn{N_j(w)} is the abundance density of species \eqn{j} and
\eqn{N_R(w)} is the abundance density of resource.
The overall prefactor \eqn{\gamma_i(w)} determines the predation power of the
predator. It could be interpreted as a search volume and is set with the
\code{\link[=setSearchVolume]{setSearchVolume()}} function. The predation kernel
\eqn{\phi(w,w_p)} is set with the \code{\link[=setPredKernel]{setPredKernel()}} function. The
species interaction matrix \eqn{\theta_{ij}} is set with \code{\link[=setInteraction]{setInteraction()}}
and the resource interaction vector \eqn{\theta_{ip}} is taken from the
\code{interaction_resource} column in \code{params@species_params}.
}

\section{Details}{

The encounter rate is multiplied by \eqn{1-f_0} to obtain the consumption
rate, where \eqn{f_0} is the feeding level calculated with
\code{\link[=getFeedingLevel]{getFeedingLevel()}}. This is used by the \code{\link[=project]{project()}} function for performing
simulations.

The function returns values also for sizes outside the size-range of the
species. These values should not be used, as they are meaningless.

If your model contains additional components that you added with
\code{\link[=setComponent]{setComponent()}} and for which you specified an \code{encounter_fun} function then
the encounters of these components will be included in the returned value.
}

\section{Extension hook}{

\code{projectEncounter()} is the S3 generic used by extension-aware projections.
Extension packages can add methods for their extension classes and call
\code{NextMethod()} to compose encounter-rate changes. The \code{MizerParams} method
contains the standard mizer calculation and is also exported as
\code{mizerEncounter()} for compatibility.
}

\section{Your own encounter function}{

By default \code{\link[=getEncounter]{getEncounter()}} calls \code{\link[=mizerEncounter]{mizerEncounter()}} on models without
extensions. However you can replace this with your own alternative encounter
function. If your function is called \code{"myEncounter"} then you register it in
a MizerParams object \code{params} with

\if{html}{\out{<div class="sourceCode">}}\preformatted{params <- setRateFunction(params, "Encounter", "myEncounter")
}\if{html}{\out{</div>}}

Your function will then be called instead of \code{\link[=mizerEncounter]{mizerEncounter()}}, with the
same arguments.
}

\examples{
encounter <- getEncounter(NS_params)
str(encounter)
}
\seealso{
Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFlux]{getFlux()}},
\code{\link[=getFluxGradient]{getFluxGradient()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getRates]{getRates()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
