% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getEncounter}
\alias{getEncounter}
\title{Get encounter rate}
\usage{
getEncounter(params, n = params@initial_n, n_pp = params@initial_n_pp,
  B = params@initial_B)
}
\arguments{
\item{params}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}
}
\value{
A two dimensional array (predator species x predator size)
}
\description{
Calculates the rate \eqn{E_i(w)} at which a predator of species \eqn{i} and
weight \eqn{w} encounters food (grams/year).
}
\section{Predation encounter}{

The encounter rate has contributions from the encounter of fish prey, of
plankton, and of other resoruces. The contribution from fish and plankton
is determined by summing over all prey species and the resource spectrum and
then integrating over all prey sizes \eqn{w_p}, weighted by predation kernel 
\eqn{\phi(w,w_p)}:
\deqn{
E_{e.i}(w) = \gamma_i(w) \int 
\left( \theta_{ip} N_R(w_p) + \sum_{j} \theta_{ij} N_j(w_p) \right) 
\phi_i(w,w_p) w_p \, dw_p.
}{\gamma_i(w) \int 
( \theta_{ip} N_R(w_p) + \sum_{j} \theta_{ij} N_j(w_p) ) 
\phi_i(w,w_p) w_p dw_p.}
Here \eqn{N_j(w)} is the abundance density of species \eqn{j} and
\eqn{N_R(w)} is the abundance density of plankton.
The overall prefactor \eqn{\gamma_i(w)} determines the predation power of the
predator. It could be interpreted as a search volume and is changed with the
\code{\link{setSearchVolume}} function. The predation kernel
\eqn{\phi(w,w_p)} is changed with the \code{\link{setPredKernel}} function. The
species interaction matrix \eqn{\theta_{ij}} and the plankton interaction
vector \eqn{\theta_{ip}} are changed with \code{\link{setInteraction}}.
}

\section{Resource encounter}{

In addition to the contribution from predation on fish prey and plankton,
the food encounter rate may have a contribution from unstructured resource
components. This takes the form
\deqn{E_{u.i} = \sum_d \rho_{id}(w) B_d.}
where \eqn{B_d} is the biomass of the d-th unstructured resource component
and \eqn{\rho_{id}(w)} is a parameter that therefore determines the rate at
which a predator of species \eqn{i} and size \eqn{w} encounters biomass from
the d-th unstructured resource component. This is changed with
\code{\link{setResourceEncounter}}.
}

\section{Details}{

The total encounter rate is the sum of the contribution from fish and
plankton and the contribution from unstructured resources, if any:
\deqn{E_i(w)=E_{e.i}(w)+E_{u.i}(w).}

The encounter rate is multiplied by \eqn{1-f_0} to obtain the consumption rate,
where \eqn{f_0} is the feeding level calculated with \code{\link{getFeedingLevel}}.
This is used by the \code{\link{project}} function for performing simulations.

The function returns values also for sizes outside the size-range of the
species. These values should not be used, as they are meaningless.
}

\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
# Run simulation with constant fishing effort for all gears for 20 years
sim <- project(params, t_max = 20, effort = 0.5)
n <- sim@n[21, , ]
n_pp <- sim@n_pp[21, ]
getEncounter(params, n, n_pp)
}
}
\seealso{
Other rate functions: \code{\link{getEGrowth}},
  \code{\link{getEReproAndGrowth}},
  \code{\link{getERepro}}, \code{\link{getFMortGear}},
  \code{\link{getFMort}}, \code{\link{getFeedingLevel}},
  \code{\link{getMort}}, \code{\link{getPlanktonMort}},
  \code{\link{getPredMort}}, \code{\link{getPredRate}},
  \code{\link{getRDD}}, \code{\link{getRDI}},
  \code{\link{getRates}}
}
\concept{rate functions}
