% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getERepro}
\alias{getERepro}
\title{Get energy rate available for reproduction}
\usage{
getERepro(params, n = params@initial_n, n_pp = params@initial_n_pp,
  B = params@initial_B, e = getEReproAndGrowth(params, n = n, n_pp =
  n_pp, B = B))
}
\arguments{
\item{params}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{e}{The energy available for reproduction and growth (optional). A
matrix of size no. species x no. size bins. If not supplied, is calculated
internally using \code{\link{getEReproAndGrowth}}.}
}
\value{
A two dimensional array (prey species x prey size)
}
\description{
Calculates the energy rate (grams/year) available by species and size for
reproduction after metabolism and movement have been accounted for:
\eqn{\psi_i(w)E_{r.i}(w)}. Used by the \code{project} function for performing
simulations.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getERepro(params,sim@n[21,,],sim@n_pp[21,])
}
}
\seealso{
\code{\link{getERepro}} and \code{\link{getEReproAndGrowth}}.

Other rate functions: \code{\link{getEGrowth}()},
  \code{\link{getEReproAndGrowth}()},
  \code{\link{getEncounter}()},
  \code{\link{getFMortGear}()}, \code{\link{getFMort}()},
  \code{\link{getFeedingLevel}()}, \code{\link{getMort}()},
  \code{\link{getPlanktonMort}()},
  \code{\link{getPredMort}()}, \code{\link{getPredRate}()},
  \code{\link{getRDD}()}, \code{\link{getRDI}()},
  \code{\link{getRates}()}
}
\concept{rate functions}
