% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getEReproAndGrowth}
\alias{getEReproAndGrowth}
\title{Get energy rate available for reproduction and growth}
\usage{
getEReproAndGrowth(
  object,
  n = initialN(object),
  n_pp = initialNResource(object),
  n_other = initialNOther(object),
  t = 0,
  ...
)
}
\arguments{
\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{...}{Unused}

\item{drop}{If \code{TRUE} then any dimension of length 1 will be removed
from the returned array.}
}
\value{
\itemize{
\item \code{MizerParams}: An \code{ArraySpeciesBySize} object (species x size) with the
energy rate \eqn{E_{r.i}(w)} available for growth and reproduction
(grams/year).
\item \code{MizerSim}: An \code{ArrayTimeBySpeciesBySize} object (time step x species x
size) with the energy rate at every time step. If \code{drop = TRUE} then
dimensions of length 1 will be removed.
}
}
\description{
Calculates the energy rate \eqn{E_{r.i}(w)} (grams/year) available for
reproduction and growth after metabolism and movement have been accounted
for.
}
\section{Your own energy rate function}{

By default \code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}} calls \code{\link[=mizerEReproAndGrowth]{mizerEReproAndGrowth()}}. However you
can replace this with your own alternative energy rate function. If
your function is called \code{"myEReproAndGrowth"} then you register it in a
MizerParams object \code{params} with

\if{html}{\out{<div class="sourceCode">}}\preformatted{params <- setRateFunction(params, "EReproAndGrowth", "myEReproAndGrowth")
}\if{html}{\out{</div>}}

Your function will then be called instead of \code{\link[=mizerEReproAndGrowth]{mizerEReproAndGrowth()}}, with
the same arguments.
}

\examples{
\donttest{
params <- NS_params
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
e <- getEReproAndGrowth(params, n = N(sim)[15, , ],
                        n_pp = NResource(sim)[15, ], t = 15)
# Rate at this time for Sprat of size 2g
e["Sprat", "2"]
}
}
\seealso{
The part of this energy rate that is invested into growth is
calculated with \code{\link[=getEGrowth]{getEGrowth()}} and the part that is invested into
reproduction is calculated with \code{\link[=getERepro]{getERepro()}}.

Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFlux]{getFlux()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getRates]{getRates()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
