% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getEReproAndGrowth}
\alias{getEReproAndGrowth}
\title{Get energy after metabolism and movement}
\usage{
getEReproAndGrowth(object, n = object@initial_n,
  n_pp = object@intitial_n_pp, B = object@initial_B,
  feeding_level = getFeedingLevel(object, n = n, n_pp = n_pp, B = B))
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{feeding_level}{The current feeding level (optional). A matrix of size
no. species x no. size bins. If not supplied, is calculated internally
using the \code{getFeedingLevel()} method.}
}
\value{
A two dimensional array (species x size)
}
\description{
Calculates the energy rate available by species and size for reproduction and
growth after metabolism and movement have been accounted for: \eqn{E_{r.i}(w)}.
Used by the \code{project} method for performing simulations.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getEReproAndGrowth(params,sim@n[21,,],sim@n_pp[21,])
}
}
\seealso{
\code{\link{project}} and \code{\link{getFeedingLevel}}.
}
