% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getEReproAndGrowth}
\alias{getEReproAndGrowth}
\title{Get energy rate available for reproduction and growth}
\usage{
getEReproAndGrowth(object, n = object@initial_n,
  n_pp = object@initial_n_pp, B = object@initial_B,
  encounter = getEncounter(object, n = n, n_pp = n_pp, B = B),
  feeding_level = getFeedingLevel(object, n = n, n_pp = n_pp, B = B,
  encounter = encounter))
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{encounter}{The encounter rate matrix (optional) of dimension no.
species x no. size bins. If not passed in, it is calculated internally
using the \code{\link{getEncounter}} function.}

\item{feeding_level}{The current feeding level (optional). A matrix of size
no. species x no. size bins. If not supplied, is calculated internally
using the \code{\link{getFeedingLevel}} function.}
}
\value{
A two dimensional array (species x size)
}
\description{
Calculates the energy rate (grams/year) available by species and size for
reproduction and growth after metabolism and movement have been accounted
for: \eqn{E_{r.i}(w)}. Used by the \code{project} function for performing
simulations.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getEReproAndGrowth(params,sim@n[21,,],sim@n_pp[21,])
}
}
\seealso{
\code{\link{project}} and \code{\link{getFeedingLevel}}.
}
