% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getEReproAndGrowth}
\alias{getEReproAndGrowth}
\title{Get energy rate available for reproduction and growth}
\usage{
getEReproAndGrowth(params, n = params@initial_n,
  n_pp = params@initial_n_pp, B = params@initial_B,
  encounter = getEncounter(params, n = n, n_pp = n_pp, B = B),
  feeding_level = getFeedingLevel(params, n = n, n_pp = n_pp, B = B,
  encounter = encounter))
}
\arguments{
\item{params}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{encounter}{The encounter rate matrix (optional) of dimension no.
species x no. size bins. If not passed in, it is calculated internally
using the \code{\link{getEncounter}} function.}

\item{feeding_level}{The current feeding level (optional). A matrix of size
no. species x no. size bins. If not supplied, is calculated internally
using the \code{\link{getFeedingLevel}} function.}
}
\value{
A two dimensional array (species x size) holding
\deqn{E_{r.i}(w) = \max(0, \alpha_i\, (1 - {\tt feeding\_level}_i(w))\, 
                           {\tt encounter}_i(w) - {\tt metab}_i(w)).}{
  E_{r.i}(w) = max(0, \alpha_i * (1 - feeding_level_i(w)) * 
                      encounter_i(w) - metab_i(w)).}
The assimilation rate \eqn{\alpha_i} is taken from the species parameter
data frame in \code{params}. The metabolic rate \code{metab} is taken from 
\code{params}.
}
\description{
Calculates the energy rate (grams/year) available by species and size for
reproduction and growth after metabolism and movement have been accounted
for.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getEReproAndGrowth(params,sim@n[21,,],sim@n_pp[21,])
}
}
\seealso{
\code{\link{getERepro}} and \code{\link{getEGrowth}}

Other rate functions: \code{\link{getEGrowth}},
  \code{\link{getERepro}}, \code{\link{getEncounter}},
  \code{\link{getFMortGear}}, \code{\link{getFMort}},
  \code{\link{getFeedingLevel}}, \code{\link{getMort}},
  \code{\link{getPlanktonMort}}, \code{\link{getPredMort}},
  \code{\link{getPredRate}}, \code{\link{getRDD}},
  \code{\link{getRDI}}, \code{\link{getRates}}
}
\concept{rate functions}
