% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getESpawning}
\alias{getESpawning}
\title{Alias for getERepro}
\usage{
getESpawning(object, n, n_pp, e = getEReproAndGrowth(object, n = n, n_pp =
  n_pp))
}
\arguments{
\item{object}{A \code{MizerParams} object.}

\item{n}{A matrix of species abundance (species x size).}

\item{n_pp}{A vector of the plankton abundance by size.}

\item{e}{The energy available for reproduction and growth (optional). A
matrix of size no. species x no. size bins. If not supplied, is calculated
internally using the \code{getEReproAndGrowth()} method.}
}
\value{
A two dimensional array (prey species x prey size)
}
\description{
An alias provided for backward compatibility with mizer version <= 1.0
}
\seealso{
\code{\link{project}} and \code{\link{getEReproAndGrowth}}.
}
