% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getFlux}
\alias{getFlux}
\title{Get flux into size bins}
\usage{
getFlux(
  object,
  n = initialN(object),
  n_pp = initialNResource(object),
  n_other = initialNOther(object),
  t = 0,
  ...
)
}
\arguments{
\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{...}{Unused}

\item{drop}{If \code{TRUE} then any dimension of length 1 will be removed
from the returned array.}
}
\value{
\itemize{
\item \code{MizerParams}: An \code{ArraySpeciesBySize} object (species x size) with the
flux of individuals entering each size class (numbers/year).
\item \code{MizerSim}: An \code{ArrayTimeBySpeciesBySize} object (time step x species x
size) with the flux at every time step. If \code{drop = TRUE} then dimensions
of length 1 will be removed.
}
}
\description{
Calculates the flux \eqn{J_i(w)} (numbers/year) entering each size class
from the one below it. This is composed of an advective flux from somatic
growth and a diffusive flux from the redistribution of individuals.
}
\details{
At the recruitment size, the flux is simply the recruitment rate
\eqn{R_{dd,i}} (see \code{\link[=getRDD]{getRDD()}}). For sizes below the recruitment size
the flux is zero.
}
\examples{
\donttest{
params <- NS_params
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the flux at a particular time step
flux <- getFlux(params, n = N(sim)[15, , ], n_pp = NResource(sim)[15, ], t = 15)
# Flux for Sprat of size 2g
flux["Sprat", "2"]
}
}
\seealso{
\code{\link[=getEGrowth]{getEGrowth()}}, \code{\link[=getRDD]{getRDD()}}

Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getRates]{getRates()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
