% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getFlux}
\alias{getFlux}
\title{Get flux into size bins}
\usage{
getFlux(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  t = 0,
  ...
)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{...}{Unused}
}
\value{
A \code{ArraySpeciesBySize} object (species x size) with the flux of individuals
entering each size class (numbers/year).
}
\description{
Calculates the flux \eqn{J_i(w)} (numbers/year) entering each size class
from the one below it. This is composed of an advective flux from somatic
growth and a diffusive flux from the redistribution of individuals.
}
\details{
At the recruitment size, the flux is simply the recruitment rate
\eqn{R_{dd,i}} (see \code{\link[=getRDD]{getRDD()}}). For sizes below the recruitment size
the flux is zero.
}
\examples{
\donttest{
params <- NS_params
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the flux at a particular time step
flux <- getFlux(params, n = N(sim)[15, , ], n_pp = NResource(sim)[15, ], t = 15)
# Flux for Sprat of size 2g
flux["Sprat", "2"]
}
}
\seealso{
\code{\link[=getEGrowth]{getEGrowth()}}, \code{\link[=getRDD]{getRDD()}}

Other rate functions: 
\code{\link{getEGrowth}()},
\code{\link{getERepro}()},
\code{\link{getEReproAndGrowth}()},
\code{\link{getEncounter}()},
\code{\link{getFMort}()},
\code{\link{getFMortGear}()},
\code{\link{getFeedingLevel}()},
\code{\link{getMort}()},
\code{\link{getPredMort}()},
\code{\link{getPredRate}()},
\code{\link{getRDD}()},
\code{\link{getRDI}()},
\code{\link{getRates}()},
\code{\link{getResourceMort}()}
}
\concept{rate functions}
