% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getGrowthCurves}
\alias{getGrowthCurves}
\title{Get growth curves giving weight as a function of age}
\usage{
getGrowthCurves(object, max_age = 24)
}
\arguments{
\item{object}{MizerSim or MizerParams object}

\item{max_age}{The age up to which the weight is to be plotted. Default is 20}
}
\value{
A ggplot2 object
}
\description{
If given a \linkS4class{MizerSim} object, uses the growth rates at the final
time of a simulation to calculate the size at age. If given a
\linkS4class{MizerParams} object, uses the initial growth rates instead.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
getGrowthCurves(params)
sim <- project(params, effort=1, t_max=20, t_save = 2)
getGrowthCurves(sim, max_age = 24)
}
}
