% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getM2Background}
\alias{getM2Background}
\title{Alias for getPlanktonMort}
\usage{
getM2Background(object, n, n_pp, pred_rate = getPredRate(object, n = n, n_pp =
  n_pp))
}
\arguments{
\item{object}{A \code{MizerParams} object.}

\item{n}{A matrix of species abundance (species x size).}

\item{n_pp}{A vector of the plankton abundance by size.}

\item{pred_rate}{An array of predation rates of dimension no. sp x no.
community size bins x no. of size bins in whole spectra (i.e. community +
plankton, the w_full slot). The array is optional. If it is not provided
it is calculated by the \code{getPredRate()} method.}
}
\value{
A vector of mortality rate by plankton size.
}
\description{
An alias provided for backward compatibility with mizer version <= 1.0
}
\seealso{
\code{\link{project}} and \code{\link{getPredMort}}.
}
