% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getM2Background}
\alias{getM2Background}
\title{Alias for getPlanktonMort}
\usage{
getM2Background(object, n, n_pp, B = 0, pred_rate = getPredRate(object,
  n = n, n_pp = n_pp, B = B))
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{pred_rate}{An array of predation rates of dimension no. sp x no.
community size bins x no. of size bins in whole spectra (i.e. community +
plankton, the w_full slot). The array is optional. If it is not provided
it is calculated by the \code{getPredRate()} method.}
}
\value{
A vector of mortality rate by plankton size.
}
\description{
An alias provided for backward compatibility with mizer version <= 1.0
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get plankton mortality at one time step
n <- sim@n[21,,]
n_pp <- sim@n_pp[21,]
getPlanktonMort(params,n,n_pp)
}
}
\seealso{
\code{\link{project}} and \code{\link{getPredMort}}.
}
