% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getMort}
\alias{getMort}
\title{Get total mortality rate}
\usage{
getMort(params, n = params@initial_n, n_pp = params@initial_n_pp,
  B = params@initial_B, effort = 1, m2 = getPredMort(params, n = n,
  n_pp = n_pp, B = B))
}
\arguments{
\item{params}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{effort}{A numeric vector of the effort by gear or a single numeric
effort value which is used for all gears. Default 1.}

\item{m2}{A two dimensional array of predation mortality (optional). Has
dimensions no. sp x no. size bins in the community. If not supplied is
calculated using the \code{\link{getPredMort}} function.}
}
\value{
A two dimensional array (prey species x prey size).
}
\description{
Calculates the total mortality rate \eqn{\mu_i(w)}  (in units 1/year) on each
species by size from predation mortality, background mortality and fishing
mortality for a single time step.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the total mortality at a particular time step
getMort(params,sim@n[21,,],sim@n_pp[21,],effort=0.5)
}
}
\seealso{
\code{\link{getPredMort}}, \code{\link{getFMort}}

Other rate functions: \code{\link{getEGrowth}},
  \code{\link{getEReproAndGrowth}},
  \code{\link{getERepro}}, \code{\link{getEncounter}},
  \code{\link{getFMortGear}}, \code{\link{getFMort}},
  \code{\link{getFeedingLevel}},
  \code{\link{getPlanktonMort}}, \code{\link{getPredMort}},
  \code{\link{getPredRate}}, \code{\link{getRDD}},
  \code{\link{getRDI}}, \code{\link{getRates}}
}
\concept{rate functions}
