% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{getN}
\alias{getN}
\title{Calculate the number of individuals within a size range}
\usage{
getN(object, ...)
}
\arguments{
\item{object}{An object of class \code{MizerParams} or \code{MizerSim}.}

\item{...}{
  Arguments passed on to \code{\link{get_size_range_array}}
  \describe{
    \item{\code{min_w}}{Smallest weight in size range. Defaults to smallest weight in
the model.}
    \item{\code{max_w}}{Largest weight in size range. Defaults to largest weight in the
model.}
    \item{\code{min_l}}{Smallest length in size range. If supplied, this takes
precedence over \code{min_w}.}
    \item{\code{max_l}}{Largest length in size range. If supplied, this takes precedence
over \code{max_w}.}
  }}
}
\value{
If called with a MizerParams object, a named vector with the numbers
for each species in the model. If called with a MizerSim object, a
\code{ArrayTimeBySpecies} object (time x species) containing the numbers at
each time step for all species.
}
\description{
Calculates the number of individuals within user-defined size limits. The
default option is to use the whole size range. You can specify minimum and
maximum weight or lengths for the species. Lengths take precedence over
weights (i.e. if both min_l and min_w are supplied, only min_l will be used)
}
\examples{
numbers <- getN(NS_sim)
numbers["1972", "Herring"]
# The above gave a huge number, because that included all the larvae.
# The number of Herrings between 10g and 1kg is much smaller.
numbers <- getN(NS_sim, min_w = 10, max_w = 1000)
numbers["1972", "Herring"]
}
\seealso{
Other summary functions:
\code{\link[=getBiomass]{getBiomass()}},
\code{\link[=getDiet]{getDiet()}},
\code{\link[=getGrowthCurves]{getGrowthCurves()}},
\code{\link[=getSSB]{getSSB()}},
\code{\link[=getSteadyResidual]{getSteadyResidual()}},
\code{\link[=getTrophicLevel]{getTrophicLevel()}},
\code{\link[=getTrophicLevelBySpecies]{getTrophicLevelBySpecies()}},
\code{\link[=getYield]{getYield()}},
\code{\link[=getYieldGear]{getYieldGear()}},
\code{\link[=sizeIntegral]{sizeIntegral()}}
}
\concept{summary functions}
\concept{summary_function}
