% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{getParams}
\alias{getParams}
\title{Extract the parameter object underlying a simulation}
\usage{
getParams(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
The MizerParams object that was used to run the simulation
}
\description{
Extract the parameter object underlying a simulation
}
\examples{
# This will be identical to the params object that was used to create the
# simulation
sim <- project(NS_params, t_max = 1)
identical(getParams(sim), NS_params)
}
