% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getPhiPrey}
\alias{getPhiPrey}
\title{Get available energy}
\usage{
getPhiPrey(object, n, n_pp, ...)
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the background abundance by size}

\item{...}{Other arguments (currently unused)}
}
\value{
A two dimensional array (predator species x predator size)
equal to \code{getEncounter(object, n, n_pp) / getSearchVolume(object)}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This is deprecated and is no longer used by the mizer project() method.
Calculates the amount \eqn{E_{a,i}(w)} of food exposed to each predator as
a function of predator size.
}
\seealso{
\code{\link[=project]{project()}}
}
\concept{deprecated}
