% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getPhiPrey}
\alias{getPhiPrey}
\title{Alias for getAvailEnergy}
\usage{
getPhiPrey(object, n, n_pp)
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}
}
\value{
A two dimensional array (predator species x predator size)
}
\description{
An alias provided for backward compatibility with mizer version <= 1.0
}
\details{
This method is used by the \code{\link{project}} method for
performing simulations.
}
\seealso{
\code{\link{project}}
}
