% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getPhiPrey}
\alias{getPhiPrey}
\title{Get available energy}
\usage{
getPhiPrey(object, n, n_pp, ...)
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the background abundance by size}

\item{...}{Other arguments (currently unused)}
}
\value{
A two dimensional array (predator species x predator size)
}
\description{
This is deprecated and is no longer used by the mizer project() method.
Calculates the amount \eqn{E_{a,i}(w)} of food exposed to each predator as
a function of predator size.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
n <- sim@n[21,,]
n_pp <- sim@n_pp[21,]
getPhiPrey(params,n,n_pp)
}
}
\seealso{
\code{\link[=project]{project()}}
}
