% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getPredMort}
\alias{getPredMort}
\title{Get total predation mortality rate}
\usage{
getPredMort(object, n, n_pp, n_other, time_range, drop = TRUE, ...)
}
\arguments{
\item{object}{A \code{MizerParams} object or a \code{MizerSim} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{time_range}{A numeric or character vector of times. Only the range of
values matters, so all saved times between \code{min(time_range)} and
\code{max(time_range)} are selected.}

\item{drop}{If \code{TRUE} then any dimension of length 1 will be removed
from the returned array.}

\item{...}{Unused}
}
\value{
\itemize{
\item \code{MizerParams}: An \code{ArraySpeciesBySize} object (prey species x prey size)
with the predation mortality rates.
\item \code{MizerSim}: An \code{ArrayTimeBySpeciesBySize} object (time step x prey species
x prey size) with the predation mortality at every time step.
If \code{drop = TRUE} then dimensions of length 1 will be removed.
}
}
\description{
Calculates the total predation mortality rate \eqn{\mu_{p,i}(w_p)} (in units
of 1/year) on each prey species by prey size:
\deqn{\mu_{p.i}(w_p) = \sum_j {\tt pred\_rate}_j(w_p)\, \theta_{ji}.}{
  \mu_{p.i}(w_p) = \sum_j pred_rate_j(w_p) \theta_{ji}.}
The predation rate \code{pred_rate} is returned by \code{\link[=getPredRate]{getPredRate()}}.
}
\section{Your own predation mortality function}{

By default \code{\link[=getPredMort]{getPredMort()}} calls \code{\link[=mizerPredMort]{mizerPredMort()}}. However you can
replace this with your own alternative predation mortality function. If
your function is called \code{"myPredMort"} then you register it in a MizerParams
object \code{params} with

\if{html}{\out{<div class="sourceCode">}}\preformatted{params <- setRateFunction(params, "PredMort", "myPredMort")
}\if{html}{\out{</div>}}

Your function will then be called instead of \code{\link[=mizerPredMort]{mizerPredMort()}}, with the
same arguments.
}

\examples{
\donttest{
params <- NS_params
# Predation mortality in initial state
M2 <- getPredMort(params)
str(M2)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get predation mortality at one time step
M2 <- getPredMort(params, n = N(sim)[15, , ], n_pp = NResource(sim)[15, ])
# Get predation mortality at all saved time steps
M2 <- getPredMort(sim)
str(M2)
# Get predation mortality over the years 15 - 20
M2 <- getPredMort(sim, time_range = c(15, 20))
}
}
\seealso{
Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFlux]{getFlux()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getRates]{getRates()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
