% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getPredRate}
\alias{getPredRate}
\title{Get predation rate}
\usage{
getPredRate(object, ...)
}
\arguments{
\item{object}{A \linkS4class{MizerParams} or \linkS4class{MizerSim} object.}

\item{...}{Additional arguments that depend on the class of \code{object}.

\strong{For a \linkS4class{MizerParams} object:}
\describe{
\item{\code{n}}{A matrix of species abundances (species x size). Defaults to
the initial abundances stored in \code{object}.}
\item{\code{n_pp}}{A vector of the resource abundance by size. Defaults to the
initial resource abundance stored in \code{object}.}
\item{\code{n_other}}{A named list of the abundances of other dynamical
components. Defaults to the initial values stored in \code{object}.}
\item{\code{t}}{The time for which to do the calculation. Defaults to 0.}
}

\strong{For a \linkS4class{MizerSim} object:}
\describe{
\item{\code{time_range}}{The time range over which to return the rates. Either
a vector of values, a vector of min and max time, or a single value.
Defaults to the whole time range of the simulation.}
\item{\code{drop}}{If \code{TRUE} then any dimension of length 1 is removed from the
returned array.}
}}
}
\value{
\itemize{
\item \code{MizerParams}: An \code{ArraySpeciesBySize} object (predator species x prey
size), where the prey size runs over fish community plus resource spectrum.
\item \code{MizerSim}: An \code{ArrayTimeBySpeciesBySize} object (time step x predator
species x prey size) with the predation rates at every time step.
If \code{drop = TRUE} then dimensions of length 1 will be removed.
}
}
\description{
Calculates the potential rate (in units 1/year) at which a prey individual of
a given size \eqn{w} is killed by predators from species \eqn{j}. In formulas
\deqn{{\tt pred\_rate}_j(w_p) = \int \phi_j(w,w_p) (1-f_j(w))
  \gamma_j(w) N_j(w) \, dw.}{pred_rate_j(w_p) = \int\phi_i(w,w_p) (1-f_i(w))
  \gamma_i(w) N_i(w) dw.}
This potential rate is used in \code{\link[=getPredMort]{getPredMort()}} to
calculate the realised predation mortality rate on the prey individual.
}
\section{Your own predation rate function}{

By default \code{\link[=getPredRate]{getPredRate()}} calls \code{\link[=mizerPredRate]{mizerPredRate()}}. However you can
replace this with your own alternative predation rate function. If
your function is called \code{"myPredRate"} then you register it in a MizerParams
object \code{params} with

\if{html}{\out{<div class="sourceCode">}}\preformatted{params <- setRateFunction(params, "PredRate", "myPredRate")
}\if{html}{\out{</div>}}

Your function will then be called instead of \code{\link[=mizerPredRate]{mizerPredRate()}}, with
the same arguments.
}

\examples{
\donttest{
params <- NS_params
# Predation rate in initial state
pred_rate <- getPredRate(params)
str(pred_rate)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the feeding level at one time step
pred_rate <- getPredRate(params, n = N(sim)[15, , ],
                         n_pp = NResource(sim)[15, ], t = 15)
}
}
\seealso{
Other rate functions: 
\code{\link{getDiffusion}()},
\code{\link{getEGrowth}()},
\code{\link{getERepro}()},
\code{\link{getEReproAndGrowth}()},
\code{\link{getEncounter}()},
\code{\link{getFMort}()},
\code{\link{getFMortGear}()},
\code{\link{getFeedingLevel}()},
\code{\link{getFlux}()},
\code{\link{getFluxGradient}()},
\code{\link{getMort}()},
\code{\link{getPredMort}()},
\code{\link{getRDD}()},
\code{\link{getRDI}()},
\code{\link{getRates}()},
\code{\link{getResourceMort}()}
}
\concept{rate functions}
