% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getPredRate}
\alias{getPredRate}
\title{Get predation rate}
\usage{
getPredRate(params, n = params@initial_n, n_pp = params@initial_n_pp,
  B = params@initial_B, feeding_level = getFeedingLevel(params, n = n,
  n_pp = n_pp, B = B))
}
\arguments{
\item{params}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{feeding_level}{The current feeding level (optional). A matrix of size
no. species x no. size bins. If not supplied, is calculated internally
using the \code{\link{getFeedingLevel}} function.}
}
\value{
A two dimensional array (predator species x prey size), 
  where the prey size runs over fish community plus plankton spectrum.
}
\description{
Calculates the potential rate (in units 1/year) at which a prey individual of
a given size \eqn{w} is killed by predators from species \eqn{j}. In formulas
\deqn{{\tt pred\_rate}_j(w_p) = \int \phi_j(w,w_p) (1-f_j(w)) 
  \gamma_j(w) N_j(w) \, dw.}{pred_rate_j(w_p) = \int\phi_i(w,w_p) (1-f_i(w)) 
  \gamma_i(w) N_i(w) dw.}
This potential rate is used in the function \code{\link{getPredMort}} to
calculate the realised predation mortality rate on the prey individual.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the feeding level at one time step
n <- sim@n[21,,]
n_pp <- sim@n_pp[21,]
getPredRate(params,n,n_pp)
}
}
\seealso{
Other rate functions: \code{\link{getEGrowth}()},
  \code{\link{getEReproAndGrowth}()},
  \code{\link{getERepro}()}, \code{\link{getEncounter}()},
  \code{\link{getFMortGear}()}, \code{\link{getFMort}()},
  \code{\link{getFeedingLevel}()}, \code{\link{getMort}()},
  \code{\link{getPlanktonMort}()},
  \code{\link{getPredMort}()}, \code{\link{getRDD}()},
  \code{\link{getRDI}()}, \code{\link{getRates}()}
}
\concept{rate functions}
