% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getPredRate}
\alias{getPredRate}
\title{Get predation rate}
\usage{
getPredRate(object, n, n_pp, feeding_level = getFeedingLevel(object, n =
  n, n_pp = n_pp))
}
\arguments{
\item{object}{A \code{MizerParams} object.}

\item{n}{A matrix of species abundance (species x size).}

\item{n_pp}{A vector of the plankton abundance by size.}

\item{feeding_level}{The current feeding level (optional). A matrix of size
no. species x no. size bins. If not supplied, is calculated internally
using the \code{getFeedingLevel()} method.}
}
\value{
A two dimensional array (predator species x prey size), 
  where the prey size runs over fish community plus plankton spectrum.
}
\description{
Calculates the potential rate at which a prey individual of a given size 
\eqn{w} is killed by predators from species \eqn{i}. 
In formulas \deqn{\int\phi_i(w_p/w) (1-f_i(w)) \gamma_i w^q N_i(w) dw}
This potential rate is used in the function \code{\link{getPredMort}} to
calculate the realised predation mortality rate on the prey individual.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the feeding level at one time step
n <- sim@n[21,,]
n_pp <- sim@n_pp[21,]
getPredRate(params,n,n_pp)
}
}
\seealso{
\code{\link{getPredMort}} and
  \code{\link{getFeedingLevel}}
}
