% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getRates}
\alias{getRates}
\title{Get all rates}
\usage{
getRates(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  effort,
  t = 0,
  ...
)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{effort}{The effort for each fishing gear}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{...}{Unused}
}
\value{
List of rates.
}
\description{
Calls other rate functions in sequence and collects the results in a list.

\code{projectRates()} is an S3 generic used by extension-aware
projections to calculate all rates. Models without extensions keep using
\code{\link[=mizerRates]{mizerRates()}} directly. The base method mirrors \code{\link[=mizerRates]{mizerRates()}} but calls
migrated projection hooks directly, starting with \code{\link[=projectEncounter]{projectEncounter()}}.
}
\details{
By default this function returns a list with the following components:
\itemize{
\item encounter from \code{\link[=mizerEncounter]{mizerEncounter()}}
\item feeding_level from \code{\link[=mizerFeedingLevel]{mizerFeedingLevel()}}
\item e from \code{\link[=mizerEReproAndGrowth]{mizerEReproAndGrowth()}}
\item e_repro from \code{\link[=mizerERepro]{mizerERepro()}}
\item e_growth from \code{\link[=mizerEGrowth]{mizerEGrowth()}}
\item pred_rate from \code{\link[=mizerPredRate]{mizerPredRate()}}
\item pred_mort from \code{\link[=mizerPredMort]{mizerPredMort()}}
\item f_mort from \code{\link[=mizerFMort]{mizerFMort()}}
\item mort from \code{\link[=mizerMort]{mizerMort()}}
\item rdi from \code{\link[=mizerRDI]{mizerRDI()}}
\item rdd from \code{\link[=BevertonHoltRDD]{BevertonHoltRDD()}}
\item resource_mort from \code{\link[=mizerResourceMort]{mizerResourceMort()}}
}
However you can replace any of these rate functions by your own rate
function if you wish, see \code{\link[=setRateFunction]{setRateFunction()}} for details.
}
\examples{
rates <- getRates(NS_params)
names(rates)
identical(rates$encounter, getEncounter(NS_params))
}
\seealso{
Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFlux]{getFlux()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
