% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getRates}
\alias{getRates}
\title{Get all rates}
\usage{
getRates(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  effort,
  t = 0,
  ...
)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{effort}{The effort for each fishing gear}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{...}{Unused}
}
\description{
Calls other rate functions in sequence and collects the results in a list.
The rates returned are encounter, feeding level, energy for growth and
reproduction, predation rate, predation mortality, and resource mortality.
The purpose of this function is to provide a convenient way to get all the
rates at once, and to ensure that they are all calculated at the same time
step with the same inputs. The rates are returned in a list with the same
names as the rate functions that calculate them, so for example the encounter
rate is returned in the list element named "encounter" and is calculated with
the getEncounter() function.
}
\details{
When mizer needs to calculate the rates during a simulation it does not use
this function but instead the faster \code{projectRates()}.
}
\examples{
rates <- getRates(NS_params)
names(rates)
identical(rates$encounter, getEncounter(NS_params))
}
\seealso{
Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFlux]{getFlux()}},
\code{\link[=getMort]{getMort()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{rate functions}
