% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getRates}
\alias{getRates}
\title{Get all rates}
\usage{
getRates(
  params,
  n = params@initial_n,
  n_pp = params@initial_n_pp,
  n_other = params@initial_n_other,
  t = 0,
  effort = params@initial_effort,
  rates_fns = lapply(params@rates_funcs, get)
)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the plankton abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The current time}

\item{effort}{The effort for each fishing gear}

\item{rates_fns}{Named list of the functions to call to calculate the rates.
Note that this list holds the functions themselves, not their names.}
}
\value{
A list with the following components:
  \itemize{
    \item encounter from \code{\link{getEncounter}}
    \item feeding_level from \code{\link{getFeedingLevel}}
    \item pred_rate from \code{\link{getPredRate}}
    \item pred_mort from \code{\link{getPredMort}}
    \item mort from \code{\link{getMort}}
    \item plankton_mort from \code{\link{getPlanktonMort}}
    \item e from \code{\link{getEReproAndGrowth}}
    \item e_repro from \code{\link{getERepro}}
    \item e_growth from \code{\link{getEGrowth}}
    \item rdi from \code{\link{getRDI}}
    \item rdd from \code{\link{getRDD}}
  }
}
\description{
Calls all the other rate functions in sequence and collects the results in a
list.
}
\seealso{
Other rate functions: 
\code{\link{getEGrowth}()},
\code{\link{getEReproAndGrowth}()},
\code{\link{getERepro}()},
\code{\link{getEncounter}()},
\code{\link{getFMortGear}()},
\code{\link{getFMort}()},
\code{\link{getFeedingLevel}()},
\code{\link{getMort}()},
\code{\link{getPlanktonMort}()},
\code{\link{getPredMort}()},
\code{\link{getPredRate}()},
\code{\link{getRDD}()},
\code{\link{getRDI}()}
}
\concept{rate functions}
