% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getRates}
\alias{getRates}
\title{Get all rates}
\usage{
getRates(object, n = object@initial_n, n_pp = object@initial_n_pp,
  B = object@initial_B, effort = 0, sex_ratio = 0.5)
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{effort}{The effort of each fishing gear}

\item{sex_ratio}{Proportion of the population that is female. Default value
is 0.5.}
}
\value{
A list with the following components:
  \itemize{
    \item encounter from \code{\link{getEncounter}}
    \item feeding_level from \code{\link{getFeedingLevel}}
    \item pred_rate from \code{\link{getPredRate}}
    \item pred_mort from \code{\link{getPredMort}}
    \item mort from \code{\link{getMort}}
    \item plankton_mort from \code{\link{getPlanktonMort}}
    \item e from \code{\link{getEReproAndGrowth}}
    \item e_repro from \code{\link{getERepro}}
    \item e_growth from \code{\link{getEGrowth}}
    \item rdi from \code{\link{getRDI}}
    \item rdd from \code{\link{getRDD}}
  }
}
\description{
Calls all the other rate functions in sequence and collects the results in a
list.
}
