% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getRDD}
\alias{getRDD}
\title{Get density dependent recruitment}
\usage{
getRDD(object, n, n_pp, sex_ratio = 0.5, rdi = getRDI(object, n = n,
  n_pp = n_pp, sex_ratio = sex_ratio))
}
\arguments{
\item{object}{An \code{MizerParams} object}

\item{n}{A matrix of species abundance (species x size)}

\item{n_pp}{A vector of the background abundance by size}

\item{rdi}{A vector of density independent recruitment for each species. 
If not specified rdi is calculated internally using
the \code{\link{getRDI}} method.}
}
\value{
A numeric vector the length of the number of species.
}
\description{
Calculates the density dependent recruitment (total egg production) \eqn{R_i}
for each species. This is the flux entering the smallest size class of each
species. The density dependent recruitment is the density independent
recruitment after it has been put through the density dependent
stock-recruitment relationship function. This method is used by the
\code{project} method for performing simulations.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getRDD(params,sim@n[21,,],sim@n_pp[21,])
}
}
