% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getRDD}
\alias{getRDD}
\title{Get density dependent recruitment}
\usage{
getRDD(object, n = object@initial_n, n_pp = object@initial_n_pp,
  B = object@initial_B, sex_ratio = 0.5, rdi = getRDI(object, n = n,
  n_pp = n_pp, B = B, sex_ratio = sex_ratio))
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{sex_ratio}{Proportion of the population that is female. Default value
is 0.5.}

\item{rdi}{A vector of density independent recruitment for each species. 
If not specified rdi is calculated internally using
the \code{\link{getRDI}} method.}
}
\value{
A numeric vector the length of the number of species.
}
\description{
Calculates the density dependent recruitment (total egg production) \eqn{R_i}
for each species. This is the flux entering the smallest size class of each
species. The density dependent recruitment is the density independent
recruitment after it has been put through the density dependent
stock-recruitment relationship function.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getRDD(params,sim@n[21,,],sim@n_pp[21,])
}
}
