% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getRDI}
\alias{getRDI}
\title{Get density independent recruitment}
\usage{
getRDI(object, n, n_pp, B = 0, e_repro = getERepro(object, n = n, n_pp
  = n_pp, B = B), sex_ratio = 0.5)
}
\arguments{
\item{object}{An \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{A vector of biomasses of unstructured resource components}

\item{e_repro}{The energy available for reproduction (optional). A matrix of
size no. species x no. size bins. If not supplied, is calculated internally
using the \code{\link{getERepro}} method.}

\item{sex_ratio}{Proportion of the population that is female. Default value
is 0.5.}
}
\value{
A numeric vector the length of the number of species
}
\description{
Calculates the density independent recruitment (total egg production)
\eqn{R_{p.i}} before density dependence, by species. Used by the
\code{project} method for performing simulations.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the recruitment at a particular time step
getRDI(params,sim@n[21,,],sim@n_pp[21,])
}
}
\seealso{
\code{\link{project}}
}
