% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{getRegisteredExtensions}
\alias{getRegisteredExtensions}
\title{Get the registered mizer extension chain}
\usage{
getRegisteredExtensions()
}
\value{
A named character vector giving the maximal extension chain
registered for this R session.
}
\description{
Get the registered mizer extension chain
}
\seealso{
Other extensions:
\code{\link[=NOther]{NOther()}},
\code{\link[=coerceToExtensionClass]{coerceToExtensionClass()}},
\code{\link[=initialNOther<-]{initialNOther<-()}},
\code{\link[=registerExtension]{registerExtension()}},
\code{\link[=registerExtensions]{registerExtensions()}},
\code{\link[=resetMizerSession]{resetMizerSession()}},
\code{\link[=setComponent]{setComponent()}},
\code{\link[=setRateFunction]{setRateFunction()}}
}
\concept{extensions}
