% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{getRegisteredExtensions}
\alias{getRegisteredExtensions}
\title{Get the registered mizer extension chain}
\usage{
getRegisteredExtensions()
}
\value{
A named character vector giving the maximal extension chain
registered for this R session.
}
\description{
Get the registered mizer extension chain
}
\seealso{
"Using mizer extension packages":
\code{vignette("using-extension-packages", package = "mizer")}

Other extension tools: 
\code{\link{NOther}()},
\code{\link{clearExtensionChain}()},
\code{\link{coerceToExtensionClass}()},
\code{\link{initialNOther<-}()},
\code{\link{recordExtension}()},
\code{\link{registerExtension}()},
\code{\link{registerExtensions}()},
\code{\link{setComponent}()},
\code{\link{setRateFunction}()}
}
\concept{extension tools}
