% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerExtensions.R
\name{getRegisteredExtensions}
\alias{getRegisteredExtensions}
\title{Get the registered mizer extension chain}
\usage{
getRegisteredExtensions()
}
\value{
A named character vector giving the maximal extension chain
registered for this R session.
}
\description{
Get the registered mizer extension chain
}
