% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRequiredRDD.R
\name{getRequiredRDD}
\alias{getRequiredRDD}
\title{Determine reproduction rate needed for initial egg abundance}
\usage{
getRequiredRDD(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A vector of reproduction rates for all species
}
\description{
Determine reproduction rate needed for initial egg abundance
}
