% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{getSimParams}
\alias{getSimParams}
\title{Extract the projection parameters used to produce a simulation}
\usage{
getSimParams(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
A named list of projection parameters.
}
\description{
Returns the named list of arguments passed to \code{\link[=project]{project()}} or
\code{\link[=projectToSteady]{projectToSteady()}} when producing this \code{MizerSim} object, such as
\code{method} and \code{dt}. Returns an empty list for simulations produced by
older versions of mizer.
}
