% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{getTrophicLevelBySpecies}
\alias{getTrophicLevelBySpecies}
\title{Get mean trophic level of each species}
\usage{
getTrophicLevelBySpecies(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  w_R = 1e-10,
  beta_R = 1000,
  ...
)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object.}

\item{n}{A matrix of species abundances (species x size). Defaults to
the initial abundances stored in \code{params}.}

\item{n_pp}{A vector of the resource abundance by size. Defaults to the
initial resource abundance stored in \code{params}.}

\item{n_other}{A named list of the abundances of other dynamical
components. Defaults to the initial values stored in \code{params}.}

\item{w_R}{An average size (in grams) of primary producers in the resource
spectrum, used to set the size-dependent resource trophic level. Defaults
to \code{1e-10}.}

\item{beta_R}{An average predator/prey mass ratio for the resource spectrum,
used to set the size-dependent resource trophic level. Must be greater than
\code{1}. Defaults to \code{1000}.}

\item{...}{Unused}
}
\value{
A named vector with the mean trophic level for each species.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculates the consumption-rate-weighted mean trophic level of each species,
defined as
\deqn{
  T_i = \frac{\int r_i(w)\,N_i(w)\,T_i(w)\,dw}
             {\int r_i(w)\,N_i(w)\,dw},
}
where \eqn{r_i(w) = (1 - f_i(w))\,E_i(w)} is the consumption rate of an
individual of species \eqn{i} at weight \eqn{w}, \eqn{N_i(w)} is the
abundance density, and \eqn{T_i(w)} is the size-resolved trophic level
from \code{\link[=getTrophicLevel]{getTrophicLevel()}}. As in \code{\link[=getTrophicLevel]{getTrophicLevel()}}, the resource is given a
size-dependent trophic level controlled by the \code{w_R} and \code{beta_R} arguments.
}
\examples{
getTrophicLevelBySpecies(NS_params)
}
\seealso{
\code{\link[=getTrophicLevel]{getTrophicLevel()}}

Other summary functions:
\code{\link[=getBiomass]{getBiomass()}},
\code{\link[=getDiet]{getDiet()}},
\code{\link[=getGrowthCurves]{getGrowthCurves()}},
\code{\link[=getN]{getN()}},
\code{\link[=getSSB]{getSSB()}},
\code{\link[=getTrophicLevel]{getTrophicLevel()}},
\code{\link[=getYield]{getYield()}},
\code{\link[=getYieldGear]{getYieldGear()}},
\code{\link[=sizeIntegral]{sizeIntegral()}}
}
\concept{summary functions}
\concept{summary_function}
