% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{getYieldGear}
\alias{getYieldGear}
\title{Calculate the rate at which biomass of each species is fished by each gear}
\usage{
getYieldGear(sim)
}
\arguments{
\item{sim}{An object of class \code{MizerSim}.}
}
\value{
An array (time x gear x species) containing the yield rate in
grams per year.
}
\description{
This yield rate is given in grams per year. It is calculated at each time
step saved in the MizerSim object.
}
\details{
For details of how the yield rate is defined see the help page of
\code{\link[=getYield]{getYield()}}.
}
\examples{
yield <- getYieldGear(NS_sim)
yield["1972", "Herring", "Herring"]
# (In this example MizerSim object each species was set up with its own gear)
}
\seealso{
\code{\link[=getYield]{getYield()}}

Other summary functions: 
\code{\link{getBiomass}()},
\code{\link{getDiet}()},
\code{\link{getGrowthCurves}()},
\code{\link{getN}()},
\code{\link{getSSB}()},
\code{\link{getYield}()}
}
\concept{summary functions}
\concept{summary_function}
