% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_functions.R
\name{getZ}
\alias{getZ}
\title{Alias for \code{getMort()}}
\usage{
getZ(object, ...)
}
\arguments{
\item{object}{A \linkS4class{MizerParams} or \linkS4class{MizerSim} object.}

\item{...}{Additional arguments that depend on the class of \code{object}.

\strong{For a \linkS4class{MizerParams} object:}
\describe{
\item{\code{n}}{A matrix of species abundances (species x size). Defaults to
the initial abundances stored in \code{object}.}
\item{\code{n_pp}}{A vector of the resource abundance by size. Defaults to the
initial resource abundance stored in \code{object}.}
\item{\code{n_other}}{A named list of the abundances of other dynamical
components. Defaults to the initial values stored in \code{object}.}
\item{\code{effort}}{A numeric vector of the effort by gear or a single
numeric effort value which is used for all gears. Defaults to the
initial effort stored in \code{object}.}
\item{\code{t}}{The time for which to do the calculation. Defaults to 0.}
}

\strong{For a \linkS4class{MizerSim} object:}
\describe{
\item{\code{time_range}}{The time range over which to return the rates. Either
a vector of values, a vector of min and max time, or a single value.
Defaults to the whole time range of the simulation.}
\item{\code{drop}}{If \code{TRUE} then any dimension of length 1 is removed from the
returned array.}
}}
}
\value{
\itemize{
\item \code{MizerParams}: An \code{ArraySpeciesBySize} object (species x size) with the
total mortality rates.
\item \code{MizerSim}: An \code{ArrayTimeBySpeciesBySize} object (time step x species x
size) with the total mortality rates at every time step. If \code{drop = TRUE}
then dimensions of length 1 will be removed.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
An alias provided for backward compatibility with mizer version <= 1.0
}
\details{
If your model contains additional components that you added with
\code{\link[=setComponent]{setComponent()}} and for which you specified a \code{mort_fun} function then
the mortality inflicted by these components will be included in the returned
value.
}
\section{Your own mortality function}{

By default \code{\link[=getMort]{getMort()}} calls \code{\link[=mizerMort]{mizerMort()}}. However you can
replace this with your own alternative mortality function. If
your function is called \code{"myMort"} then you register it in a MizerParams
object \code{params} with

\if{html}{\out{<div class="sourceCode">}}\preformatted{params <- setRateFunction(params, "Mort", "myMort")
}\if{html}{\out{</div>}}

Your function will then be called instead of \code{\link[=mizerMort]{mizerMort()}}, with the
same arguments.
}

\examples{
\donttest{
params <- NS_params
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the total mortality at a particular time step
mort <- getMort(params, n = N(sim)[15, , ], n_pp = NResource(sim)[15, ],
                t = 15, effort = 0.5)
# Mortality rate at this time for Sprat of size 2g
mort["Sprat", "2"]
}
}
\seealso{
\code{\link[=getPredMort]{getPredMort()}}, \code{\link[=getFMort]{getFMort()}}

Other rate functions:
\code{\link[=getDiffusion]{getDiffusion()}},
\code{\link[=getEGrowth]{getEGrowth()}},
\code{\link[=getERepro]{getERepro()}},
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}},
\code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getFMort]{getFMort()}},
\code{\link[=getFMortGear]{getFMortGear()}},
\code{\link[=getFeedingLevel]{getFeedingLevel()}},
\code{\link[=getFlux]{getFlux()}},
\code{\link[=getFluxGradient]{getFluxGradient()}},
\code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getPredRate]{getPredRate()}},
\code{\link[=getRDD]{getRDD()}},
\code{\link[=getRDI]{getRDI()}},
\code{\link[=getRates]{getRates()}},
\code{\link[=getResourceMort]{getResourceMort()}}
}
\concept{deprecated}
