% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{getZ}
\alias{getZ}
\title{Alias for getMort}
\usage{
getZ(object, n, n_pp, effort, m2 = getPredMort(object, n = n, n_pp = n_pp))
}
\arguments{
\item{object}{A \code{MizerParams} object.}

\item{n}{A matrix of species abundance (species x size).}

\item{n_pp}{A vector of the background abundance by size.}

\item{effort}{A numeric vector of the effort by gear or a single numeric
effort value which is used for all gears.}

\item{m2}{A two dimensional array of predation mortality (optional). Has
dimensions no. sp x no. size bins in the community. If not supplied is
calculated using the \code{getPredMort()} method.}
}
\value{
A two dimensional array (prey species x prey size).
}
\description{
An alias provided for backward compatibility with mizer version <= 1.0
}
\seealso{
\code{\link{getPredMort}}, \code{\link{getFMort}}
}
