% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues.R
\name{initialN<-}
\alias{initialN<-}
\alias{initialN}
\title{Initial values for fish spectra}
\usage{
initialN(params) <- value

initialN(object)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A matrix with dimensions species x size holding the initial
number densities for the fish spectra.}

\item{object}{An object of class MizerParams or MizerSim}
}
\description{
Values used as starting values for simulations with \code{project()}.
}
