% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{initialNOther<-}
\alias{initialNOther<-}
\alias{initialNOther}
\title{Initial values for other ecosystem components}
\usage{
initialNOther(params) <- value

initialNOther(object)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A named list with the initial values of other ecosystem
components}

\item{object}{An object of class MizerParams or MizerSim}
}
\description{
Values used as starting values for simulations with \code{project()}.
}
