% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInitialValues.R
\name{initialNResource<-}
\alias{initialNResource<-}
\alias{initialNResource}
\title{Initial value for resource spectrum}
\usage{
initialNResource(params) <- value

initialNResource(object)
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A vector with the initial number densities for the resource
spectrum}

\item{object}{An object of class MizerParams or MizerSim}
}
\description{
Value used as starting value for simulations with \code{project()}.
}
