\docType{data}
\name{inter}
\alias{inter}
\title{Sample interaction data set}
\description{
  Interaction data for 12 species in the North Sea.
}
\details{
  This *.csv file can be read into R using the
  \code{read.csv} function. By then transforming the
  resulting data frame into a matrix the data can be used
  by the \code{\link{MizerParams}} constructor.

  The interaction data represents the interaction between
  pairs of species in the model and can be interpreted as a
  proxy for spatial interaction. See the package vignette
  for more details.
}
\references{
  The North Sea paper (Blanchard et al)
}
\seealso{
  \code{\link{MizerParams}}
}
\keyword{datasets}
\keyword{inter}

