% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inter}
\alias{inter}
\title{Example interaction matrix for the North Sea example}
\format{A 12 x 12 matrix.}
\source{
{Blanchard et al.}
}
\usage{
inter
}
\description{
The interaction coefficient between predators and preys in the North Sea.
}
\keyword{datasets}
