% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_funcs.R
\name{knife_edge}
\alias{knife_edge}
\title{Weight based knife-edge selectivity function}
\usage{
knife_edge(w, knife_edge_size, ...)
}
\arguments{
\item{w}{The size of the individual.}

\item{knife_edge_size}{The weight at which the knife-edge operates.}

\item{...}{Unused}
}
\description{
A knife-edge selectivity function where weights greater or equal to
\code{knife_edge_size} are selected.
}
