% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_kernel_funcs.R
\name{lognormal_pred_kernel}
\alias{lognormal_pred_kernel}
\title{Lognormal predation kernel}
\usage{
lognormal_pred_kernel(ppmr, beta, sigma)
}
\arguments{
\item{ppmr}{A vector of predator/prey size ratios}

\item{beta}{The preferred predator/prey size ratio}

\item{sigma}{The width parameter of the log-normal kernel}
}
\value{
A vector giving the value of the predation kernel at each of the
predator/prey mass ratios in the \code{ppmr} argument.
}
\description{
This is the most commonly-used predation kernel. The log of the predator/prey
mass ratio is normally distributed.
}
\details{
Writing the predator mass as \eqn{w} and the prey mass as \eqn{w_p},
the feeding kernel is given as
\deqn{\phi_i(w, w_p) =
\exp \left[ \frac{-(\ln(w / w_p / \beta_i))^2}{2\sigma_i^2} \right]
}{\phi_i(w/w_p) = exp(-(ln(w/w_p/\beta_i))^2/(2\sigma_i^2))}
if \eqn{w/w_p} is larger than 1 and zero otherwise. Here \eqn{\beta_i} is the
preferred predator-prey mass ratio and \eqn{\sigma_i} determines the width of
the kernel. These two parameters need to be given in the species parameter
dataframe in the columns \code{beta} and \code{sigma}.

This function is called from \code{\link[=setPredKernel]{setPredKernel()}} to set up the
predation kernel slots in a MizerParams object.
}
\examples{
params <- NS_params
plot(w_full(params), getPredKernel(params)["Cod", 10, ], type="l", log="x")
# The restriction that the kernel is zero for w/w_p < 1 is more
# noticeable for larger sigma
species_params(params)$sigma <- 4
plot(w_full(params), getPredKernel(params)["Cod", 10, ], type="l", log="x")
}
\seealso{
\code{\link[=setPredKernel]{setPredKernel()}}

Other predation kernel:
\code{\link[=box_pred_kernel]{box_pred_kernel()}},
\code{\link[=gaussian_mixture_pred_kernel]{gaussian_mixture_pred_kernel()}},
\code{\link[=power_law_pred_kernel]{power_law_pred_kernel()}},
\code{\link[=truncated_lognormal_pred_kernel]{truncated_lognormal_pred_kernel()}}
}
\concept{predation kernel}
