% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_kernel_funcs.R
\name{lognormal_pred_kernel}
\alias{lognormal_pred_kernel}
\title{Lognormal predation kernel}
\usage{
lognormal_pred_kernel(ppmr, sp, params)
}
\arguments{
\item{ppmr}{A vector of predator/prey size ratios}

\item{sp}{The index of the predator species}

\item{params}{A MizerParams object}
}
\value{
A vector giving the value of the predation kernel at each of the
predator/prey mass ratios in the \code{ppmr} argument.
}
\description{
This is the most commonly-used predation kernel. The log of the predator/prey
mass ratio is normally distributed.
}
\details{
Writing the predator mass as \eqn{w} and the prey mass as \eqn{w_p},
the feeding kernel is given as
\deqn{\phi_i(w, w_p) = 
\exp \left[ \dfrac{-(\ln(w / w_p / \beta_i))^2}{2\sigma_i^2} \right]
}{\phi_i(w/w_p) = exp(-(ln(w/w_p/\beta_i))^2/(2\sigma_i^2))}
if \eqn{w/w_p} is between 1 and \eqn{\beta_i\exp(3\sigma_i)}{\beta_i exp(3\sigma_i)}
and zero otherwise.
Here \eqn{\beta_i} is the preferred predator-prey mass ratio and \eqn{\sigma_i}
determines the width of the kernel.
These two parameters need to be given in the species parameter dataframe in
the colunns \code{beta} and \code{sigma}.

This function is called from \code{\link{setPredKernel}} to set up the
predation kernel slots in a \code{MizerParams} object. Associated to this
function there is the function \code{\link{lognormal_max_ppmr}} that
returns the largest allowed predator/prey mass ratio, i.e.,
\eqn{\beta_i\exp(3\sigma_i)}{\beta_i exp(3\sigma_i)}.
}
