% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{matchBiomasses}
\alias{matchBiomasses}
\title{Match biomasses to observations}
\usage{
matchBiomasses(params, species = NULL)
}
\arguments{
\item{params}{A MizerParams object}

\item{species}{The species to be affected. Optional. By default all observed
biomasses will be matched. A vector of species names, or a numeric vector
with the species indices, or a logical vector indicating for each species
whether it is to be affected (TRUE) or not.}
}
\value{
A MizerParams object
}
\description{
Match biomasses to observations
}
\examples{
params <- NS_params
species_params(params)$biomass_observed <- 
    c(0.8, 61, 12, 35, 1.6, 20, 10, 7.6, 135, 60, 30, 78)
species_params(params)$biomass_cutoff <- 10
params <- calibrateBiomass(params)
params <- matchBiomasses(params)
lotBiomassObservedVsModel(params)
}
