% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{mizerFMortGear}
\alias{mizerFMortGear}
\title{Get the fishing mortality needed to project
standard mizer model}
\usage{
mizerFMortGear(params, effort)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{effort}{A vector with the effort for each fishing gear.}
}
\value{
A three dimensional array (gear x species x size) with the fishing
mortality.
}
\description{
Calculates the fishing mortality rate \eqn{F_{g,i,w}} by gear, species and
size.
This is a helper function for \code{\link[=mizerFMort]{mizerFMort()}}.
}
\note{
Here: fishing mortality = catchability x selectivity x effort.
}
\seealso{
\code{\link[=setFishing]{setFishing()}}

Other mizer rate functions:
\code{\link[=mizerEGrowth]{mizerEGrowth()}},
\code{\link[=mizerERepro]{mizerERepro()}},
\code{\link[=mizerEReproAndGrowth]{mizerEReproAndGrowth()}},
\code{\link[=mizerEncounter]{mizerEncounter()}},
\code{\link[=mizerFMort]{mizerFMort()}},
\code{\link[=mizerFeedingLevel]{mizerFeedingLevel()}},
\code{\link[=mizerMort]{mizerMort()}},
\code{\link[=mizerPredMort]{mizerPredMort()}},
\code{\link[=mizerPredRate]{mizerPredRate()}},
\code{\link[=mizerRDI]{mizerRDI()}},
\code{\link[=mizerRates]{mizerRates()}},
\code{\link[=mizerResourceMort]{mizerResourceMort()}}
}
\concept{mizer rate functions}
