% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MizerParams-class.r
\docType{class}
\name{MizerParams-class}
\alias{MizerParams,data.frame,matrix,matrix-method}
\alias{MizerParams,data.frame,matrix,missing-method}
\alias{MizerParams,data.frame,missing,missing-method}
\alias{MizerParams,numeric,missing,missing-method}
\alias{MizerParams-class}
\title{A class to hold the parameters for a size based model.}
\usage{
\S4method{MizerParams}{numeric,missing,missing}(object, min_w = 0.001,
  max_w = 1000, no_w = 100, min_w_pp = 1e-10, no_w_pp = round(no_w) *
  0.3, species_names = 1:object, gear_names = species_names)

\S4method{MizerParams}{data.frame,matrix,matrix}(object, interaction,
  interaction_pp, n = 2/3, p = 0.7, q = 0.8, r_pp = 10, kappa = 1e+11,
  lambda = (2 + q - n), w_pp_cutoff = 10, max_w = max(object$w_inf) * 1.1,
  f0 = 0.6, z0pre = 0.6, z0exp = n - 1, ...)

\S4method{MizerParams}{data.frame,missing,missing}(object, interaction,
  interaction_pp, ...)

\S4method{MizerParams}{data.frame,matrix,missing}(object, interaction,
  interaction_pp, ...)
}
\arguments{
\item{object}{Data frame holding species parameters}

\item{interaction}{Matrix of predator--prey interaction strengths}

\item{interaction_pp}{Matrix of predator--resource interaction strengths}
}
\description{
These parameters include the model species, their life history parameters and
the size ranges of the model.
}
\details{
\code{MizerParams} objects can be created using a range of
\code{MizerParams} constructor methods.

Dynamic simulations are performed using the \code{\link{project}} method on
objects of this class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{MizerParams}: Only really used to make MizerParams of the right
size and shouldn't be used by user

\item \code{MizerParams}: Constructor that takes the species_params data.frame
and the interaction matrix

\item \code{MizerParams}: If interaction and interaction_pp are missing, make
one of the right size and fill with 1s

\item \code{MizerParams}: If interaction_pp is missing, make one for a single
resource
}}
\section{Slots}{

\describe{
\item{\code{w}}{A numeric vector of size bins used for the community (i.e. fish) part
of the model. These are usually spaced on a log10 scale}

\item{\code{dw}}{The absolute difference between the size bins specified in the w
slot. A vector the same length as the w slot. The final value is the same
as the second to last value}

\item{\code{w_full}}{A numeric vector of size bins used for the whole model (i.e. the
community and background spectra) . These are usually spaced on a log10
scale}

\item{\code{dw_full}}{The absolute difference between the size bins specified in the
w_full slot. A vector the same length as the w_full slot. The final value
is the same as the second to last value}

\item{\code{psi}}{An array (species x size) that holds the allocation to reproduction
for each species at size}

\item{\code{intake_max}}{An array (species x size) that holds the maximum intake for
each species at size}

\item{\code{search_vol}}{An array (species x size) that holds the search volume for
each species at size}

\item{\code{activity}}{An array (species x size) that holds the activity for each
species at size}

\item{\code{std_metab}}{An array (species x size) that holds the standard metabolism
for each species at size}

\item{\code{pred_kernel}}{An array (species x predator size x prey size) that holds
the predation coefficient of each predator at size on each prey size}

\item{\code{rr_pp}}{A vector the same length as the w_full slot. The size specific
growth rate of the background spectrum}

\item{\code{cc_pp}}{A vector the same length as the w_full slot. The size specific
carrying capacity of the background spectrum}

\item{\code{species_params}}{A data.frame to hold the species specific parameters
(see the mizer vignette, Table 2, for details)}

\item{\code{interaction}}{The species specific interaction matrix.}

\item{\code{interaction_pp}}{The species specific interaction matrix for resources}

\item{\code{srr}}{Function to calculate the realised (density dependent) recruitment.
Has two arguments which are rdi and species_params}

\item{\code{selectivity}}{An array (gear x species x w) that holds the selectivity of
each species by gear and species size}

\item{\code{catchability}}{An array (gear x species) that holds the catchability of
  each species by each gear}
}}
\note{
The \code{MizerParams} class is fairly complex with a large number of
  slots, many of which are multidimensional arrays. The dimensions of these
  arrays is strictly enforced so that \linkS4class{MizerParams} objects are
  consistent in terms of number of species and number of size classes.

  Although it is possible to build a \code{MizerParams} object by hand it is
  not recommended and several constructors are available.

  The \code{MizerParams} class does not hold any dynamic information, e.g.
  abundances or harvest effort through time. These are held in
  \linkS4class{MizerSim} objects.
}
\seealso{
\code{\link{project}} \linkS4class{MizerSim}
}

