% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\docType{class}
\name{MizerParams-class}
\alias{MizerParams-class}
\title{A class to hold the parameters for a size based model.}
\description{
MizerParams objects can be created using a range of constructor functions.
}
\details{
Dynamic simulations are performed using the \code{\link{project}} method on
objects of this class.
}
\section{Slots}{

\describe{
\item{\code{w}}{The size grid for the fish part of the spectrum. An increasing
vector of weights (in grams) running from the smallest egg size to the
largest asymptotic size.}

\item{\code{dw}}{The spacing in the size grid. So dw[i] = w[i+1] - w[i]. A vector 
the same length as the w_full slot. The last entry is not determined by
the w slot but represents the size of the last size bin.}

\item{\code{w_full}}{The size grid for the full size range including the resource
spectrum. An increasing vector of weights (in grams) running from the
smallest resource size to the largest asymptotic size of fish. The
last entries of the vector have to be equal to the content of the w slot.}

\item{\code{dw_full}}{The spacing in the full size grid. 
So dw_full[i] = w_full[i+1] - w_full[i]. The last entries have to be
equal to the content of the dw slot.}

\item{\code{w_min_idx}}{A vector holding the index of the weight of the egg size
of each species}

\item{\code{psi}}{An array (species x size) that holds the allocation to reproduction
for each species at size, \eqn{\psi_i(w)}}

\item{\code{intake_max}}{An array (species x size) that holds the maximum intake for
each species at size.}

\item{\code{search_vol}}{An array (species x size) that holds the search volume for
each species at size.}

\item{\code{metab}}{An array (species x size) that holds the metabolism
for each species at size.}

\item{\code{mu_b}}{An array (species x size) that holds the background death 
\eqn{\mu_{b.i}(w)}}

\item{\code{pred_kernel}}{An array (species x predator size x prey size) that holds
the predation coefficient of each predator at size on each prey size. If
this is NA then the following two slots will be used.}

\item{\code{ft_pred_kernel_e}}{An array (species x log of predator/prey size ratio)
that holds the Fourier transform of the feeding kernel in a form
appropriate for evaluating the available energy integral. If this is NA
then the \code{pred_kernel} will be used to calculate the available 
energy integral.}

\item{\code{ft_pred_kernel_p}}{An array (species x log of predator/prey size ratio)
that holds the Fourier transform of the feeding kernel in a form
appropriate for evaluating the predation mortality integral. If this is NA
then the \code{pred_kernel} will be used to calculate the integral.}

\item{\code{rr_pp}}{A vector the same length as the w_full slot. The size specific
growth rate of the plankton spectrum.}

\item{\code{cc_pp}}{A vector the same length as the w_full slot. The size specific
carrying capacity of the plankton spectrum.}

\item{\code{sc}}{The community abundance of the scaling community}

\item{\code{species_params}}{A data.frame to hold the species specific parameters
(see the mizer vignette, Table 2, for details)}

\item{\code{interaction}}{The species specific interaction matrix, \eqn{\theta_{ij}}}

\item{\code{srr}}{Function to calculate the realised (density dependent) recruitment.
Has two arguments which are rdi and species_params}

\item{\code{selectivity}}{An array (gear x species x w) that holds the selectivity of
each gear for species and size, \eqn{S_{g,i,w}}}

\item{\code{catchability}}{An array (gear x species) that holds the catchability of
each species by each gear, \eqn{Q_{g,i}}}

\item{\code{initial_n}}{An array (species x size) that holds abundance of each species
at each weight at our candidate steady state solution.}

\item{\code{initial_n_pp}}{A vector the same length as the w_full slot that describes
the plankton abundance at each weight.}

\item{\code{n}}{Exponent of maximum intake rate.}

\item{\code{p}}{Exponent of metabolic cost.}

\item{\code{lambda}}{Exponent of resource spectrum.}

\item{\code{q}}{Exponent for volumetric search rate.}

\item{\code{f0}}{Initial feeding level.}

\item{\code{kappa}}{Magnitude of resource spectrum.}

\item{\code{A}}{Abundance multipliers.}

\item{\code{linecolour}}{A named vector of colour values, named by species. Used 
to give consistent colours to species in plots.}

\item{\code{linetype}}{A named vector of linetypes, named by species. Used 
to give consistent colours to species in plots.}
}}

\note{
The \linkS4class{MizerParams} class is fairly complex with a large number of
  slots, many of which are multidimensional arrays. The dimensions of these
  arrays is strictly enforced so that \code{MizerParams} objects are
  consistent in terms of number of species and number of size classes.
  
  Although it is possible to build a \code{MizerParams} object by hand it is
  not recommended and several constructors are available.
  
  The \code{MizerParams} class does not hold any dynamic information, e.g.
  abundances or harvest effort through time. These are held in
  \linkS4class{MizerSim} objects.
}
\seealso{
\code{\link{project}} \code{\link{MizerSim}}
}
