% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies_params.R
\name{MizerParams}
\alias{MizerParams}
\title{Alias for set_multispecies_model}
\usage{
MizerParams(species_params, interaction = matrix(1, nrow =
  nrow(species_params), ncol = nrow(species_params)), no_w = 100,
  min_w = 0.001, max_w = NA, min_w_pp = NA, no_w_pp = NA,
  n = 2/3, p = 0.7, q = 0.8, r_pp = 10, kappa = 1e+11,
  lambda = (2 + q - n), w_pp_cutoff = 10, f0 = 0.6, z0pre = 0.6,
  z0exp = n - 1, pred_kernel_type = "lognormal", store_kernel = (no_w
  <= 100), plankton_dynamics = plankton_semichemostat,
  interaction_p = rep(1, nrow(species_params)), rho = NULL,
  resource_dynamics = list(), resource_params = list(),
  srr = srrBevertonHolt)
}
\arguments{
\item{species_params}{A data frame of species-specific parameter values.}

\item{interaction}{Interaction matrix of the species (predator by prey). By
default all interactions between species are set to 1.}

\item{no_w}{The number of size bins in the consumer spectrum.}

\item{min_w}{Sets the size of the eggs of all species for which this is not
given in the \code{w_min} column of the \code{species_params} dataframe.}

\item{max_w}{The largest size of the consumer spectrum. By default this is
set to the largest w_inf specified in the species_params data frame.}

\item{min_w_pp}{The smallest size of the plankton spectrum. By default this
is set to the smallest value at which any of the consumers can feed.}

\item{no_w_pp}{Obsolete argument that is no longer used because the number
of plankton size bins is determined because all size bins have to
be logarithmically equally spaced.}

\item{n}{Scaling of the intake. Default value is 2/3.}

\item{p}{Scaling of the standard metabolism. Default value is 0.7.}

\item{q}{Exponent of the search volume. Default value is 0.8.}

\item{r_pp}{Growth rate of the primary productivity. Default is 10 g/year.}

\item{kappa}{Carrying capacity of the plankton spectrum. Default value is
1e11.}

\item{lambda}{Exponent of the plankton spectrum. Default value is (2+q-n).}

\item{w_pp_cutoff}{The upper cut off size of the plankton spectrum. 
Default is 10 g.}

\item{f0}{Average feeding level. Used to calculated \code{h} and \code{gamma}
if those are not columns in the species data frame. Also requires
\code{k_vb} (the von Bertalanffy K parameter) to be a column in the species
data frame. If \code{h} and \code{gamma} are supplied then this argument is
ignored. Default is 0.6.}

\item{z0pre}{If \code{z0}, the mortality from other sources, is not a column
in the species data frame, it is calculated as z0pre * w_inf ^ z0exp.
Default value is 0.6.}

\item{z0exp}{If \code{z0}, the mortality from other sources, is not a column
in the species data frame, it is calculated as \code{z0pre * w_inf ^ z0exp}.
Default value is \code{n-1}.}

\item{pred_kernel_type}{A string that determines which predation kernel
function is used. Default is "lognormal".}

\item{store_kernel}{A boolean flag that determines whether the full
predation kernel is stored. If FALSE, only its Fourier transforms are stored.
The default is TRUE if the number of size bins is no larger than 100 and
FALSE otherwise.}

\item{plankton_dynamics}{Function that determines plankton dynamics by
calculating the plankton spectrum at the next time step from the current
state. The default is \code{"\link{plankton_semichemostat}"}.}

\item{interaction_p}{Vector specifying for each species its interaction with
plankton, similar to what the interaction matrix does for the interaction
with other species. Entries should be numbers between 0 and 1. The 
default is a vector of 1s.}

\item{rho}{An array (species x resource) that gives the
rate \eqn{\rho_{id}} that determines the rate at which species \eqn{i}
encounters biomass of resource \eqn{d}.}

\item{resource_dynamics}{A named list of functions that determine the
dynamics of the unstructured resources by calculating their biomasses at
the next time step from the current state. See
\code{\link{resource_dynamics}} for details. An empty list if the model
does not have unstructured resources.}

\item{resource_params}{A list of parameters needed by the
\code{resource_dynamics} functions. An empty list if no parameters are
needed.}

\item{srr}{The stock recruitment function. Default is
\code{\link{srrBevertonHolt}}.}
}
\value{
An object of type \linkS4class{MizerParams}
}
\description{
An alias provided for backward compatibility with mizer version <= 1.0
}
\examples{
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
}
\seealso{
\code{\link{project}} \linkS4class{MizerSim}
}
