% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{MizerParams}
\alias{MizerParams}
\title{Alias for multispeciesParams}
\usage{
MizerParams(object, interaction, min_w = 0.001, max_w = NA,
  no_w = 100, min_w_pp = NA, no_w_pp = NA, n = 2/3, p = 0.7,
  q = 0.8, r_pp = 10, kappa = 1e+11, lambda = (2 + q - n),
  w_pp_cutoff = 10, f0 = 0.6, z0pre = 0.6, z0exp = n - 1,
  store_kernel = (no_w <= 100))
}
\arguments{
\item{object}{A data frame of species specific parameter values (see notes
below).}

\item{interaction}{Optional argument to specify the interaction matrix of the
species (predator by prey). If missing a default interaction is used where
all interactions between species are set to 1. Note that any dimnames of
the interaction matrix argument are ignored by the constructor. The
dimnames of the interaction matrix in the returned \code{MizerParams}
object are taken from the species names in the \code{species_params} slot.
This means that the order of the columns and rows of the interaction matrix
argument should be the same as the species name in the
\code{species_params} slot.}

\item{min_w}{The smallest size of the community spectrum. This is ignored if
the egg sizes of the species are given in the species_params data frame.}

\item{max_w}{Obsolete. The largest fish size is now set to the largest
w_inf specified in the species_params data frame.}

\item{no_w}{The number of size bins in the community spectrum.}

\item{min_w_pp}{The smallest size of the plankton spectrum.}

\item{no_w_pp}{Obsolete argument that is no longer used because the number
of plankton size bins is determined because all size bins have to
be logarithmically equally spaced.}

\item{n}{Scaling of the intake. Default value is 2/3.}

\item{p}{Scaling of the standard metabolism. Default value is 0.7.}

\item{q}{Exponent of the search volume. Default value is 0.8.}

\item{r_pp}{Growth rate of the primary productivity. Default value is 10.}

\item{kappa}{Carrying capacity of the resource spectrum. Default
value is 1e11.}

\item{lambda}{Exponent of the resource spectrum. Default value is
(2+q-n).}

\item{w_pp_cutoff}{The cut off size of the plankton spectrum.
Default value is 10.}

\item{f0}{Average feeding level. Used to calculated \code{h} and
\code{gamma} if those are not columns in the species data frame. Also
requires \code{k_vb} (the von Bertalanffy K parameter) to be a column
in the species data frame. If \code{h} and \code{gamma} are supplied
then this argument is ignored. Default is 0.6..}

\item{z0pre}{If \code{z0}, the mortality from other sources, is not
a column in the species data frame, it is calculated as 
z0pre * w_inf ^ z0exp. Default value is 0.6.}

\item{z0exp}{If \code{z0}, the mortality from other sources, is not
a column in the species data frame, it is calculated as 
z0pre * w_inf ^ z0exp. Default value is n-1.}

\item{store_kernel}{A boolean flag that determines whether the full
feeding kernel is stored. If FALSE, only its Fourier transforms are stored.
The default is TRUE if the number of size bins is no larger than 100 and
FALSE otherwise.}
}
\value{
An object of type \code{MizerParams}
}
\description{
An alias provided for backward compatibility with mizer version <= 1.0
}
\examples{
data(NS_species_params_gears)
data(inter)
params <- multispeciesParams(NS_species_params_gears, inter)
}
\seealso{
\code{\link{project}} \linkS4class{MizerSim}
}
