% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MizerSim-class.r
\docType{class}
\name{MizerSim-class}
\alias{MizerSim,MizerParams-method}
\alias{MizerSim-class}
\title{MizerSim}
\usage{
\S4method{MizerSim}{MizerParams}(object, t_dimnames = NA, t_max = 100,
  t_save = 1)
}
\description{
A class that holds the results of projecting a \linkS4class{MizerParams}
object through time.
}
\details{
\code{MizerSim} objects are created by using the \code{\link{project}} method
on an object of type \code{MizerParams}.

There are several plotting methods available to explore the contents of a
\code{MizerSim} object. See the package vignette for more details.
}
\section{Methods (by generic)}{
\itemize{
\item \code{MizerSim}: 
}}
\section{Slots}{

\describe{
\item{\code{params}}{An object of type \linkS4class{MizerParams}.}

\item{\code{n}}{Array that stores the projected community population abundances by
time, species and size}

\item{\code{effort}}{Array that stores the fishing effort through time by time and
gear}

\item{\code{n_pp}}{Array that stores the projected background population by time,
  resource and size}
}}
\seealso{
\code{\link{project}} \code{\link{MizerParams}}
}

