\docType{class}
\name{MizerSim-class}
\alias{MizerSim-class}
\title{MizerSim}
\description{
  A class that holds the results of a projecting an
  \code{MizerParams} object through time
}
\details{
  \describe{ \item{params}{An object of type
  \code{MizerParams}} \item{n}{Array to store the projected
  community populations by time, species and size}
  \item{effort}{Array that stores the fishing effort
  through time by time and gear. SEE DETAILS}
  \item{n_pp}{Array to store the projected background
  population by time and size} }
}

