% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\docType{class}
\name{MizerSim-class}
\alias{MizerSim-class}
\title{A class to hold the results of a simulation}
\description{
A class that holds the results of projecting a \linkS4class{MizerParams}
object through time using the \code{\link{project}} function.
}
\details{
A new \code{MizerSim} object can be created with the \code{\link{MizerSim}}
constructor, but you will never have to do that because the object is
created automatically by the \code{\link{project}} function when needed.

The arrays all have named dimensions. The names of the "time" dimension are
numeric and denote the time in years. The names of the "sp" dimension are the
same as the species name in the order specified in the species_params data
frame. The names of the "gear" dimension are the names of the gears, in the
same order as specified when setting up the \code{MizerParams} object.

There are several \code{link{summary_functions}} and
\code{\link{plotting_functions}} available to explore the contents of a
\code{MizerSim} object.
}
\section{Slots}{

\describe{
\item{\code{params}}{An object of type \linkS4class{MizerParams}.}

\item{\code{n}}{Three-dimensional array (time x species x size) that stores the 
projected community number densities.}

\item{\code{n_pp}}{An array (time x size) that stores the projected plankton number
densities.}

\item{\code{n_other}}{A list array (time x component) that stores the projected
values for other ecosystem components.}

\item{\code{effort}}{An array (time x gear) that stores the fishing effort by time and 
gear.}
}}

