% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\docType{class}
\name{MizerSim-class}
\alias{MizerSim-class}
\title{A class to hold the results of a simulation}
\description{
A class that holds the results of projecting a \linkS4class{MizerParams}
object through time using the \code{\link{project}} function.
}
\details{
A new \code{MizerSim} object can be created with the \code{\link{MizerSim}}
constructor, but you will never have to do that because the object is
created automatically by the \code{\link{project}} function when needed.

The arrays all have named dimensions. The names of the "time" dimension are
numeric and denote the time in years. The names of the "sp" dimension are
the same as the species name in the order specified in the species_params
data frame. The names of the "gear" and "res" dimension are the names of the 
gears and resources respectively, in the same order as specified when setting
up the \code{MizerParams} object.

There are several \code{link{summary_functions}} and
\code{\link{plotting_functions}} available to explore the contents of a
\code{MizerSim} object.
}
\section{Slots}{

\describe{
\item{\code{params}}{An object of type \linkS4class{MizerParams}.}

\item{\code{n}}{Array that stores the projected community population abundances by
time, species and size}

\item{\code{n_pp}}{Array that stores the projected plankton abundance by time and
size}

\item{\code{B}}{Array that stores biomasses of unstructured resources by time and
resource}

\item{\code{effort}}{Array that stores the fishing effort by time and gear}
}}

