\docType{methods}
\name{MizerSim}
\alias{MizerSim}
\alias{MizerSim,MizerParams-method}
\title{Constructors for objects of \code{MizerSim} class}
\arguments{
  \item{object}{something}
}
\value{
  An object of type \code{MizerSim}
}
\description{
  A range of constructors for the \code{MizerSim} class.
  Some are more user friendly than others and some are
  really only used internally to help construct
  \code{MizerSim objects of the right dimension
}
\note{
  or details or something
}
\examples{
x <- 1 # something here
}

