% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{MizerSim}
\alias{MizerSim}
\title{Constructor for the \code{MizerSim} class}
\usage{
MizerSim(params, t_dimnames = NA, t_max = 100, t_save = 1)
}
\arguments{
\item{params}{a \linkS4class{MizerParams} object}

\item{t_dimnames}{Numeric vector that is used for the time dimensions of the
slots. Default = NA.}

\item{t_max}{The maximum time step of the simulation. Only used if t_dimnames
= NA. Default value = 100.}

\item{t_save}{How often should the results of the simulation be stored. Only
used if t_dimnames = NA. Default value = 1.}
}
\value{
An object of type \linkS4class{MizerSim}
}
\description{
A constructor for the \code{MizerSim} class. This is used by
\code{\link[=project]{project()}} to create \code{MizerSim} objects of the right
dimensions. It is not necessary for users to use this constructor.
}
