% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{N}
\alias{N}
\alias{NPlankton}
\title{Time series of size spectra}
\usage{
N(sim)

NPlankton(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
For \code{N()}: A three-dimensional array (time x species x size) with the
number density of consumers

For \code{NPlankton()}: An array (time x size) with the number density of plankton
}
\description{
Fetch the simulation results for the size spectra over time.
}
