% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{n}
\alias{n}
\title{Time series of consumer size spectra}
\usage{
n(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
A three-dimensional array (time x species x size) with the number
  density of consumers
}
\description{
Fetch the simulation results for the consumer size spectra over time.
}
