% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMultispeciesParams.R
\name{newMultispeciesParams}
\alias{newMultispeciesParams}
\title{Set up parameters for a general multispecies model}
\usage{
newMultispeciesParams(
  species_params,
  interaction = NULL,
  no_w = 100,
  min_w = 0.001,
  max_w = NA,
  min_w_pp = NA,
  pred_kernel = NULL,
  search_vol = NULL,
  intake_max = NULL,
  metab = NULL,
  p = 0.7,
  ext_mort = NULL,
  z0pre = 0.6,
  z0exp = n - 1,
  ext_encounter = NULL,
  maturity = NULL,
  repro_prop = NULL,
  RDD = "BevertonHoltRDD",
  kappa = 1e+11,
  n = 2/3,
  resource_rate = 10,
  resource_capacity = kappa,
  lambda = 2.05,
  w_pp_cutoff = 10,
  resource_dynamics = "resource_semichemostat",
  gear_params = NULL,
  selectivity = NULL,
  catchability = NULL,
  initial_effort = NULL,
  second_order_w = FALSE,
  info_level = default_info_level(),
  z0 = deprecated(),
  r_pp = deprecated()
)
}
\arguments{
\item{species_params}{A data frame of species-specific parameter values.}

\item{interaction}{Optional interaction matrix of the species (predator
species x prey species). By default all entries are 1. See "Setting
interaction matrix" section below.}

\item{no_w}{The number of size bins in the consumer spectrum.}

\item{min_w}{Sets the size of the eggs of all species for which this is not
given in the \code{w_min} column of the \code{species_params} dataframe.}

\item{max_w}{The largest size of the consumer spectrum. By default this is
set to the largest \code{w_max} specified in the \code{species_params} data
frame.}

\item{min_w_pp}{The smallest size of the resource spectrum. By default this
is set to the smallest value at which any of the consumers can feed.}

\item{pred_kernel}{Optional. An array (species x predator size x prey size)
that holds the predation coefficient of each predator at size on each prey
size. If not supplied, a default is set as described in section "Setting
predation kernel".}

\item{search_vol}{Optional. An array (species x size) holding the search volume
for each species at size. If not supplied, a default is set as described in
the section "Setting search volume".}

\item{intake_max}{Optional. An array (species x size) holding the maximum
intake rate for each species at size. If not supplied, a default is set as
described in the section "Setting maximum intake rate".}

\item{metab}{Optional. An array (species x size) holding the metabolic rate
for each species at size. If not supplied, a default is set as described in
the section "Setting metabolic rate".}

\item{p}{The allometric metabolic exponent. This can be overruled for
individual species by including a \code{p} column in the \code{species_params}.}

\item{ext_mort}{Optional. An array (species x size) holding the external
mortality rate.  If not supplied, a default is set as described in the
section "Setting external mortality rate".}

\item{z0pre}{If \code{z0}, the mortality from other sources, is not a column
in the species data frame, it is calculated as z0pre * w_inf ^ z0exp.
Default value is 0.6.}

\item{z0exp}{If \code{z0}, the mortality from other sources, is not a column in
the species data frame, it is calculated as \code{z0pre * w_inf ^ z0exp}.
Default value is \code{n-1}.}

\item{ext_encounter}{Optional. An array (species x size) holding the external
encounter rate.  If not supplied, a default is calculated from the \code{E_ext}
and \code{n} species parameters as described in the section "Setting external
encounter rate".}

\item{maturity}{Optional. An array (species x size) that holds the proportion
of individuals of each species at size that are mature. If not supplied, a
default is set as described in the section "Setting reproduction".}

\item{repro_prop}{Optional. An array (species x size) that holds the
proportion of the energy available for growth and reproduction that a mature
individual allocates to reproduction for each species at size. If not
supplied, a default is set as described in the section "Setting
reproduction".}

\item{RDD}{The name of the function calculating the density-dependent
reproduction rate from the density-independent rate. Defaults to
"\code{\link[=BevertonHoltRDD]{BevertonHoltRDD()}}".}

\item{kappa}{The coefficient \eqn{\kappa} of the resource carrying capacity
power law \eqn{c_R(w) = \kappa\, w^{-\lambda}}, which also sets the initial
resource abundance. See \code{\link[=resource_params]{resource_params()}}.}

\item{n}{The allometric growth exponent. This can be overruled for individual
species by including a \code{n} column in the \code{species_params}.}

\item{resource_rate}{Optional. A vector of per-capita resource birth
rate for each size class or a single number giving the coefficient in the
power-law for this rate, see "Setting resource dynamics" below.
Must be strictly positive.}

\item{resource_capacity}{Optional. Vector of resource intrinsic carrying
capacities or coefficient in the power-law for the capacity, see
"Setting resource dynamics" below.
The resource capacity must not be smaller than the resource abundance.}

\item{lambda}{Used to set power-law exponent for resource capacity if the
\code{resource_capacity} argument is given as a single number.}

\item{w_pp_cutoff}{The upper cut off size of the resource spectrum power law
used when \code{resource_capacity} is given as a single number. When changing
\code{w_pp_cutoff} without providing \code{resource_capacity}, the cutoff can only be
decreased. In that case, both the carrying capacity and the initial
resource abundance will be cut off at the new value. To increase the
cutoff, you must also provide the \code{resource_capacity} for the extended
range.}

\item{resource_dynamics}{Optional. Name of the function that determines the
resource dynamics by calculating the resource spectrum at the next time
step from the current state.}

\item{gear_params}{A data frame with gear-specific parameter values.}

\item{selectivity}{Optional. An array (gear x species x size) that holds the
selectivity of each gear for species and size, \eqn{S_{g,i,w}}.}

\item{catchability}{Optional. An array (gear x species) that holds the catchability of
each species by each gear, \eqn{Q_{g,i}}.}

\item{initial_effort}{Optional. A number or a named numeric vector specifying
the fishing effort. If a number, the same effort is used for all gears. If
a vector, must be named by gear.}

\item{second_order_w}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Selects the
second-order numerical scheme for the new model. Accepts the same values as
the \code{\link[=second_order_w]{second_order_w()}} setter: a single logical (\code{TRUE} switches on both
second-order flux and bin-averaging), a single flux scheme name
(\code{"upwind"}, \code{"van_leer"} or \code{"centred"}), or a named vector with entries
\code{flux} and/or \code{bin_average}. The \code{bin_average} choice is applied \emph{before}
the resource and abundance power laws are constructed, so they are built
bin-averaged from the start (unlike setting \code{\link[=second_order_w]{second_order_w()}} on an
existing object). The \code{flux} scheme governs time projection only, so the
robust first-order upwind scheme is used for the construction-time
steady-state solve and the chosen scheme is then activated for the returned
model. Defaults to \code{FALSE} (the first-order behaviour of previous mizer).}

\item{info_level}{Controls the amount of information messages that are shown
when the function sets default values for parameters. Higher levels lead
to more messages, \code{info_level = 0} gives silence. The default is taken
from the \code{mizer_info_level} option, see \code{\link[=default_info_level]{default_info_level()}}.}

\item{z0}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{ext_mort} instead. Not to
be confused with the species_parameter \code{z0}.}

\item{r_pp}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. Use \code{resource_rate} argument
instead.}
}
\value{
An object of type \linkS4class{MizerParams}
}
\description{
Sets up a multi-species size spectrum model by filling all slots in the
\linkS4class{MizerParams} object based on user-provided or default
parameters. There is a long list of arguments, but almost
all of them have sensible default values. The only required argument is
the \code{species_params} data frame. All arguments are described in more
details in the sections below the list.
}
\section{Species parameters}{

The only essential argument is a data frame that contains the species
parameters. The data frame is arranged species by parameter, so each column
of the parameter data frame is a parameter and each row has the values of the
parameters for one of the species in the model.

There are two essential columns that must be included in the species
parameter data.frame and that do not have default values: the
\code{species} column that should hold strings with the names of the
species and the \code{w_inf} column with the von Bertalanffy asymptotic sizes of
the species in grams. (You could alternatively specify the corresponding
length in cm in an \code{l_inf} column.) The computational upper size boundary
\code{w_max} is not essential; if it is missing it defaults to \code{1.5 * w_inf}. For
backwards compatibility, if \code{w_inf} is missing it is taken from the
\code{w_repro_max} or \code{w_max} column instead.

The \verb{species_params dataframe} also needs to contain the parameters needed
by any predation kernel function (size selectivity function). This will
be mentioned in the appropriate sections below.

For all other species parameters, mizer will calculate default values if they
are not included in the species parameter data frame. They will be
automatically added when the \code{MizerParams} object is created. For these
parameters you can also specify values for only some species and leave the
other entries as NA and the missing values will be set to the defaults.
So the \code{species_params} data frame saved in the returned MizerParams object
will differ from the one you supply because it will have the missing
species parameters filled in with default values.

If you are not happy with any of the species parameter values used you can
always change them later with \code{\link[=species_params<-]{species_params<-()}}.

All the parameters will be mentioned in the following sections.
}

\section{Setting initial values}{

The initial values for the species number densities are set using the
function \code{get_initial_n()}. These are quite arbitrary and not very close to
the steady state abundances. We intend to improve this in the future.

The initial resource number density \eqn{N_R(w)} is set to a power law with
coefficient \code{kappa} (\eqn{\kappa}) and exponent \code{-lambda} (\eqn{-\lambda}):
\deqn{N_R(w) = \kappa\, w^{-\lambda}}{c_R(w) = \kappa w^{-\lambda}}
for all \eqn{w} less than \code{w_pp_cutoff} and zero for sizes at or above
\code{w_pp_cutoff}.
}

\section{Size grid}{

A size grid is created so that
the log-sizes are equally spaced. The spacing is chosen so that there will be
\code{no_w} fish size bins, with the smallest starting at \code{min_w} and the largest
starting at \code{max_w}. For the resource spectrum there is a larger set of
bins containing additional bins below
\code{min_w}, with the same log size. The number of extra bins is such that
\code{min_w_pp} comes to lie within the smallest bin.
}

\section{Units in mizer}{

Mizer uses grams to measure weight, centimetres to measure lengths, and
years to measure time.

Mizer is agnostic about whether abundances are given as
\enumerate{
\item numbers per area,
\item numbers per volume or
\item total numbers for the entire study area.
}

You should make the choice most convenient for your application and then
stick with it. If you make choice 1 or 2 you will also have to choose a unit
for area or volume. Your choice will then determine the units for some of
the parameters. This will be mentioned when the parameters are discussed in
the sections below.

Your choice will also affect the units of the quantities you may want to
calculate with the model. For example, the yield will be in grams/year/m^2 in
case 1 if you choose m^2 as your measure of area, in grams/year/m^3 in case 2
if you choose m^3 as your unit of volume, or simply grams/year in case 3. The
same comment applies for other measures, like total biomass, which will be
grams/area in case 1, grams/volume in case 2 or simply grams in case 3. When
mizer puts units on axes in plots, it will choose the units appropriate for
case 3. So for example in \code{\link[=plotBiomass]{plotBiomass()}} it gives the unit as grams.

You can convert between these choices. For example, if you use case 1, you
need to multiply with the area of the ecosystem to get the total quantity.
If you work with case 2, you need to multiply by both area and the thickness
of the productive layer. In that respect, case 2 is a bit cumbersome. The
function \code{\link[=scaleModel]{scaleModel()}} is useful to change the units you are using.
}

\section{Setting interaction matrix}{


You do not need to specify an interaction matrix. If you do not, then the
predator-prey interactions are purely determined by the size of predator
and prey and totally independent of the species of predator and prey.

The interaction matrix \eqn{\theta_{ij}} modifies the interaction of each
pair of species in the model. This can be used for example to allow for
different spatial overlap among the species.
The values in the interaction matrix are used to scale the encountered food
and predation mortality (see on the website \href{https://sizespectrum.org/mizer/articles/model_description.html#sec:pref}{the section on predator-prey encounter rate}
and on \href{https://sizespectrum.org/mizer/articles/model_description.html#mortality}{predation mortality}).
The first index refers to the predator species and the second to the prey
species.

The interaction matrix is used when calculating the food encounter rate in
\code{\link[=getEncounter]{getEncounter()}} and the predation mortality rate in \code{\link[=getPredMort]{getPredMort()}}. Its
entries are dimensionless numbers. If all the values in the interaction
matrix are equal then predator-prey interactions are determined entirely by
size-preference.

This function checks that the supplied interaction matrix is valid and then
stores it in the \code{interaction} slot of the \code{params} object.

The order of the columns and rows of the \code{interaction} argument should be
the same as the order in the species params data frame in the \code{params}
object. If you supply a named array then the function will check the order
and message if it is different before ignoring the supplied dimnames. If
you supply only column names then these are also used as the row names. One
way of creating your own interaction
matrix is to enter the data using a spreadsheet program and saving it as a
.csv file. The data can then be read into R using the command \code{read.csv()}.

The interaction of the species with the resource are set via a column
\code{interaction_resource} in the \code{species_params} data frame. By default this
column is set to all 1s.
}

\section{Setting predation kernel}{

\strong{Kernel dependent on predator to prey size ratio}

If the \code{pred_kernel} argument is not supplied, then this function sets a
predation kernel that depends only on the ratio of predator mass to prey
mass, not on the two masses independently. The shape of that kernel is then
determined by the \code{pred_kernel_type} column in species_params.

The default for \code{pred_kernel_type} is "lognormal". This will call the function
\code{\link[=lognormal_pred_kernel]{lognormal_pred_kernel()}} to calculate the predation kernel.
Alternative pred_kernel types are "box", implemented by \code{\link[=box_pred_kernel]{box_pred_kernel()}},
"power_law", implemented by \code{\link[=power_law_pred_kernel]{power_law_pred_kernel()}}, and
"gaussian_mixture", implemented by \code{\link[=gaussian_mixture_pred_kernel]{gaussian_mixture_pred_kernel()}}. These
functions require certain species parameters in the species_params data
frame. For the lognormal kernel these are \code{beta} and \code{sigma}, for the box
kernel they are \code{ppmr_min} and \code{ppmr_max}, and for the Gaussian mixture they
are the list-columns \code{kernel_p}, \code{kernel_mean}, and \code{kernel_sd}. They are
explained in the help pages for the kernel functions. Except for \code{beta} and
\code{sigma}, no defaults are set for these parameters. If they are missing from
the species_params data frame then mizer will issue an error message.

You can use any other string for \code{pred_kernel_type}. If for example you
choose "my" then you need to define a function \code{my_pred_kernel} that you can
model on the existing functions like \code{\link[=lognormal_pred_kernel]{lognormal_pred_kernel()}}.

When using a kernel that depends on the predator/prey size ratio only, mizer
does not need to store the entire three dimensional array in the MizerParams
object. Such an array can be very big when there is a large number of size
bins. Instead, mizer only needs to store two two-dimensional arrays that hold
Fourier transforms of the feeding kernel function that allow the encounter
rate and the predation rate to be calculated very efficiently. However, if
you need the full three-dimensional array you can calculate it with the
\code{\link[=pred_kernel]{pred_kernel()}} function.

\strong{Kernel dependent on both predator and prey size}

If you want to work with a feeding kernel that depends on predator mass and
prey mass independently, you can specify the full feeding kernel as a
three-dimensional array (predator species x predator size x prey size).

You should use this option only if a kernel dependent only on the
predator/prey mass ratio is not appropriate. Using a kernel dependent on
predator/prey mass ratio only allows mizer to use fast Fourier transform
methods to significantly reduce the running time of simulations.

The order of the predator species in \code{pred_kernel} should be the same
as the order in the species params dataframe in the \code{params} object. If you
supply a named array then the function will check the order and warn if it is
different.
}

\section{Setting search volume}{

The search volume \eqn{\gamma_i(w)} of an individual of species \eqn{i}
and weight \eqn{w} multiplies the predation kernel when
calculating the encounter rate in \code{\link[=getEncounter]{getEncounter()}} and the
predation rate in \code{\link[=getPredRate]{getPredRate()}}.

The name "search volume" is a bit misleading, because \eqn{\gamma_i(w)} does
not have units of volume. It is simply a parameter that determines the rate
of predation. Its units depend on your choice, see section "Units in mizer".
If you have chosen to work with total abundances, then it is a rate with units
1/year. If you have chosen to work with abundances per m^2 then it has units
of m^2/year. If you have chosen to work with abundances per m^3 then it has
units of m^3/year.

If the \code{search_vol} argument is not supplied, then the search volume is
set to
\deqn{\gamma_i(w) = \gamma_i w^q_i.}
The values of \eqn{\gamma_i} (the search volume at 1g) and \eqn{q_i} (the
allometric exponent of the search volume) are taken from the \code{gamma} and
\code{q} columns in the species parameter dataframe. If the \code{gamma}
column is not supplied in the species parameter dataframe, a default is
calculated by the \code{\link[=get_gamma_default]{get_gamma_default()}} function. If the \code{q} column is not
supplied, a default of \code{lambda - 2 + n} is used. Note that only
for predators of size \eqn{w = 1} gram is the value of the species parameter
\eqn{\gamma_i} the same as the value of the search volume \eqn{\gamma_i(w)}.

If the \code{search_vol} slot has a comment and \code{reset = FALSE}, then a
recalculation from the species parameters is suppressed and a message is
issued if the recalculated values would differ from the stored ones.
}

\section{Setting maximum intake rate}{

The maximum intake rate \eqn{h_i(w)} of an individual of species \eqn{i} and
weight \eqn{w} determines the feeding level, calculated with
\code{\link[=getFeedingLevel]{getFeedingLevel()}}. It is measured in grams/year.

If the \code{intake_max} argument is not supplied, then the maximum intake
rate is set to \deqn{h_i(w) = h_i w^{n_i}.}
The values of \eqn{h_i} (the maximum intake rate of an individual of size 1
gram) and \eqn{n_i} (the allometric exponent for the intake rate) are taken
from the \code{h} and \code{n} columns in the species parameter dataframe. If
the \code{h} column is not supplied in the species parameter dataframe, it is
calculated by the \code{\link[=get_h_default]{get_h_default()}} function. If the \code{n} column is not
supplied, a default of \eqn{n_i = 3/4} is used.

If \eqn{h_i} is set to \code{Inf}, fish of species i will consume all encountered
food.

If the \code{intake_max} slot has a comment and \code{reset = FALSE}, then a
recalculation from the species parameters is suppressed and a message is
issued if the recalculated values would differ from the stored ones.
}

\section{Setting metabolic rate}{

The metabolic rate is subtracted from the energy income rate to calculate
the rate at which energy is available for growth and reproduction, see
\code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}}. It is measured in grams/year.

If the \code{metab} argument is not supplied, then for each species the
metabolic rate \eqn{k(w)} for an individual of size \eqn{w} is set to
\deqn{k(w) = k_s w^p + k w,}
where \eqn{k_s w^p} represents the rate of standard metabolism and \eqn{k w}
is the rate at which energy is expended on activity and movement. The values
of \eqn{k_s}, \eqn{p} and \eqn{k} are taken from the \code{ks}, \code{p} and
\code{k} columns in the species parameter dataframe. If any of these
parameters are not supplied, the defaults are \eqn{k = 0}, \eqn{p = n} and
\deqn{k_s = f_c h \alpha w_{mat}^{n-p},}{k_s = f_c * h * alpha * w_mat^(n - p),}
where \eqn{f_c} is the critical feeding level taken from the \code{fc} column
in the species parameter data frame. If the critical feeding level is not
specified, a default of \eqn{f_c = 0.2} is used.

If the \code{metab} slot has a comment and \code{reset = FALSE}, then a recalculation
from the species parameters is suppressed and a message is issued if the
recalculated values would differ from the stored ones.
}

\section{Setting external mortality rate}{

The external mortality is all the mortality that is not due to fishing or
predation by predators included in the model. The external mortality could be
due to predation by predators that are not explicitly included in the model
(e.g. mammals or seabirds) or due to other causes like illness. It is a rate
with units 1/year.

The \code{ext_mort} argument allows you to specify an external mortality rate
that depends on species and body size. You can see an example of this in
the Examples section of the help page for \code{\link[=setExtMort]{setExtMort()}}.

If the \code{ext_mort} argument is not supplied, then the external mortality is
taken from the species parameters as
\deqn{\mu_{ext.i}(w) = z_{0.i} + z_{ext.i} w^{d_i}.}{mu_ext,i(w) = z0_i + z_ext,i w ^ d_i.}
The value of the constant \eqn{z_0} for each species is taken from the \code{z0}
column of the species parameter data frame, if that column exists.
Otherwise it is calculated as
\deqn{z_{0.i} = {\tt z0pre}_i\, w_{inf}^{\tt z0exp}.}{z_{0.i} = z0pre_i w_{inf}^{z0exp}.}
Missing values of \code{z_ext} are set to 0 and missing values of \code{d} are set to
\code{n - 1}.

By default the power law is evaluated at the left bin edges \eqn{w_j}
(point sampling). If the \code{bin_average} entry of the \code{second_order_w} slot is
\code{TRUE} (see \code{\link[=second_order_w]{second_order_w()}}), then the \eqn{z_{ext} w^d} term is instead
replaced by its exact average over each bin \eqn{[w_j, w_{j+1}]},
\deqn{\frac{z_{ext}}{\Delta w_j}\int_{w_j}^{w_{j+1}} w^d\, dw
  = z_{ext}\,\frac{w_{j+1}^{d+1} - w_j^{d+1}}{(d+1)\,\Delta w_j},}
(with the limiting form \eqn{z_{ext}\ln(w_{j+1}/w_j)/\Delta w_j} when
\eqn{d = -1}). This is the consistent choice in the finite-volume scheme,
where the external mortality multiplies the bin-averaged abundance. The
bin-averaging is applied only to the auto-calculated power-law default; a
user-supplied \code{ext_mort} array is left untouched.
}

\section{Setting external encounter rate}{

The external encounter rate is the rate at which a predator encounters
food that is not explicitly modelled. It is a rate with units mass/year.

The \code{ext_encounter} argument allows you to specify an external encounter rate
that depends on species and body size. You can see an example of this in
the Examples section of the help page for \code{\link[=setExtEncounter]{setExtEncounter()}}.

If the \code{ext_encounter} argument is not supplied, then the external encounter
rate is calculated as a power law:
\deqn{E_{ext.i}(w) = E_{ext.i}\, w^{n_i}.}{E_{ext.i}(w) = E_{ext.i} * w^n_i.}
The coefficient \eqn{E_{ext.i}} is taken from the \code{E_ext} column of the
species parameter data frame, which defaults to 0. The exponent \eqn{n_i} is
taken from the \code{n} column of the species parameter data frame.

If the \code{ext_encounter} slot has a comment and \code{reset = FALSE}, then a
recalculation from the species parameters is suppressed and a message is
issued if the recalculated values would differ from the stored ones.
}

\section{Setting external diffusion rate}{

The external diffusion rate allows you to impose additional diffusion
beyond the predation-driven diffusion that can be internally modelled by mizer.

The \code{ext_diffusion} argument allows you to specify a diffusion rate that
depends on species and body size.

If the \code{ext_diffusion} argument is not supplied, then the external diffusion
rate is calculated as a power law:
\deqn{D_{ext.i}(w) = D_{ext.i}\, w^{n_i+1}.}{
  D_{ext.i}(w) = D_{ext.i} * w^(n_i+1).}
The coefficient \eqn{D_{ext.i}} is taken from the \code{D_ext} column of the
species parameter data frame, which defaults to 0. The exponent
\eqn{n_i + 1} uses the \code{n} column of the species parameter data frame.

If the \code{ext_diffusion} slot has a comment and \code{reset = FALSE}, then a
recalculation from the species parameters is suppressed and a message is
issued if the recalculated values would differ from the stored ones.
}

\section{Setting reproduction}{


For each species and at each size, the proportion \eqn{\psi}{psi} of the
available energy
that is invested into reproduction is the product of two factors: the
proportion \code{maturity} of individuals that are mature and the proportion
\code{repro_prop} of the energy available to a mature individual that is
invested into reproduction. There is a size \code{w_repro_max} at which a typical
mature individual invests all of its available energy into reproduction.
This is \strong{not} a hard ceiling on size: not all individuals are mature at
\code{w_repro_max}, and diffusion in the growth process allows some individuals to
grow beyond it, so fish larger than \code{w_repro_max} can exist. If you have not
specified the \code{w_repro_max} column in the species parameter data frame, then
the von Bertalanffy asymptotic size \code{w_inf} is used instead.

\subsection{Maturity ogive}{
If the the proportion of individuals that are mature is not supplied via
the \code{maturity} argument, then it is set to a sigmoidal
maturity ogive that changes from 0 to 1 at around the maturity size:
\deqn{{\tt maturity}(w) = \left[1+\left(\frac{w}{w_{mat}}\right)^{-U}\right]^{-1}.}{
  maturity(w) = [1+(w/w_mat)^(-U)]^(-1)}
(To avoid clutter, we are not showing the species index in the equations,
although each species has its own maturity ogive.)
The maturity weights are taken from the \code{w_mat} column of the
species_params data frame. Any missing maturity weights are set to 1/4 of the
asymptotic size in the \code{w_inf} column.

The exponent \eqn{U} determines the steepness of the maturity ogive. By
default it is chosen as \eqn{U = 10}, however this can be overridden by
including a column \code{w_mat25} in the species parameter dataframe that
specifies the weight at which 25\% of individuals are mature, which sets
\eqn{U = \log(3) / \log(w_{mat} / w_{mat25}).}{U = log(3) / log(w_mat /
w_mat25).}

The sigmoidal function given above would strictly reach 0 only
asymptotically and thus have some (negligible) amount of reproduction at
arbitrarily small size.
For computational simplicity, any proportion smaller than
\code{1e-8} is set to \code{0}.
}

\subsection{Investment into reproduction}{
If the the energy available to a mature individual that is
invested into reproduction is not supplied via the \code{repro_prop} argument,
it is set to the allometric form
\deqn{{\tt repro\_prop}(w) =
  \min\left(\left(\dfrac{w}{w_{\tt{repro\_max}}}\right)^{m-n},1\right).}{
  repro_prop(w) = min((w/w_repro_max)^(m - n),1).}
Here \eqn{n} is the scaling exponent of the energy income rate. Hence
the exponent \eqn{m} determines the scaling of the investment into
reproduction for mature individuals. By default it is chosen to be
\eqn{m = 1} so that the rate at which energy is invested into reproduction
scales linearly with the size. This default can be overridden by including a
column \code{m} in the species parameter dataframe. The sizes \eqn{w_{repro\_max}}
are taken from the \code{w_repro_max} column in the species parameter data frame,
if it exists, or otherwise from the \code{w_inf} column.

The total proportion of energy invested into reproduction of an individual
of size \eqn{w} is then
\deqn{\psi(w) = {\tt maturity}(w){\tt repro\_prop}(w)}{
  psi(w) = maturity(w) * repro_prop(w)}
In mizer edition 1, at sizes above \code{w_repro_max} the value of \eqn{\psi}
is additionally forced to 1, so that all available energy is invested into
reproduction and growth stops. In edition 2 and above this forcing is not
applied, and \eqn{\psi} is determined entirely by the maturity ogive and the
reproductive proportion.
}

\subsection{Reproductive efficiency}{
The reproductive efficiency \eqn{\epsilon}, i.e., the proportion of energy allocated to
reproduction that results in egg biomass, is set through the \code{erepro}
column in the species_params data frame. If that is not provided, the default
is set to 1 (which you will want to override). The offspring biomass divided
by the egg biomass gives the rate of egg production, returned by
\code{\link[=getRDI]{getRDI()}}:
\deqn{R_{di} = \frac{\epsilon}{2 w_{min}} \int N(w)  E_r(w) \psi(w) \, dw}{R_di = (\epsilon/(2 w_min)) \int N(w)  E_r(w) \psi(w) dw}
}

\subsection{Density dependence}{
The stock-recruitment relationship is an emergent phenomenon in mizer, with
several sources of density dependence. Firstly, the amount of energy invested
into reproduction depends on the energy income of the spawners, which is
density-dependent due to competition for prey. Secondly, the proportion of
larvae that grow up to recruitment size depends on the larval mortality,
which depends on the density of predators, and on larval growth rate, which
depends on density of prey.

Finally, to encode all the density dependence in the stock-recruitment
relationship that is not already included in the other two sources of density
dependence, mizer puts the the density-independent rate of egg production
through a density-dependence function. The result is returned by
\code{\link[=getRDD]{getRDD()}}. The name of the density-dependence function is
specified by the \code{RDD} argument. The default is the Beverton-Holt
function \code{\link[=BevertonHoltRDD]{BevertonHoltRDD()}}, which requires an \code{R_max} column
in the species_params data frame giving the maximum egg production rate. If
this column does not exist, it is initialised to \code{Inf}, leading to no
density-dependence. Other functions provided by mizer are
\code{\link[=RickerRDD]{RickerRDD()}} and \code{\link[=SheperdRDD]{SheperdRDD()}} and you can easily use
these as models for writing your own functions.
}
}

\section{Setting fishing}{


\strong{Gears}

In \code{mizer}, fishing mortality is imposed on species by fishing gears. The
total per-capita fishing mortality (1/year) is obtained by summing over the
mortality from all gears,
\deqn{\mu_{f.i}(w) = \sum_g F_{g,i}(w),}
where the fishing mortality \eqn{F_{g,i}(w)} imposed by gear \eqn{g} on
species \eqn{i} at size \eqn{w} is calculated as:
\deqn{F_{g,i}(w) = S_{g,i}(w) Q_{g,i} E_{g},}
where \eqn{S} is the selectivity by species, gear and size, \eqn{Q} is the
catchability by species and gear and \eqn{E} is the fishing effort by gear.

\strong{Selectivity}

The selectivity at size of each gear for each species is saved as a three
dimensional array (gear x species x size). Each entry has a range between 0
(that gear is not selecting that species at that size) to 1 (that gear is
selecting all individuals of that species of that size). This three
dimensional array can be specified explicitly via the \code{selectivity}
argument, but usually mizer calculates it from the \code{gear_params} slot of
the MizerParams object.

To allow the calculation of the \code{selectivity} array, the \code{gear_params} slot
must be a data frame with one row for each gear-species combination. So if
for example a gear can select three species, then that gear contributes three
rows to the \code{gear_params} data frame, one for each species it can select. The
data frame must have columns \code{gear}, holding the name of the gear, \code{species},
holding the name of the species, and \code{sel_func}, holding the name of the
function that calculates the selectivity curve. Some selectivity functions
are included in the package: \code{knife_edge()}, \code{sigmoid_length()},
\code{double_sigmoid_length()}, and \code{sigmoid_weight()}.
Users are able to write their own size-based selectivity function. The first
argument to the function must be \code{w} and the function must return a vector of
the selectivity (between 0 and 1) at size.

Each selectivity function may have parameters. Values for these
parameters must be included as columns in the gear parameters data.frame.
The names of the columns must exactly match the names of the corresponding
arguments of the selectivity function. For example, the default selectivity
function is \code{knife_edge()} that a has sudden change of selectivity from 0 to 1
at a certain size. In its help page you can see that the \code{knife_edge()}
function has arguments \code{w} and \code{knife_edge_size}. The first argument, \code{w}, is
size (the function calculates selectivity at size). All selectivity functions
must have \code{w} as the first argument. The values for the other arguments must
be found in the gear parameters data.frame. So for the \code{knife_edge()}
function there should be a \code{knife_edge_size} column. Because \code{knife_edge()}
is the default selectivity function, the \code{knife_edge_size} argument has a
default value = \code{w_mat}.

The most commonly-used selectivity function is \code{sigmoid_length()}. It has a
smooth transition from 0 to 1 at a certain size. The \code{sigmoid_length()}
function has the two parameters \code{l50} and \code{l25} that are the lengths in cm at
which 50\% or 25\% of the fish are selected by the gear. If you choose this
selectivity function then the \code{l50} and \code{l25} columns must be included in the
gear parameters data.frame.

In case each species is only selected by one gear, the columns of the
\code{gear_params} data frame can alternatively be provided as columns of the
\code{species_params} data frame, if this is more convenient for the user to set
up. Mizer will then copy these columns over to create the \code{gear_params} data
frame when it creates the MizerParams object. However changing these columns
in the species parameter data frame later will not update the \code{gear_params}
data frame.

\strong{Catchability}

Catchability is used as an additional factor to make the link between gear
selectivity, fishing effort and fishing mortality. For example, it can be set
so that an effort of 1 gives a desired fishing mortality. In this way effort
can then be specified relative to a 'base effort', e.g. the effort in a
particular year.

Catchability is stored as a two dimensional array (gear x species). This can
either be provided explicitly via the \code{catchability} argument, or the
information can be provided via a \code{catchability} column in the \code{gear_params}
data frame.

In the case where each species is selected by only a single gear, the
\code{catchability} column can also be provided in the \code{species_params} data
frame. Mizer will then copy this over to the \code{gear_params} data frame when
the MizerParams object is created.

\strong{Effort}

The initial fishing effort is stored in the \code{MizerParams} object. If it is
not supplied, it is set to zero. The initial effort can be overruled when
the simulation is run with \code{project()}, where it is also possible to specify
an effort that varies through time.
}

\section{Setting resource dynamics}{


The \code{resource_dynamics} argument allows you to choose the resource dynamics
function. By default, mizer uses a semichemostat model to describe the
resource dynamics in each size class independently. This semichemostat
dynamics is implemented by the function \code{\link[=resource_semichemostat]{resource_semichemostat()}}. You can
change that to use a logistic model implemented by \code{\link[=resource_logistic]{resource_logistic()}} or
you can use \code{\link[=resource_constant]{resource_constant()}} which keeps the resource constant or you
can write your own function.

Both the \code{\link[=resource_semichemostat]{resource_semichemostat()}} and the \code{\link[=resource_logistic]{resource_logistic()}} dynamics
are parametrised in terms of a size-dependent birth rate \eqn{r_R(w)} and a
size-dependent capacity \eqn{c_R}. The help pages of these functions give
the details.

The \code{resource_rate} argument can be a vector (with the same length as
\code{w_full(params)}) specifying the intrinsic resource birth rate for each size
class. Alternatively it can be a single number that is used as the
coefficient in a power law: then the intrinsic birth rate \eqn{r_R(w)} at
size \eqn{w} is set to
\deqn{r_R(w) = r_R w^{n-1}.}
The power-law exponent \eqn{n} is taken from the \code{n} argument.

The \code{resource_capacity} argument can be a vector specifying the intrinsic
resource carrying capacity for each size class. Alternatively it can be a
single number that is used as the coefficient in a truncated power
law: then the intrinsic carrying capacity \eqn{c_R(w)} at size \eqn{w}
is set to
\deqn{c_R(w) = c_R\, w^{-\lambda}}{c_R(w) = c_R w^{-\lambda}}
for all \eqn{w} less than \code{w_pp_cutoff} and zero for larger sizes.
The power-law exponent \eqn{\lambda} is taken from the \code{lambda} argument.

The values for \code{lambda}, \code{n} and \code{w_pp_cutoff} are stored in a list
in the \code{resource_params} slot of the MizerParams object so that they can be
re-used automatically in the future. If you specify \code{resource_rate} or
\code{resource_capacity} as a single number, that coefficient is likewise stored,
as \code{r_pp} and \code{kappa} respectively. That list can be accessed with
\code{\link[=resource_params]{resource_params()}}.
}

\examples{
params <- newMultispeciesParams(NS_species_params)
}
\seealso{
Other functions for setting up models:
\code{\link[=newCommunityParams]{newCommunityParams()}},
\code{\link[=newSingleSpeciesParams]{newSingleSpeciesParams()}},
\code{\link[=newTraitParams]{newTraitParams()}}
}
\concept{functions for setting up models}
