% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{NOther}
\alias{NOther}
\alias{finalNOther}
\title{Time series of other components}
\usage{
NOther(sim)

finalNOther(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
For \code{NOther}: A list array indexed by time and component that stores the projected
values for other ecosystem components.

For \code{finalNOther}: A named list holding the values of other ecosystem components at the
end of the simulation
}
\description{
Fetch the simulation results for other components over time.
}
\seealso{
Other extension tools: 
\code{\link{clearExtensionChain}()},
\code{\link{coerceToExtensionClass}()},
\code{\link{getRegisteredExtensions}()},
\code{\link{initialNOther<-}()},
\code{\link{recordExtension}()},
\code{\link{registerExtension}()},
\code{\link{registerExtensions}()},
\code{\link{setComponent}()},
\code{\link{setRateFunction}()}
}
\concept{extension tools}
