% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankton_dynamics.R
\name{plankton_semichemostat}
\alias{plankton_semichemostat}
\title{Project plankton using semichemostat model}
\usage{
plankton_semichemostat(params, n, n_pp, B, rates, dt, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the plankton abundance by size}

\item{B}{The biomass of detritus}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step}

\item{...}{Unused}
}
\value{
Vector containing plankton spectrum at next timestep
}
\description{
Project plankton using semichemostat model
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter,
                             plankton_dyn = plankton_semichemostat)
}
}
