% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayResourceBySize-class.R
\name{plot.ArrayResourceBySize}
\alias{plot.ArrayResourceBySize}
\title{Plot method for \code{ArrayResourceBySize} objects}
\usage{
\method{plot}{ArrayResourceBySize}(
  x,
  return_data = FALSE,
  log_x = TRUE,
  log_y = TRUE,
  log = NULL,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  y_ticks = 6,
  ...
)
}
\arguments{
\item{x}{An \code{ArrayResourceBySize} object.}

\item{return_data}{If \code{TRUE}, return the data frame instead of the
plot.}

\item{log_x}{If \code{TRUE}, use a log10 x-axis. Default is \code{TRUE}.}

\item{log_y}{If \code{TRUE}, use a log10 y-axis. Default is \code{TRUE}.}

\item{log}{Character string specifying which axes should use log10
scales, in the same form as the base \code{\link[=plot]{plot()}} argument. For example,
\code{"x"}, \code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and
\code{log_y}.}

\item{wlim}{A numeric vector of length two providing lower and upper
limits for the weight (x) axis. Use \code{NA} to refer to the existing
minimum or maximum.}

\item{ylim}{A numeric vector of length two providing lower and upper
limits for the value (y) axis. Use \code{NA} to refer to the existing
minimum or maximum.}

\item{y_ticks}{The approximate number of ticks desired on the y axis.}

\item{...}{Unused.}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a
data frame is returned.
}
\description{
See \code{\link[=plot]{plot()}} for an overview of the mizer plotting system and the
arguments shared by all of its methods.
}
\examples{
\donttest{
plot(getResourceMort(NS_params))
plot(initialNResource(NS_params))
}
}
\keyword{internal}
