% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R
\name{plot.ArraySpeciesBySize}
\alias{plot.ArraySpeciesBySize}
\title{Plot a ArraySpeciesBySize object}
\usage{
\method{plot}{ArraySpeciesBySize}(
  x,
  species = NULL,
  all.sizes = FALSE,
  highlight = NULL,
  return_data = FALSE,
  log_x = TRUE,
  log_y = FALSE,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  ...
)
}
\arguments{
\item{x}{An \code{ArraySpeciesBySize} object.}

\item{species}{Character vector of species to include. \code{NULL} (default) means
all species.}

\item{all.sizes}{If \code{FALSE} (default), values outside a species' size range
(\code{w_min} to \code{w_max}) are removed.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{return_data}{If \code{TRUE}, return the data frame instead of the plot.}

\item{log_x}{If \code{TRUE} (default), use a log10 x-axis.}

\item{log_y}{If \code{TRUE}, use a log10 y-axis. Default is \code{FALSE}. Useful for
quantities that span several orders of magnitude (e.g. encounter rates,
search volumes).}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the weight (x) axis. Use \code{NA} to refer to the existing minimum or
maximum.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the value (y) axis. Use \code{NA} to refer to the existing minimum or
maximum.}

\item{...}{Further arguments (currently unused).}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the variables 'w', 'value', 'Species' is returned.
}
\description{
Plots the value against size for each species, using species colours and
linetypes from the MizerParams object stored in the \code{params} attribute.
}
\examples{
\donttest{
plot(getEncounter(NS_params))
plot(getFeedingLevel(NS_params), species = c("Cod", "Herring"))
}
}
