% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R
\name{plot.ArraySpeciesBySize}
\alias{plot.ArraySpeciesBySize}
\title{Plot method for \code{ArraySpeciesBySize} objects}
\usage{
\method{plot}{ArraySpeciesBySize}(
  x,
  species = NULL,
  all.sizes = FALSE,
  highlight = NULL,
  return_data = FALSE,
  log_x = TRUE,
  log_y = FALSE,
  log = NULL,
  wlim = c(NA, NA),
  llim = c(NA, NA),
  ylim = c(NA, NA),
  size_axis = c("w", "l"),
  total = FALSE,
  background = TRUE,
  y_ticks = 6,
  ...
)
}
\arguments{
\item{x}{An \code{ArraySpeciesBySize} object.}

\item{species}{Character vector of species to include. \code{NULL}
(default) means all species.}

\item{all.sizes}{If \code{FALSE} (default), values outside a species' size
range (\code{w_min} to \code{w_max}) are removed.}

\item{highlight}{Name or vector of names of the species to be
highlighted.}

\item{return_data}{If \code{TRUE}, return the data frame instead of the
plot.}

\item{log_x}{If \code{TRUE}, use a log10 x-axis. Default is \code{TRUE}.}

\item{log_y}{If \code{TRUE}, use a log10 y-axis. Default is \code{FALSE}.}

\item{log}{Character string specifying which axes should use log10
scales, in the same form as the base \code{\link[=plot]{plot()}} argument. For example,
\code{"x"}, \code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and
\code{log_y}.}

\item{wlim}{A numeric vector of length two providing lower and upper
limits for the weight (x) axis. Use \code{NA} to refer to the existing
minimum or maximum.}

\item{llim}{A numeric vector of length two providing lower and upper
limits for the length (x) axis when \code{size_axis = "l"}. Use \code{NA} to
refer to the existing minimum or maximum.}

\item{ylim}{A numeric vector of length two providing lower and upper
limits for the value (y) axis. Use \code{NA} to refer to the existing
minimum or maximum.}

\item{size_axis}{Whether to plot size as weight (\code{"w"}, default) or
length (\code{"l"}), using the allometric weight-length relationship.}

\item{total}{A boolean value that determines whether the total over
all selected species is plotted as well. Default is \code{FALSE}.}

\item{background}{A boolean value that determines whether background
species are included. Ignored if the model does not contain background
species. Default is \code{TRUE}.}

\item{y_ticks}{The approximate number of ticks desired on the y axis.}

\item{...}{Unused.}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a
data frame is returned.
}
\description{
See \code{\link[=plot]{plot()}} for an overview of the mizer plotting system and the
arguments shared by all of its methods.
}
\examples{
\donttest{
plot(getEncounter(NS_params))
plot(getFeedingLevel(NS_params), species = c("Cod", "Herring"))
plot(getPredMort(NS_params), species = c("Cod", "Herring"),
     size_axis = "l")
}
}
\keyword{internal}
