% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R
\name{plot.ArraySpeciesBySize}
\alias{plot.ArraySpeciesBySize}
\title{Plot a ArraySpeciesBySize object}
\usage{
\method{plot}{ArraySpeciesBySize}(
  x,
  params = NULL,
  species = NULL,
  all.sizes = FALSE,
  highlight = NULL,
  return_data = FALSE,
  log_x = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{ArraySpeciesBySize} object.}

\item{params}{A \code{MizerParams} object. Used for species colours, linetypes,
and size ranges. If \code{NULL}, a basic plot is produced.}

\item{species}{Character vector of species to include. \code{NULL} (default) means
all species.}

\item{all.sizes}{If \code{FALSE} (default), values outside a species' size range
(\code{w_min} to \code{w_max}) are removed. Only effective when \code{params} is provided.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{return_data}{If \code{TRUE}, return the data frame instead of the plot.}

\item{log_x}{If \code{TRUE} (default), use a log10 x-axis.}

\item{...}{Further arguments (currently unused).}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the variables 'w', 'value', 'Species' is returned.
}
\description{
Plots the value against size for each species, using species colours and
linetypes from the MizerParams object.
}
\examples{
\donttest{
plot(getEncounter(NS_params), NS_params)
plot(getFeedingLevel(NS_params), NS_params,
     species = c("Cod", "Herring"))
}
}
