% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R,
%   R/ArrayTimeBySpecies-class.R, R/ArrayTimeBySpeciesBySize-class.R,
%   R/ArrayResourceBySize-class.R
\name{plot}
\alias{plot}
\alias{plot.ArraySpeciesBySize}
\alias{plot.ArrayTimeBySpecies}
\alias{plot.ArrayTimeBySpeciesBySize}
\alias{plot.ArrayResourceBySize}
\alias{plot.ArrayTimeByResourceBySize}
\title{Plot mizer arrays}
\arguments{
\item{x}{An \code{ArraySpeciesBySize}, \code{ArrayTimeBySpecies}, or
\code{ArrayTimeBySpeciesBySize} object.}

\item{...}{\strong{Arguments used by all methods:}
\describe{
\item{\code{species}}{Character vector of species to include. \code{NULL}
(default) means all species.}
\item{\code{highlight}}{Name or vector of names of the species to be
highlighted.}
\item{\code{total}}{A boolean value that determines whether the total over
all selected species is plotted as well. Default is \code{FALSE}.}
\item{\code{background}}{A boolean value that determines whether background
species are included. Ignored if the model does not contain background
species. Default is \code{TRUE}.}
\item{\code{return_data}}{If \code{TRUE}, return the data frame instead of the
plot.}
\item{\code{log_x}}{If \code{TRUE}, use a log10 x-axis. Default is \code{TRUE} for size
spectra and \code{FALSE} for time series.}
\item{\code{log_y}}{If \code{TRUE}, use a log10 y-axis. Default is \code{FALSE} for
\code{ArraySpeciesBySize} and \code{TRUE} for \code{ArrayTimeBySpecies}.}
\item{\code{log}}{Character string specifying which axes should use log10
scales, in the same form as the base \code{\link[=plot]{plot()}} argument. For example,
\code{"x"}, \code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and
\code{log_y}.}
\item{\code{ylim}}{A numeric vector of length two providing lower and upper
limits for the value (y) axis. Use \code{NA} to refer to the existing
minimum or maximum.}
\item{\code{y_ticks}}{The approximate number of ticks desired on the y axis.}
}

\strong{For \code{ArraySpeciesBySize} and \code{ArrayTimeBySpeciesBySize} methods:}
\describe{
\item{\code{all.sizes}}{If \code{FALSE} (default), values outside a species' size
range (\code{w_min} to \code{w_max}) are removed.}
\item{\code{wlim}}{A numeric vector of length two providing lower and upper
limits for the weight (x) axis. Use \code{NA} to refer to the existing
minimum or maximum.}
\item{\code{llim}}{A numeric vector of length two providing lower and upper
limits for the length (x) axis when \code{size_axis = "l"}. Use \code{NA} to
refer to the existing minimum or maximum.}
\item{\code{size_axis}}{Whether to plot size as weight (\code{"w"}, default) or
length (\code{"l"}), using the allometric weight-length relationship.}
}

\strong{For \code{ArrayTimeBySpecies} methods:}
\describe{
\item{\code{tlim}}{A numeric vector of length two providing lower and upper
limits for the time axis, e.g. \code{c(1980, 2000)}. Use \code{NA} to apply no
limit at that end. Default is \code{c(NA, NA)}.}
}

\strong{For \code{ArrayTimeBySpeciesBySize} methods:}
\describe{
\item{\code{time}}{The time to display. Default (\code{NULL}) is the final time
step.}
}}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame is returned. \code{\link[=plotHover]{plotHover()}} returns a plotly object.
}
\description{
Many mizer functions return values that depend on species and either size or
time. \code{plot()} creates a ggplot2 figure with one line for each species
showing the values against size or against time (depending on the type of
output). \code{\link[=plotHover]{plotHover()}} creates an interactive version of the same figure.
}
\details{
This works because the mizer functions that give values that depend on
species and size return an \code{ArraySpeciesBySize} object and those that
give values that depend on species and time return an \code{ArrayTimeBySpecies}
object. These objects have attributes that store the name of the value,
its units, and a reference to the \code{MizerParams} object that the value was
computed from. This allows the plots to be automatically labelled and
coloured appropriately.

To compare two mizer arrays in a single plot, use \code{\link[=plot2]{plot2()}}. To show the
relative difference between two arrays, use \code{\link[=plotRelative]{plotRelative()}}.
}
\examples{
\donttest{
plot(getEncounter(NS_params))
plot(getFeedingLevel(NS_params), species = c("Cod", "Herring"))
plot(getPredMort(NS_params), species = c("Cod", "Herring"),
     size_axis = "l")
}
\donttest{
plot(getBiomass(NS_sim))
plot(getBiomass(NS_sim), species = c("Cod", "Herring"), total = TRUE)
plot(getYield(NS_sim), species = c("Cod", "Herring"))
}
\donttest{
plot(getFMort(NS_sim), time = 2010)
}
\donttest{
plot(getResourceMort(NS_params))
plot(initialNResource(NS_params))
}
\donttest{
plot(NResource(NS_sim))
}
}
\seealso{
Other plotting functions: 
\code{\link{addPlot}()},
\code{\link{animate.ArrayTimeBySpeciesBySize}()},
\code{\link{plot2}()},
\code{\link{plotBiomass}()},
\code{\link{plotCDF}()},
\code{\link{plotCDF2}()},
\code{\link{plotDiet}()},
\code{\link{plotFMort}()},
\code{\link{plotFeedingLevel}()},
\code{\link{plotGrowthCurves}()},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link{plotPredMort}()},
\code{\link{plotRelative}()},
\code{\link{plotSpectra}()},
\code{\link{plotSpectra2}()},
\code{\link{plotSpectraRelative}()},
\code{\link{plotYield}()},
\code{\link{plotYieldGear}()},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
