% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R,
%   R/ArrayTimeBySpecies-class.R, R/ArrayTimeBySpeciesBySize-class.R
\name{plot}
\alias{plot}
\alias{plot.ArraySpeciesBySize}
\alias{ggplotly.ArraySpeciesBySize}
\alias{plot.ArrayTimeBySpecies}
\alias{ggplotly.ArrayTimeBySpecies}
\alias{plot.ArrayTimeBySpeciesBySize}
\alias{ggplotly.ArrayTimeBySpeciesBySize}
\title{Plotting outputs}
\usage{
\method{plot}{ArraySpeciesBySize}(
  x,
  species = NULL,
  all.sizes = FALSE,
  highlight = NULL,
  return_data = FALSE,
  log_x = TRUE,
  log_y = FALSE,
  log = NULL,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  y_ticks = 6,
  ...
)

\method{ggplotly}{ArraySpeciesBySize}(x, ...)

\method{plot}{ArrayTimeBySpecies}(
  x,
  species = NULL,
  start_time = NULL,
  end_time = NULL,
  y_ticks = 6,
  ylim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  highlight = NULL,
  log_x = FALSE,
  log_y = TRUE,
  log = NULL,
  return_data = FALSE,
  ...
)

\method{ggplotly}{ArrayTimeBySpecies}(x, ...)

\method{plot}{ArrayTimeBySpeciesBySize}(
  x,
  species = NULL,
  time = NULL,
  all.sizes = FALSE,
  highlight = NULL,
  return_data = FALSE,
  log_x = TRUE,
  log_y = FALSE,
  log = NULL,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  y_ticks = 6,
  ...
)

\method{ggplotly}{ArrayTimeBySpeciesBySize}(x, ...)
}
\arguments{
\item{x}{An \code{ArraySpeciesBySize} or \code{ArrayTimeBySpecies} object.}

\item{species}{Character vector of species to include. \code{NULL} (default) means
all species.}

\item{all.sizes}{If \code{FALSE} (default), values outside a species' size range
(\code{w_min} to \code{w_max}) are removed. Only applies to \code{ArraySpeciesBySize}.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{return_data}{If \code{TRUE}, return the data frame instead of the plot.}

\item{log_x}{If \code{TRUE}, use a log10 x-axis. Default is \code{TRUE} for size
spectra and \code{FALSE} for time series.}

\item{log_y}{If \code{TRUE}, use a log10 y-axis. Default is \code{FALSE} for
\code{ArraySpeciesBySize} and \code{TRUE} for \code{ArrayTimeBySpecies}.}

\item{log}{Character string specifying which axes should use log10 scales,
in the same form as the base \code{\link[=plot]{plot()}} argument. For example, \code{"x"},
\code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and \code{log_y}.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the weight (x) axis. Use \code{NA} to refer to the existing minimum or
maximum. Only applies to \code{ArraySpeciesBySize}.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the value (y) axis. Use \code{NA} to refer to the existing minimum or
maximum.}

\item{total}{A boolean value that determines whether the total over all
selected species is plotted as well. Default is \code{FALSE}.}

\item{background}{A boolean value that determines whether background species
are included. Ignored if the model does not contain background species.
Default is \code{TRUE}.}

\item{y_ticks}{The approximate number of ticks desired on the y axis.}

\item{...}{Further arguments (currently unused).}

\item{start_time}{The first time to be plotted. Default (\code{NULL}) is the
beginning of the time series. Only applies to \code{ArrayTimeBySpecies}.}

\item{end_time}{The last time to be plotted. Default (\code{NULL}) is the end of
the time series. Only applies to \code{ArrayTimeBySpecies}.}

\item{time}{The time to display. Default (\code{NULL}) is the final time step.
Only applies to \code{ArrayTimeBySpeciesBySize}.}
}
\value{
\code{plot()} returns a ggplot2 object, unless \code{return_data = TRUE}, in
which case a data frame is returned.

\code{ggplotly()} returns a plotly object.
}
\description{
Many mizer functions return values that depend on species and either size or
time. \code{plot()} creates a ggplot2 figure with one line for each species
showing the values against size or against time (depending on the type of
output). \code{ggplotly()} creates an interactive version of the same figure.
}
\details{
This works because the mizer functions that give values that depend on
species and size return an \code{ArraySpeciesBySize} object and those that
give values that depend on species and time return an \code{ArrayTimeBySpecies}
object. These objects have attributes that store the name of the value,
its units, and a reference to the \code{MizerParams} object that the value was
computed from. This allows the plots to be automatically labelled and
coloured appropriately.
}
\examples{
\donttest{
plot(getEncounter(NS_params))
plot(getFeedingLevel(NS_params), species = c("Cod", "Herring"))
}
\donttest{
ggplotly(getEncounter(NS_params))
}
\donttest{
plot(getBiomass(NS_sim))
plot(getBiomass(NS_sim), species = c("Cod", "Herring"), total = TRUE)
plot(getYield(NS_sim), species = c("Cod", "Herring"))
}
\donttest{
ggplotly(getBiomass(NS_sim))
}
\donttest{
plot(getFMort(NS_sim))
plot(getFMort(NS_sim), time = 2010)
}
\donttest{
ggplotly(getFMort(NS_sim))
}
}
\seealso{
Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link[=plot2]{plot2()}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotRelative]{plotRelative()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectra2]{plotSpectra2()}},
\code{\link[=plotSpectraRelative]{plotSpectraRelative()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
