% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R,
%   R/ArrayTimeBySpecies-class.R, R/ArrayTimeBySpeciesBySize-class.R,
%   R/ArrayResourceBySize-class.R
\name{plot2}
\alias{plot2}
\alias{plot2.ArraySpeciesBySize}
\alias{plot2.ArrayTimeBySpecies}
\alias{plot2.ArrayTimeBySpeciesBySize}
\alias{plot2.ArrayResourceBySize}
\alias{plot2.ArrayTimeByResourceBySize}
\title{Compare two mizer arrays in a single plot}
\usage{
plot2(
  x,
  y,
  name1 = "First",
  name2 = "Second",
  species = NULL,
  log_x,
  log_y,
  log = NULL,
  ylim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  y_ticks = 6,
  ...
)
}
\arguments{
\item{x}{The first of two compatible mizer array objects to compare.
Can be an \code{ArraySpeciesBySize}, \code{ArrayTimeBySpecies},
\code{ArrayTimeBySpeciesBySize}, \code{ArrayResourceBySize} or
\code{ArrayTimeByResourceBySize} object.}

\item{y}{The second mizer array object, compatible with \code{x}.}

\item{name1, name2}{Labels for the two objects, used in the linetype legend.}

\item{species}{Character vector of species to include. \code{NULL} (default) means
all species. A resource array holds a single spectrum, so this argument is
not used by the resource methods, which warn if it is set.}

\item{log_x}{If \code{TRUE}, use a log10 x-axis. Default is \code{TRUE} for size
spectra and \code{FALSE} for time series.}

\item{log_y}{If \code{TRUE}, use a log10 y-axis. Default is \code{FALSE} for
\code{ArraySpeciesBySize} and \code{TRUE} for \code{ArrayTimeBySpecies} and for the
resource classes.}

\item{log}{Character string specifying which axes should use log10 scales,
in the same form as the base \code{\link[=plot]{plot()}} argument. For example, \code{"x"},
\code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and \code{log_y}.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the value (y) axis. Use \code{NA} to refer to the existing minimum or
maximum.}

\item{total}{A boolean value that determines whether the total over all
selected species is plotted as well. Default is \code{FALSE}. Not used by the
resource methods, which warn if it is set.}

\item{background}{A boolean value that determines whether background species
are included. Ignored if the model does not contain background species.
Default is \code{TRUE}. Not used by the resource methods, which warn if it is
set.}

\item{y_ticks}{The approximate number of ticks desired on the y axis.}

\item{...}{Further arguments used by only some of the methods:

\strong{For the \code{ArraySpeciesBySize}, \code{ArrayTimeBySpeciesBySize},
\code{ArrayResourceBySize} and \code{ArrayTimeByResourceBySize} methods:}
\describe{
\item{\code{wlim}}{A numeric vector of length two providing lower and upper
limits for the weight (x) axis. Use \code{NA} to refer to the existing
minimum or maximum.}
}

\strong{For the \code{ArraySpeciesBySize} and \code{ArrayTimeBySpeciesBySize} methods:}
\describe{
\item{\code{all.sizes}}{If \code{FALSE} (default), values outside a species' size
range (\code{w_min} to \code{w_max}) are removed.}
\item{\code{llim}}{A numeric vector of length two providing lower and upper
limits for the length (x) axis when \code{size_axis = "l"}. Use \code{NA} to
refer to the existing minimum or maximum.}
\item{\code{size_axis}}{Whether to plot size as weight (\code{"w"}, default) or
length (\code{"l"}), using the allometric weight-length relationship. Not
available for the resource classes, because the weight-length
relationship is a species parameter.}
}

\strong{For \code{ArrayTimeBySpecies} methods:}
\describe{
\item{\code{tlim}}{A numeric vector of length two providing lower and upper
limits for the time axis, e.g. \code{c(1980, 2000)}. Use \code{NA} to apply no
limit at that end. Default is \code{c(NA, NA)}.}
}

\strong{For the \code{ArrayTimeBySpeciesBySize} and \code{ArrayTimeByResourceBySize}
methods:}
\describe{
\item{\code{time}}{The time to display. Default (\code{NULL}) is the final time
step.}
}}
}
\value{
A ggplot2 object.
}
\description{
\code{plot2()} compares two compatible mizer array objects in a single ggplot.
Colours identify species or groups, and linetype identifies which object
the values came from.
}
\examples{
\donttest{
plot2(getEncounter(NS_params), getEncounter(NS_params))
plot2(getResourceMort(NS_params), getResourceMort(NS_params))
}
}
\seealso{
Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate]{animate()}},
\code{\link{plot}},
\code{\link[=plotBifurcation]{plotBifurcation()}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotRelative]{plotRelative()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectra2]{plotSpectra2()}},
\code{\link[=plotSpectraRelative]{plotSpectraRelative()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
