% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R,
%   R/ArrayTimeBySpecies-class.R, R/ArrayTimeBySpeciesBySize-class.R
\name{plot2}
\alias{plot2}
\alias{plot2.ArraySpeciesBySize}
\alias{plot2.ArrayTimeBySpecies}
\alias{plot2.ArrayTimeBySpeciesBySize}
\title{Compare two mizer array objects in one plot}
\usage{
plot2(x, y, ...)

\method{plot2}{ArraySpeciesBySize}(
  x,
  y,
  name1 = "First",
  name2 = "Second",
  species = NULL,
  all.sizes = FALSE,
  log_x = TRUE,
  log_y = FALSE,
  log = NULL,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  y_ticks = 6,
  ...
)

\method{plot2}{ArrayTimeBySpecies}(
  x,
  y,
  name1 = "First",
  name2 = "Second",
  species = NULL,
  start_time = NULL,
  end_time = NULL,
  y_ticks = 6,
  ylim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  log_x = FALSE,
  log_y = TRUE,
  log = NULL,
  ...
)

\method{plot2}{ArrayTimeBySpeciesBySize}(
  x,
  y,
  name1 = "First",
  name2 = "Second",
  species = NULL,
  time = NULL,
  all.sizes = FALSE,
  log_x = TRUE,
  log_y = FALSE,
  log = NULL,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  y_ticks = 6,
  ...
)
}
\arguments{
\item{x, y}{Two compatible mizer array objects of the same class.}

\item{...}{Further arguments (currently unused).}

\item{name1, name2}{Labels for the two objects, used in the linetype legend.}

\item{species}{Character vector of species to include. \code{NULL} (default) means
all species.}

\item{all.sizes}{If \code{FALSE} (default), values outside a species' size range
(\code{w_min} to \code{w_max}) are removed. Only applies to \code{ArraySpeciesBySize}.}

\item{log_x}{If \code{TRUE}, use a log10 x-axis. Default is \code{TRUE} for size
spectra and \code{FALSE} for time series.}

\item{log_y}{If \code{TRUE}, use a log10 y-axis. Default is \code{FALSE} for
\code{ArraySpeciesBySize} and \code{TRUE} for \code{ArrayTimeBySpecies}.}

\item{log}{Character string specifying which axes should use log10 scales,
in the same form as the base \code{\link[=plot]{plot()}} argument. For example, \code{"x"},
\code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and \code{log_y}.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the weight (x) axis. Use \code{NA} to refer to the existing minimum or
maximum. Only applies to \code{ArraySpeciesBySize}.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the value (y) axis. Use \code{NA} to refer to the existing minimum or
maximum.}

\item{total}{A boolean value that determines whether the total over all
selected species is plotted as well. Default is \code{FALSE}.}

\item{background}{A boolean value that determines whether background species
are included. Ignored if the model does not contain background species.
Default is \code{TRUE}.}

\item{y_ticks}{The approximate number of ticks desired on the y axis.}

\item{start_time}{The first time to be plotted. Default (\code{NULL}) is the
beginning of the time series. Only applies to \code{ArrayTimeBySpecies}.}

\item{end_time}{The last time to be plotted. Default (\code{NULL}) is the end of
the time series. Only applies to \code{ArrayTimeBySpecies}.}

\item{time}{The time to display. Default (\code{NULL}) is the final time step.
Only applies to \code{ArrayTimeBySpeciesBySize}.}
}
\value{
A ggplot2 object.
}
\description{
\code{plot2()} compares two compatible mizer array objects in a single ggplot.
Colours identify species or groups, and linetype identifies which object the
values came from.
}
\examples{
\donttest{
enc <- getEncounter(NS_params)
plot2(enc, enc, name1 = "Original", name2 = "Changed")
plot2(getBiomass(NS_sim), getBiomass(NS_sim), species = "Cod")
}
}
\seealso{
Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotRelative]{plotRelative()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectra2]{plotSpectra2()}},
\code{\link[=plotSpectraRelative]{plotSpectraRelative()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
