% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotComparisonDataFrame}
\alias{plotComparisonDataFrame}
\title{Make a plot comparing two data frames}
\usage{
plotComparisonDataFrame(
  frame1,
  frame2,
  params,
  name1 = "First",
  name2 = "Second",
  xlab = waiver(),
  ylab = waiver(),
  xtrans = "identity",
  ytrans = "identity",
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  y_ticks = 6,
  highlight = NULL,
  legend_var = "Legend",
  size_axis = NULL
)
}
\arguments{
\item{frame1, frame2}{Data frames sharing the same first three variables (x, y
and grouping variable). The names of \code{frame1} are imposed on \code{frame2}.}

\item{params}{A MizerParams object, used for the line colours.}

\item{name1, name2}{Labels for the two data frames, used in the linetype
legend.}

\item{xlab, ylab}{Labels for the x and y axes.}

\item{xtrans, ytrans}{Transformations for the x and y axes, e.g. \code{"log10"} or
\code{"identity"}.}

\item{xlim, ylim}{Numeric vectors of length two giving the axis limits. Use
\code{NA} to refer to the existing minimum or maximum.}

\item{y_ticks}{The approximate number of ticks desired on the y axis.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{legend_var}{Name of the variable used in the legend and to determine the
line colour.}

\item{size_axis}{Optional. If non-NULL, the x-axis is converted to weight
(\code{"w"}) or length (\code{"l"}).}
}
\value{
A \code{mizer_plot} (ggplot2) object.
}
\description{
Used internally by the comparison plotting functions such as \code{\link[=plotSpectra2]{plotSpectra2()}}
and \code{\link[=plotCDF2]{plotCDF2()}}. The two data frames are combined and drawn with colour
identifying the species or group and linetype identifying the object.
}
\keyword{internal}
