% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot biomass or abundance distribution}
\usage{
plotDistribution(params, sp)
}
\description{
Where \code{\link{plotSpectra}} plots the density function, this function 
plots the cummulative distribution function. Work in progress.
}
