% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFeedingLevel}
\alias{plotFeedingLevel}
\title{Plot the feeding level of species by size}
\usage{
plotFeedingLevel(sim, species = dimnames(sim@n)$sp,
  time_range = max(as.numeric(dimnames(sim@n)$time)), print_it = FALSE,
  ...)
}
\arguments{
\item{sim}{An object of class \linkS4class{MizerSim}.}

\item{species}{Name or vector of names of the species to be plotted. By
default all species are plotted.}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step.}

\item{print_it}{Display the plot, or just return the ggplot2 object.
Defaults to FALSE}

\item{...}{Other arguments to pass to \code{getFeedingLevel} method}
}
\value{
A ggplot2 object
}
\description{
After running a projection, plot the feeding level of each species by size. 
The feeding level is averaged over the specified time range (a single value
for the time range can be used).
}
\examples{

data(NS_species_params_gears)
data(inter)
params <- suppressMessages(set_multispecies_model(NS_species_params_gears, inter))
sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
plotFeedingLevel(sim)
plotFeedingLevel(sim, time_range = 10:20, species = c("Cod", "Herring"))

}
\seealso{
\code{\link{getFeedingLevel}}
}
