% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFMort}
\alias{plotFMort}
\alias{plotFMort.MizerSim}
\alias{plotFMort.MizerParams}
\alias{plotlyFMort}
\title{Plot total fishing mortality of each species by size}
\usage{
plotFMort(object, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{...}{.}

\item{species}{The species to be selected. Optional. By default all target
species are selected. A vector of species names, or a
numeric vector with the species indices, or a logical vector indicating for
each species whether it is to be selected (TRUE) or not.}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step. Ignored when called with a \linkS4class{MizerParams}
object.}

\item{all.sizes}{If TRUE, then fishing mortality is plotted also for sizes
outside a species' size range. Default FALSE.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the w axis. Use NA for the default: the lower default is
\code{min(params@w) / 100} when \code{resource = TRUE} (to show some resource below
the fish grid) or \code{min(params@w)} when \code{resource = FALSE}; the upper
default is \code{max(params@w_full)}. Data is filtered to this range and the
axis limits are set accordingly.}

\item{llim}{A numeric vector of length two providing lower and upper limits
for the length axis when \code{size_axis = "l"}. Use \code{NA} to auto-scale to the
data range. Data is filtered to this range and the axis limits are set
accordingly.}

\item{size_axis}{Whether to plot size as weight (\code{"w"}, default) or length
(\code{"l"}), using the allometric weight-length relationship.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is FALSE}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the three variables 'w' (or 'l' if \code{size_axis = "l"}), 'value',
'Species' is returned.
}
\description{
After running a projection, plot the total fishing mortality of each species
by size. The total fishing mortality is averaged over the specified time
range (a single value for the time range can be used to plot a single time
step).
}
\examples{
\donttest{
params <-  NS_params
sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
plotFMort(sim)
plotFMort(sim, highlight = c("Cod", "Haddock"))

# Returning the data frame
fr <- plotFMort(sim, return_data = TRUE)
str(fr)
}
}
\seealso{
\link{plotting_functions}, \code{\link[=getFMort]{getFMort()}}

Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
