% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plotGrowthCurves}
\alias{plotGrowthCurves}
\alias{plotGrowthCurves,MizerSim-method}
\title{Plot growth curves giving weight as a function of age}
\usage{
plotGrowthCurves(sim, ...)

\S4method{plotGrowthCurves}{MizerSim}(sim,
  species = as.character(sim@params@species_params$species), max_age = 20,
  percentage = FALSE, print_it = TRUE, ...)
}
\arguments{
\item{sim}{MizerSim object}

\item{species}{Vector of names of the species to be plotted. The default is
the list of all species.}

\item{max_age}{The age up to which the weight is to be plotted. Default is 20}

\item{percentage}{Boolean value. If TRUE, the size is shown as a percentage
of the maximal size.}

\item{print_it}{Display the plot, or just return the ggplot2 object.
Defaults to TRUE}
}
\value{
A ggplot2 object
}
\description{
Uses the growth rates at the final time of a simulation to calculate
the size at age.

Plot growth curves using a \code{MizerSim} object.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
sim <- project(params, effort=1, t_max=20, t_save = 2)
plotGrowthCurves(sim, percentage = TRUE)
plotGrowthCurves(sim, species = "Cod", max_age = 24)
}
}
