% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGrowthCurves}
\alias{plotGrowthCurves}
\alias{plotGrowthCurves.MizerSim}
\alias{plotGrowthCurves.MizerParams}
\alias{plotlyGrowthCurves}
\title{Plot growth curves}
\usage{
plotGrowthCurves(
  object,
  species = NULL,
  max_age = 20,
  percentage = FALSE,
  species_panel = FALSE,
  highlight = NULL,
  size_at_age = NULL,
  return_data = FALSE,
  log_x = FALSE,
  log_y = FALSE,
  log = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{species}{The species to be selected. Optional. By default all target
species are selected. A vector of species names, or a numeric vector with
the species indices, or a logical vector indicating for each species
whether it is to be selected (TRUE) or not.}

\item{max_age}{The age up to which to run the growth curve. Default is 20.}

\item{percentage}{Boolean value. If TRUE, the size is given as a percentage
of the maximal size.}

\item{species_panel}{If TRUE (default), and \code{percentage = FALSE}, display all
species as facets. Otherwise puts all species into a single panel.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{size_at_age}{A data frame with observed size at age data to be plotted
on top of growth curve graphs. Should contain columns \code{species} (species
name as used in the model), \code{age} (in years) and either \code{weight} (in grams)
or \code{length} (in cm). If both \code{weight} and \code{length} are provided, only
\code{weight} is used.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default is FALSE.}

\item{log_x}{If \code{TRUE}, use a log10 x-axis. Default is \code{FALSE}.}

\item{log_y}{If \code{TRUE}, use a log10 y-axis. Default is \code{FALSE}.}

\item{log}{Character string specifying which axes should use log10 scales,
in the same form as the base \code{\link[=plot]{plot()}} argument. For example, \code{"x"},
\code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and \code{log_y}.}

\item{...}{Unused.}
}
\value{
A ggplot2 object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The growth curves represent the average age of all the living fish of a
species as a function of their size. So it would be natural to plot size
on the x-axis. But to follow the usual convention from age-based models, we
plot size on the y-axis and age on the x-axis.
}
\details{
In each panel for a single species, a horizontal line is included that
indicate the maturity size of the species and a vertical line indicating its
maturity age.

If size at age data is passed via the \code{size_at_age} argument, this is plotted
on top of the growth curve. When comparing this to the growth curves, you
need to remember that the growth curves should only represent the average
age at each size. So a scatter in the x-direction around the curve is to be
expected.

If the species parameters contain the variables \code{a} and \code{b} for length to
weight conversion and the von Bertalanffy parameter \code{k_vb}, \code{w_inf} (and
optionally \code{t0}), then the von Bertalanffy growth curve is superimposed in
black. Note that the von Bertalanffy curve (which approximates the average
length at each age) should not be compared directly to the mizer growth
curves (which approximate the average age at each length).
}
\examples{
\donttest{
params <-  NS_params
sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
plotGrowthCurves(sim, percentage = TRUE)
plotGrowthCurves(sim, species = "Cod", max_age = 24)
plotGrowthCurves(sim, species_panel = TRUE)

# Returning the data frame
fr <- plotGrowthCurves(sim, return_data = TRUE)
str(fr)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate]{animate()}},
\code{\link{plot}},
\code{\link[=plot2]{plot2()}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotRelative]{plotRelative()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectra2]{plotSpectra2()}},
\code{\link[=plotSpectraRelative]{plotSpectraRelative()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link[=plotYieldVsF]{plotYieldVsF()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
