% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R,
%   R/ArrayTimeBySpecies-class.R, R/ArrayTimeBySpeciesBySize-class.R, R/plots.R
\name{plotHover.ArraySpeciesBySize}
\alias{plotHover.ArraySpeciesBySize}
\alias{plotHover.ArrayTimeBySpecies}
\alias{plotHover.ArrayTimeBySpeciesBySize}
\alias{plotHover}
\alias{plotHover.mizer_plot}
\title{Create a hover-enabled plotly plot from a mizer object}
\usage{
\method{plotHover}{ArraySpeciesBySize}(
  x,
  width = NULL,
  height = NULL,
  tooltip = "all",
  dynamicTicks = FALSE,
  layerData = 1,
  originalData = TRUE,
  source = "A",
  ...
)

\method{plotHover}{ArrayTimeBySpecies}(
  x,
  width = NULL,
  height = NULL,
  tooltip = "all",
  dynamicTicks = FALSE,
  layerData = 1,
  originalData = TRUE,
  source = "A",
  ...
)

\method{plotHover}{ArrayTimeBySpeciesBySize}(
  x,
  width = NULL,
  height = NULL,
  tooltip = "all",
  dynamicTicks = FALSE,
  layerData = 1,
  originalData = TRUE,
  source = "A",
  ...
)

plotHover(x, ...)

\method{plotHover}{mizer_plot}(
  x = ggplot2::last_plot(),
  ...,
  tooltip = attr(x, "mizer_tooltip") \%||\% "all"
)
}
\arguments{
\item{x}{A \code{mizer_plot}, \code{ArraySpeciesBySize}, \code{ArrayTimeBySpecies}, or
\code{ArrayTimeBySpeciesBySize} object.}

\item{width, height, tooltip, dynamicTicks, layerData, originalData, source}{Arguments passed to \code{\link[plotly:ggplotly]{plotly::ggplotly()}}.}

\item{...}{Arguments passed to the corresponding \code{plot()} method for mizer
array objects, or to \code{\link[plotly:ggplotly]{plotly::ggplotly()}} for \code{mizer_plot} objects.}
}
\value{
A plotly object.
}
\description{
Creates an interactive plotly version of a mizer plot. Can be called on any
mizer array object (such as those returned by \code{\link[=getEncounter]{getEncounter()}},
\code{\link[=getBiomass]{getBiomass()}}, etc.) or on any \code{mizer_plot} object returned by the named
plot functions such as \code{\link[=plotBiomass]{plotBiomass()}}, \code{\link[=plotSpectra]{plotSpectra()}}, etc. In the first
case it is a shorthand for \code{plotly::ggplotly(plot(x, ...))}.
}
\examples{
\donttest{
plotHover(getEncounter(NS_params))
}
\donttest{
plotHover(getBiomass(NS_sim))
}
\donttest{
plotHover(getFMort(NS_sim))
}
}
\seealso{
\code{\link[=plot]{plot()}}, \code{\link[=plotBiomass]{plotBiomass()}}, \code{\link[=plotSpectra]{plotSpectra()}}, \link{plotting_functions}
}
\concept{plotting functions}
