% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotlyGrowthCurves}
\alias{plotlyGrowthCurves}
\title{Plot growth curves giving weight as a function of age with plotly}
\usage{
plotlyGrowthCurves(object, species, max_age = 20, percentage = FALSE,
  highlight = NULL)
}
\arguments{
\item{object}{MizerSim or MizerParams object}

\item{species}{Name or vector of names of the species to be included. By
default all species are included.}

\item{max_age}{The age up to which to run the growth curve. Default is 20.}

\item{percentage}{Boolean value. If TRUE, the size is given as a percentage
of the maximal size.}

\item{highlight}{Name or vector of names of the species to be highlighted.}
}
\value{
A ggplot2 object
}
\description{
If given a \linkS4class{MizerSim} object, uses the growth rates at the final
time of a simulation to calculate the size at age. If given a
\linkS4class{MizerParams} object, uses the initial growth rates instead.
}
\details{
When the growth curve for only a single species is plotted, horizontal
lines are included that indicate the maturity size and the maximum size for 
that species. If furthermore the species parameters contain the variables
a and b for length to weight conversion and the von Bertalanffy parameter
k_vb, then the von Bertalanffy growth curve is superimposed in black.
}
\seealso{
Other plotting functions: \code{\link{displayFrames}},
  \code{\link{plot,MizerSim,missing-method}},
  \code{\link{plotBiomass}}, \code{\link{plotDiet}},
  \code{\link{plotFMort}}, \code{\link{plotFeedingLevel}},
  \code{\link{plotGrowthCurves}},
  \code{\link{plotPredMort}}, \code{\link{plotSpectra}},
  \code{\link{plotYieldGear}}, \code{\link{plotYield}},
  \code{\link{plotlyBiomass}}, \code{\link{plotlyFMort}},
  \code{\link{plotlyFeedingLevel}},
  \code{\link{plotlyPredMort}},
  \code{\link{plotlySpectra}},
  \code{\link{plotlyYieldGear}}, \code{\link{plotlyYield}},
  \code{\link{plotting_functions}}
}
\concept{plotting functions}
