% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMizerParams}
\alias{plotMizerParams}
\alias{plot.MizerParams}
\title{Summary plot for \code{MizerParams} objects}
\usage{
\method{plot}{MizerParams}(x, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{MizerParams}}

\item{...}{Arguments passed on to the individual plotting functions
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}}, \code{\link[=plotSpectra]{plotSpectra()}} and \code{\link[=plotPredMort]{plotPredMort()}}. The arguments
that describe the plotted spectrum (\code{power}, \code{biomass}, \code{per_log_size} and
\code{resource}) go to \code{\link[=plotSpectra]{plotSpectra()}} only, because they mean nothing to a
rate.}
}
\value{
A viewport object
}
\description{
Produces 3 plots in the same window: abundance spectra, feeding
level and predation mortality of each species against size. This method just
puts the plots generated by \code{\link[=plotFeedingLevel]{plotFeedingLevel()}}, \code{\link[=plotPredMort]{plotPredMort()}} and
\code{\link[=plotSpectra]{plotSpectra()}} all in one window.
}
\examples{
\donttest{
params <-  NS_params
plot(params)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate]{animate()}},
\code{\link{plot}},
\code{\link[=plot2]{plot2()}},
\code{\link[=plotBifurcation]{plotBifurcation()}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotRelative]{plotRelative()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectra2]{plotSpectra2()}},
\code{\link[=plotSpectraRelative]{plotSpectraRelative()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
