% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMizerParams}
\alias{plotMizerParams}
\alias{plot.MizerParams}
\title{Summary plot for \code{MizerParams} objects}
\usage{
\method{plot}{MizerParams}(x, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{MizerParams}}

\item{...}{Arguments passed on to the individual plotting functions
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},\code{\link[=plotSpectra]{plotSpectra()}},\code{\link[=plotPredMort]{plotPredMort()}}}
}
\value{
A viewport object
}
\description{
Produces 3 plots in the same window: abundance spectra, feeding
level and predation mortality of each species against size. This method just
puts the plots generated by \code{\link[=plotFeedingLevel]{plotFeedingLevel()}}, \code{\link[=plotPredMort]{plotPredMort()}} and
\code{\link[=plotSpectra]{plotSpectra()}} all in one window.
}
\examples{
\donttest{
params <-  NS_params
plot(params)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions: 
\code{\link{addPlot}()},
\code{\link{animate.ArrayTimeBySpeciesBySize}()},
\code{\link{plot}()},
\code{\link{plot2}()},
\code{\link{plotBiomass}()},
\code{\link{plotCDF}()},
\code{\link{plotCDF2}()},
\code{\link{plotDiet}()},
\code{\link{plotFMort}()},
\code{\link{plotFeedingLevel}()},
\code{\link{plotGrowthCurves}()},
\code{\link{plotMizerSim}},
\code{\link{plotPredMort}()},
\code{\link{plotRelative}()},
\code{\link{plotSpectra}()},
\code{\link{plotSpectra2}()},
\code{\link{plotSpectraRelative}()},
\code{\link{plotYield}()},
\code{\link{plotYieldGear}()},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
