% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMizerSim}
\alias{plotMizerSim}
\alias{plot.MizerSim}
\title{Summary plot for \code{MizerSim} objects}
\usage{
\method{plot}{MizerSim}(x, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{MizerSim}}

\item{...}{Arguments passed on to the individual plotting functions
\code{\link[=plotBiomass]{plotBiomass()}}, \code{\link[=plotFeedingLevel]{plotFeedingLevel()}}, \code{\link[=plotSpectra]{plotSpectra()}}, \code{\link[=plotPredMort]{plotPredMort()}}
and \code{\link[=plotFMort]{plotFMort()}}. The arguments that describe the plotted spectrum
(\code{power}, \code{biomass}, \code{per_log_size} and \code{resource}) go to \code{\link[=plotSpectra]{plotSpectra()}}
only, because they mean nothing to a rate.}
}
\value{
A viewport object
}
\description{
After running a projection, produces 5 plots in the same window: feeding
level, abundance spectra, predation mortality and fishing mortality of each
species by size; and biomass of each species through time. This method just
puts the plots generated by \code{\link[=plotBiomass]{plotBiomass()}}, \code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotSpectra]{plotSpectra()}}, \code{\link[=plotPredMort]{plotPredMort()}} and \code{\link[=plotFMort]{plotFMort()}} all in one window.
}
\examples{
\donttest{
params <-  NS_params
sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
plot(sim)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate]{animate()}},
\code{\link{plot}},
\code{\link[=plot2]{plot2()}},
\code{\link[=plotBifurcation]{plotBifurcation()}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotRelative]{plotRelative()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectra2]{plotSpectra2()}},
\code{\link[=plotSpectraRelative]{plotSpectraRelative()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
