% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMizerSim}
\alias{plotMizerSim}
\alias{plot.MizerSim}
\title{Summary plot for \code{MizerSim} objects}
\usage{
\method{plot}{MizerSim}(x, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{MizerSim}}

\item{...}{Arguments passed on to the individual plotting functions
\code{\link[=plotBiomass]{plotBiomass()}}, \code{\link[=plotFeedingLevel]{plotFeedingLevel()}}, \code{\link[=plotSpectra]{plotSpectra()}}, \code{\link[=plotPredMort]{plotPredMort()}}
and \code{\link[=plotFMort]{plotFMort()}}.}
}
\value{
A viewport object
}
\description{
After running a projection, produces 5 plots in the same window: feeding
level, abundance spectra, predation mortality and fishing mortality of each
species by size; and biomass of each species through time. This method just
puts the plots generated by \code{\link[=plotBiomass]{plotBiomass()}}, \code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotSpectra]{plotSpectra()}}, \code{\link[=plotPredMort]{plotPredMort()}} and \code{\link[=plotFMort]{plotFMort()}} all in one window.
}
\examples{
\donttest{
params <-  NS_params
sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
plot(sim)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions: 
\code{\link{addPlot}()},
\code{\link{animate.ArrayTimeBySpeciesBySize}()},
\code{\link{plot}()},
\code{\link{plot2}()},
\code{\link{plotBiomass}()},
\code{\link{plotCDF}()},
\code{\link{plotCDF2}()},
\code{\link{plotDiet}()},
\code{\link{plotFMort}()},
\code{\link{plotFeedingLevel}()},
\code{\link{plotGrowthCurves}()},
\code{\link{plotMizerParams}},
\code{\link{plotPredMort}()},
\code{\link{plotRelative}()},
\code{\link{plotSpectra}()},
\code{\link{plotSpectra2}()},
\code{\link{plotSpectraRelative}()},
\code{\link{plotYield}()},
\code{\link{plotYieldGear}()},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
