% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySpeciesBySize-class.R,
%   R/ArrayTimeBySpecies-class.R, R/ArrayTimeBySpeciesBySize-class.R
\name{plotRelative}
\alias{plotRelative}
\alias{plotRelative.ArraySpeciesBySize}
\alias{plotRelative.ArrayTimeBySpecies}
\alias{plotRelative.ArrayTimeBySpeciesBySize}
\title{Plot the relative difference between two mizer array objects}
\usage{
plotRelative(x, y, ...)

\method{plotRelative}{ArraySpeciesBySize}(
  x,
  y,
  species = NULL,
  all.sizes = FALSE,
  log_x = TRUE,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  size_axis = c("w", "l"),
  total = FALSE,
  background = TRUE,
  ...
)

\method{plotRelative}{ArrayTimeBySpecies}(
  x,
  y,
  species = NULL,
  start_time = NULL,
  end_time = NULL,
  ylim = c(NA, NA),
  total = FALSE,
  background = TRUE,
  log_x = FALSE,
  ...
)

\method{plotRelative}{ArrayTimeBySpeciesBySize}(
  x,
  y,
  species = NULL,
  time = NULL,
  all.sizes = FALSE,
  log_x = TRUE,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  size_axis = c("w", "l"),
  total = FALSE,
  background = TRUE,
  ...
)
}
\arguments{
\item{x, y}{Two compatible mizer array objects of the same class.}

\item{...}{Further arguments (currently unused).}

\item{species}{Character vector of species to include. \code{NULL} (default) means
all species.}

\item{all.sizes}{If \code{FALSE} (default), values outside a species' size range
(\code{w_min} to \code{w_max}) are removed. Only applies to \code{ArraySpeciesBySize}.}

\item{log_x}{If \code{TRUE}, use a log10 x-axis. Default is \code{TRUE} for size
spectra and \code{FALSE} for time series.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the weight (x) axis. Use \code{NA} to refer to the existing minimum or
maximum. Only applies to \code{ArraySpeciesBySize}.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the value (y) axis. Use \code{NA} to refer to the existing minimum or
maximum.}

\item{size_axis}{Whether to plot size as weight (\code{"w"}, default) or length
(\code{"l"}), using the allometric weight-length relationship.}

\item{total}{A boolean value that determines whether the total over all
selected species is plotted as well. Default is \code{FALSE}.}

\item{background}{A boolean value that determines whether background species
are included. Ignored if the model does not contain background species.
Default is \code{TRUE}.}

\item{start_time}{The first time to be plotted. Default (\code{NULL}) is the
beginning of the time series. Only applies to \code{ArrayTimeBySpecies}.}

\item{end_time}{The last time to be plotted. Default (\code{NULL}) is the end of
the time series. Only applies to \code{ArrayTimeBySpecies}.}

\item{time}{The time to display. Default (\code{NULL}) is the final time step.
Only applies to \code{ArrayTimeBySpeciesBySize}.}
}
\value{
A ggplot2 object.
}
\description{
\code{plotRelative()} plots the difference between two compatible mizer array
objects relative to their average. If the values in the first object are
\eqn{N_1} and the values in the second are \eqn{N_2}, it plots
\deqn{2 (N_2 - N_1) / (N_1 + N_2).}
}
\examples{
\donttest{
enc <- getEncounter(NS_params)
plotRelative(enc, enc, species = "Cod")
plotRelative(getBiomass(NS_sim), getBiomass(NS_sim), species = "Cod")
}
}
\seealso{
Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plot2]{plot2()}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectra2]{plotSpectra2()}},
\code{\link[=plotSpectraRelative]{plotSpectraRelative()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
