% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSpectra}
\alias{plotSpectra}
\alias{plotSpectra.MizerSim}
\alias{plotSpectra.MizerParams}
\alias{plotSpectra2}
\alias{plotlySpectra2}
\alias{plotSpectraRelative}
\alias{plotlySpectraRelative}
\alias{plotlySpectra}
\title{Plot abundance and biomass spectra and compare between objects}
\usage{
plotSpectra(object, ...)

plotSpectra2(
  object1,
  object2,
  name1 = "First",
  name2 = "Second",
  species = NULL,
  wlim = c(NA, NA),
  llim = c(NA, NA),
  ylim = c(NA, NA),
  power = 1,
  total = FALSE,
  resource = TRUE,
  background = TRUE,
  log_x = TRUE,
  log_y = TRUE,
  log = NULL,
  size_axis = c("w", "l"),
  ...
)

plotSpectraRelative(
  object1,
  object2,
  species = NULL,
  wlim = c(NA, NA),
  llim = c(NA, NA),
  ylim = c(NA, NA),
  power = 1,
  total = FALSE,
  resource = TRUE,
  background = TRUE,
  log_x = TRUE,
  size_axis = c("w", "l"),
  ...
)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{...}{For \code{plotSpectra2()} and \code{plotSpectraRelative()}, additional
arguments passed to \code{\link[=plotSpectra]{plotSpectra()}} for preparing the spectra data, for
example \code{time_range} or \code{geometric_mean} for \code{MizerSim} objects.
Otherwise unused.}

\item{species}{The species to be selected. Optional. By default all target
species are selected. A vector of species names, or a
numeric vector with the species indices, or a logical vector indicating for
each species whether it is to be selected (TRUE) or not.}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step. Ignored when called with a \linkS4class{MizerParams}
object.}

\item{geometric_mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
If TRUE then the average of the abundances over the
time range is a geometric mean instead of the default arithmetic mean.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the w axis. Use NA for the default: the lower default is
\code{min(params@w) / 100} when \code{resource = TRUE} (to show some resource below
the fish grid) or \code{min(params@w)} when \code{resource = FALSE}; the upper
default is \code{max(params@w_full)}. Data is filtered to this range and the
axis limits are set accordingly.}

\item{llim}{A numeric vector of length two providing lower and upper limits
for the length axis when \code{size_axis = "l"}. Use \code{NA} to auto-scale to the
data range. Data is filtered to this range and the axis limits are set
accordingly.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the y axis. Use NA to auto-scale to the data range. Values below 1e-20
are always filtered out from the data regardless of \code{ylim[1]}. Data above
\code{ylim[2]} is filtered and the upper axis limit is set accordingly. For
\code{plotSpectraRelative()}, the y axis is the relative difference and \code{NA}
refers to the existing minimum or maximum.}

\item{power}{The abundance is plotted as the number density times the weight
raised to \code{power}. The default \code{power = 1} gives the biomass
density, whereas \code{power = 2} gives the biomass density with respect
to logarithmic size bins.}

\item{biomass}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Only used if \code{power} argument is missing. Then
\code{biomass = TRUE} is equivalent to \code{power=1} and
\code{biomass = FALSE} is equivalent to \code{power=0}}

\item{total}{A boolean value that determines whether the total over all
species in the system is plotted as well. Note that even if the plot
only shows a selection of species, the total is including all species.
Default is FALSE.}

\item{resource}{A boolean value that determines whether resource is included.
Default is TRUE.}

\item{background}{A boolean value that determines whether background species
are included. Ignored if the model does not contain background species.
Default is TRUE.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{log_x}{If \code{TRUE} (default), use a log10 x-axis.}

\item{log_y}{If \code{TRUE} (default), use a log10 y-axis. Not used by
\code{plotSpectraRelative()}.}

\item{log}{Character string specifying which axes should use log10 scales,
in the same form as the base \code{\link[=plot]{plot()}} argument. For example, \code{"x"},
\code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and \code{log_y}.}

\item{size_axis}{Whether to plot size as weight (\code{"w"}, default) or length
(\code{"l"}), using the allometric weight-length relationship.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is FALSE}

\item{object1}{First \code{MizerParams} or \code{MizerSim} object.}

\item{object2}{Second \code{MizerParams} or \code{MizerSim} object.}

\item{name1, name2}{Labels for the two objects, used in the linetype legend.}
}
\value{
\code{plotSpectra()} returns a ggplot2 object, unless \code{return_data = TRUE}, in which case a data frame with the four variables 'w' (or 'l' if
\code{size_axis = "l"}), 'value', 'Species', 'Legend' is returned.

\code{plotSpectra2()} and \code{plotSpectraRelative()} return a ggplot2 object.

\code{plotlySpectra()}, \code{plotlySpectra2()} and \code{plotlySpectraRelative()} return
plotly objects.
}
\description{
\code{plotSpectra()} plots the number density multiplied by a power of the
weight, with the power specified by the \code{power} argument. When called with a
\link{MizerSim} object, the abundance is averaged over the specified time range
(a single value for the time range can be used to plot a single time step).
When called with a \link{MizerParams} object the initial abundance is plotted.
}
\details{
\code{plotSpectra2()} compares the abundance spectra from two \code{MizerParams} or
\code{MizerSim} objects in a single plot. Colours identify species or groups and
linetype identifies the object.

\code{plotSpectraRelative()} plots the difference between the spectra relative to
their average. If we denote the number density from the first object as
\eqn{N_1(w)} and that from the second object as \eqn{N_2(w)}, then this
plot shows
\deqn{2 (N_2(w) - N_1(w)) / (N_2(w) + N_1(w)).}
Note that it does not matter whether the relative difference is calculated
for number density, biomass density, or biomass density in log weight,
because the factors of \eqn{w} by which the densities differ cancel out in
the relative difference.

\code{plotlySpectra()}, \code{plotlySpectra2()} and \code{plotlySpectraRelative()} are the
interactive plotly versions.
}
\examples{
\donttest{
params <-  NS_params
sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
plotSpectra(sim)
plotSpectra(sim, wlim = c(1e-6, NA))
plotSpectra(sim, time_range = 10:20)
plotSpectra(sim, time_range = 10:20, power = 0)
plotSpectra(sim, species = c("Cod", "Herring"), power = 1)
plotSpectra(sim, species = c("Cod", "Herring"), size_axis = "l")

# Returning the data frame
fr <- plotSpectra(sim, return_data = TRUE)
str(fr)

sim1 <- project(NS_params, t_max = 10, progress_bar = FALSE)
sim2 <- project(NS_params, effort = 0.5, t_max = 10, progress_bar = FALSE)
plotSpectra2(sim1, sim2, "Original", "Effort = 0.5")
plotSpectraRelative(sim1, sim2)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
