% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plotSpectra}
\alias{plotSpectra}
\alias{plotSpectra,MizerSim-method}
\title{Plot the abundance spectra of fish species and plankton}
\usage{
plotSpectra(sim, ...)

\S4method{plotSpectra}{MizerSim}(sim,
  species = as.character(sim@params@species_params$species),
  time_range = max(as.numeric(dimnames(sim@n)$time)),
  min_w = min(sim@params@w)/100, type = 1, biomass = TRUE,
  print_it = TRUE, total = FALSE, ...)
}
\arguments{
\item{sim}{An object of class \code{MizerSim}.}

\item{...}{Other arguments (currently unused)}

\item{species}{Name or vector of names of the species to be plotted. By
default all species are plotted.}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step.}

\item{min_w}{Minimum weight to be plotted (useful for truncating the
background spectrum). Default value is a hundredth of the minimum size
value of the community.}

\item{type}{Determines whether the abundance is represented in terms of
number density (type = 0) or biomass density (type = 1) or biomass in
logarithmically-spaced size bins (type = 2)}

\item{biomass}{Obsolete. Only used if \code{type} argument is missing. Then
\code{biomass = TRUE} is equivalent to \code{type=1} and 
\code{biomass = FALSE} is equivalent to \code{type=0}}

\item{print_it}{Display the plot, or just return the ggplot2 object.
Defaults to TRUE}

\item{total}{A boolean value that determines whether the total over all
species in the system is plotted as well. Default is FALSE}
}
\value{
A ggplot2 object
}
\description{
After running a projection, the spectra of the abundance of each species and
the plankton can be plotted. The abundance is averaged over the specified
time range (a single value for the time range can be used to plot a single
time step). Depending on the \code{type} argument, the abundance can be in
terms of number density (type = 0) or biomass density (type = 1) or biomass
in logarithmically-spaced size bins (type = 2).

Plot the abundance spectra using a \code{MizerSim} object.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
sim <- project(params, effort=1, t_max=20, t_save = 2)
plotSpectra(sim)
plotSpectra(sim, min_w = 1e-6)
plotSpectra(sim, time_range = 10:20)
plotSpectra(sim, time_range = 10:20, type = 0)
plotSpectra(sim, species = c("Cod", "Herring"), type = 1)
}
}
