% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSpectra2}
\alias{plotSpectra2}
\alias{plotlySpectra2}
\title{Compare two size spectra in the same plot}
\usage{
plotSpectra2(
  object1,
  object2,
  name1 = "First",
  name2 = "Second",
  power = 1,
  log_x = TRUE,
  log_y = TRUE,
  log = NULL,
  llim = c(NA, NA),
  size_axis = c("w", "l"),
  ...
)

plotlySpectra2(
  object1,
  object2,
  name1 = "First",
  name2 = "Second",
  power = 1,
  log_x = TRUE,
  log_y = TRUE,
  log = NULL,
  llim = c(NA, NA),
  size_axis = c("w", "l"),
  ...
)
}
\arguments{
\item{object1}{First \code{MizerParams} or \code{MizerSim} object.}

\item{object2}{Second \code{MizerParams} or \code{MizerSim} object.}

\item{name1, name2}{Labels for the two objects, used in the linetype legend.}

\item{power}{The abundance is plotted as the number density times the weight
raised to \code{power}. The default \code{power = 1} gives the biomass
density, whereas \code{power = 2} gives the biomass density with respect
to logarithmic size bins.}

\item{log_x}{If \code{TRUE} (default), use a log10 x-axis.}

\item{log_y}{If \code{TRUE} (default), use a log10 y-axis.}

\item{log}{Character string specifying which axes should use log10 scales,
in the same form as the base \code{\link[=plot]{plot()}} argument. For example, \code{"x"},
\code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and \code{log_y}.}

\item{llim}{A numeric vector of length two providing lower and upper limits
for the length axis when \code{size_axis = "l"}. Use \code{NA} to auto-scale to the
data range. Data is filtered to this range and the axis limits are set
accordingly.}

\item{size_axis}{Whether to plot size as weight (\code{"w"}, default) or length
(\code{"l"}), using the allometric weight-length relationship.}

\item{...}{Arguments passed to \code{\link[=plotSpectra]{plotSpectra()}} for preparing the spectra
data, for example \code{species}, \code{time_range}, \code{wlim}, \code{ylim}, \code{resource},
\code{background} or \code{total}.}
}
\value{
A ggplot2 object.

\code{plotlySpectra2()} returns a plotly object.
}
\description{
\code{plotSpectra2()} compares the abundance spectra from two \code{MizerParams} or
\code{MizerSim} objects in a single plot. Colours identify species or groups and
linetype identifies the object.
}
\examples{
\donttest{
sim1 <- project(NS_params, t_max = 10, progress_bar = FALSE)
sim2 <- project(NS_params, effort = 0.5, t_max = 10, progress_bar = FALSE)
plotSpectra2(sim1, sim2, "Original", "Effort = 0.5")
}
}
\seealso{
Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plot2]{plot2()}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotRelative]{plotRelative()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectraRelative]{plotSpectraRelative()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
