% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSpectraRelative}
\alias{plotSpectraRelative}
\alias{plotlySpectraRelative}
\title{Plot the relative difference between two spectra}
\usage{
plotSpectraRelative(object1, object2, log_x = TRUE, ylim = c(NA, NA), ...)

plotlySpectraRelative(object1, object2, log_x = TRUE, ylim = c(NA, NA), ...)
}
\arguments{
\item{object1}{First \code{MizerParams} or \code{MizerSim} object.}

\item{object2}{Second \code{MizerParams} or \code{MizerSim} object.}

\item{log_x}{If \code{TRUE} (default), use a log10 x-axis.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the relative difference (y) axis. Use \code{NA} to refer to the existing
minimum or maximum.}

\item{...}{Arguments passed to \code{\link[=plotSpectra]{plotSpectra()}} for preparing the spectra
data, for example \code{species}, \code{time_range}, \code{wlim}, \code{resource},
\code{background} or \code{total}.}
}
\value{
A ggplot2 object.
}
\description{
\code{plotSpectraRelative()} plots the difference between the spectra relative to
their average. If we denote the number density from the first object as
\eqn{N_1(w)} and that from the second object as \eqn{N_2(w)}, then this plot
shows
\deqn{2 (N_2(w) - N_1(w)) / (N_2(w) + N_1(w)).}
}
\details{
The individual spectra are calculated by \code{\link[=plotSpectra]{plotSpectra()}}, to which all
additional arguments are passed. For example, you can determine a time range
over which to average simulation results via \code{time_range}. See
\code{\link[=plotSpectra]{plotSpectra()}} for more options.

Note that it does not matter whether the relative difference is calculated
for number density, biomass density, or biomass density in log weight,
because the factors of \eqn{w} by which the densities differ cancel out in
the relative difference.
}
\examples{
\donttest{
sim1 <- project(NS_params, t_max = 10, progress_bar = FALSE)
sim2 <- project(NS_params, effort = 0.5, t_max = 10, progress_bar = FALSE)
plotSpectraRelative(sim1, sim2)
}
}
\seealso{
Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plot2]{plot2()}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotCDF2]{plotCDF2()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotRelative]{plotRelative()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotSpectra2]{plotSpectra2()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
