% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotting_functions}
\alias{plotting_functions}
\title{Description of the plotting functions}
\description{
Mizer provides a range of plotting functions for visualising the results
of running a simulation, stored in a MizerSim object, or the initial state
stored in a MizerParams object.
}
\details{
The quickest way to make a standard plot is often to call \code{plot()} directly.
mizer provides \code{plot()} methods for \code{MizerSim} and \code{MizerParams} objects, and
also for the array classes returned by many summary and rate functions:
\itemize{
\item \verb{plot(<MizerSim>)} produces a five-panel summary plot with
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}}, \code{\link[=plotBiomass]{plotBiomass()}}, \code{\link[=plotPredMort]{plotPredMort()}}, \code{\link[=plotFMort]{plotFMort()}} and
\code{\link[=plotSpectra]{plotSpectra()}}.
\item \verb{plot(<MizerParams>)} produces a three-panel summary plot with
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}}, \code{\link[=plotPredMort]{plotPredMort()}} and \code{\link[=plotSpectra]{plotSpectra()}} from the initial
state.
\item \verb{plot(<ArrayTimeBySpecies>)} plots any time-by-species array, such as
those returned by \code{\link[=getBiomass]{getBiomass()}}, \code{\link[=getSSB]{getSSB()}}, \code{\link[=getYield]{getYield()}} and \code{\link[=getN]{getN()}} on
a \code{MizerSim}, as lines of value against time.
\item \verb{plot(<ArraySpeciesBySize>)} plots any species-by-size array, such as
those returned by \code{\link[=getEncounter]{getEncounter()}}, \code{\link[=getFeedingLevel]{getFeedingLevel()}}, \code{\link[=getPredMort]{getPredMort()}},
\code{\link[=getFMort]{getFMort()}} and \code{\link[=getSearchVolume]{getSearchVolume()}}, as lines of value against body size.
\item \verb{plot(<ArrayTimeBySpeciesBySize>)} plots a time slice from a
time-by-species-by-size array, such as those returned by \code{\link[=getFMort]{getFMort()}} or
\code{\link[=getPredMort]{getPredMort()}} on a \code{MizerSim}.
}

The same array objects can be passed to \code{\link[=ggplotly]{ggplotly()}} to produce interactive
versions, for example \code{ggplotly(getBiomass(sim))} or
\code{ggplotly(getEncounter(params))}. To add another compatible array to an
existing ggplot, use \code{\link[=addPlot]{addPlot()}}. This is useful for comparing two simulations
or two parameter sets on the same axes. To visualise how spectra or rates
change through time, use \code{\link[=animate]{animate()}} on a \code{MizerSim} or an
\code{ArrayTimeBySpeciesBySize} object.

The named plotting functions give more specialised control. This table shows
the available named plotting functions.
\tabular{ll}{
Plot \tab Description \cr
\code{\link[=plotBiomass]{plotBiomass()}} \tab Plots the total biomass of each species through time. A time range to be plotted can be specified. The size range of the community can be specified in the same way as for \code{\link[=getBiomass]{getBiomass()}}. \cr
\code{\link[=plotYield]{plotYield()}} \tab Plots the total yield of each species across all fishing gears against time. \cr
\code{\link[=plotYieldGear]{plotYieldGear()}} \tab Plots the total yield of each species by gear against time. \cr
\code{\link[=plotSpectra]{plotSpectra()}} \tab Plots the abundance (biomass or numbers) spectra of each species and the background community. It is possible to specify a minimum size which is useful for truncating the plot. \cr
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}} \tab Plots the feeding level of each species against size. \cr
\code{\link[=plotPredMort]{plotPredMort()}} \tab Plots the predation mortality of each species against size. \cr
\code{\link[=plotFMort]{plotFMort()}} \tab Plots the total fishing mortality of each species against size. \cr
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}} \tab Plots the size as a function of age. \cr
\code{\link[=plotDiet]{plotDiet()}} \tab Plots the diet composition at size for a given predator species. \cr
\code{\link[=plotBiomassObservedVsModel]{plotBiomassObservedVsModel()}} \tab Compares observed biomass with model biomass. \cr
\code{\link[=plotYieldObservedVsModel]{plotYieldObservedVsModel()}} \tab Compares observed yield with model yield. \cr
\code{\link[=animate]{animate()}} \tab Animates spectra or rate arrays through time. The older \code{\link[=animateSpectra]{animateSpectra()}} name is retained as an alias. \cr
}

The static plotting functions use ggplot2 and return a ggplot object. This
means that you can manipulate the plot further after its creation using the
ggplot grammar of graphics. Many named plot functions also have a plotly
counterpart, for example \code{\link[=plotlyBiomass]{plotlyBiomass()}} or \code{\link[=plotlySpectra]{plotlySpectra()}}, for
interactive exploration.

While most plot functions take their data from a MizerSim object, some of
those that make plots representing data at a single time can also take their
data from the initial values in a MizerParams object.

Where plots show results for species, the line colour and line type for each
species are specified by the \code{linecolour} and \code{linetype} slots in
the MizerParams object. These were either taken from a default palette
hard-coded into \code{\link[=emptyParams]{emptyParams()}} or they were specified by the user
in the species parameters dataframe used to set up the MizerParams object.
The \code{linecolour} and \code{linetype} slots hold named vectors, named by
the species. They can be overwritten by the user at any time.

Most plots allow the user to select to show only a subset of species,
specified as a vector in the \code{species} argument to the plot function.

The ordering of the species in the legend is the same as the ordering in
the species parameter data frame.
}
\examples{
\donttest{
sim <- NS_sim

# Generic plot methods
plot(sim)
plot(getBiomass(sim), species = c("Cod", "Herring"))
ggplotly(getBiomass(sim))

# Named plot functions
plotFeedingLevel(sim)

# Plotting only a subset of species
plotFeedingLevel(sim, species = c("Cod", "Herring"))

# Adding another compatible array to an existing plot
p <- plot(getBiomass(sim), species = "Cod")
addPlot(p, getBiomass(sim), species = "Herring", linetype = "dashed")

# Specifying new colours and linetypes for some species
sim@params@linetype["Cod"] <- "dashed"
sim@params@linecolour["Cod"] <- "red"
plotFeedingLevel(sim, species = c("Cod", "Herring"))

# Manipulating the plot
library(ggplot2)
p <- plotFeedingLevel(sim)
p <- p + geom_hline(aes(yintercept = 0.7))
p <- p + theme_bw()
p
}
}
\seealso{
\link{summary_functions}, \link{indicator_functions}

Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotYield]{plotYield()}},
\code{\link[=plotYieldGear]{plotYieldGear()}}
}
\concept{plotting functions}
