% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotting_functions}
\alias{plotting_functions}
\title{Plotting functions}
\description{
Mizer provides a range of plotting functions for visualising the results
of running a simulation, stored in a MizerSim object.
}
\details{
This table shows the available plotting functions.
\tabular{ll}{
  Plot \tab Description \cr
  \code{\link{plotBiomass}} \tab Plots the total biomass of each species through time. A time range to be plotted can be specified. The size range of the community can be specified in the same way as the method getBiomass(). \cr
  \code{\link{plotSpectra}} \tab Plots the abundance (biomass or numbers) spectra of each species and the background community. It is possible to specify a minimum size which is useful for truncating the plot. \cr
  \code{\link{plotFeedingLevel}} \tab Plots the feeding level of each species against size. \cr
  \code{\link{plotM2}} \tab Plots the predation mortality of each species against size. \cr
  \code{\link{plotFMort}} \tab Plots the total fishing mortality of each species against size. \cr
  \code{\link{plotYield}} \tab Plots the total yield of each species across all fishing gears against time. \cr
  \code{\link{plotYieldGear}} \tab Plots the total yield of each species by gear against time. \cr
  \code{\link{plot}} \tab Produces 5 plots (plotFeedingLevel(), plotBiomass(), plotM2(), plotFMort() and plotSpectra()) in the same window as a summary. \cr
}

These functions use the ggplot2 package and return the plot as a ggplot
object. This means that you can manipulate the plot further after its 
creation using the ggplot grammar of graphics.
}
\seealso{
\code{\link{summary_functions}}
}
