% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotYield}
\alias{plotYield}
\alias{plotYield.MizerSim}
\alias{plotlyYield}
\title{Plot the total yield of species through time}
\usage{
plotYield(object, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim}}

\item{...}{For \code{plotSpectra2()} and \code{plotSpectraRelative()}, additional
arguments passed to \code{\link[=plotSpectra]{plotSpectra()}} for preparing the spectra data, for
example \code{time_range} or \code{geometric_mean} for \code{MizerSim} objects.
Otherwise unused.}

\item{sim2}{An optional second object of class \linkS4class{MizerSim}. If
this is provided its yields will be shown on the same plot in bolder lines.}

\item{species}{The species to be selected. Optional. By default all target
species are selected. A vector of species names, or a
numeric vector with the species indices, or a logical vector indicating for
each species whether it is to be selected (TRUE) or not.}

\item{total}{A boolean value that determines whether the total over all
species in the system is plotted as well. Note that even if the plot
only shows a selection of species, the total is including all species.
Default is FALSE.}

\item{log_x}{If \code{TRUE}, use a log10 x-axis. Default is \code{FALSE}.}

\item{log_y}{If \code{TRUE}, use a log10 y-axis. Default is \code{TRUE}.}

\item{log}{Character string specifying which axes should use log10 scales,
in the same form as the base \code{\link[=plot]{plot()}} argument. For example, \code{"x"},
\code{"y"}, \code{"xy"} or \code{""}. If supplied, this overrides \code{log_x} and \code{log_y}.
For backward compatibility, \code{TRUE} and \code{FALSE} are interpreted as setting
only \code{log_y}.}

\item{ylim}{A numeric vector of length two providing lower and upper limits
for the y axis. Use NA to auto-scale to the data range. Values below 1e-20
are always filtered out from the data regardless of \code{ylim[1]}. Data above
\code{ylim[2]} is filtered and the upper axis limit is set accordingly. For
\code{plotSpectraRelative()}, the y axis is the relative difference and \code{NA}
refers to the existing minimum or maximum.}

\item{tlim}{A numeric vector of length two providing lower and upper limits
for the time axis, e.g. \code{c(1980, 2000)}. Use \code{NA} to apply no limit at
that end. Default is \code{c(NA, NA)}.}

\item{highlight}{Name or vector of names of the species to be highlighted.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is FALSE}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the three variables 'Year', 'Yield', 'Species' is returned.
}
\description{
After running a projection, the total yield of each species across all
fishing gears can be plotted against time. The yield is obtained with
\code{\link[=getYield]{getYield()}}.
}
\examples{
\donttest{
params <- NS_params
sim <- project(params, effort = 1, t_max = 20, t_save = 0.2, progress_bar = FALSE)
plotYield(sim)
plotYield(sim, species = c("Cod", "Herring"), total = TRUE)

# Comparing with yield from twice the effort
sim2 <- project(params, effort=2, t_max=20, t_save = 0.2, progress_bar = FALSE)
plotYield(sim, sim2, species = c("Cod", "Herring"), log = FALSE)

# Returning the data frame
fr <- plotYield(sim, return_data = TRUE)
str(fr)
}
}
\seealso{
\link{plotting_functions},  \code{\link[=getYield]{getYield()}}

Other plotting functions:
\code{\link[=addPlot]{addPlot()}},
\code{\link[=animate.ArrayTimeBySpeciesBySize]{animate.ArrayTimeBySpeciesBySize()}},
\code{\link{plot}},
\code{\link[=plotBiomass]{plotBiomass()}},
\code{\link[=plotCDF]{plotCDF()}},
\code{\link[=plotDiet]{plotDiet()}},
\code{\link[=plotFMort]{plotFMort()}},
\code{\link[=plotFeedingLevel]{plotFeedingLevel()}},
\code{\link[=plotGrowthCurves]{plotGrowthCurves()}},
\code{\link{plotMizerParams}},
\code{\link{plotMizerSim}},
\code{\link[=plotPredMort]{plotPredMort()}},
\code{\link[=plotSpectra]{plotSpectra()}},
\code{\link[=plotYieldGear]{plotYieldGear()}},
\code{\link{plotting_functions}}
}
\concept{plotting functions}
