% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_kernel_funcs.R
\name{power_law_pred_kernel}
\alias{power_law_pred_kernel}
\title{Power-law predation kernel}
\usage{
power_law_pred_kernel(ppmr, kernel_exp, kernel_l_l, kernel_u_l, kernel_l_r,
  kernel_u_r)
}
\arguments{
\item{ppmr}{A vector of predator/prey size ratios at which to evaluate the
predation kernel.}

\item{kernel_exp}{x}

\item{kernel_l_l}{x}

\item{kernel_u_l}{x}

\item{kernel_l_r}{x}

\item{kernel_u_r}{x}
}
\value{
A vector giving the value of the predation kernel at each of the
  predator/prey mass ratios in the \code{ppmr} argument.
}
\description{
A predation kernel where the predator/prey mass ratio follows a power-law,
with sigmoidal cut-offs at large and small ratios.
}
\details{
Writing the predator mass as \eqn{w} and the prey mass as \eqn{w_p}, the
feeding kernel is given by
\deqn{.} 
The parameters need to be given in the
species parameter dataframe in the columns \code{ppmr_min} and
\code{ppmr_max}.
}
