% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project size spectrum forward in time}
\usage{
project(params, sim = NULL, effort = 0, t_max = 100, dt = 0.1,
  t_save = 1, initial_n = params@initial_n,
  initial_n_pp = params@initial_n_pp, initial_B = params@initial_B,
  progress_bar = TRUE, ...)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{sim}{Optional \linkS4class{MizerSim} object. If supplied, then the
initial values for the abundances of species and of plankton and the
resource biomasses are taken from the final time of that simulation.}

\item{effort}{The effort of each fishing gear through time. See notes below.}

\item{t_max}{The maximum time the projection runs for. The default value is
100. However, this argument is not needed if an array is used for the
\code{effort} argument, in which case this argument is ignored. See notes
below.}

\item{dt}{Time step of the solver. The default value is 0.1.}

\item{t_save}{The frequency with which the output is stored. The default
value is 1. Must be an integer multiple of dt.}

\item{initial_n}{The initial abundances of species. Ignored if a \code{sim}
argument is supplied. A matrix with dimensions species x size. The order of
species must be the same as in the \code{MizerParams} argument. By default
is set to the \code{initial_n} slot of the \code{params} argument.}

\item{initial_n_pp}{The initial abundances of plankton. Ignored if a
\code{sim} argument is supplied. A numeric vector of the same length as the
\code{w_full} slot of the \code{MizerParams} argument. By default is set to
the \code{initial_n_pp} slot of the \code{params} argument.}

\item{initial_B}{The initial biomasses of the unstructured resources. It
should be a vector with one entry for each resource, with the names the
same as the names of the list in the \code{resource_dynamics} slot of the
\code{MizerParams}. By default is set to the \code{initial_B} slot of the
\code{params} argument.}

\item{progress_bar}{Either a boolean value to determine whether a progress
bar should be shown in the console of a shiny progress object to implement 
a progress bar in a shiny app}

\item{...}{Currently unused.}
}
\value{
An object of class \linkS4class{MizerSim}.
}
\description{
Runs the size spectrum model simulation.
The function returns an object of type
\linkS4class{MizerSim} that can then be explored with a range of
\code{\link{summary_functions}} and \code{\link{plotting_functions}}
plotting methods.
}
\note{
The \code{effort} argument specifies the level of fishing effort during
the simulation. It can be specified in three different ways: \itemize{ \item
A single numeric value. This specifies the effort of all fishing gears which
is constant through time (i.e. all the gears have the same constant effort). 
\item A numerical vector which has the same length as the number of fishing
gears. The vector must be named and the names must correspond to the gear
names in the \code{MizerParams} object. The values in the vector specify the
constant fishing effort of each of the fishing gears, i.e. the effort is
constant through time but each gear may have a different fishing effort. 
\item A numerical array with dimensions time step x gear. This specifies the
fishing effort of each gear at each time step.  The first dimension, time,
must be named numerically and contiguously. The second dimension of the array
must be named and the names must correspond to the gear names in the
\code{MizerParams} argument. The value for the effort for a particular time
is used during the interval from that time to the next time in the array.}

If effort is specified as an array then the smallest time in the array is 
used as the initial time for the simulation. Otherwise the initial time is
set to 0. Also, if the effort is an array then the \code{t_max} argument is 
ignored and the maximum simulation time is the largest time of the effort
array.

The \code{initial_n} argument is a matrix with dimensions species x size. 
It specifies the abundances of the species at the initial time. The
order of species must be the same as in the \code{MizerParams} argument. If
the initial population is not specified, the argument is set by default by
the \code{\link{get_initial_n}} function which is set up for a North Sea model.
}
\examples{
\dontrun{
# Data set with different fishing gears
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# With constant fishing effort which is different for each gear
effort <- c(Industrial = 0, Pelagic = 1, Beam = 0.5, Otter = 0.5)
sim <- project(params, t_max = 20, effort = effort)
# With fishing effort that varies through time for each gear
gear_names <- c("Industrial","Pelagic","Beam","Otter")
times <- seq(from = 1, to = 10, by = 1)
effort_array <- array(NA, dim = c(length(times), length(gear_names)),
    dimnames = list(time = times, gear = gear_names))
effort_array[,"Industrial"] <- 0.5
effort_array[,"Pelagic"] <- seq(from = 1, to = 2, length = length(times))
effort_array[,"Beam"] <- seq(from = 1, to = 0, length = length(times))
effort_array[,"Otter"] <- seq(from = 1, to = 0.5, length = length(times))
sim <- project(params, effort = effort_array)
}
}
\seealso{
\code{\link{MizerParams}}, \code{\link{summary_functions}} and 
  \code{\link{plotting_functions}}
}
