% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_n.R
\name{project_n}
\alias{project_n}
\title{Advance abundance density by one time step}
\usage{
project_n(params, r, n, dt, a, b, S, idx, w_min_idx_array_ref, no_sp, no_w)
}
\description{
Advance abundance density by one time step
}
