% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_simple}
\alias{project_simple}
\title{Project abundances by a given number of time steps into the future}
\usage{
project_simple(
  params,
  n,
  n_pp,
  n_other,
  t,
  dt,
  steps,
  effort,
  resource_dynamics_fn,
  other_dynamics_fns,
  rates_fns
)
}
\arguments{
\item{params}{A MizerParams object.}

\item{n}{Number density at start of simulation.}

\item{n_pp}{Resource number density at start of simulation.}

\item{n_other}{Abundances of other components at start of simulation.}

\item{t}{Time at the start of the simulation.}

\item{dt}{Size of time step.}

\item{steps}{The number of time steps by which to project.}

\item{effort}{The fishing effort to be used throughout the simulation.}

\item{resource_dynamics_fn}{The function (not its name) for the resource
dynamics.}

\item{other_dynamics_fns}{List with the functions (not their names) for the
dynamics of the other components.}

\item{rates_fns}{List with the functions (not their names) for calculating
the rates.}
}
\value{
List with the final values of \code{n}, \code{n_pp} and \code{n_other}.
}
\description{
This is an internal function used by the user-facing \code{project()} function.
Of interest only potentially to mizer extension authors.
}
