% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{removeBackgroundSpecies}
\alias{removeBackgroundSpecies}
\title{Remove all background species}
\usage{
removeBackgroundSpecies(params)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}
}
\value{
A \linkS4class{MizerParams} object with background species removed
}
\description{
Removes all species that have been marked as background species with
\code{\link[=markBackground]{markBackground()}}.
}
\seealso{
\code{\link[=markBackground]{markBackground()}}
}
