% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\docType{methods}
\name{removeSpecies}
\alias{removeSpecies}
\alias{removeSpecies,MizerParams-method}
\title{Remove species from an ecosystem}
\usage{
removeSpecies(params, species, ...)

\S4method{removeSpecies}{MizerParams}(params, species, ...)
}
\arguments{
\item{params}{A mizer params object for the original system.}

\item{species}{A vector of the names of the species 
to be deleted or a boolean vector indicating for each species 
whether it is to be removed (TRUE) or not.}
}
\value{
An object of type \linkS4class{MizerParams}
}
\description{
This method simply removes all entries from the MizerParams object that 
refer to the selected species. It does not recalculate the initial 
abundances.

Remove species from a \code{MizerParams} object.
}
