% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_species.R
\name{renameSpecies}
\alias{renameSpecies}
\alias{renameSpecies.MizerParams}
\alias{renameSpecies.default}
\title{Rename species}
\usage{
renameSpecies(params, ...)

\method{renameSpecies}{MizerParams}(params, replace)

\method{renameSpecies}{default}(params, ...)
}
\arguments{
\item{params}{A mizer params object}

\item{replace}{A named character vector, with new names as values, and old
names as names.}
}
\value{
An object of type \linkS4class{MizerParams}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Changes the names of species in a MizerParams object. This involves for
example changing the species dimension names of rate arrays appropriately.
}
\examples{
replace <- c(Cod = "Kabeljau", Haddock = "Schellfisch")
params <- renameSpecies(NS_params, replace)
species_params(params)$species
}
