% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setResource.R
\name{resource_params}
\alias{resource_params}
\alias{resource_params<-}
\title{Resource parameters}
\usage{
resource_params(params)

resource_params(params) <- value
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A named list of resource parameters.}
}
\description{
These functions allow you to get or set the resource parameters stored in
a MizerParams object.
The resource parameters are stored as a named list with the slot names
\code{r_pp}, \code{kappa}, \code{lambda}, \code{n}, \code{w_pp_cutoff}. For their meaning see
\code{\link[=setResource]{setResource()}}. If you change these parameters then this will
recalculate the resource rate and the resource capacity, unless you have
protected these with comments.
}
\seealso{
Other functions for setting parameters: 
\code{\link{gear_params}()},
\code{\link{setExtMort}()},
\code{\link{setFishing}()},
\code{\link{setInitialValues}()},
\code{\link{setInteraction}()},
\code{\link{setMaxIntakeRate}()},
\code{\link{setMetabolicRate}()},
\code{\link{setParams}()},
\code{\link{setPredKernel}()},
\code{\link{setReproduction}()},
\code{\link{setResource}()},
\code{\link{setSearchVolume}()},
\code{\link{species_params}()}
}
\concept{functions for setting parameters}
