% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveParams.R
\name{saveParams}
\alias{saveParams}
\alias{readParams}
\alias{saveSim}
\alias{readSim}
\title{Save and restore mizer objects}
\usage{
saveParams(params, file)

readParams(file, install_extensions = FALSE)

saveSim(sim, file)

readSim(file, install_extensions = FALSE)
}
\arguments{
\item{params}{A MizerParams object}

\item{file}{The name of the file or a connection where the object is saved
to or read from.}

\item{install_extensions}{Logical. Should \code{\link[=readParams]{readParams()}} or \code{\link[=readSim]{readSim()}}
attempt to install missing extension packages before registering the saved
extension chain?}

\item{sim}{A MizerSim object}
}
\value{
\code{saveParams()} and \code{saveSim()} return NULL invisibly.
\code{readParams()} returns a MizerParams object. \code{readSim()} returns a MizerSim
object.
}
\description{
\code{saveParams()} saves a MizerParams object to a file. This can then be
restored with \code{readParams()}. \code{saveSim()} and \code{readSim()} provide the same
lifecycle for MizerSim objects.
}
\details{
Issues a warning if the model you are saving relies on some custom functions.
Before saving a model you may want to set its metadata with
\code{\link[=setMetadata]{setMetadata()}}.
}
\examples{
# Save params to a temporary file and read them back
tmp <- tempfile(fileext = ".rds")
saveParams(NS_params, file = tmp)
params <- readParams(tmp)
identical(params, NS_params)

# Save and read back a simulation
tmp2 <- tempfile(fileext = ".rds")
saveSim(NS_sim, file = tmp2)
sim <- readSim(tmp2)
identical(sim, NS_sim)
}
\seealso{
"Using mizer extension packages":
\code{vignette("using-extension-packages", package = "mizer")}
}
