% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveParams.R
\name{saveParams}
\alias{saveParams}
\alias{readParams}
\title{Save a MizerParams object}
\usage{
saveParams(params, file)

readParams(file)
}
\arguments{
\item{params}{A MizerParams object}

\item{file}{The name of the file or a connection where the MizerParams object
is saved to or read from.}
}
\description{
\code{This}saveParams()`` saves a MizerParams object including all functions that
are referenced in the MizerParams object. This can then be loaded back with
\code{loadParams()}.
}
