% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{set_multispecies_model}
\alias{set_multispecies_model}
\title{Deprecated obsolete function for setting up multispecies parameters}
\usage{
set_multispecies_model(
  species_params,
  interaction = matrix(1, nrow = nrow(species_params), ncol = nrow(species_params)),
  min_w_pp = 1e-10,
  max_w = max(species_params$w_inf) * 1.1,
  n = 2/3,
  q = 0.8,
  f0 = 0.6,
  kappa = 1e+11,
  lambda = 2 + q - n,
  ...
)
}
\arguments{
\item{species_params}{A data frame of species-specific parameter values.}

\item{interaction}{Interaction matrix of the species (predator by prey).
Entries should be numbers between 0 and 1. See "Setting interactions"
section below.}

\item{min_w_pp}{The smallest size of the plankton spectrum. By default this
is set to the smallest value at which any of the consumers can feed.}

\item{max_w}{The largest size of the consumer spectrum. By default this is
set to the largest w_inf specified in the species_params data frame.}

\item{n}{Scaling exponent of the intrinsic birth rate}

\item{q}{Allometric exponent of search volume}

\item{f0}{Expected average feeding level. Used to set \code{gamma}, the
coefficient in the search rate. The default value is 0.6. Ignored if 
\code{gamma} is given explicitly.}

\item{kappa}{Coefficient of the intrinsic carrying capacity}

\item{lambda}{Scaling exponent of the intrinsic carrying capacity}

\item{...}{Unused}
}
\description{
This function has been deprecated in favour of the function
\code{\link{newMultispeciesParams}} that sets better default values.
}
\seealso{
Other deprecated functions: 
\code{\link{set_community_model}()},
\code{\link{set_trait_model}()}
}
\concept{deprecated functions}
