% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setColours.R
\name{setColours}
\alias{setColours}
\alias{getColours}
\title{Set line colours to be used in mizer plots}
\usage{
setColours(params, colours)

getColours(params)
}
\arguments{
\item{params}{A MizerParams object}

\item{colours}{A named list or named vector of line colours.}
}
\value{
\code{setColours}: The MizerParams object with updated line colours

\code{getColours()}: A named vector of colours
}
\description{
Colours for names that already had a colour set will be overwritten by
the colour you specify. Colours for names that did not yet have a colour
will be appended to the list of colours.
}
\examples{
params <- setColours(NS_params, list("Resource" = "red", "Total" = "#0000ff"))
species_params(params)["Cod", "linecolour"] <- "black"
plotSpectra(params, total = TRUE)
getColours(params)
}
