% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{setComponent}
\alias{setComponent}
\alias{getComponent}
\title{Add a dynamical ecosystem component}
\usage{
setComponent(
  params,
  component,
  initial_value,
  encounter_fun,
  mortality_fun,
  dynamics_fun,
  component_params
)

getComponent(params, component)
}
\arguments{
\item{params}{A MizerParams object}

\item{component}{Name of the component of interest. If missing, a list of
all components will be returned.}

\item{initial_value}{Initial value of the component}

\item{encounter_fun}{Name of function to calculate contribution to encounter
rate}

\item{mortality_fun}{Name of function to calculate contribution to the
predation mortality rate.}

\item{dynamics_fun}{Name of function to calculate value at the next time step}

\item{component_params}{Named list of parameters needed by the component
functions.}
}
\value{
For `setComponent`: The updated MizerParams object

For \code{getComponent}: A list with the entries \code{initial_value}, \code{dynamics_fun},
\code{encounter_fun}, \code{morality_fun}, \code{component_params}. If \code{component} is
missing, then a list of lists for all components is returned.
}
\description{
Add a dynamical ecosystem component

Get information about other ecosystem components
}
