% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDiffusion.R
\name{setDiffusion}
\alias{setDiffusion}
\alias{diffusion}
\alias{diffusion<-}
\title{Set diffusion rate}
\usage{
setDiffusion(params, diffusion = NULL, ...)

diffusion(params)

diffusion(params) <- value
}
\arguments{
\item{params}{MizerParams}

\item{diffusion}{Optional. An array (species x size) holding the diffusion rate. If
not supplied, the diffusion rate is left unchanged. Initially it is set to 0.}

\item{...}{Unused}

\item{value}{diffusion}
}
\value{
\code{setDiffusion()}: A MizerParams object with updated diffusion rate.

\code{diffusion()}: An array (species x size) with the diffusion rate.
}
\description{
Set diffusion rate
}
\section{Setting diffusion rate}{

The diffusion rate governs the redistribution of abundance density over body size
beyond deterministic growth dynamics.

The \code{diffusion} argument allows you to specify a diffusion rate that depends on
species and body size.
}

\seealso{
Other functions for setting parameters: 
\code{\link{gear_params}()},
\code{\link{setExtEncounter}()},
\code{\link{setExtMort}()},
\code{\link{setFishing}()},
\code{\link{setInitialValues}()},
\code{\link{setInteraction}()},
\code{\link{setMaxIntakeRate}()},
\code{\link{setMetabolicRate}()},
\code{\link{setParams}()},
\code{\link{setPredKernel}()},
\code{\link{setReproduction}()},
\code{\link{setSearchVolume}()},
\code{\link{species_params}()}
}
\concept{functions for setting parameters}
