% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setExtDiffusion.R
\name{setExtDiffusion}
\alias{setExtDiffusion}
\alias{setExtDiffusion.MizerParams}
\alias{ext_diffusion}
\alias{ext_diffusion<-}
\title{Set external diffusion rate}
\usage{
setExtDiffusion(params, ext_diffusion = NULL, reset = FALSE, ...)

\method{setExtDiffusion}{MizerParams}(params, ext_diffusion = NULL, reset = FALSE, ...)

ext_diffusion(params)

ext_diffusion(params) <- value
}
\arguments{
\item{params}{MizerParams}

\item{ext_diffusion}{Optional. An array (species x size) holding the
external diffusion rate. If not supplied, the external diffusion rate is
left unchanged. Initially it is set to 0.}

\item{reset}{Unused. Included for interface consistency with other setter
functions.}

\item{...}{Unused}

\item{value}{ext_diffusion}
}
\value{
\code{setExtDiffusion()}: A MizerParams object with updated external
diffusion rate.

\code{ext_diffusion()}: An \code{ArraySpeciesBySize} object (species x size)
with the external diffusion rate.
}
\description{
Set external diffusion rate
}
\section{Setting external diffusion rate}{

The external diffusion rate allows you to impose additional diffusion of
abundance density over body size beyond the predation-driven diffusion
computed by \code{\link[=mizerDiffusion]{mizerDiffusion()}}.

The \code{ext_diffusion} argument allows you to specify a diffusion rate that
depends on species and body size.
}

\seealso{
Other functions for setting parameters: 
\code{\link{gear_params}()},
\code{\link{setExtEncounter}()},
\code{\link{setExtMort}()},
\code{\link{setFishing}()},
\code{\link{setInitialValues}()},
\code{\link{setInteraction}()},
\code{\link{setMaxIntakeRate}()},
\code{\link{setMetabolicRate}()},
\code{\link{setParams}()},
\code{\link{setPredKernel}()},
\code{\link{setReproduction}()},
\code{\link{setSearchVolume}()},
\code{\link{species_params}()}
}
\concept{functions for setting parameters}
