% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setExtDiffusion.R
\name{setExtDiffusion}
\alias{setExtDiffusion}
\alias{setExtDiffusion.MizerParams}
\alias{ext_diffusion}
\alias{ext_diffusion<-}
\title{Set external diffusion rate}
\usage{
setExtDiffusion(params, ext_diffusion = NULL, reset = FALSE, ...)

ext_diffusion(params)

ext_diffusion(params) <- value
}
\arguments{
\item{params}{MizerParams}

\item{ext_diffusion}{Optional. An array (species x size) holding the
external diffusion rate. If not supplied, a default is calculated from the
\code{D_ext} and \code{n} species parameters as described in the section "Setting
external diffusion rate".}

\item{reset}{If set to TRUE, then the external diffusion rate will be reset to the
value calculated from the species parameters, even if it was previously
overwritten with a custom value. If set to FALSE (default) then a
recalculation from the species parameters will take place only if no
custom value has been set.}

\item{...}{Unused}

\item{value}{ext_diffusion}
}
\value{
\code{setExtDiffusion()}: A MizerParams object with updated external
diffusion rate.

\code{ext_diffusion()}: An \code{ArraySpeciesBySize} object (species x size)
with the external diffusion rate.
}
\description{
Set external diffusion rate
}
\section{Setting external diffusion rate}{

The external diffusion rate allows you to impose additional diffusion
beyond the predation-driven diffusion that can be internally modelled by mizer.

The \code{ext_diffusion} argument allows you to specify a diffusion rate that
depends on species and body size.

If the \code{ext_diffusion} argument is not supplied, then the external diffusion
rate is calculated as a power law:
\deqn{D_{ext.i}(w) = D_{ext.i}\, w^{n_i+1}.}{
  D_{ext.i}(w) = D_{ext.i} * w^(n_i+1).}
The coefficient \eqn{D_{ext.i}} is taken from the \code{D_ext} column of the
species parameter data frame, which defaults to 0. The exponent
\eqn{n_i + 1} uses the \code{n} column of the species parameter data frame.

If the \code{ext_diffusion} slot has a comment and \code{reset = FALSE}, then a
recalculation from the species parameters is suppressed and a message is
issued if the recalculated values would differ from the stored ones.
}

\seealso{
Other functions for setting parameters: 
\code{\link{gear_params}},
\code{\link{setExtEncounter}()},
\code{\link{setExtMort}()},
\code{\link{setFishing}()},
\code{\link{setInteraction}()},
\code{\link{setMaxIntakeRate}()},
\code{\link{setMetabolicRate}()},
\code{\link{setParams}()},
\code{\link{setPredKernel}()},
\code{\link{setReproduction}()},
\code{\link{setSearchVolume}()},
\code{\link{species_params}},
\code{\link{use_predation_diffusion}()}
}
\concept{functions for setting parameters}
