% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setExtMort.R
\name{setExtMort}
\alias{setExtMort}
\alias{ext_mort}
\alias{ext_mort<-}
\title{Set external mortality rate}
\usage{
setExtMort(
  params,
  ext_mort = NULL,
  z0pre = 0.6,
  z0exp = params@resource_params$n - 1,
  reset = FALSE,
  z0 = deprecated(),
  ...
)

ext_mort(params)

ext_mort(params) <- value
}
\arguments{
\item{params}{MizerParams}

\item{ext_mort}{Optional. An array (species x size) holding the external
mortality rate.  If not supplied, a default is set as described in the
section "Setting external mortality rate".}

\item{z0pre}{If \code{z0}, the mortality from other sources, is not present in
\code{\link[=given_species_params]{given_species_params()}}, it is calculated as
\code{z0pre * w_inf ^ z0exp}. Default value is 0.6.}

\item{z0exp}{The exponent used with \code{z0pre} to calculate non-given \code{z0}.
Default value is \code{n - 1}.}

\item{reset}{If set to TRUE, then the external mortality rate will be reset
to the value calculated from the species parameters, even if it was
previously overwritten with a custom value. If set to FALSE (default) then
a recalculation from the species parameters will take place only if no
custom value has been set.}

\item{z0}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{ext_mort} instead. Not to
be confused with the species_parameter \code{z0}.}

\item{...}{Unused}

\item{value}{ext_mort}
}
\value{
\code{setExtMort()}: A MizerParams object with updated external mortality
rate.

\code{ext_mort()}: An \code{ArraySpeciesBySize} object (species x size) with
the external mortality.
}
\description{
You will usually not need to call this function directly. Instead change
the \code{z0}, \code{z_ext} and \code{d} species parameters with
\verb{given_species_params(params) <-} and let mizer recalculate the external
mortality rate for you. Call \code{setExtMort()} directly only if you want to
impose a different functional form for the size dependence of the external
mortality. See \code{vignette("guide-change-parameters")} for a full
explanation of when to reach for which level of the model.
}
\section{Setting external mortality rate}{

The external mortality is all the mortality that is not due to fishing or
predation by predators included in the model. The external mortality could be
due to predation by predators that are not explicitly included in the model
(e.g. mammals or seabirds) or due to other causes like illness. It is a rate
with units 1/year.

The \code{ext_mort} argument allows you to specify an external mortality rate
that depends on species and body size. You can see an example of this in
the Examples section of the help page for \code{\link[=setExtMort]{setExtMort()}}.

If the \code{ext_mort} argument is not supplied, then the external mortality is
taken from the species parameters as
\deqn{\mu_{ext.i}(w) = z_{0.i} + z_{ext.i} w^{d_i}.}{mu_ext,i(w) = z0_i + z_ext,i w ^ d_i.}
The value of the constant \eqn{z_0} for each species is taken from the \code{z0}
column of \code{\link[=given_species_params]{given_species_params()}} if it is present there. Otherwise it is
recalculated, even if a value from an earlier calculation is still present
in \code{species_params}, as
\deqn{z_{0.i} = {\tt z0pre}_i\, w_{inf}^{\tt z0exp}.}{z_{0.i} = z0pre_i w_{inf}^{z0exp}.}
When \code{z0pre} or \code{z0exp} is supplied explicitly and used to calculate
non-given \code{z0}, the resulting values are recorded in
\code{\link[=given_species_params]{given_species_params()}}. Values calculated from the defaults
\code{z0pre = 0.6} and \code{z0exp = n - 1} are not recorded there. If either argument
is supplied but cannot be used because \code{z0} is given for every species or
because \code{ext_mort} was supplied, a warning is issued.
Missing values of \code{z_ext} are set to 0 and missing values of \code{d} are set to
\code{n - 1}.

By default the power law is evaluated at the left bin edges \eqn{w_j}
(point sampling). If the \code{bin_average} entry of the \code{second_order_w} slot is
\code{TRUE} (see \code{\link[=second_order_w]{second_order_w()}}), then the \eqn{z_{ext} w^d} term is instead
replaced by its exact average over each bin \eqn{[w_j, w_{j+1}]},
\deqn{\frac{z_{ext}}{\Delta w_j}\int_{w_j}^{w_{j+1}} w^d\, dw
  = z_{ext}\,\frac{w_{j+1}^{d+1} - w_j^{d+1}}{(d+1)\,\Delta w_j},}
(with the limiting form \eqn{z_{ext}\ln(w_{j+1}/w_j)/\Delta w_j} when
\eqn{d = -1}). This is the consistent choice in the finite-volume scheme,
where the external mortality multiplies the bin-averaged abundance. The
bin-averaging is applied only to the auto-calculated power-law default; a
user-supplied \code{ext_mort} array is left untouched.
}

\examples{
params <- newMultispeciesParams(NS_species_params)

#### Setting allometric death rate #######################

# Set coefficient for each species. Here we choose 0.1 for each species
z0pre <- rep(0.1, nrow(species_params(params)))

# Multiply by power of size with exponent, here chosen to be -1/4
# The outer() function makes it an array species x size
allo_mort <- outer(z0pre, w(params)^(-1/4))

# Change the external mortality rate in the params object
ext_mort(params) <- allo_mort
}
\seealso{
Other functions for setting parameters:
\code{\link[=gear_params]{gear_params()}},
\code{\link[=setExtDiffusion]{setExtDiffusion()}},
\code{\link[=setExtEncounter]{setExtEncounter()}},
\code{\link[=setFishing]{setFishing()}},
\code{\link[=setInteraction]{setInteraction()}},
\code{\link[=setMaxIntakeRate]{setMaxIntakeRate()}},
\code{\link[=setMetabolicRate]{setMetabolicRate()}},
\code{\link[=setParams]{setParams()}},
\code{\link[=setPredKernel]{setPredKernel()}},
\code{\link[=setReproduction]{setReproduction()}},
\code{\link[=setSearchVolume]{setSearchVolume()}},
\code{\link[=species_params]{species_params()}},
\code{\link[=use_predation_diffusion]{use_predation_diffusion()}}
}
\concept{functions for setting parameters}
