% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setFishing}
\alias{setFishing}
\title{Set fishing parameters}
\usage{
setFishing(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
MizerParams object
}
\description{
Set fishing parameters
}
\section{Setting fishing}{

Needs to be documented
}

