% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setIntakeMax}
\alias{setIntakeMax}
\title{Set maximum intake rate}
\usage{
setIntakeMax(params, intake_max = NULL, n = params@n)
}
\arguments{
\item{params}{MizerParams}

\item{intake_max}{Optional. An array (species x size) holding the maximum
intake rate for each species at size. If not supplied, a default is set as
described in the section "Setting maximum intake rate".}

\item{n}{Scaling exponent of the intake rate.}
}
\value{
A \code{MizerParams} object
}
\description{
Set maximum intake rate
}
\section{Setting maximum intake rate}{

The maximum intake rate \eqn{h_i(w)} of an individual of species \eqn{i} and
weight \eqn{w} determines the feeding level, calculated with
\code{\link{getFeedingLevel}}. It is measured in grams/year.

If the \code{intake_max} argument is not supplied, then the maximum intake
rate is set to \deqn{h_i(w) = h_i w^n.} 
The values of \eqn{h_i} (the maximum intake rate of an individual of size
1 gram) are taken from the \code{h} column in the
species parameter dataframe. If the \code{h} column is not supplied in the
species parameter dataframe, it is calculated by the 
\code{\link{get_h_default}} function, using \code{f0} and the \code{k_vb}
column, if they are supplied.

If \eqn{h_i} is set to \code{Inf}, fish will consume all encountered food.
}

\examples{
\dontrun{
params <- NS_params
params@species_params$h[3] <- 35
params <- setIntakeMax(params)
}
}
\seealso{
Other functions for setting parameters: 
\code{\link{setBMort}()},
\code{\link{setFishing}()},
\code{\link{setInitial}()},
\code{\link{setInteraction}()},
\code{\link{setMetab}()},
\code{\link{setParams}()},
\code{\link{setPlankton}()},
\code{\link{setPredKernel}()},
\code{\link{setReproduction}()},
\code{\link{setSearchVolume}()}
}
\concept{functions for setting parameters}
