% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setInteraction}
\alias{setInteraction}
\title{Set species interation matrix}
\usage{
setInteraction(params, interaction)
}
\arguments{
\item{params}{MizerParams object}

\item{interaction}{Interaction matrix of the species (predator by prey). By
default all interactions between species are set to 1.}
}
\value{
MizerParams object
}
\description{
Checks that the supplied interaction matrix is valid and then stores it in
the \code{interaction} slot of the params object before returning that 
object.
}
\details{
Any dimnames of the interaction matrix argument are ignored by the
constructor. The dimnames of the interaction matrix in the returned
\code{MizerParams} object are taken from the species names in the
\code{species_params} slot. This means that the order of the columns and rows
of the interaction matrix argument should be the same as the species name in
the \code{species_params} slot.
}
