% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setInteraction}
\alias{setInteraction}
\title{Set species interation matrix}
\usage{
setInteraction(params, interaction)
}
\arguments{
\item{params}{MizerParams object}

\item{interaction}{The interaction matrix of the species (predator by prey)}
}
\value{
MizerParams object
}
\description{
Checks that the supplied interaction matrix is valid and then stores it in
the \code{interaction} slot of the params object before returning that 
object.
}
