% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setInteraction}
\alias{setInteraction}
\title{Set species interation matrix and plankton interaction vector}
\usage{
setInteraction(params, interaction = NULL)
}
\arguments{
\item{params}{MizerParams object}

\item{interaction}{Interaction matrix of the species (predator by prey).
Entries should be numbers between 0 and 1. See "Setting interactions"
section below.}
}
\value{
MizerParams object
}
\description{
Set species interation matrix and plankton interaction vector
}
\section{Setting interactions}{


The species interaction matrix \eqn{\theta_{ij}} is used when calculating the food encounter rate in
\code{\link{getEncounter}} and the predation rate in
\code{\link{getPredRate}}. This function checks that the supplied interaction
matrix is valid and then stores it in the \code{interaction} slot of the params
object before returning that object.

Any dimnames of the interaction matrix argument are ignored by the
constructor. The dimnames of the interaction matrix in the returned
\code{MizerParams} object are taken from the species names in the
\code{species_params} slot. This means that the order of the columns and rows
of the interaction matrix argument should be the same as the species name in
the \code{species_params} slot.
}

