% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setInteraction}
\alias{setInteraction}
\title{Set species interation matrix}
\usage{
setInteraction(params, interaction = NULL)
}
\arguments{
\item{params}{MizerParams object}

\item{interaction}{Interaction matrix of the species (predator by prey).
Entries should be numbers between 0 and 1. See "Setting interactions"
section below.}
}
\value{
MizerParams object
}
\description{
Set species interation matrix
}
\section{Setting interactions}{


The species interaction matrix \eqn{\theta_{ij}}, is used when calculating the
food encounter rate in \code{\link{getEncounter}} and the predation mortality rate in
\code{\link{getPredMort}}. Its entries are dimensionless numbers between
0 and 1 that characterise the strength at which predator species \eqn{i}
predates on prey species \eqn{j}. 

This function checks that the supplied interaction
matrix is valid and then stores it in the \code{interaction} slot of the
params object before returning that object.

The order of the columns and rows of the \code{interaction} argument should be the 
same as the order in the species params dataframe in the \code{params} object.
If you supply a named array then the function will check the order and warn 
if it is different.

The interaction of the species with the plankton are set via a column
\code{interaction_p} in the \code{species_params} data frame. Again the entries
have to be numbers between 0 and 1. By default this column is set to all
1s.
}

\examples{
\dontrun{
params <- set_trait_model()
interaction <- params@interaction
interaction[1, 3] <- 0
params <- setInteraction(params, interaction)
}
}
\seealso{
Other functions for setting parameters: \code{\link{setBMort}},
  \code{\link{setFishing}}, \code{\link{setInitial}},
  \code{\link{setIntakeMax}}, \code{\link{setMetab}},
  \code{\link{setParams}}, \code{\link{setPlankton}},
  \code{\link{setPredKernel}},
  \code{\link{setReproduction}},
  \code{\link{setResourceDynamics}},
  \code{\link{setResourceEncounter}},
  \code{\link{setSearchVolume}}
}
\concept{functions for setting parameters}
