% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setLinetypes}
\alias{setLinetypes}
\alias{getLinetypes}
\title{Set linetypes to be used in mizer plots}
\usage{
setLinetypes(params, linetypes)

getLinetypes(params)
}
\arguments{
\item{params}{A MizerParams object}

\item{linetypes}{A named list or named vector of linetypes.}
}
\value{
The MizerParams object with updated linetypes
}
\description{
Set linetypes to be used in mizer plots
}
\examples{
params <- NS_params
params <- setLinetypes(params, list("Cod" = "solid"))
plotSpectra(params)
getLinetypes(params)
}
