% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setMetab}
\alias{setMetab}
\title{Set metabolic rate}
\usage{
setMetab(params, metab = NULL, p)
}
\arguments{
\item{params}{MizerParams}

\item{metab}{Optional. An array (species x size) holding the metabolic rate
for each species at size. If not supplied, a default is set as described in
the section "Setting metabolic rate".}

\item{p}{Scaling exponent of the standard metabolic rate.}
}
\value{
MizerParams
}
\description{
Sets the rate at which energy is used for metabolism and activity
}
\section{Setting metabolic rate}{

To be documented
}

