% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setMetab}
\alias{setMetab}
\title{Set metabolic rate}
\usage{
setMetab(params, metab = NULL, p = params@p)
}
\arguments{
\item{params}{MizerParams}

\item{metab}{Optional. An array (species x size) holding the metabolic rate
for each species at size. If not supplied, a default is set as described in
the section "Setting metabolic rate".}

\item{p}{Scaling exponent of the standard metabolic rate.}
}
\value{
MizerParams
}
\description{
Sets the rate at which energy is used for metabolism and activity
}
\section{Setting metabolic rate}{

The metabolic rate is subtracted from the energy income rate to calculate
the rate at which energy is available for growth and reproduction, see
\code{\link{getEReproAndGrowth}}. It is measured in grams/year.

If the \code{metab} argument is not supplied, then the metabolic
rate \eqn{k_i(w)} for an individual of species \eqn{i} and size \eqn{w}
is set to \deqn{k_i(w) = ks_i\, w^p + k_i\, w,}{k_i(w) = ks_i w^p + k_i w}
where \eqn{ks_i w^p} represents the rate of standard metabolism and 
\eqn{k_i w} is the rate at which energy is expended on activity and movement.
The values of \eqn{ks_i} and \eqn{k_i} are taken from the \code{ks} and
\code{k} columns in the
species parameter dataframe. If these parameters are not supplied, the
defaults are \eqn{ks_i = 0.2 h_i} and \eqn{k_i = 0}, where \eqn{h_i} is the
coefficient of the maximum intake rate and is taken from the species
parameter data frame in \code{params}.
}

\examples{
\dontrun{
params <- NS_params
# Change activity coefficient for species 3
params@species_params$k[3] <- 8
params <- setMetab(params)
}
}
\seealso{
Other functions for setting parameters: 
\code{\link{setBMort}()},
\code{\link{setFishing}()},
\code{\link{setInitial}()},
\code{\link{setIntakeMax}()},
\code{\link{setInteraction}()},
\code{\link{setParams}()},
\code{\link{setPlankton}()},
\code{\link{setPredKernel}()},
\code{\link{setReproduction}()},
\code{\link{setResourceDynamics}()},
\code{\link{setResourceEncounter}()},
\code{\link{setSearchVolume}()}
}
\concept{functions for setting parameters}
