% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMetadata.R
\name{setMetadata}
\alias{setMetadata}
\alias{getMetadata}
\title{Set metadata for a model}
\usage{
setMetadata(
  params,
  title = NULL,
  description = NULL,
  authors = NULL,
  orcid = NULL
)

getMetadata(params)
}
\arguments{
\item{params}{The MizerParams object for the model}

\item{title}{A string with the title for the model}

\item{description}{A string with a description of the model}

\item{authors}{A vector of strings with the author names}

\item{orcid}{A named vector of strings where each name is an author name and
each value is an orcid id.}
}
\value{
The MizerParams object with updated metadata

\code{getMetadata()}: A list with the entries \code{title}, \code{description},
\code{authors}, \code{orcid}, \code{version}, \code{extensions}, \code{time_created} and
\code{time_modified}.
}
\description{
Set metadata for a model
}
\section{Setting model metadata}{


The metadata of a model consists of the following slots:
\itemize{
\item \code{title} A string with the title for the model
\item \code{description} A string with a description of the model
\item \code{authors} A vector of strings with the author names
\item \code{orcid} A named vector of strings where each name is an author name and
each value is an orcid id.
\item \code{version} The version string of the mizer version that last modified the
model as returned by packageVersion("mizer")
\item \code{extensions} A named vector of strings where each name is the name of
and extension package needed to run the model and each value is a string
giving the information that the remotes package needs to install the
correct version of the extension package, see https://remotes.r-lib.org/
\item \code{time_created} A POSIXct date-time object with the creation time
\item \code{time_modified} A POSIXct date-time object with the last modified time
}

The \code{version}, \code{time_created} \code{time_modified} and \code{extensions} fields are
set automatically by the appropriate functions. The other fields are set
via \code{setMetadata()}.
}

