% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setPlankton}
\alias{setPlankton}
\title{Set up plankton}
\usage{
setPlankton(params, kappa = params@kappa, lambda = params@lambda,
  r_pp = 10, w_pp_cutoff = 10, plankton_dynamics = NULL)
}
\arguments{
\item{params}{A MizerParams object}

\item{kappa}{Carrying capacity of the plankton spectrum.}

\item{lambda}{Exponent of the plankton spectrum.}

\item{r_pp}{Growth rate of the primary productivity. Default is 10 g/year.}

\item{w_pp_cutoff}{The upper cut off size of the plankton spectrum. 
Default is 10 g.}

\item{plankton_dynamics}{Function that determines plankton dynamics by
calculating the plankton spectrum at the next time step from the current
state.}
}
\value{
A MizerParams object
}
\description{
Set up plankton
}
\section{Setting plankton dynamics}{

Still need to document
}

\seealso{
Other functions for setting parameters: \code{\link{setBMort}()},
  \code{\link{setFishing}()}, \code{\link{setInitial}()},
  \code{\link{setIntakeMax}()},
  \code{\link{setInteraction}()}, \code{\link{setMetab}()},
  \code{\link{setParams}()}, \code{\link{setPredKernel}()},
  \code{\link{setReproduction}()},
  \code{\link{setResourceDynamics}()},
  \code{\link{setResourceEncounter}()},
  \code{\link{setSearchVolume}()}
}
\concept{functions for setting parameters}
