% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{setPredKernel}
\alias{setPredKernel}
\title{Set the predation kernel to allow size-dependent PPMR}
\usage{
setPredKernel(params, pred_kernel)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{pred_kernel}{An array (species x predator size x prey size) 
that holds the predation coefficient of each predator at size on 
each prey size. The dimensions are thus no_sp, no_w, no_w_full.}
}
\value{
A \linkS4class{MizerParams} object
}
\description{
Allows you to specify the full feeding kernel as a three-dimensional
array (predator species x predator size x prey size), without the
restriction that the kernel is a function of the ratio of predator size
to prey size only that is imposed by the \code{\link{defaultPredKernel}}.
}
\details{
You should use this function only if a kernel dependent only on the
predator/prey mass ratio is not appropriate. Otherwise you should use
\code{\link{defaultPredKernel}} which is easier to use and, more importantly,
will lead to much reduced runtimes for simulations because the efficient fast
Fourier transform methods can be used.

One way to set up a model with size-dependent predator/prey mass ratio
(PPMR) is to first use one of the standard set up functions to create
a MizerParams object and then to call this function with that object 
and an array holding the predation kernel.

The order of the predator species in the predation kernel should be the same
as the order in the species params dataframe. If you supply a named array
then the function will check the order and warn if it is different.
}
\examples{
\dontrun{
## Set up a MizerParams object
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)

## Create a predation kernel
beta <- params@species_params$beta
sigma <- params@species_params$sigma
w <- params@w
w_full <- params@w_full
pk = array(beta, dim = c(length(beta),length(w), length(w_full)))
pk <- exp(-0.5 * sweep(log(sweep(sweep(pk, 3, w_full, "*") ^ -1, 2, w, "*")),
                       1, sigma, "/") ^ 2)
pk <- sweep(pk, c(2, 3), combn(w_full, 1, function(x, w) x < w, w = w), "*")

## Create a new MizerParams object to use the new predation kernel
params_new <- setPredKernel(params, pred_kernel = pk)
}
}
