% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{setRateFunction}
\alias{setRateFunction}
\alias{getRateFunction}
\title{Set own rate function to replace mizer rate function}
\usage{
setRateFunction(params, rate = "Rates", fun)

getRateFunction(params, rate = "Rates")
}
\arguments{
\item{params}{A \code{MizerParams} object}

\item{rate}{Name of the rate for which a new function is to be set.}

\item{fun}{Name of the function to use to calculate the rate.}
}
\description{
At each time step during a simulation with the \code{\link[=project]{project()}} function, mizer
needs to calculate the instantaneous values of the various rates. By
default it calls the \code{\link[=mizerRates]{mizerRates()}} function which creates a list with the
following components:
\itemize{
\item \code{encounter} from \code{\link[=mizerEncounter]{mizerEncounter()}}
\item \code{feeding_level} from \code{\link[=mizerFeedingLevel]{mizerFeedingLevel()}}
\item \code{pred_rate} from \code{\link[=mizerPredRate]{mizerPredRate()}}
\item \code{pred_mort} from \code{\link[=mizerPredMort]{mizerPredMort()}}
\item \code{fishing_mort} from \code{\link[=mizerFMort]{mizerFMort()}}
\item \code{mort} from \code{\link[=mizerMort]{mizerMort()}}
\item \code{plankton_mort} from \code{\link[=mizerPlanktonMort]{mizerPlanktonMort()}}
\item \code{e} from \code{\link[=mizerEReproAndGrowth]{mizerEReproAndGrowth()}}
\item \code{e_repro} from \code{\link[=mizerERepro]{mizerERepro()}}
\item \code{e_growth} from \code{\link[=mizerEGrowth]{mizerEGrowth()}}
\item \code{rdi} from \code{\link[=mizerRDI]{mizerRDI()}}
\item \code{rdd} from \code{\link[=BevertonHoltRDD]{BevertonHoltRDD()}}
}
}
\details{
You can modify these in two ways.
}
