% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setReproduction.R
\name{setReproduction}
\alias{setReproduction}
\alias{maturity}
\alias{maturity<-}
\alias{repro_prop}
\alias{repro_prop<-}
\alias{psi}
\title{Set reproduction parameters}
\usage{
setReproduction(
  params,
  maturity = NULL,
  repro_prop = NULL,
  reset = FALSE,
  RDD = NULL,
  ...
)

maturity(params)

maturity(params) <- value

repro_prop(params)

repro_prop(params) <- value

psi(params)
}
\arguments{
\item{params}{A MizerParams object}

\item{maturity}{Optional. An array (species x size) that holds the proportion
of individuals of each species at size that are mature. If not supplied, a
default is set as described in the section "Setting reproduction".}

\item{repro_prop}{Optional. An array (species x size) that holds the
proportion of the energy available for growth and reproduction that a mature
individual allocates to reproduction for each species at size. If not
supplied, a default is set as described in the section "Setting
reproduction".}

\item{reset}{If set to TRUE, then both \code{maturity} and \code{repro_prop} will be
reset to the value calculated from the species parameters, even if they
were previously overwritten with custom values. If set to FALSE (default)
then a recalculation from the species parameters will take place only if no
custom values have been set.}

\item{RDD}{The name of the function calculating the density-dependent
reproduction rate from the density-independent rate. Defaults to
"\code{\link[=BevertonHoltRDD]{BevertonHoltRDD()}}".}

\item{...}{Unused}

\item{value}{The desired new value for the respective parameter.}
}
\value{
\code{setReproduction()}: A MizerParams object with updated reproduction
parameters.

\code{maturity()}: An \code{ArraySpeciesBySize} object (species x size) that
holds the proportion
of individuals of each species at size that are mature.

\code{repro_prop()}: An \code{ArraySpeciesBySize} object (species x size) that
holds the proportion
of the energy available for growth and reproduction that a mature
individual allocates to reproduction for each species at size. For sizes
where the maturity proportion is zero, also the reproduction proportion is
returned as zero.
}
\description{
Sets the proportion of the total energy available for reproduction and growth
that is invested into reproduction as a function of the size of the
individual and sets additional density dependence. You will usually not
need to call this function directly. Instead change the \code{w_mat},
\code{w_mat25}, \code{w_repro_max} and \code{m} species parameters with
\verb{given_species_params(params) <-} and let mizer recalculate the maturity
ogive and reproduction allocation for you. Call \code{setReproduction()}
directly only if you want to impose different functional forms for these.
See \code{vignette("guide-change-parameters")} for a full explanation of
when to reach for which level of the model.
}
\section{Setting reproduction}{


For each species and at each size, the proportion \eqn{\psi}{psi} of the
available energy
that is invested into reproduction is the product of two factors: the
proportion \code{maturity} of individuals that are mature and the proportion
\code{repro_prop} of the energy available to a mature individual that is
invested into reproduction. There is a size \code{w_repro_max} at which a typical
mature individual invests all of its available energy into reproduction.
This is \strong{not} a hard ceiling on size: not all individuals are mature at
\code{w_repro_max}, and diffusion in the growth process allows some individuals to
grow beyond it, so fish larger than \code{w_repro_max} can exist. If you have not
specified the \code{w_repro_max} column in the species parameter data frame, then
the von Bertalanffy asymptotic size \code{w_inf} is used instead.

\subsection{Maturity ogive}{
If the the proportion of individuals that are mature is not supplied via
the \code{maturity} argument, then it is set to a sigmoidal
maturity ogive that changes from 0 to 1 at around the maturity size:
\deqn{{\tt maturity}(w) = \left[1+\left(\frac{w}{w_{mat}}\right)^{-U}\right]^{-1}.}{
  maturity(w) = [1+(w/w_mat)^(-U)]^(-1)}
(To avoid clutter, we are not showing the species index in the equations,
although each species has its own maturity ogive.)
The maturity weights are taken from the \code{w_mat} column of the
species_params data frame. Any missing maturity weights are set to 1/4 of the
asymptotic size in the \code{w_inf} column.

The exponent \eqn{U} determines the steepness of the maturity ogive. By
default it is chosen as \eqn{U = 10}, however this can be overridden by
including a column \code{w_mat25} in the species parameter dataframe that
specifies the weight at which 25\% of individuals are mature, which sets
\eqn{U = \log(3) / \log(w_{mat} / w_{mat25}).}{U = log(3) / log(w_mat /
w_mat25).}

The sigmoidal function given above would strictly reach 0 only
asymptotically and thus have some (negligible) amount of reproduction at
arbitrarily small size.
For computational simplicity, any proportion smaller than
\code{1e-8} is set to \code{0}.
}

\subsection{Investment into reproduction}{
If the the energy available to a mature individual that is
invested into reproduction is not supplied via the \code{repro_prop} argument,
it is set to the allometric form
\deqn{{\tt repro\_prop}(w) =
  \min\left(\left(\dfrac{w}{w_{\tt{repro\_max}}}\right)^{m-n},1\right).}{
  repro_prop(w) = min((w/w_repro_max)^(m - n),1).}
Here \eqn{n} is the scaling exponent of the energy income rate. Hence
the exponent \eqn{m} determines the scaling of the investment into
reproduction for mature individuals. By default it is chosen to be
\eqn{m = 1} so that the rate at which energy is invested into reproduction
scales linearly with the size. This default can be overridden by including a
column \code{m} in the species parameter dataframe. The sizes \eqn{w_{repro\_max}}
are taken from the \code{w_repro_max} column in the species parameter data frame,
if it exists, or otherwise from the \code{w_inf} column.

The total proportion of energy invested into reproduction of an individual
of size \eqn{w} is then
\deqn{\psi(w) = {\tt maturity}(w){\tt repro\_prop}(w)}{
  psi(w) = maturity(w) * repro_prop(w)}
In mizer edition 1, at sizes above \code{w_repro_max} the value of \eqn{\psi}
is additionally forced to 1, so that all available energy is invested into
reproduction and growth stops. In edition 2 and above this forcing is not
applied, and \eqn{\psi} is determined entirely by the maturity ogive and the
reproductive proportion.
}

\subsection{Reproductive efficiency}{
The reproductive efficiency \eqn{\epsilon}, i.e., the proportion of energy allocated to
reproduction that results in egg biomass, is set through the \code{erepro}
column in the species_params data frame. If that is not provided, the default
is set to 1 (which you will want to override). The offspring biomass divided
by the egg biomass gives the rate of egg production, returned by
\code{\link[=getRDI]{getRDI()}}:
\deqn{R_{di} = \frac{\epsilon}{2 w_{min}} \int N(w)  E_r(w) \psi(w) \, dw}{R_di = (\epsilon/(2 w_min)) \int N(w)  E_r(w) \psi(w) dw}
}

\subsection{Density dependence}{
The stock-recruitment relationship is an emergent phenomenon in mizer, with
several sources of density dependence. Firstly, the amount of energy invested
into reproduction depends on the energy income of the spawners, which is
density-dependent due to competition for prey. Secondly, the proportion of
larvae that grow up to recruitment size depends on the larval mortality,
which depends on the density of predators, and on larval growth rate, which
depends on density of prey.

Finally, to encode all the density dependence in the stock-recruitment
relationship that is not already included in the other two sources of density
dependence, mizer puts the the density-independent rate of egg production
through a density-dependence function. The result is returned by
\code{\link[=getRDD]{getRDD()}}. The name of the density-dependence function is
specified by the \code{RDD} argument. The default is the Beverton-Holt
function \code{\link[=BevertonHoltRDD]{BevertonHoltRDD()}}, which requires an \code{R_max} column
in the species_params data frame giving the maximum egg production rate. If
this column does not exist, it is initialised to \code{Inf}, leading to no
density-dependence. Other functions provided by mizer are
\code{\link[=RickerRDD]{RickerRDD()}} and \code{\link[=SheperdRDD]{SheperdRDD()}} and you can easily use
these as models for writing your own functions.
}
}

\examples{
\donttest{
# Plot maturity and reproduction ogives for Cod in North Sea model
mat <- maturity(NS_params)["Cod", ]
rp <- repro_prop(NS_params)["Cod", ]
df <- data.frame(Size = w(NS_params),
                 Reproduction = rp,
                 Maturity = mat,
                 Total = mat * rp)
dff <- reshape2::melt(df, id.vars = "Size",
            variable.name = "Type",
            value.name = "Proportion")
library(ggplot2)
ggplot(dff) + geom_line(aes(x = Size, y = Proportion, colour = Type))
}
}
\seealso{
Other functions for setting parameters:
\code{\link[=gear_params]{gear_params()}},
\code{\link[=setExtDiffusion]{setExtDiffusion()}},
\code{\link[=setExtEncounter]{setExtEncounter()}},
\code{\link[=setExtMort]{setExtMort()}},
\code{\link[=setFishing]{setFishing()}},
\code{\link[=setInteraction]{setInteraction()}},
\code{\link[=setMaxIntakeRate]{setMaxIntakeRate()}},
\code{\link[=setMetabolicRate]{setMetabolicRate()}},
\code{\link[=setParams]{setParams()}},
\code{\link[=setPredKernel]{setPredKernel()}},
\code{\link[=setSearchVolume]{setSearchVolume()}},
\code{\link[=species_params]{species_params()}},
\code{\link[=use_predation_diffusion]{use_predation_diffusion()}}
}
\concept{functions for setting parameters}
