% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBevertonHolt.R
\name{setRmax}
\alias{setRmax}
\title{Alias for setBevertonHolt}
\usage{
setRmax(params, R_factor)
}
\arguments{
\item{params}{A MizerParams object}

\item{R_factor}{The factor by which the maximum reproduction rate should be
higher than the initial-state reproduction rate}
}
\value{
A MizerParams object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

An alias provided for backward compatibility with mizer version <= 2.0.4
}
\concept{deprecated}
