% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{setRmax}
\alias{setRmax}
\title{Set maximum recruitment}
\usage{
setRmax(params, rfac)
}
\arguments{
\item{params}{A MizerParams object}

\item{rfac}{The factor by which the maximum recruitment should be higher than
the initial-state recruitment}
}
\value{
A MizerParams object
}
\description{
Takes a MizerParams object with trivial stock recruitment function and sets 
Beverton-Holt stock recruitment with a maximum recruitment that is a chosen
factor \code{rfac} higher than the iniital-state recruitment.
}
