% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSearchVolume.R
\name{setSearchVolume}
\alias{setSearchVolume}
\alias{search_vol}
\alias{search_vol<-}
\title{Set search volume}
\usage{
setSearchVolume(params, search_vol = NULL, reset = FALSE, ...)

search_vol(params)

search_vol(params) <- value
}
\arguments{
\item{params}{MizerParams}

\item{search_vol}{Optional. An array (species x size) holding the search volume
for each species at size. If not supplied, a default is set as described in
the section "Setting search volume".}

\item{reset}{If set to TRUE, then the search volume will be reset to the
value calculated from the species parameters, even if it was previously
overwritten with a custom value. If set to FALSE (default) then a
recalculation from the species parameters will take place only if no custom
value has been set.}

\item{...}{Unused}

\item{value}{search_vol}
}
\value{
\code{setSearchVolume()}: A MizerParams object with updated search volume.

\code{search_vol()}: An \code{ArraySpeciesBySize} object (species x size)
holding the search volume.
}
\description{
You will usually not need to call this function directly. Instead change
the \code{gamma} and \code{q} species parameters with \verb{given_species_params(params) <-}
and let mizer recalculate the search volume for you. Call \code{setSearchVolume()}
directly only if you want to impose a different functional form for the size
dependence of the search volume. See \code{vignette("guide-change-parameters")}
for a full explanation of when to reach for which level of the model.
}
\section{Setting search volume}{

The search volume \eqn{\gamma_i(w)} of an individual of species \eqn{i}
and weight \eqn{w} multiplies the predation kernel when
calculating the encounter rate in \code{\link[=getEncounter]{getEncounter()}} and the
predation rate in \code{\link[=getPredRate]{getPredRate()}}.

The name "search volume" is a bit misleading, because \eqn{\gamma_i(w)} does
not have units of volume. It is simply a parameter that determines the rate
of predation. Its units depend on your choice, see section "Units in mizer".
If you have chosen to work with total abundances, then it is a rate with units
1/year. If you have chosen to work with abundances per m^2 then it has units
of m^2/year. If you have chosen to work with abundances per m^3 then it has
units of m^3/year.

If the \code{search_vol} argument is not supplied, then the search volume is
set to
\deqn{\gamma_i(w) = \gamma_i w^q_i.}
The values of \eqn{\gamma_i} (the search volume at 1g) and \eqn{q_i} (the
allometric exponent of the search volume) are taken from the \code{gamma} and
\code{q} columns in the species parameter dataframe. If the \code{gamma}
column is not supplied in the species parameter dataframe, a default is
calculated by the \code{\link[=get_gamma_default]{get_gamma_default()}} function. If the \code{q} column is not
supplied, a default of \code{lambda - 2 + n} is used. Note that only
for predators of size \eqn{w = 1} gram is the value of the species parameter
\eqn{\gamma_i} the same as the value of the search volume \eqn{\gamma_i(w)}.

If the \code{search_vol} slot has a comment and \code{reset = FALSE}, then a
recalculation from the species parameters is suppressed and a message is
issued if the recalculated values would differ from the stored ones.
}

\examples{
# Inspect the current search volume
search_vol(NS_params)["Cod", 1:5]

# Double the search volume for all species
sv <- search_vol(NS_params) * 2
params <- setSearchVolume(NS_params, search_vol = sv)
search_vol(params)["Cod", 1:5]
}
\seealso{
Other functions for setting parameters:
\code{\link[=gear_params]{gear_params()}},
\code{\link[=setExtDiffusion]{setExtDiffusion()}},
\code{\link[=setExtEncounter]{setExtEncounter()}},
\code{\link[=setExtMort]{setExtMort()}},
\code{\link[=setFishing]{setFishing()}},
\code{\link[=setInteraction]{setInteraction()}},
\code{\link[=setMaxIntakeRate]{setMaxIntakeRate()}},
\code{\link[=setMetabolicRate]{setMetabolicRate()}},
\code{\link[=setParams]{setParams()}},
\code{\link[=setPredKernel]{setPredKernel()}},
\code{\link[=setReproduction]{setReproduction()}},
\code{\link[=species_params]{species_params()}},
\code{\link[=use_predation_diffusion]{use_predation_diffusion()}}
}
\concept{functions for setting parameters}
