% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{setSearchVolume}
\alias{setSearchVolume}
\title{Set search volume}
\usage{
setSearchVolume(params, gamma = NULL)
}
\arguments{
\item{params}{MizerParams}

\item{gamma}{Optional. An array (species x size) holding the search volume
for each species at size. If not supplied, a default is set as described in
the Details.}
}
\value{
MizerParams
}
\description{
Sets the search volume \eqn{\gamma_i(w)} of an individual of species \eqn{i}
and weight \eqn{w}. The search volume multiplies the predation kernel when
calculating the encounter rate and the predation rate.
}
\details{
If the \code{gamma} argument is not supplied, then it is set to
\deqn{\gamma_i(w) = \gamma_i w^q}. The values of \eqn{gamma_i} are taken from
the \code{gamma} column in the species parameter dataframe. If the \code{gamma}
column is not supplied in the species parameter dataframe, it is calculated
from \code{f0, h, beta, sigma, lambda} and \code{kappa}.
}
