% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_funcs.R
\name{sigmoid_weight}
\alias{sigmoid_weight}
\title{Weight based sigmoidal selectivity function}
\usage{
sigmoid_weight(w, sigmoidal_weight, sigmoidal_sigma, ...)
}
\arguments{
\item{w}{The size of the individual.}

\item{sigmoidal_weight}{The weight at which the knife-edge operates.}

\item{sigmoidal_sigma}{The width of the selection function}

\item{...}{Unused}
}
\description{
A sigmoidal selectivity function with 50\% selectivity at
weight \code{sigmoidal_weight} and width \code{sigmoidal_sigma}.
}
