% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_params.R
\name{species_params}
\alias{species_params}
\alias{species_params.MizerParams}
\alias{species_params.MizerSim}
\alias{species_params.data.frame}
\alias{species_params.species_params}
\alias{species_params<-}
\alias{species_params<-.MizerParams}
\alias{is.species_params}
\alias{given_species_params}
\alias{given_species_params.MizerParams}
\alias{given_species_params.MizerSim}
\alias{given_species_params.data.frame}
\alias{given_species_params.given_species_params}
\alias{is.given_species_params}
\alias{given_species_params<-}
\alias{given_species_params<-.MizerParams}
\alias{calculated_species_params}
\title{Species parameters}
\usage{
species_params(object, ...)

species_params(object, recalculate = TRUE) <- value

is.species_params(x)

given_species_params(object, ...)

is.given_species_params(x)

given_species_params(object) <- value

calculated_species_params(params)
}
\arguments{
\item{object}{A MizerParams object, a MizerSim object or a data frame}

\item{...}{Other arguments passed to methods.}

\item{recalculate}{Whether \verb{species_params<-()} should re-derive the
calculated species parameters and recalculate all the rates that depend on
the species parameters. Defaults to \code{TRUE}. See the section "Setting
species parameters without recalculation" below before setting it to
\code{FALSE}.}

\item{value}{A data frame with the new species parameters.}

\item{x}{An object to test with \code{is.species_params()} or
\code{is.given_species_params()}.}

\item{params}{A MizerParams object.}
}
\value{
\code{species_params()}: Data frame containing all species parameters
currently stored in the model.

\verb{species_params<-()}: Updates the \code{given_species_params} with any
parameters you have changed, and then recalculates the full species
parameter table and the model parameters. With \code{recalculate = FALSE} it
only does the recording and stores the parameters you supplied, see the
section "Setting species parameters without recalculation" below.

\code{given_species_params()}: Data frame containing the species parameter
values that were supplied explicitly by the user.

\verb{given_species_params<-()}: An alternative to \verb{species_params<-()} that
also triggers a recalculation of other parameters. The only difference is
that \verb{given_species_params<-()} issues warnings when a parameter is
changed whose effect is overridden by another parameter that has already
been given. This is especially useful during interactive use.

\code{calculated_species_params()}: Data frame containing only those species
parameter entries that are not explicit user input. Columns that would
consist entirely of \code{NA} values are dropped.

\code{is.species_params()} returns \code{TRUE} if \code{x} is a \code{species_params}
object, \code{FALSE} otherwise.

\code{is.given_species_params()} returns \code{TRUE} if \code{x} is a
\code{given_species_params} object, \code{FALSE} otherwise.
}
\description{
These functions allow you to get or set the species-specific parameters
stored in a MizerParams object.
}
\details{
There are a lot of species parameters and we will list them all below, but
most of them have sensible default values. The only required columns are
\code{species} for the species name and \code{w_inf} for its von Bertalanffy
asymptotic size. However if you have information about the values of other
parameters then you should provide them.

Three species parameters describe maximum sizes and play distinct roles:
\itemize{
\item \code{w_inf} is the von Bertalanffy asymptotic size of an average individual.
It is the required maximum-size parameter and is used to set default values
for \code{w_max}, \code{w_repro_max} and \code{w_mat}.
\item \code{w_repro_max} is the size at which a typical mature individual invests all
of its available energy into reproduction, see \code{\link[=setReproduction]{setReproduction()}}. It is
not a hard ceiling on size and defaults to \code{w_inf}.
\item \code{w_max} is purely a computational boundary: it sets the upper end of the
size grid and the range of plots. It defaults to \code{1.5 * w_inf}. For
backwards compatibility, if \code{w_inf} is not supplied it is taken from
\code{w_repro_max} or \code{w_max} instead.
}

Mizer distinguishes between the species parameters that you have given
explicitly and the species parameters that have been calculated by mizer or
set to default values. You can retrieve the given species parameters with
\code{given_species_params()} and the calculated ones with
\code{calculated_species_params()}. You get all species_params with
\code{species_params()}.

When you change species parameters with \verb{species_params<-()}, mizer
automatically detects which parameters you have changed. It records these
changed parameters in \code{given_species_params} so that they are protected
against being overwritten by future recalculations. It then triggers a
re-calculation of the calculated species parameters.

There are some species parameters that are used to set up the
size-dependent parameters that are used in the mizer model:
\itemize{
\item \code{gamma} and \code{q} are used to set the search volume, see \code{\link[=setSearchVolume]{setSearchVolume()}}.
\item \code{h} and \code{n} are used to set the maximum intake rate, see \code{\link[=setMaxIntakeRate]{setMaxIntakeRate()}}.
\item \code{k}, \code{ks} and \code{p} are used to set activity and basic metabolic rate,
see \code{\link[=setMetabolicRate]{setMetabolicRate()}}.
\item \code{z0}, \code{z_ext} and \code{d} are used to set the external mortality rate, see
\code{\link[=setExtMort]{setExtMort()}}.
\item \code{E_ext} and \code{n} are used to set the external encounter rate, see
\code{\link[=setExtEncounter]{setExtEncounter()}}.
\item \code{D_ext} and \code{n} are used to set the external diffusion rate, see
\code{\link[=setExtDiffusion]{setExtDiffusion()}}.
\item \code{w_mat}, \code{w_mat25}, \code{w_repro_max} and \code{m} are used to set the allocation to
reproduction, see \code{\link[=setReproduction]{setReproduction()}}.
\item \code{pred_kernel_type} specifies the shape of the predation kernel. The default
is a "lognormal", for other options see the "Setting predation kernel"
section in the help for \code{\link[=setPredKernel]{setPredKernel()}}.
\item \code{beta} and \code{sigma} are parameters of the lognormal predation kernel, see
\code{\link[=lognormal_pred_kernel]{lognormal_pred_kernel()}}. There will be other parameters if you are
using other predation kernel functions.
}

When you change one of the above species parameters using
\verb{species_params<-()} or \verb{given_species_params<-()}, the new value will be
used to update the corresponding size-dependent rates automatically, unless
you have set those size-dependent rates manually, in which case the
corresponding species parameters will be ignored.

There are some species parameters that are used directly in the model
rather than being used for setting up size-dependent parameters:
\itemize{
\item \code{alpha} is the assimilation efficiency, the proportion of the consumed
biomass that can be used for growth, metabolism and reproduction, see
the help for \code{\link[=getEReproAndGrowth]{getEReproAndGrowth()}}.
\item \code{w_min} is the egg size.
\item \code{interaction_resource} sets the interaction strength with the resource,
see "Predation encounter" section in the help for \code{\link[=getEncounter]{getEncounter()}}.
\item \code{erepro} is the reproductive efficiency, the proportion of the energy
invested into reproduction that is converted to egg biomass, see
\code{\link[=getRDI]{getRDI()}}.
\item \code{R_max} is the parameter in the Beverton-Holt density dependence added to
the reproduction, see \code{\link[=setBevertonHolt]{setBevertonHolt()}}. There will be other such
parameters if you use other density dependence functions, see the
"Density dependence" section in the help for \code{\link[=setReproduction]{setReproduction()}}.
}

Two parameters are used only by functions that need to convert between
weight and length:
\itemize{
\item \code{a} and \code{b} are the parameters in the allometric weight-length
relationship \eqn{w = a l ^ b}.
}

If you have supplied the \code{a} and \code{b} parameters, then you can replace weight
parameters like \code{w_inf}, \code{w_max}, \code{w_mat}, \code{w_mat25}, \code{w_repro_max} and
\code{w_min} by their corresponding length parameters \code{l_inf}, \code{l_max}, \code{l_mat},
\code{l_mat25}, \code{l_repro_max} and \code{l_min}.

You can also keep both, and change either of them later. Mizer keeps the two
consistent by the rule that the one you gave last wins, and if you gave both
at the same time the weight wins. So on a model set up with lengths you can
still set \code{w_mat} with \verb{species_params<-()} and mizer will update \code{l_mat} to
match, and if you set \code{l_mat} it will update \code{w_mat} as always. When you
supply a length and a weight together that do not agree, mizer uses the
weight and warns you that it has changed the length to match.

The rule is applied when a species parameter data frame is put into a model.
A data frame that you have taken out of a model and are editing on its own
is left exactly as you write it: no conversions, no checks and no warnings
until you assign it back with \verb{species_params<-()}, which is when mizer can
tell which values you changed. A data frame that was never in a model, for
example one you pass to \code{\link[=validSpeciesParams]{validSpeciesParams()}}, carries no such history, so
a length and a weight that disagree there count as given at the same time
and the weight wins.

The parameters that are only used to calculate default values for other
parameters are:
\itemize{
\item \code{f0} is the feeding level and is used to get a default value for the
coefficient of the search volume \code{gamma}, see \code{\link[=get_gamma_default]{get_gamma_default()}}.
\item \code{fc} is the critical feeding level below which the species can not
maintain itself. This is used to get a default value for the coefficient
\code{ks} of the metabolic rate, see \code{\link[=get_ks_default]{get_ks_default()}}.
\item \code{age_mat} is the age at maturity and is used to get a default value for
the coefficient \code{h} of the maximum intake rate, see \code{\link[=get_h_default]{get_h_default()}}.
\item If \code{age_mat} is not supplied, mizer used the von Bertalanffy parameters
\code{k_vb}, \code{w_inf} and \code{t0} as well as the weight-length exponent \code{b} to
determine it. This is unreliable and is therefore not recommended.
}

Changing these parameters with \verb{species_params<-()} will trigger a
recalculation of the downstream parameters, provided they are not protected
by being explicitly given.

There are other species parameters that are used in tuning the model to
observations:
\itemize{
\item \code{biomass_observed} and \code{biomass_cutoff} allow you to specify for each
species the total observed biomass above some cutoff size. This is
used by \code{\link[=calibrateBiomass]{calibrateBiomass()}} and \code{\link[=matchBiomasses]{matchBiomasses()}}.
\item \code{yield_observed} allows you to specify for each
species the total annual fisheries yield. This is
used by \code{\link[=calibrateYield]{calibrateYield()}} and \code{\link[=matchYields]{matchYields()}}.
}

Finally there are two species parameters that control the way the species are
represented in plots:
\itemize{
\item \code{linecolour} specifies the colour and can be any valid R colour value.
\item \code{linetype} specifies the line type ("solid", "dashed", "dotted", "dotdash",
"longdash", "twodash" or "blank")
}

Other species-specific information that is related to how the species is
fished is specified in a gear parameter data frame, see \code{\link[=gear_params]{gear_params()}}.
However in the case where each species is caught by only a single gear,
this information can also optionally be provided as species parameters and
\code{\link[=newMultispeciesParams]{newMultispeciesParams()}} will transfer them to the \code{gear_params} data frame.
However changing these parameters later in the species parameter data frames
will have no effect.

You are allowed to include additional columns in the species parameter
data frames. They will simply be ignored by mizer but will be stored in the
MizerParams object, in case your own code makes use of them.
}
\section{Setting species parameters without recalculation}{

\code{species_params(params, recalculate = FALSE) <- value} records the values you
changed among the given species parameters, so that they are not calculated
away later, and stores \code{value} as the species parameters. It then stops
there: the calculated species parameters are not re-derived from the given
ones, no missing parameters are filled in with their default values, and none
of the size-dependent rates are recalculated. Your species parameters are
stored as you supplied them, after the same checks and length-to-weight
conversions that writing into the \code{species_params} slot would trigger.

This is for code that has worked out a species parameter \emph{together with} the
rate array that the parameter determines, for example an optimiser that fits
\code{ks} and the matching \code{metab}, or \code{z_ext} and the matching \code{mu_b}. There the
recalculation is not just wasted work but would overwrite the rates the
caller has just set.

The object you get back is only as consistent as you make it. Mizer will not
check that the species parameters you supplied agree with the rate arrays in
the object, nor that they agree with the other species parameters that are
normally derived from them. Unless you are setting the affected rates
yourself, use the default \code{recalculate = TRUE}.
}

\seealso{
\code{\link[=validSpeciesParams]{validSpeciesParams()}}, \code{\link[=setParams]{setParams()}}

Other functions for setting parameters:
\code{\link[=gear_params]{gear_params()}},
\code{\link[=setExtDiffusion]{setExtDiffusion()}},
\code{\link[=setExtEncounter]{setExtEncounter()}},
\code{\link[=setExtMort]{setExtMort()}},
\code{\link[=setFishing]{setFishing()}},
\code{\link[=setInteraction]{setInteraction()}},
\code{\link[=setMaxIntakeRate]{setMaxIntakeRate()}},
\code{\link[=setMetabolicRate]{setMetabolicRate()}},
\code{\link[=setParams]{setParams()}},
\code{\link[=setPredKernel]{setPredKernel()}},
\code{\link[=setReproduction]{setReproduction()}},
\code{\link[=setSearchVolume]{setSearchVolume()}},
\code{\link[=use_predation_diffusion]{use_predation_diffusion()}}
}
\concept{functions for setting parameters}
