% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{species_params}
\alias{species_params}
\alias{species_params<-}
\title{Species parameter data frame}
\usage{
species_params(params)

species_params(params) <- value
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A data frame with the species parameters}
}
\description{
This is the right place to document the use of species parameters in mizer.
}
