% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spectra_y_label}
\alias{spectra_y_label}
\title{Y-axis label for a size-spectrum plot}
\usage{
spectra_y_label(power)
}
\arguments{
\item{power}{The power of weight that the abundance was multiplied by.}
}
\value{
A character string for the y-axis label.
}
\description{
Y-axis label for a size-spectrum plot
}
\keyword{internal}
