% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{srrNone}
\alias{srrNone}
\title{Identity stock-recruitment function}
\usage{
srrNone(rdi, species_params)
}
\arguments{
\item{rdi}{Vector of density independent egg production rates \eqn{R_p} for
all species.}

\item{species_params}{A species parameter dataframe. Not used.}
}
\value{
Vector of density-dependent larvae production rates.
}
\description{
Simply returns its \code{rdi} argument.
}
\seealso{
Other stock-recruitment functions: \code{\link{srrBevertonHolt}()},
  \code{\link{srrRicker}()}, \code{\link{srrSheperd}()}
}
\concept{stock-recruitment functions}
