% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{steady}
\alias{steady}
\title{Tune params object to be at steady state}
\usage{
steady(params, rfac = Inf, effort = 0)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{rfac}{A number that determines the strength of the non-linearity in
the Beverton-Holt stock-recruitment relationship. The maximal recruitment
will be set to rfac times the normal steady-state recruitment.
Default value is 10.}

\item{effort}{The fishing effort. Default is 0}
}
\description{
Tune params object to be at steady state
}
