% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{steady}
\alias{steady}
\title{Tune params object to be at steady state}
\usage{
steady(params, effort = 0, t_max = 50)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{effort}{The fishing effort. Default is 0}
}
\description{
This is done by running the dynamics for a specified number of years
while keeping recruitment constant to reach steady state. Then
the reproductive efficiencies are retuned to achieve that level of
recruitment.
}
