% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{tuneParams}
\alias{tuneParams}
\title{Launch gadget for tuning parameters}
\usage{
tuneParams(p, catchdist = NULL)
}
\arguments{
\item{p}{MizerParams object to tune. If missing, the gadget tries to recover
information from log files left over from aborted previous runs.}

\item{catchdist}{Data frame holding observed catch distribution with
columns \code{length} (the start of each size bin in cm), \code{catch}
(the number of individuals of a particular species caught in this size bin)
and \code{species} (the name of the species).}
}
\value{
The tuned MizerParams object
}
\description{
Launch gadget for tuning parameters
}
