% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{tuneParams}
\alias{tuneParams}
\title{Launch app for tuning model parameters}
\usage{
tuneParams()
}
\description{
Launch app for tuning model parameters
}
\seealso{
Other functions for changing parameters: \code{\link{changeBMort}},
  \code{\link{changeFishing}}, \code{\link{changeInitial}},
  \code{\link{changeIntakeMax}},
  \code{\link{changeInteraction}},
  \code{\link{changeMetab}}, \code{\link{changeParams}},
  \code{\link{changePlankton}},
  \code{\link{changePredKernel}},
  \code{\link{changeReproduction}},
  \code{\link{changeResourceDynamics}},
  \code{\link{changeResourceEncounter}},
  \code{\link{changeSearchVolume}}
}
\concept{functions for changing parameters}
