% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{tuneParams}
\alias{tuneParams}
\title{Launch shiny gadget for tuning parameters}
\usage{
tuneParams(p, catch = NULL, stomach = NULL)
}
\arguments{
\item{p}{MizerParams object to tune. If missing, the gadget tries to recover
information from log files left over from aborted previous runs.}

\item{catch}{Data frame holding binned observed catch data. The data can
be binned either into length bins or weight bins. In the former the data
frame should have columns \code{length} and \code{dl} holding the start of
the size bins in cm and the width of the size bins in cm respectively. In
the latter case the data frame should have columns \code{weight} and
\code{dw} holding the start of the size bins in frams and the width of the
size bins in grams. The data frame also needs to have the columns
\code{species} (the name of the species), \code{catch} (the number of
individuals of a particular species caught in a size bin).}

\item{stomach}{Data frame holding observations of prey items in predator
stomachs. The required columns are 
\itemize{
\item \code{species} holding the name of the predator species,
\item \code{wpredator} with the weight in grams of the predator,
\item \code{wprey} with the weight of the prey item.
}
In case prey items of the same weight have been aggregated in the data
frame then there should be a column \code{Nprey} saying how many prey 
items have been aggregated in each row.}
}
\value{
The tuned MizerParams object
}
\description{
The shiny gadget has sliders for the model parameters and tabs with a
variety of plots to visualise the steady-state
}
