% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{updateInitial}
\alias{updateInitial}
\title{Update the initial abundances}
\usage{
updateInitial(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
The MizerParams object with updated \code{initial_n} and 
  \code{initial_n_pp} slots.
}
\description{
Recalculates the steady-state abundances in a fixed background
given by the current abundances, keeping the abundances fixed in the
smallest size class for each species. Then readjusts the \code{erepro}
values.
}
