% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{upgradeParams}
\alias{upgradeParams}
\title{Upgrade MizerParams object from earlier mizer versions}
\usage{
upgradeParams(params)
}
\arguments{
\item{params}{An old MizerParams object to be upgraded}
}
\value{
The upgraded MizerParams object
}
\description{
Uses set_multispecies_model to create a new MizerParams object using the
parameters extracted from the old MizerParams object.
}
