% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgradeParams}
\alias{upgradeParams}
\title{Upgrade MizerParams object from earlier mizer versions}
\usage{
upgradeParams(params)
}
\arguments{
\item{params}{An old MizerParams object to be upgraded}
}
\value{
The upgraded MizerParams object
}
\description{
This function is called from \code{\link[=validParams]{validParams()}}. You should never need to call
it directly.
}
\seealso{
\code{\link[=validParams]{validParams()}}
}
\concept{helper}
\keyword{internal}
