% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{upgradeParams}
\alias{upgradeParams}
\title{Upgrade MizerParams object from earlier mizer versions}
\usage{
upgradeParams(params)
}
\arguments{
\item{params}{An old MizerParams object to be upgraded}
}
\value{
The upgraded MizerParams object
}
\description{
Uses set_multispecies_model to create a new MizerParams object using the
parameters extracted from the old MizerParams object.
}
\details{
If you only have a serialised version of the old object, for example
created via `saveRDS()`, and you get an error when trying to read it in
with `readRDS()` then unfortunately you will need to install the old version
of mizer first to read the params object into your workspace, then switch
to the current version and then call `upgradeParams()`. You can then save
the new version again with `saveRDS()`.
}
