% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{upgradeSim}
\alias{upgradeSim}
\title{Upgrade MizerSim object from earlier mizer versions}
\usage{
upgradeSim(sim)
}
\arguments{
\item{sim}{An old MizerSim object to be upgraded}
}
\value{
The upgraded MizerSim object
}
\description{
Occasionally during the development of new features for mizer, the MizerSim
class or the MizerParams class gains extra slots. MizerSim objects created in
older versions of mizer are then no longer valid in the new version because
of the missing slots. This function adds the missing slots and fills them
with default values. It calls `upgradeParams()` to upgrade the MizerParams
object inside the MizerSim object.
}
\details{
If you only have a serialised version of the old object, for example created
via `saveRDS()`, and you get an error when trying to read it in with
`readRDS()` then unfortunately you will need to install the old version of
mizer first to read the params object into your workspace, then switch to the
current version and then call `upgradeSim()`. You can then save the new
version again with `saveRDS()`.
}
