% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgradeSim}
\alias{upgradeSim}
\title{Upgrade MizerSim object from earlier mizer versions}
\usage{
upgradeSim(sim)
}
\arguments{
\item{sim}{An old MizerSim object to be upgraded}
}
\value{
The upgraded MizerSim object
}
\description{
Upgrade MizerSim object from earlier mizer versions
}
\concept{helper}
\keyword{internal}
