% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady.R
\name{valid_species_arg}
\alias{valid_species_arg}
\title{Helper function to assure validity of species argument}
\usage{
valid_species_arg(params, species)
}
\arguments{
\item{params}{A MizerParams object}

\item{species}{A vector of the names of the species to be affected or a
boolean vector indicating for each species whether it is to be affected
(TRUE) or not. By default all species are affected.}
}
\description{
Helper function to assure validity of species argument
}
