% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{validParams}
\alias{validParams}
\title{Validate MizerParams object and upgrade if necessary}
\usage{
validParams(params)
}
\arguments{
\item{params}{The MizerParams object to validate}
}
\value{
A valid MizerParams object
}
\description{
Validate MizerParams object and upgrade if necessary
}
