% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim-class.R
\name{validSim}
\alias{validSim}
\title{Validate MizerParams object and upgrade if necessary}
\usage{
validSim(sim)
}
\arguments{
\item{params}{The MizerParams object to validate}
}
\value{
A valid MizerParams object
}
\description{
Checks that the given MizerParams object is valid and upgrades it if
necessary by calling \code{\link[=upgradeParams]{upgradeParams()}}.
}
\details{
Besides upgrading if necessary, the only change that may be made to the
given MizerParams object is that the \code{w_min_idx} slot is recalculated.
}
