% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_params.R
\name{validSpeciesParams}
\alias{validSpeciesParams}
\title{Check validity of species parameters and set defaults for missing but
required parameters}
\usage{
validSpeciesParams(species_params)
}
\arguments{
\item{species_params}{The user-supplied species parameter data frame}
}
\value{
A valid species parameter data frame
}
\description{
Check validity of species parameters and set defaults for missing but
required parameters
}
\concept{("helper")}
