% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{w}
\alias{w}
\alias{w_full}
\alias{dw}
\alias{dw_full}
\title{Size bins}
\usage{
w(params)

w_full(params)

dw(params)

dw_full(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
\code{w()} returns a vector with the sizes at the start of each size bin
of the community spectrum.

\code{w_full()} returns a vector with the sizes at the start of each size bin
of the resource spectrum, which typically starts at smaller sizes than
the community spectrum.

\code{dw()} returns a vector with the widths of the size bins of the
community spectrum.

\code{dw_full()} returns a vector with the widths of the size bins of the
resource spectrum.
}
\description{
Functions to fetch information about the size bins used in the model
described by \code{params}.
}
\details{
TODO: Give more details about how mizer discretises the size
}
