% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerParams-class.R
\name{w}
\alias{w}
\alias{w_full}
\alias{dw}
\alias{dw_full}
\title{Size bins}
\usage{
w(params)

w_full(params)

dw(params)

dw_full(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\description{
This is a good place to explain how mizer discretises the size
}
