% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{wrapper_functions}
\alias{wrapper_functions}
\title{Functions used for setting up models}
\description{
The functions defined in the file wrapper functions set up MizerParams
objects for various kinds of size-spectrum models.
}
\section{List of Functions}{

In this list we relate the functions in this file to the sections in
the mizer vignette where the corresponding model is described.
\tabular{llll}{
Function name \tab Description \tab Section in vignette\cr
\code{\link[=newCommunityParams]{newCommunityParams()}} \tab Community model \tab 5 \cr
\code{\link[=newTraitParams]{newTraitParams()}} \tab Trait-based model \tab 6 \cr
}

The file also contains functions for adding, removing or rescaling species.
}

